\docType{package}
\name{arrow}
\alias{arrow}
\alias{arrow-package}
\title{Arrow: functional, composable R.}
\description{
Arrow is a
}
\section{Additional Documentation}{
  Arrow is a big, idiosyncratic (but hopefully logical!)
  library, so it might be a good idea to browse some of the
  additional help files.

  \itemize{ \item{\link{help_arrow_naming}: }{Details of
  arrow's naming conventions.}
  \item{\link{help_arrow_best_practices}: }{A guide to
  writing idiomatic Arrow code.}
  \item{\link{help_arrow_gotchas}: }{A guide to avoiding
  easy-to-make mistakes in Arrow.} }
}

