% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Fix}
\alias{Fix}
\title{An efficient implementation of xFix, lacking safe-guards and some features.
This is used internally when a function is called with only some of its arguments.}
\usage{
Fix(FN, SYM1, SYM2, SYM3, PRE1, PRE2, PRE3, FINAL)
}
\description{
param FN   a symbol, used as a reference to the function to fix.
}
\details{
param SYM1 a symbol, the name of the first parametre.
param SYM2 a symbol, the name of the first parametre. Optional.
param SYM3 a symbol, the name of the first parametre. Optional.

param PRE1 an expression. Input checks that are run exclusively
on the first argument. Optional.

param PRE2 an expression. Input checks that are run exclusively
on the second argument. Optional.

param PRE3 an expression. Input checks that are run exclusively
on the third argument. Optional.

param PRE an expression. Input checks that need more than one
argument at once (checking if two lengths are equal for example). Optional.

param FINAL an expression. An arbitrary code block to run before running
the function body.
}

