% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/values.R
\docType{data}
\name{Na}
\alias{False}
\alias{Na}
\alias{Null}
\alias{True}
\title{Special Values}
\format{\preformatted{ logi NA
}}
\usage{
Na

Null

True

False
}
\description{
These are aliases to values already in R.
}
\keyword{datasets}

