% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{x.}
\alias{!=.xlambda_builder}
\alias{$.xlambda_builder}
\alias{&&.xlambda_builder}
\alias{&.xlambda_builder}
\alias{*.xlambda_builder}
\alias{+.xlambda_builder}
\alias{-.xlambda_builder}
\alias{/.xlambda_builder}
\alias{:.xlambda_builder}
\alias{::.xlambda_builder}
\alias{<.xlambda_builder}
\alias{<=.xlambda_builder}
\alias{==.xlambda_builder}
\alias{>.xlambda_builder}
\alias{>=.xlambda_builder}
\alias{@.xlambda_builder}
\alias{\%/\%.xlambda_builder}
\alias{\%\%.xlambda_builder}
\alias{^.xlambda_builder}
\alias{x.}
\alias{|.xlambda_builder}
\alias{||.xlambda_builder}
\alias{~.xlambda_builder}
\title{x.}
\format{\preformatted{ list()
 - attr(*, "class")= chr "xlambda_builder"
}}
\usage{
x.

\\method{\\\%\\\%}{xlambda_builder}(x, y)

\\method{\\\%/\\\%}{xlambda_builder}(x, y)

"::.xlambda_builder"(x, y)

\\method{$}{xlambda_builder}(x, y)

"@.xlambda_builder"(x, y)

\\method{^}{xlambda_builder}(x, y)

":.xlambda_builder"(x, y)

\\method{*}{xlambda_builder}(x, y)

\\method{/}{xlambda_builder}(x, y)

\\method{-}{xlambda_builder}(x, y)

\\method{+}{xlambda_builder}(x, y)

\\method{>}{xlambda_builder}(x, y)

\\method{>=}{xlambda_builder}(x, y)

\\method{<}{xlambda_builder}(x, y)

\\method{<=}{xlambda_builder}(x, y)

\\method{==}{xlambda_builder}(x, y)

\\method{!=}{xlambda_builder}(x, y)

\\method{&}{xlambda_builder}(x, y)

"&&.xlambda_builder"(x, y)

\\method{|}{xlambda_builder}(x, y)

"||.xlambda_builder"(x, y)

"~.xlambda_builder"(x, y)
}
\description{
Wildcard
}
\keyword{datasets}

