% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xAddKeys}
\alias{xAddKeys}
\alias{xAddKeys...}
\title{xAddKeys}
\usage{
xAddKeys(strs, coll)

xAddKeys...(strs, ...)
}
\arguments{
\item{strs}{a collection of strings. The names to add to the
   input collection.}

\item{coll}{a collection. The collection to add names to.}
}
\value{
A named list.
}
\description{
Add names to a collection.
}
\details{
\bold{xAddKeys} is similar to the in-place assignment
    function \bold{names<-}, except that it is not an in-place
    assignment function.
}
\section{Type Signature}{

   |character| -> |any| -> [any]
}

\section{Corner Cases}{

   Returns the empty list if \bold{coll} is length-zero. Both
   duplicated and length-zero names are allowed.
}
\examples{

}
\seealso{
Other key_functions: \code{\link{xPluck}},
  \code{\link{xPluck_}}; \code{\link{xUnzipIndices}},
  \code{\link{xUnzipIndices_}}; \code{\link{xUnzipKeys}},
  \code{\link{xUnzipKeys_}}; \code{\link{xZipKeys}},
  \code{\link{xZipKeys_}}

Other key_functions: \code{\link{xPluck}},
  \code{\link{xPluck_}}; \code{\link{xUnzipIndices}},
  \code{\link{xUnzipIndices_}}; \code{\link{xUnzipKeys}},
  \code{\link{xUnzipKeys_}}; \code{\link{xZipKeys}},
  \code{\link{xZipKeys_}}
}

