% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xAmend}
\alias{xAmend}
\title{xAmend}
\usage{
xAmend(rexp, str1, str2)
}
\arguments{
\item{rexp}{a string. The regular expression used to select parts
    of a string to replace.}

\item{str1}{a string. The replacement for selected parts of a string.}

\item{str2}{a string. The string to replace parts of.}
}
\value{
a character vector.
}
\description{
Replace regular expression matches in a string with another string.
}
\section{|character| -> |character| -> |character| -> <character>}{

}

\section{Corner Cases}{

     the empty character vector is returned for when \bold{rexp},
     \bold{str1}, or \bold{str2} is used.
}
\examples{

# 1. remove enclosing brackets.

xAmend('[()]', '', '(09:34)')

# "09:34"
}
\seealso{
Other text_processing_functions: \code{\link{xCharsOf}};
  \code{\link{xExplode}}; \code{\link{xFromChars}},
  \code{\link{xFromChars_}}; \code{\link{xFromLines}},
  \code{\link{xFromLines_}}; \code{\link{xFromWords}},
  \code{\link{xFromWords_}}; \code{\link{xImplode}},
  \code{\link{xImplode_}}; \code{\link{xIsMatch}};
  \code{\link{xNotMatch}}; \code{\link{xRead}};
  \code{\link{xCarve}}, \code{\link{xCarve_}};
  \code{\link{xToChars}}; \code{\link{xToLines}};
  \code{\link{xToWords}}; \code{\link{xWrite}}
}

