% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xAnyOf}
\alias{xAnyOf}
\alias{xAnyOf_}
\title{xAnyOf}
\usage{
xAnyOf(pred, coll)

xAnyOf_(pred, ...)
}
\arguments{
\item{pred}{a predicate. The function used to test each element of
   the input collection.}

\item{coll}{a collection. The collection to test each element of.}

\item{...}{see above.}
}
\value{
A boolean value.
}
\description{
Is a predicate true for any element of a collection?
}
\section{Type Signature}{

    (any -> &lt;logical>) -> |any| -> &lt;logical>
}

\section{Corner Cases}{

   If coll is length zero then logical(0) is returned.
}
\examples{

# 1. Using any to test if there are any true values in a vector.

xAnyOf(xI, c(True, True, False, True))

# True

# 2. Does a function name with more that 25 characters exist
#    in the base R libraries?

xAnyOf(name := nchar(name) > 25, ls('package:base'))

# several matches exist, including the terse 'suppressPackageStartupMessages'
}
\seealso{
Other quantifier_functions: \code{\link{xAllOf}},
  \code{\link{xAllOf_}}; \code{\link{xExists}},
  \code{\link{xExists_}}; \code{\link{xForall}},
  \code{\link{xForall_}}; \code{\link{xNoneOf}},
  \code{\link{xNoneOf_}}; \code{\link{xPoll}},
  \code{\link{xPoll_}}
}

