\name{xAsComplex}
\alias{xAsComplex}
\alias{xAsComplex...}
\title{xAsComplex}
\usage{
xAsComplex(comps)

xAsComplex...(...)
}
\arguments{
  \item{comps}{a collection of complex values.}

  \item{...}{see above.}
}
\value{
A complex vector.
}
\description{
Convert a collection to a complex vector.
}
\section{Variadic Functions

\link{ help_variadic }}{

}

