% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xAsDataFrame}
\alias{xAsDataFrame}
\alias{xAsDataFrame_}
\title{xAsDataFrame}
\usage{
xAsDataFrame(colls)

xAsDataFrame_(...)
}
\arguments{
\item{colls}{a collection of collections. The collection
   to convert to a data frame.}

\item{...}{see above.}
}
\value{
Returns a data.frame.
}
\description{
Convert a collection of collections to a data.frame
}
\details{
data.frames are not a beloved part of kea,
   but they are ubiquitous in the base language.
   Kea uses collections of collections instead of
   data.frames, but data.frames are more common in the
   base language and importantly in libraries like ggplot2.
}
\section{Corner Cases}{

   \bold{xAsDataFrame} tried hard to avoid the crazy corner
   cases of \bold{data.frame}, but data frames are kludgey.

   Character vectors are not converted into factors,
   unlike the base functions.

  Rownames are added to the data.frame if the names of the
  collections inside colls are identical.
  If the rownames are different, an error is thrown.

  If row names or column names are omitted they default to
  indices.
}

\section{Stability 1 - \bold{Uncertain}}{


This feature may be removed or altered significantly in future versions.
If you find this feature useful, please leave an issue at https://github.com/rgrannell1/kea/issues.
}

