\name{xAsInteger}
\alias{xAsInteger}
\alias{xAsInteger...}
\title{xAsInteger}
\usage{
xAsInteger(nums)

xAsInteger...(...)
}
\arguments{
  \item{nums}{a collection of boolean values.}

  \item{...}{see above.}
}
\value{
An integer vector.
}
\description{
Convert a collection to a integer vector.
}
\section{Variadic Functions

\link{ help_variadic }}{

}

