% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xAsList}
\alias{xAsList}
\alias{xAsList_}
\title{xAsList}
\usage{
xAsList(coll)

xAsList_(...)
}
\arguments{
\item{coll}{a collection. The collection to convert to a list.}

\item{...}{see above.}
}
\value{
a list.
}
\description{
Convert a collection to a list.
}
\section{Type Signature}{

    |any| -> [any]
}

\section{Corner Cases}{

    xAsList preserves the names of its input collection.
}
\examples{

# 1. Convert a list of character values to a character vector.

xAsCharacter(list('this', 'is', 'much', 'words'))

# c('this', 'is', 'much', 'words')

# 2. This does NOT work

# xAsCharacter(1L)
#
# xAsCharacter only unlists character vectors, it doesn't convert
# numbers to strings.
}
\seealso{
Other container_conversion_functions: \code{\link{xAsCharacter}},
  \code{\link{xAsCharacter_}}; \code{\link{xAsComplex}},
  \code{\link{xAsComplex_}}; \code{\link{xAsDouble}},
  \code{\link{xAsDouble_}}; \code{\link{xAsInteger}},
  \code{\link{xAsInteger_}}; \code{\link{xAsLogical}},
  \code{\link{xAsLogical_}}; \code{\link{xAsRaw}},
  \code{\link{xAsRaw_}}
}

