\name{xAsLogical}
\alias{xAsLogical}
\alias{xAsLogical...}
\title{xAsLogical}
\usage{
xAsLogical(bools)

xAsLogical...(...)
}
\arguments{
  \item{bools}{a collection of boolean values.}

  \item{...}{see above.}
}
\description{
Convert a collection to a logical vector.
}
\section{Variadic Functions

\link{ help_variadic }}{

}

