\name{xAsRaw}
\alias{xAsRaw}
\alias{xAsRaw...}
\title{xAsRaw}
\usage{
xAsRaw(raws)

xAsRaw...(...)
}
\arguments{
  \item{raws}{a collection of raw values.}

  \item{...}{see above.}
}
\value{
A raw vector.
}
\description{
Convert a collection to a raw vector.
}
\section{Variadic Functions

\link{ help_variadic }}{

}

