\name{xAtCol}
\alias{xAtCol}
\alias{xAtCol...}
\title{xAtCol}
\usage{
xAtCol(num, colls)

xAtCol...(num, ...)
}
\arguments{
  \item{num}{a whole number. The number of the column to
  select.}

  \item{colls}{a collection of collections. The collection
  of rows to select a column from.}

  \item{...}{see above.}
}
\value{
A list.
}
\description{
Select a column of values from a collection of rows.
}
\examples{

NULL
}
\seealso{
Other selection_functions: \code{\link{xAt}},
\code{\link{xAt...}}; \code{\link{xDropWhile}},
\code{\link{xDropWhile...}}; \code{\link{xDrop}},
\code{\link{xDrop...}}; \code{\link{xFirstOf}},
\code{\link{xFirstOf...}}; \code{\link{xFourthOf}},
\code{\link{xFourthOf...}}; \code{\link{xGetKey}};
\code{\link{xInitOf}}, \code{\link{xInitOf...}};
\code{\link{xLastOf}}, \code{\link{xLastOf...}};
\code{\link{xPluck}}, \code{\link{xPluck...}};
\code{\link{xRestOf}}, \code{\link{xRestOf...}};
\code{\link{xSecondOf}}, \code{\link{xSecondOf...}};
\code{\link{xSlice}}, \code{\link{xSlice...}};
\code{\link{xTakeWhile}}, \code{\link{xTakeWhile...}};
\code{\link{xTake}}, \code{\link{xTake...}};
\code{\link{xThirdOf}}, \code{\link{xThirdOf...}}
}

