\name{xAtColumn}
\alias{xAtColumn}
\alias{xAtColumn...}
\title{xAtColumn}
\usage{
xAtColumn(num, colls)

xAtColumn...(num, ...)
}
\arguments{
  \item{num}{a whole number. The number of the column to
  select.}

  \item{colls}{a collection of collections. The collection
  of rows to select a column from.}

  \item{...}{see above.}
}
\value{
A list.
}
\description{
Select a column of values from a collection of rows.
}
\examples{

NULL
}

