% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xAtKey}
\alias{xAtKey}
\alias{xAtKey_}
\title{xAtKey}
\usage{
xAtKey(str, coll)

xAtKey_(str, ...)
}
\arguments{
\item{str}{a string. The key to select.}

\item{coll}{a collection.}

\item{...}{see above}
}
\value{
an arbitrary value.
}
\description{
Select a value from a collection using a key.
}
\section{Type Signature}{

    |character| -> |any| -> any
}

\section{Corner Cases}{

   If \bold{str} occurs multiple times in \bold{coll} the first match is returned.
}
\examples{

NULL
}
\seealso{
Other selection_functions: \code{\link{xAtCol}},
  \code{\link{xAtCol_}}; \code{\link{xAt}},
  \code{\link{xAt_}}; \code{\link{xDropWhile}},
  \code{\link{xDropWhile_}}; \code{\link{xDrop}},
  \code{\link{xDrop_}}; \code{\link{xFirstOf}},
  \code{\link{xFirstOf_}}; \code{\link{xFourthOf}},
  \code{\link{xFourthOf_}}; \code{\link{xInitOf}},
  \code{\link{xInitOf_}}; \code{\link{xLastOf}},
  \code{\link{xLastOf_}}; \code{\link{xMaxBy}},
  \code{\link{xMaxBy_}}; \code{\link{xMinBy}},
  \code{\link{xMinBy_}}; \code{\link{xPluck}},
  \code{\link{xPluck_}}; \code{\link{xRestOf}},
  \code{\link{xRestOf_}}; \code{\link{xSecondOf}},
  \code{\link{xSecondOf_}}; \code{\link{xSlice}},
  \code{\link{xSlice_}}; \code{\link{xTakeWhile}},
  \code{\link{xTakeWhile_}}; \code{\link{xTake}},
  \code{\link{xTake_}}; \code{\link{xThirdOf}},
  \code{\link{xThirdOf_}}
}

