\name{xByCols}
\alias{xByCols}
\title{$ xByCols}
\usage{
x_(  ) $ xByCols()
}
\value{
a list of lists.
}
\description{
Convert a matrix or data.frame to a collection of columns.
}
\seealso{
Other exclusively_methods: \code{\link{xByColnames}};
\code{\link{xByLevels}}; \code{\link{xByRows}};
\code{\link{xColUnit}}; \code{\link{xExecute}};
\code{\link{xRowUnit}}
}

