% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xByRowkeys}
\alias{xByRowkeys}
\title{xByRowkeys}
\usage{
x_(  ) $ xByRowkeys()
}
\arguments{
\item{tab}{a matrix or data frame. The tabular structure to get the rownames of.}
}
\value{
A kiwi containing a character vector.
}
\description{
Convert a matrix or data.frame to a collection of column names.
}
\details{
\bold{xByRowkeys} returns the column names of a data frame or matrix.
}
\section{Corner Cases}{

}
\seealso{
Other methods: \code{\link{xByColkeys}};
  \code{\link{xByCols}}; \code{\link{xByLevels}},
  \code{\link{xByLevels}}, \code{\link{xByValues}};
  \code{\link{xByRows}}; \code{\link{xElemsByCols}};
  \code{\link{xElemsByRows}}; \code{\link{xExecute}};
  \code{\link{xTap}}; \code{\link{x_}}, \code{\link{x__}}
}

