% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xByRownames}
\alias{xByRowkeys}
\alias{xByRownames}
\title{$ xByRownames}
\usage{
x_(  ) $ xByRownames()
}
\value{
An kiwi containing a character vector.
}
\description{
Convert a matrix or data.frame to a collection of column names.
}
\details{
\bold{xByRownames} returns the column names of a data frame or matrix.
}
\seealso{
Other methods: \code{\link{xByColkeys}};
  \code{\link{xByCols}}; \code{\link{xByLevels}},
  \code{\link{xByLevels}}, \code{\link{xByValues}};
  \code{\link{xByRows}}; \code{\link{xElemsByCols}};
  \code{\link{xElemsByRows}}; \code{\link{xExecute}};
  \code{\link{xTap}}; \code{\link{x_}}, \code{\link{x__}}
}

