% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xByValues}
\alias{xByValues}
\title{xByValues}
\usage{
xByValues(fact)
}
\arguments{
\item{fact}{a factor. The factor to get the underlying values from.}
}
\value{
A kea object containing a list.
}
\description{
Convert a factor to a collection of the factor's data.
}
\section{Corner Cases}{

}
\seealso{
Other methods: \code{\link{xByColkeys}};
  \code{\link{xByCols}}; \code{\link{xByLevels}};
  \code{\link{xByRowkeys}}; \code{\link{xByRows}};
  \code{\link{xElemsByCols}}; \code{\link{xElemsByRows}};
  \code{\link{xExecute}}; \code{\link{xTap}};
  \code{\link{x_}}, \code{\link{x__}}
}

