% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xCharsOf}
\alias{xCharsOf}
\title{xCharsOf}
\usage{
xCharsOf(str)
}
\arguments{
\item{str}{a length-one character vector. The string to
    get the numbers of characters in.}
}
\value{
An integer.
}
\description{
Get the number of characters in a string.
}
\section{Type Signature}{

    |character| -> &lt;character>
}

\section{Corner Cases}{

   none.
}
\examples{

# 1. A toy example.

xCharsOf('string?')

# 7
}
\seealso{
Other text_processing_functions: \code{\link{xAmend}};
  \code{\link{xExplode}}; \code{\link{xFromChars}},
  \code{\link{xFromChars_}}; \code{\link{xFromLines}},
  \code{\link{xFromLines_}}; \code{\link{xFromWords}},
  \code{\link{xFromWords_}}; \code{\link{xImplode}},
  \code{\link{xImplode_}}; \code{\link{xIsMatch}};
  \code{\link{xNotMatch}}; \code{\link{xRead}};
  \code{\link{xSliceString}}, \code{\link{xSliceString_}};
  \code{\link{xToChars}}; \code{\link{xToLines}};
  \code{\link{xToWords}}; \code{\link{xWrite}}
}

