% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xChoose}
\alias{xChoose}
\alias{xChoose_}
\title{Enumerate all ways of choosing several elements from a collection.}
\usage{
xChoose(num, coll)

xChoose_(num, ...)
}
\arguments{
\item{num}{a nonnegative whole number. The number of elements to choose
   from \bold{coll}.}

\item{coll}{a collection. The collection to draw elements from.}

\item{...}{see above.}
}
\value{
A list of lists, with each list containing \bold{num} elements.
}
\description{
Enumerate all ways of choosing several elements from a collection.
}
\details{
\bold{xChoose} enumerates all ways of choosing \bold{num} distinct
elements from a larger collection, where order doesn't matter.

The number of ways of choosing \bold{num} elements from a collection
is given by the base R function \code{choose(num, length(coll))}.
}
\section{Corner Cases}{

     Returns the empty list if \bold{num} is zero.
}
\examples{

NULL

}
\seealso{
Other combinatoric_functions: \code{\link{xCycle}},
  \code{\link{xCycle_}}; \code{\link{xPowerSetOf}},
  \code{\link{xPowerSetOf_}}; \code{\link{xProdSetOf}},
  \code{\link{xProdSetOf_}}; \code{\link{xReorder}},
  \code{\link{xReorder_}}
}

