\name{xCollapse}
\alias{xCollapse}
\title{xCollapse}
\usage{
xCollapse(str, strs)
}
\arguments{
  \item{str}{a string to use as a delimiter.}

  \item{strs}{several character vectors.}
}
\value{
a length-one character vector.
}
\description{
Concatenate a character vector into a single string using a
delimiter.
}
\seealso{
Other character_vector_functions: \code{\link{xChars}}
}

