\name{xColUnit}
\alias{xColUnit}
\title{$ xColUnit}
\usage{
x_(  ) $ xColUnit()
}
\value{
An arrow object containing an empty matrix or data.frame.
}
\description{
Return the concatenative unit for a matrix or data.frame
column.
}
\seealso{
Other methods: \code{\link{xByColnames}};
\code{\link{xByCols}}; \code{\link{xByLevels}},
\code{\link{xByLevels}}; \code{\link{xByRows}};
\code{\link{xElemsByCols}}; \code{\link{xElemsByRows}};
\code{\link{xExecute}}; \code{\link{xRowUnit}};
\code{\link{xTap}}; \code{\link{xTranspose}};
\code{\link{x_}}
}

