% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xCross.R
\name{xCross}
\alias{xCross}
\alias{xCross_}
\title{xCross}
\usage{
xCross(colls)

xCross_(...)
}
\arguments{
\item{colls}{a collection of collections. The collections to
use to take the Cartesian product.}

\item{...}{see above.}
}
\value{
A list of collections, with as many elements per collection as
     there are collections in \bold{colls}.
}
\description{
Get the cartesian product of several collections.
}
\section{Type Signature}{

    ||any|| -> [[any]]
}

\section{Corner Cases}{

     Returns the empty list if \bold{colls} is length-zero, or any element
     of \bold{colls} is length-zero.
}
\examples{

# 1. xCross with xMap or xReduce is
#    often a very nice replacements for loops.

xMap(
    xApply((ith : jth) := {
        ith^2 + jth^2
    }),
    xCross_(1:4, 1:4)
)

# is roughly the same as base R's_

res <- list()

for (ith in 1:4) {
    for (jth in 1:4) {
        res <- c(res, ith^2 + jth^2)
    }
}

res

# 2. Compute the standard childhood multiplication
#    table.

xCross_(1:12, 1:12)
}
\seealso{
Other combinatoric_functions: \code{\link{xChoose}},
  \code{\link{xChoose_}}; \code{\link{xCycle}},
  \code{\link{xCycle_}}; \code{\link{xPowerSetOf}},
  \code{\link{xPowerSetOf_}}

Other set_functions: \code{\link{xDistinctOf}},
  \code{\link{xDistinctOf_}}; \code{\link{xDuplicatesOf}},
  \code{\link{xDuplicatesOf_}}; \code{\link{xExclude}},
  \code{\link{xExclude_}}; \code{\link{xIntersect}},
  \code{\link{xIntersect_}}; \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsIn}},
  \code{\link{xIsIn_}}; \code{\link{xIsSubsetOf}},
  \code{\link{xIsSubsetOf_}}; \code{\link{xNotEmpty}},
  \code{\link{xNotEmpty_}}; \code{\link{xNotIn}},
  \code{\link{xNotIn_}}; \code{\link{xNotSubsetOf}},
  \code{\link{xNotSubsetOf_}}; \code{\link{xPowerSetOf}},
  \code{\link{xPowerSetOf_}}; \code{\link{xUnite}},
  \code{\link{xUnite_}}
}

