\name{xDigitsOf}
\alias{xDigitsOf}
\title{xDigitsOf}
\usage{
xDigitsOf(num)

xDigitsOf(num)
}
\arguments{
  \item{num}{a nonnegative whole number. The number to
  convert to digits.}

  \item{num}{a nonnegative whole number. The number to
  convert to digits.}
}
\value{
An integer vector.

An integer vector.
}
\description{
Get the digits composing an integer.

Get the digits composing an integer.
}
\examples{

NULL

NULL
}

