% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xDistinctOf}
\alias{xDistinctOf}
\alias{xDistinctOf_}
\title{xDistinctOf}
\usage{
xDistinctOf(coll)

xDistinctOf_(...)
}
\arguments{
\item{coll}{a collection. The collection to return distinct
   elements from.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Return the distinct elements in a collection.
}
\section{Type Signature}{

    |any| -> [any]
}

\section{Corner Cases}{

   Returns the empty list if \code{colls} is length-zero.
}
\examples{

# 1. Find the smallest number in a list.

x__(1, 2, 3, 4, 5) $ x_MinBy(xI)

# 1

# 2. CodeEvalis a string a pangram?
#    does a string contain each letter.

is_pangram <- line := {

	x_(line) $
	xToChars() $ xMap(tolower) $ xDistinctOf() $
	xSelect(
		xIsIn(coll = letters)
	) $
	x_LenOf() == 26
}

is_pangram('A quick brown fox jumps over the lazy dog')

# True

is_pangram('not pangramic thank you mam.')

# False
}
\seealso{
Other set_functions: \code{\link{xDuplicatesOf}},
  \code{\link{xDuplicatesOf_}}; \code{\link{xIntersect}},
  \code{\link{xIntersect_}}; \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsIn}},
  \code{\link{xIsIn_}}; \code{\link{xIsSubsetOf}},
  \code{\link{xIsSubsetOf_}}; \code{\link{xNotEmpty}},
  \code{\link{xNotEmpty_}}; \code{\link{xNotIn}},
  \code{\link{xNotIn_}}; \code{\link{xNotSubsetOf}},
  \code{\link{xNotSubsetOf_}}; \code{\link{xPowerSetOf}},
  \code{\link{xPowerSetOf_}}; \code{\link{xProdSetOf}},
  \code{\link{xProdSetOf_}}; \code{\link{xUnite}},
  \code{\link{xUnite_}}
}

