\name{xDuplicated}
\alias{xDuplicated}
\alias{xDuplicated...}
\title{xDuplicated}
\usage{
xDuplicated(coll)

xDuplicated...(...)
}
\arguments{
  \item{coll}{a collection.}

  \item{...}{see above.}
}
\value{
A list.
}
\description{
Return the duplicated elements in a collection.
}
\section{Corner Cases}{
  Returns the empty list if \code{colls} is length-zero.
}
\examples{

NULL

}
\seealso{
Other set_functions: \code{\link{xCombos}},
\code{\link{xCombos...}}; \code{\link{xInter}},
\code{\link{xInter...}}; \code{\link{xIsMember}},
\code{\link{xIsMember...}}; \code{\link{xNotMember}},
\code{\link{xNotMember...}}; \code{\link{xSetProd}},
\code{\link{xSetProd...}}; \code{\link{xUnion}},
\code{\link{xUnion...}}; \code{\link{xUnique}},
\code{\link{xUnique...}}
}

