% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xDuplicatesOf.R
\name{xDuplicatesOf}
\alias{xDuplicatesOf}
\alias{xDuplicatesOf_}
\title{xDuplicatesOf}
\usage{
xDuplicatesOf(coll)

xDuplicatesOf_(...)
}
\arguments{
\item{coll}{a collection. The collection to return
   duplicated elements from.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Return the duplicated elements in a collection.
}
\section{Type Signature}{

    |any| -> [any]
}

\section{Corner Cases}{

   Returns the empty list if \bold{colls} is length-zero.
}
\examples{

NULL

}
\seealso{
Other set_functions: \code{\link{xCross}},
  \code{\link{xCross_}}; \code{\link{xDistinctOf}},
  \code{\link{xDistinctOf_}}; \code{\link{xExclude}},
  \code{\link{xExclude_}}; \code{\link{xIntersect}},
  \code{\link{xIntersect_}}; \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsIn}},
  \code{\link{xIsIn_}}; \code{\link{xIsSubsetOf}},
  \code{\link{xIsSubsetOf_}}; \code{\link{xNotEmpty}},
  \code{\link{xNotEmpty_}}; \code{\link{xNotIn}},
  \code{\link{xNotIn_}}; \code{\link{xNotSubsetOf}},
  \code{\link{xNotSubsetOf_}}; \code{\link{xPowerSetOf}},
  \code{\link{xPowerSetOf_}}; \code{\link{xUnite}},
  \code{\link{xUnite_}}
}

