\name{xElemIsNa}
\alias{xElemIsNa}
\alias{xElemIsNa...}
\title{xElemIsNa}
\usage{
xElemIsNa(coll)

xElemIsNa...(...)
}
\arguments{
  \item{coll}{a collection. The collection to test each
  element of for being na.}

  \item{...}{see above.}
}
\value{
A vector of boolean values.
}
\description{
Is an element of a collection na?
}
\examples{

NULL

}

