% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xElemIsNan}
\alias{xElemIsNan}
\alias{xElemIsNan_}
\title{xElemIsNan}
\usage{
xElemIsNan(coll)

xElemIsNan_(...)
}
\arguments{
\item{coll}{a collection. The collection to test each element
   of for being nan.}

\item{...}{see above.}
}
\value{
A vector of boolean values.
}
\description{
Is an element in a collection NaN?
}
\section{Type Signature}{

    |any| -> <logical>
}

\section{Corner Cases}{

   Returns True if coll is length-zero.
}
\examples{

NULL

}
\seealso{
Other value_testing_functions: \code{\link{xElemIsFalse}},
  \code{\link{xElemIsFalse_}}; \code{\link{xElemIsNa}},
  \code{\link{xElemIsNa_}}; \code{\link{xElemIsNull}},
  \code{\link{xElemIsNull_}}; \code{\link{xElemIsTrue}},
  \code{\link{xElemIsTrue_}}; \code{\link{xElemNotFalse}},
  \code{\link{xElemNotFalse_}}; \code{\link{xElemNotNan}},
  \code{\link{xElemNotNan_}}; \code{\link{xElemNotNa}},
  \code{\link{xElemNotNa_}}; \code{\link{xElemNotNull}},
  \code{\link{xElemNotNull_}}; \code{\link{xElemNotTrue}},
  \code{\link{xElemNotTrue_}}; \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsFalse}};
  \code{\link{xIsNan}}; \code{\link{xIsNa}};
  \code{\link{xIsNull}}; \code{\link{xIsTrue}};
  \code{\link{xIs}}; \code{\link{xNotEmpty}},
  \code{\link{xNotEmpty_}}; \code{\link{xNotFalse}};
  \code{\link{xNotNan}}; \code{\link{xNotNa}};
  \code{\link{xNotNull}}; \code{\link{xNotTrue}}
}

