\name{xElemIsTrue}
\alias{xElemIsTrue}
\alias{xElemIsTrue...}
\title{xElemIsTrue}
\usage{
xElemIsTrue(coll)

xElemIsTrue...(...)
}
\arguments{
  \item{coll}{a collection.}

  \item{...}{see above.}
}
\value{
A vector of boolean values.
}
\description{
Is an element of a collection true?
}
\examples{

NULL

}

