% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xElemNotNan}
\alias{xElemNotNan}
\alias{xElemNotNan_}
\title{xElemNotNan}
\usage{
xElemNotNan(coll)

xElemNotNan_(...)
}
\arguments{
\item{coll}{a collection. The collection to test each element
   of for being non-nan.}

\item{...}{see above.}
}
\value{
A vector of boolean values.
}
\description{
Is an element of a collection not NaN?
}
\section{Corner Cases}{

   Returns the empty list if \code{coll is length-zero}.
}
\examples{

NULL

}
\seealso{
Other value_testing_functions: \code{\link{xElemIsFalse}},
  \code{\link{xElemIsFalse_}}; \code{\link{xElemIsNan}},
  \code{\link{xElemIsNan_}}; \code{\link{xElemIsNa}},
  \code{\link{xElemIsNa_}}; \code{\link{xElemIsNull}},
  \code{\link{xElemIsNull_}}; \code{\link{xElemIsTrue}},
  \code{\link{xElemIsTrue_}}; \code{\link{xElemNotFalse}},
  \code{\link{xElemNotFalse_}}; \code{\link{xElemNotNa}},
  \code{\link{xElemNotNa_}}; \code{\link{xElemNotNull}},
  \code{\link{xElemNotNull_}}; \code{\link{xElemNotTrue}},
  \code{\link{xElemNotTrue_}}; \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsFalse}};
  \code{\link{xIsNan}}; \code{\link{xIsNa}};
  \code{\link{xIsNull}}; \code{\link{xIsTrue}};
  \code{\link{xNotEmpty}}, \code{\link{xNotEmpty_}};
  \code{\link{xNotFalse}}; \code{\link{xNotNan}};
  \code{\link{xNotNa}}; \code{\link{xNotNull}};
  \code{\link{xNotTrue}}
}

