% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xElemsByCols}
\alias{xElemsByCols}
\title{$ xElemsByCols}
\usage{
x_(  ) $ xElemsByCols()
}
\value{
An arrow object containing a list.
}
\description{
Enumerate the elements of a matrix down columns.
}
\seealso{
Other methods: \code{\link{xByColkeys}};
  \code{\link{xByCols}}; \code{\link{xByLevels}},
  \code{\link{xByLevels}}; \code{\link{xByRows}};
  \code{\link{xColUnit}}; \code{\link{xElemsByRows}};
  \code{\link{xExecute}}; \code{\link{xRowUnit}};
  \code{\link{xTap}}; \code{\link{xTranspose}};
  \code{\link{x_}}
}

