\name{xElemsByRows}
\alias{xElemsByRows}
\title{$ xElemsByRows}
\usage{
x_(  ) $ xElemsByRows()
}
\value{
an arrow object containing a list.
}
\description{
Enumerate the elements of a matrix across rows.
}
\seealso{
Other exclusively_methods: \code{\link{xByColnames}};
\code{\link{xByCols}}; \code{\link{xByLevels}},
\code{\link{xByLevels}}; \code{\link{xByRows}};
\code{\link{xColUnit}}; \code{\link{xElemsByCols}};
\code{\link{xExecute}}; \code{\link{xRowUnit}};
\code{\link{xTap}}; \code{\link{xTranspose}}
}

