% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xExclude.R
\name{xExclude}
\alias{xExclude}
\alias{xExclude_}
\title{xExclude}
\usage{
xExclude(coll1, coll2)

xExclude_(coll1, ...)
}
\arguments{
\item{coll1}{a collection. The elements to exclude.}

\item{coll2}{a collection. The collection to exclude those elements from.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Get the elements of a collection not in another.
}
\section{Type Signature}{

    |any| -> |any| -> [any]
}

\section{Corner Cases}{

   Returns the empty list if \bold{coll2} is length-zero.
}
\examples{

# 1. a toy example.

xExclude(c('a', 'c'), c('a', 'b', 'c'))
}
\seealso{
Other set_functions: \code{\link{xDistinctOf}},
  \code{\link{xDistinctOf_}}; \code{\link{xDuplicatesOf}},
  \code{\link{xDuplicatesOf_}}; \code{\link{xIntersect}},
  \code{\link{xIntersect_}}; \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsIn}},
  \code{\link{xIsIn_}}; \code{\link{xIsSubsetOf}},
  \code{\link{xIsSubsetOf_}}; \code{\link{xNotEmpty}},
  \code{\link{xNotEmpty_}}; \code{\link{xNotIn}},
  \code{\link{xNotIn_}}; \code{\link{xNotSubsetOf}},
  \code{\link{xNotSubsetOf_}}; \code{\link{xPowerSetOf}},
  \code{\link{xPowerSetOf_}}; \code{\link{xProdSetOf}},
  \code{\link{xProdSetOf_}}; \code{\link{xUnite}},
  \code{\link{xUnite_}}
}

