\name{xExecute}
\alias{xExecute}
\title{x_([ any ]) $ xExecute}
\usage{
x_(  )$xExecute(fn)
}
\arguments{
  \item{fn}{a nullary function}
}
\value{
the return value of the previous method.
}
\description{
Execute a (possibly side-effectful) function before
continuing with the previous method result.
}

