% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xFalsity.R
\name{xFalsity}
\alias{xFalsity}
\title{xFalsity}
\usage{
xFalsity(...)
}
\arguments{
\item{...}{arguments to be dropped.}
}
\value{
false
}
\description{
A function that always returns false.
}
\section{Type Signature}{

   ..any -> <logical>
}

\section{Corner Cases}{

   None.
}
\examples{

# 1. Using falsity as an identity
#    with xReject.

xReject(xFalsity, 1:10)

# 1:10
}
\seealso{
Other basic_functions: \code{\link{xCapture}},
  \code{\link{xK}}; \code{\link{xClear}},
  \code{\link{xClear_}}; \code{\link{xIrrelevance}};
  \code{\link{xI}}, \code{\link{xIdentity}};
  \code{\link{xTruth}}
}

