% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xFirstAs}
\alias{xFirstAs}
\alias{xFirstAs_}
\title{xFirstAs}
\usage{
xFirstAs(val, coll)

xFirstAs_(val, ...)
}
\arguments{
\item{val}{an arbitrary value. The value to use.}

\item{coll}{a collection. The collection to set the
   first element of.}

\item{...}{see above.}
}
\value{
A collection.
}
\description{
Set the first value in a collection.
}
\section{Type Signature}{

    any -> |any| -> [any]
}

\section{Corner Cases}{

    Throws an error for collections shorter than one
    elements.
}
\examples{

NULL
}

