\name{xFlip}
\alias{xFlip}
\title{xFlip}
\usage{
xFlip(fn)
}
\arguments{
  \item{fn}{an arbitrary function.}
}
\value{
a function of the same arity as \code{fn}.
}
\description{
Reverse the formal parametres of a function.
}
\seealso{
Other higher_order_functions: \code{\link{xApply}};
\code{\link{xArity}}; \code{\link{xAsUnary}};
\code{\link{xCompose}}; \code{\link{xConst}};
\code{\link{xFormals}}; \code{\link{xIterate}};
\code{\link{xJuxtapose}}; \code{\link{xLift}};
\code{\link{xLimit}}; \code{\link{xNot}};
\code{\link{xParams}}; \code{\link{xStopWatch}};
\code{\link{xUnfold}}; \code{\link{xWait}}
}

