% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xFormalsOf}
\alias{xFormalsOf}
\title{xFormalsOf}
\usage{
xFormalsOf(fn)
}
\arguments{
\item{fn}{an arbitrary function. The function to get the parametres of.}
}
\value{
A named list, where each element's name is a parametre in \bold{fn} and each value
	  is the default value of that parametre.
}
\description{
xFormalsOf
}
\section{Type Signature}{

    function -> |any|

Get a named list of parametres and their defaults from a function or primitive function.
}

\section{Corner Cases}{

   If \bold{fn} is a primitive function a heuristic is used to obtain
   its formals. If a parametre has no default, the value of the
   corresponding element in the return value will be the empty symbol,
   which is identical to \code{quote(expr=)}.
}

\section{Stability - \bold{Uncertain}}{


This feature may be removed or altered significantly in future versions.
If you find this feature useful, please leave an issue at https://github.com/rgrannell1/kea/issues.
}
\examples{

# 1. Get the formals of a base function.

xFormalsOf(Reduce)

# list(f = , x = , init = , right = FALSE, accumulate = FALSE)
}
\seealso{
Other parametre_functions: \code{\link{xArityOf}};
  \code{\link{xFix}}, \code{\link{xFix_}};
  \code{\link{xParamsOf}}; \code{\link{xSpread}};
  \code{\link{xUnspread}}
}

