% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xFourthAs}
\alias{xFourthAs}
\alias{xFourthAs_}
\title{xFourthAs}
\usage{
xFourthAs(val, coll)

xFourthAs_(val, ...)
}
\arguments{
\item{val}{an arbitrary value. The value to set.}

\item{coll}{a collection. The collection to set the
   fourth element of.}

\item{...}{see above.}
}
\value{
A collection.
}
\description{
Set the fourth value in a collection.
}
\section{Type Signature}{

    any -> |any| -> [any]
}

\section{Corner Cases}{

    Throws an error for collections shorter than four
    elements.
}
\examples{

NULL
}

