\name{xFromChars}
\alias{xFromChars}
\alias{xFromChars...}
\title{xFromChars}
\usage{
xFromChars(strs)

xFromChars...(...)
}
\arguments{
  \item{strs}{several character vectors.}

  \item{...}{see above.}
}
\value{
A length-one character vector.
}
\description{
Collapase a character vector with empty strs as delimiters.
}
\section{Variadic Functions

\link{ help_variadic }}{

}
\seealso{
Other character_functions: \code{\link{xExplode}},
\code{\link{xExplode...}}; \code{\link{xFromLines}},
\code{\link{xFromLines...}}; \code{\link{xFromWords}},
\code{\link{xFromWords...}}; \code{\link{xImplode}},
\code{\link{xImplode...}}; \code{\link{xSubstring}},
\code{\link{xSubstring...}}; \code{\link{xToChars}};
\code{\link{xToLines}}; \code{\link{xToWords}}
}

