\name{xFromWords}
\alias{xFromWords}
\alias{xFromWords...}
\title{xFromWords}
\usage{
xFromWords(strs)

xFromWords...(...)
}
\arguments{
  \item{strs}{a collection of length-one character vectors.
  The strings to collapse into one string.}

  \item{...}{see above.}
}
\value{
A length-one character vector.
}
\description{
Concatentate a character vector with spaces as delimiters.
}
\seealso{
Other character_functions: \code{\link{xExplode}},
\code{\link{xExplode...}}; \code{\link{xFromChars}},
\code{\link{xFromChars...}}; \code{\link{xFromLines}},
\code{\link{xFromLines...}}; \code{\link{xImplode}},
\code{\link{xImplode...}}; \code{\link{xSubstring}},
\code{\link{xSubstring...}}; \code{\link{xToChars}};
\code{\link{xToLines}}; \code{\link{xToWords}}
}

