% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xGather.R
\name{xGather}
\alias{xGather}
\alias{xGather_}
\title{xGather}
\usage{
xGather(coll)

xGather_(...)
}
\arguments{
\item{coll}{a collection. The elements to group.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Group adjacent equal elements into sublists.
}
\section{Type Signature}{

    |any| -> [[any]]
}

\section{Corner Cases}{

   If \bold{coll} is a empty collection the empty list is returned.
}

\section{Stability - \bold{Experimental}}{


This feature is newly introduced or experimental, may contain bugs, and may be removed
or altered in future versions if it is not sufficiently useful. Breaking changes
are likely in the near future.

If you find this feature useful, please leave an issue at https://github.com/rgrannell1/kea/issues.
}
\examples{

# 1. A toy example

xGather(c(0, 0, 0, 1, 0, 1, 1, 0))

# xGather_(0, 0, 1, 1, 1, 0, 1, 0)




# 2. Get the length of the longest stretch of consecutive one's in a binary
#    vector (an interview question my friend was asked).

bits <- c(0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0)

x_(bits) $ xGather() $ xMap(xLenOf) $ x_MaxBy(xI)

# 5
}
\seealso{
Other reshaping_functions: \code{\link{xAmassBy}},
  \code{\link{xAmassBy_}}; \code{\link{xAppend}},
  \code{\link{xAppend_}}; \code{\link{xChunk}},
  \code{\link{xChunk_}}; \code{\link{xFlatten}},
  \code{\link{xFlatten_}}; \code{\link{xGroupBy}},
  \code{\link{xGroupBy_}}; \code{\link{xJoin}},
  \code{\link{xJoin_}}; \code{\link{xOneOf}},
  \code{\link{xOneOf_}}; \code{\link{xPrepend}},
  \code{\link{xPrepend_}}; \code{\link{xRecycle}},
  \code{\link{xRecycle_}}; \code{\link{xRepeat}},
  \code{\link{xRepeat_}}; \code{\link{xReverse}},
  \code{\link{xReverse_}}; \code{\link{xShuffle}},
  \code{\link{xShuffle_}}; \code{\link{xSortBy}},
  \code{\link{xSortBy_}}; \code{\link{xTabulate}},
  \code{\link{xTabulate_}}; \code{\link{xUnzipIndices}},
  \code{\link{xUnzipIndices_}}; \code{\link{xUnzipKeys}},
  \code{\link{xUnzipKeys_}}; \code{\link{xZipKeys}},
  \code{\link{xZipKeys_}}; \code{\link{xZip}},
  \code{\link{xZip_}}
}

