\name{xIdentity}
\alias{xI}
\alias{xIdentity}
\title{xIdentity}
\usage{
xIdentity(val)

xI(val)
}
\arguments{
  \item{val}{an arbitrary value.}
}
\value{
Returns \code{val}.
}
\description{
Return an argument without modification.
}
\examples{

NULL
}
\seealso{
Other basic_functions: \code{\link{xCapture}},
\code{\link{xK}}; \code{\link{xFalsity}};
\code{\link{xIrrelevance}}; \code{\link{xTruth}}
}

