\name{xImplode}
\alias{xImplode}
\title{xImplode}
\usage{
xImplode(str, strs)
}
\arguments{
  \item{str}{a string to use as a delimiter.}

  \item{strs}{several character vectors.}
}
\value{
a length-one character vector.
}
\description{
Concatenate a character vector into a single string using a
delimiter.
}
\seealso{
Other character_vector_functions: \code{\link{xChars}};
\code{\link{xExplode}}; \code{\link{xLines}};
\code{\link{xSubString}}; \code{\link{xUnchars}};
\code{\link{xUnlines}}; \code{\link{xUnwords}};
\code{\link{xWords}}
}

