% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xIndicesTo}
\alias{xIndicesTo}
\title{xIndicesTo}
\usage{
xIndicesTo(num)
}
\arguments{
\item{num}{a nonnegative number. The largest number in the sequence.}
}
\value{
an integer vector.
}
\description{
Create a sequence from one to another number.
}
\section{Type Signature}{

    |numeric| -> <integer>
}

\section{Corner Cases}{

    Fails when \bold{num} is infinite. If \bold{num} is zero or length-zero,
    \bold{integer(0)} is returned.
}
\examples{

# 1. get the cumulative sum of the first 7 nonnegative integers.

x_(xIndicesTo(7)) $ x_Scan('+', 0)

# list(0, 1, 3, 6, 10, 15, 21, 28)
}
\seealso{
Other key_functions: \code{\link{xAddKeys}},
  \code{\link{xAddKeys_}}; \code{\link{xIndicesOf}},
  \code{\link{xIndicesOf_}}; \code{\link{xKeysOf}},
  \code{\link{xKeysOf_}}; \code{\link{xUnzipIndices}},
  \code{\link{xUnzipIndices_}}; \code{\link{xUnzipKeys}},
  \code{\link{xUnzipKeys_}}; \code{\link{xValuesOf}},
  \code{\link{xValuesOf_}}; \code{\link{xZipKeys}},
  \code{\link{xZipKeys_}}
}

