% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xIntersect}
\alias{xIntersect}
\alias{xIntersect_}
\title{xIntersect}
\usage{
xIntersect(colls)

xIntersect_(...)
}
\arguments{
\item{colls}{a collection of collections. The collections to
   take the intersection of.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Get the set intersection of several collections.
}
\details{
The intersection of two sets is the set of elements present in
    both inputs.
}
\section{Type Signature}{

    |any| -> [any]
}

\section{Corner Cases}{

   Returns the empty list if \bold{coll} is length-zero.
}
\examples{

NULL

}
\seealso{
Other set_functions: \code{\link{xDuplicatesOf}},
  \code{\link{xDuplicatesOf_}}; \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsIn}},
  \code{\link{xIsIn_}}; \code{\link{xIsSubsetOf}},
  \code{\link{xIsSubsetOf_}}; \code{\link{xNotEmpty}},
  \code{\link{xNotEmpty_}}; \code{\link{xNotIn}},
  \code{\link{xNotIn_}}; \code{\link{xNotSubsetOf}},
  \code{\link{xNotSubsetOf_}}; \code{\link{xPowerSetOf}},
  \code{\link{xPowerSetOf_}}; \code{\link{xProdSetOf}},
  \code{\link{xProdSetOf_}}; \code{\link{xUnionOf}},
  \code{\link{xUnionOf_}}; \code{\link{xUniqueOf}},
  \code{\link{xUniqueOf_}}
}

