% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xInvoke.R
\name{xInvoke}
\alias{xInvoke}
\alias{xInvoke_}
\title{xInvoke}
\usage{
xInvoke(fn, val)

xInvoke_(fn, ...)
}
\arguments{
\item{fn}{a unary function. The function to call.}

\item{val}{an arbitrary value. The value to pass to the input funtion.}

\item{...}{see above.}
}
\value{
The return value of \bold{fn}.
}
\description{
Invoke a unary collection with a collection of arguments.
}
\details{
\bold{xInvoke} is an adaptor function somewhat like \bold{xApply}; in
    fact it was added to kea almost solely because the type signature of \bold{xApply}
    implies that this function is necessary and probably useful.

    \bold{xInvoke} allows a function that takes a single argument to be
    called with that argument, or be called with a variable number of
    arguments if \bold{xInvoke_} is used.

   It isn't usually necessary to call kea functions with \bold{xInvoke},
   since they exist with both variadic and non-variadic forms.
}
\section{Type Signature}{

    (|any| -> any) -> any -> any
}

\section{Corner Cases}{

    Fails when too many arguments are given to \bold{fn}.
}
\examples{

NULL
}
\seealso{
Other function_application_functions: \code{\link{xApply}},
  \code{\link{xApply_}}; \code{\link{xDeepMap}},
  \code{\link{xDeepMap_}}; \code{\link{xFlatMap}},
  \code{\link{xFlatMap_}}; \code{\link{xMap}},
  \code{\link{xMap_}}; \code{\link{xRelate}},
  \code{\link{xRelate_}}; \code{\link{xThread}},
  \code{\link{xThread_}}
}

