\name{xIsFalse}
\alias{xIsFalse}
\alias{xIsFalse...}
\title{xIsFalse}
\usage{
xIsFalse(coll)

xIsFalse...(...)
}
\arguments{
  \item{coll}{a collection.}

  \item{...}{see above.}
}
\value{
a vector of boolean values.
}
\description{
Is an element of a collection false?
}
\section{Variadic Functions

\link{ help_variadic }}{

}

