% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xIsIn}
\alias{xIsIn}
\alias{xIsIn_}
\title{xIsIn}
\usage{
xIsIn(val, coll)

xIsIn_(val, ...)
}
\arguments{
\item{val}{an arbitrary value. The value to test for membership
   in \bold{coll}.}

\item{coll}{a collection. The collection of elements to check against
   \bold{val}.}

\item{...}{see above.}
}
\value{
A boolean value.
}
\description{
Check if a collection contains a value.
}
\section{Type Signature}{

    any -> |any| -> &lt;logical>
}

\section{Corner Cases}{

   Various types of \bold{Na} are not-distinguished between.
   Type conversion is not carried out.
}
\examples{

# 1. A simple set membership example.

is_primary_colour <- is_primary_color <- val := {
	xIsIn_(val, 'red', 'green', 'blue')
}

is_primary_colour('red')

# True

is_primary_colour('magenta')

# False
}
\seealso{
Other set_functions: \code{\link{xDuplicatesOf}},
  \code{\link{xDuplicatesOf_}}; \code{\link{xInterOf}},
  \code{\link{xInterOf_}}; \code{\link{xIsSubset}},
  \code{\link{xIsSubset_}}; \code{\link{xNotIn}},
  \code{\link{xNotIn_}}; \code{\link{xNotSubset}},
  \code{\link{xNotSubset_}}; \code{\link{xPowerSetOf}},
  \code{\link{xPowerSetOf_}}; \code{\link{xProdSetOf}},
  \code{\link{xProdSetOf_}}; \code{\link{xUnionOf}},
  \code{\link{xUnionOf_}}; \code{\link{xUniqueOf}},
  \code{\link{xUniqueOf_}}
}

