% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xIsInfixOf}
\alias{xIsInfixOf}
\alias{xIsInfixOf_}
\title{xIsInfixOf}
\usage{
xIsInfixOf(coll1, coll2)

xIsInfixOf_(coll1, ...)
}
\arguments{
\item{coll1}{a collection. Then unbroken subsequence to test for.}

\item{coll2}{a collection. The collection to test for a subsequencematch.}

\item{...}{see above.}
}
\value{
A boolean value.
}
\description{
Is a collection an unbroken subsequence of another?
}
\section{Type Signature}{

    |any| -> |any| -> &lt;logical>
}

\section{Corner Cases}{

   If the first collection is longer than the second, False is returned.
   If either collection is length-zero, \bold{logical(0)} is returned.
}
\examples{

NULL
}

