% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xIsMatch}
\alias{xIsMatch}
\title{xIsMatch}
\usage{
xIsMatch(rexp, str)
}
\arguments{
\item{rexp}{a string. The regular expression to test against a string.}

\item{str}{a string. The string to check for a match.}

\item{...}{see above.}
}
\value{
A logical value.
}
\description{
Test if a string matches a regular expression.
}
\details{
\bold{xIsMatch} is roughly equivalent to \bold{grep} in
    base R.
}
\section{Type Signature}{

    |character| -> |character| -> <logical>
}

\section{Corner Cases}{

     the empty character vector is returned for when \bold{str} is empty.
}
\examples{

# 1. Test if a string matches a regular expression

xIsMatch('Over 90+[!]$', 'Over 9000!')

# True

# 2. Select matching strings from a list

text <-
"R version 3.1.1 (2014-07-10) -- 'Sock it to Me'
Copyright (C) 2014 The R Foundation for Statistical Computing
Platform: x86_64-pc-linux-gnu (64-bit)

R is free software and comes with ABSOLUTELY NO WARRANTY.
You are welcome to redistribute it under certain conditions.
Type 'license()' or 'licence()' for distribution details.

  Natural language support but running in an English locale

R is a collaborative project with many contributors.
Type 'contributors()' for more information and
'citation()' on how to cite R or R packages in publications."

x_(text) $ xToLines() $ x_Select(xIsMatch('licence|citation'))

# list(
#      "Type 'license()' or 'licence()' for distribution details.",
#      "'citation()' on how to cite R or R packages in publications.")
}
\seealso{
Other text_processing_functions: \code{\link{xAmend}};
  \code{\link{xCarve}}; \code{\link{xCharsOf}};
  \code{\link{xExplode}}; \code{\link{xFromChars}},
  \code{\link{xFromChars_}}; \code{\link{xFromLines}},
  \code{\link{xFromLines_}}; \code{\link{xFromWords}},
  \code{\link{xFromWords_}}; \code{\link{xImplode}},
  \code{\link{xImplode_}}; \code{\link{xNotMatch}};
  \code{\link{xRead}}; \code{\link{xToChars}};
  \code{\link{xToLines}}; \code{\link{xToWords}};
  \code{\link{xWrite}}
}

