\name{xIsNan}
\alias{xIsNan}
\title{xIsNan}
\usage{
  xIsNan(coll)
}
\arguments{
  \item{coll}{a collection}
}
\value{
  a list.
}
\description{
  Is an element in a collection NaN?
}
\section{Corner Cases}{
  returns True if coll is length-zero.
}
\seealso{
  Other collection_functions: \code{\link{xAsFunction}},
  \code{\link{xAssoc}}, \code{\link{xContains}},
  \code{\link{xDiffer}}, \code{\link{xDissoc}},
  \code{\link{xDrop}}, \code{\link{xFirst}},
  \code{\link{xFlatten}}, \code{\link{xFourth}},
  \code{\link{xInit}}, \code{\link{xInter}},
  \code{\link{xIsEmpty}}, \code{\link{xIsFalse}},
  \code{\link{xIsNa}}, \code{\link{xIsNull}},
  \code{\link{xIsTrue}}, \code{\link{xLast}},
  \code{\link{xLength}}, \code{\link{xName}},
  \code{\link{xNotFalse}}, \code{\link{xNotNa}},
  \code{\link{xNotNan}}, \code{\link{xNotTrue}},
  \code{\link{xPack}}, \code{\link{xRepeat}},
  \code{\link{xRest}}, \code{\link{xSecond}},
  \code{\link{xSegment}}, \code{\link{xShuffle}},
  \code{\link{xSplit}}, \code{\link{xTake}},
  \code{\link{xThird}}, \code{\link{xUnion}},
  \code{\link{xUnit}}
}

