% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xIsSuffixOf}
\alias{xIsSuffixOf}
\alias{xIsSuffixOf_}
\title{xIsSuffixOf}
\usage{
xIsSuffixOf(coll1, coll2)

xIsSuffixOf_(coll1, ...)
}
\arguments{
\item{coll1}{a collection. The suffix to test for.}

\item{coll2}{a collection. The collection to test for a suffix.}

\item{...}{see above.}
}
\value{
A boolean value.
}
\description{
Is a collection a suffix of another?
}
\section{Type Signature}{

    |any| -> |any| -> &lt;logical>
}

\section{Corner Cases}{

   If the first collection is longer than the second, False is returned.
   If either collection is length-zero, \bold{logical(0)} is returned.
}
\examples{

NULL
}

