\name{xIsTrue}
\alias{xIsTrue}
\alias{xIsTrue...}
\title{xIsTrue}
\usage{
xIsTrue(coll)

xIsTrue...(...)
}
\arguments{
  \item{coll}{a collection.}

  \item{...}{see above.}
}
\value{
A vector of boolean values.
}
\description{
Is an element of a collection true?
}
\section{Variadic Functions

\link{ help_variadic }}{

}

