% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xKeysOf.R
\name{xKeysOf}
\alias{xKeysOf}
\alias{xKeysOf_}
\title{xKeysOf}
\usage{
xKeysOf(coll)

xKeysOf_(...)
}
\arguments{
\item{coll}{a collection. The collection to get the names of.}

\item{...}{see above.}
}
\value{
a character vector.
}
\description{
Get the names of a collection.
}
\section{Type Signature}{

    |any| -> <character>
}

\section{Corner Cases}{

    Returns character(0) when \bold{coll} is length-zero.
}
\examples{

# 1. A simple example.

xKeysOf(list('fruit' = 'bannana', 'colour' = 'yellow'))
xKeysOf_('fruit' = 'bannana', 'colour' = 'yellow')

# c("fruit", "colour")
}
\seealso{
Other key_functions: \code{\link{xAddKeys}},
  \code{\link{xAddKeys_}}; \code{\link{xIndicesOf}},
  \code{\link{xIndicesOf_}}; \code{\link{xIndicesTo}};
  \code{\link{xLocate}}, \code{\link{xLocate_}};
  \code{\link{xUnzipIndices}},
  \code{\link{xUnzipIndices_}}; \code{\link{xUnzipKeys}},
  \code{\link{xUnzipKeys_}}; \code{\link{xValuesOf}},
  \code{\link{xValuesOf_}}; \code{\link{xZipKeys}},
  \code{\link{xZipKeys_}}
}

