% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xLastAs}
\alias{xLastAs}
\alias{xLastAs_}
\title{xLastAs}
\usage{
xLastAs(val, coll)

xLastAs_(val, ...)
}
\arguments{
\item{val}{an arbitrary value. The value to use.}

\item{coll}{a collection. The collection to set the
   last element of.}

\item{...}{see above.}
}
\value{
A collection.
}
\description{
Set the last value in a collection.
}
\section{Type Signature}{

    any -> |any| -> [any]
}

\section{Corner Cases}{

    Throws an error for collections shorter than one
    element.
}
\examples{

NULL
}

