% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xLastOf}
\alias{xLastOf}
\alias{xLastOf_}
\title{xLastOf}
\usage{
xLastOf(coll)

xLastOf_(...)
}
\arguments{
\item{coll}{a collection. The collection to return
   the last element of.}

\item{...}{see above.}
}
\value{
The value of the last element in \bold{coll}.
}
\description{
Return the last element in a collection.
}
\section{Type Signature}{

    |any| -> any
}

\section{Corner Cases}{

   Throws an error if \bold{coll} has less than one element; this is
   because any other corner case would violate the function's type-signature.
}
\examples{

# 1. extract a filename from a full path-name

my_path <- '/home/ryan/Code/kiwi.R/xMap.R'

x_(my_path) $ xExplode('/') $ x_LastOf()

# xMap.R
}
\seealso{
Other selection_functions: \code{\link{xAtCol}},
  \code{\link{xAtCol_}}; \code{\link{xAtKey}},
  \code{\link{xAtKey_}}; \code{\link{xAt}},
  \code{\link{xAt_}}; \code{\link{xDropWhile}},
  \code{\link{xDropWhile_}}; \code{\link{xDrop}},
  \code{\link{xDrop_}}; \code{\link{xFirstOf}},
  \code{\link{xFirstOf_}}; \code{\link{xFourthOf}},
  \code{\link{xFourthOf_}}; \code{\link{xInitOf}},
  \code{\link{xInitOf_}}; \code{\link{xMaxBy}},
  \code{\link{xMaxBy_}}; \code{\link{xMinBy}},
  \code{\link{xMinBy_}}; \code{\link{xRestOf}},
  \code{\link{xRestOf_}}; \code{\link{xSecondOf}},
  \code{\link{xSecondOf_}}; \code{\link{xSlice}},
  \code{\link{xSlice_}}; \code{\link{xTakeWhile}},
  \code{\link{xTakeWhile_}}; \code{\link{xTake}},
  \code{\link{xTake_}}; \code{\link{xThirdOf}},
  \code{\link{xThirdOf_}}
}

