\name{xLength}
\alias{xLength}
\title{xLength}
\usage{
  xLength(coll)
}
\arguments{
  \item{coll}{a collection}
}
\value{
  a nonnegative integer.
}
\description{
  Get the length of a collection
}
\section{Corner Cases}{
  returns 0 if \code{coll} is empty.
}
\seealso{
  Other collection_functions: \code{\link{xAsFunction}},
  \code{\link{xAssoc}}, \code{\link{xContains}},
  \code{\link{xDiffer}}, \code{\link{xDissoc}},
  \code{\link{xDrop}}, \code{\link{xFirst}},
  \code{\link{xFlatten}}, \code{\link{xFourth}},
  \code{\link{xInit}}, \code{\link{xInter}},
  \code{\link{xIsEmpty}}, \code{\link{xIsFalse}},
  \code{\link{xIsNa}}, \code{\link{xIsNan}},
  \code{\link{xIsNull}}, \code{\link{xIsTrue}},
  \code{\link{xLast}}, \code{\link{xName}},
  \code{\link{xNotFalse}}, \code{\link{xNotNa}},
  \code{\link{xNotNan}}, \code{\link{xNotTrue}},
  \code{\link{xPack}}, \code{\link{xRepeat}},
  \code{\link{xRest}}, \code{\link{xSecond}},
  \code{\link{xSegment}}, \code{\link{xShuffle}},
  \code{\link{xSplit}}, \code{\link{xTake}},
  \code{\link{xThird}}, \code{\link{xUnion}},
  \code{\link{xUnit}}
}

