% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xMaxBy}
\alias{xMaxBy}
\alias{xMaxBy_}
\title{xMaxBy}
\usage{
xMaxBy(fn, coll)

xMaxBy_(fn, ...)
}
\arguments{
\item{fn}{a unary function. The function to measure the size of
    an element.}

\item{coll}{a non-empty collection. The collection to find the
   largest value of.}

\item{...}{see above.}
}
\value{
An element from \bold{coll}.
}
\description{
Get the largest value in a collection according to a measure function.
}
\section{Type Signature}{

    (any -> &lt;number>) -> |any| -> any
}

\section{Corner Cases}{

   If an empty collection is given, an error is thrown. This is because
   is impossible to return an element from a list that has none.
}
\examples{
# 1. Select the largest row in a data-frame like struture
#    by an attribute.

GDP <- list(
	list('Saudi Arabia', 711050 * 10^6),
	list('Switzerland', 631050 * 10^6),
	list('Netherlands', 770067 * 10^6),
	list('Iran', 551588 * 10^6)
)

x_(GDP) $ x_MaxBy(xSecondOf)

# list("Netherlands", 7.70067e+11)
}
\seealso{
Other selection_functions: \code{\link{xAtCol}},
  \code{\link{xAtCol_}}; \code{\link{xAtKey}},
  \code{\link{xAtKey_}}; \code{\link{xAt}},
  \code{\link{xAt_}}; \code{\link{xDropWhile}},
  \code{\link{xDropWhile_}}; \code{\link{xDrop}},
  \code{\link{xDrop_}}; \code{\link{xFirstOf}},
  \code{\link{xFirstOf_}}; \code{\link{xFourthOf}},
  \code{\link{xFourthOf_}}; \code{\link{xInitOf}},
  \code{\link{xInitOf_}}; \code{\link{xLastOf}},
  \code{\link{xLastOf_}}; \code{\link{xMinBy}},
  \code{\link{xMinBy_}}; \code{\link{xPluck}},
  \code{\link{xPluck_}}; \code{\link{xRestOf}},
  \code{\link{xRestOf_}}; \code{\link{xSecondOf}},
  \code{\link{xSecondOf_}}; \code{\link{xSlice}},
  \code{\link{xSlice_}}; \code{\link{xTakeWhile}},
  \code{\link{xTakeWhile_}}; \code{\link{xTake}},
  \code{\link{xTake_}}; \code{\link{xThirdOf}},
  \code{\link{xThirdOf_}}
}

