\name{xNegate}
\alias{xNegate}
\title{Flip the sign of each number in a vector.}
\usage{
  xNegate(nums)
}
\arguments{
  \item{nums}{a vector of numbers.}
}
\value{
  a vector of number.
}
\description{
  xNegate
}
\section{Corner Cases}{
  If nums is empty then the unit of the collection is
  returned.
}

\section{Glossary}{
  function: a closure or primitive function.  A symbol or
  string refering to such a function may also be used.

  predicate: A function that returns a logical value
  (unary, unless stated otherwise).  A symbol or string
  refering to such a function may also be used.

  collection: A list, pairlist or vector of any type.

  arity: The number of arguments a function can accept.
  Variadic functions have an arity of +Infinity,

  number: a double, integer or complex number, unless
  stated otherwise.
}
\examples{
inst/examples/blank.R
}

