% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xNone}
\alias{xNone}
\alias{xNone_}
\title{xNone}
\usage{
xNone(pred, coll)

xNone_(pred, ...)
}
\arguments{
\item{pred}{a predicate. The function used to test each element of
   the input collection.}

\item{coll}{a collection. The collection to test each element of.}

\item{...}{see above.}
}
\value{
A boolean value.
}
\description{
Is a predicate true for none elements of a collection?
}
\section{Type Signature}{

    (any -> logical) -> |any| -> &lt;boolean>
}

\section{Corner Cases}{

   If coll is length zero then logical(0) is returned.
}
\examples{

NULL
}
\seealso{
Other quantifier_functions: \code{\link{xAll}},
  \code{\link{xAll_}}; \code{\link{xAny}},
  \code{\link{xAny_}}; \code{\link{xExists}},
  \code{\link{xExists_}}; \code{\link{xForall}},
  \code{\link{xForall_}}; \code{\link{xPoll}},
  \code{\link{xPoll_}}
}

