\name{xNotEmpty}
\alias{xNotEmpty}
\alias{xNotEmpty...}
\title{xNotEmpty}
\usage{
xNotEmpty(coll)

xNotEmpty...(...)
}
\arguments{
  \item{coll}{a collection. The value to test for having
  non-zero length.}

  \item{...}{see above.}
}
\value{
A boolean value.
}
\description{
Is a collection length-zero?
}
\examples{

NULL

}

