% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xNotIn}
\alias{xNotIn}
\alias{xNotIn_}
\title{xNotIn}
\usage{
xNotIn(val, coll)

xNotIn_(val, ...)
}
\arguments{
\item{val}{an arbitrary value. The value to test for
   non-membership in a collection.}

\item{coll}{a collection. The collection to test elements from.}

\item{...}{see above.}
}
\value{
A boolean value.
}
\description{
Check if a collection doesn't contain a value.
}
\section{Type Signature}{

    any -> |any| -> &lt;logical>
}

\section{Corner Cases}{

   The equality check used by \bold{xIsIn} has the same corner cases as \bold{xIs};
   equal integers and doubles are treated as different numbers, -0 is equal to +0.
   Returns \bold{logical(0)} when \bold{coll} is empty.
}

\section{Rcpp}{


    This function is written using C++, and should be quite performant.
}
\examples{

# 1. A toy example with the letters

xNotIn('a', LETTERS)

# False
}
\seealso{
Other set_functions: \code{\link{xDistinctOf}},
  \code{\link{xDistinctOf_}}; \code{\link{xDuplicatesOf}},
  \code{\link{xDuplicatesOf_}}; \code{\link{xIntersect}},
  \code{\link{xIntersect_}}; \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsIn}},
  \code{\link{xIsIn_}}; \code{\link{xIsSubsetOf}},
  \code{\link{xIsSubsetOf_}}; \code{\link{xNotEmpty}},
  \code{\link{xNotEmpty_}}; \code{\link{xNotSubsetOf}},
  \code{\link{xNotSubsetOf_}}; \code{\link{xPowerSetOf}},
  \code{\link{xPowerSetOf_}}; \code{\link{xProdSetOf}},
  \code{\link{xProdSetOf_}}; \code{\link{xUnite}},
  \code{\link{xUnite_}}
}

