% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xNotMatch}
\alias{xNotMatch}
\title{xNotMatch}
\usage{
xNotMatch(rexp, str)
}
\arguments{
\item{rexp}{a string. The regular expression to test against a string.}

\item{str}{a string. The string to check for a match.}

\item{...}{see above.}
}
\value{
A logical value.
}
\description{
Test if a string doesn't match a regular expression.
}
\section{Type Signature}{

    |character| -> |character| -> &lt;logical>
}

\section{Corner Cases}{

     the empty character vector is returned for when \bold{str} is empty.
}
\examples{

# 1. Test if a string doesn't match a regular expression

xNotMatch('Over 90+[!]$', 'Over 9000!')

# False
}
\seealso{
Other text_processing_functions: \code{\link{xAmend}};
  \code{\link{xCharsOf}}; \code{\link{xExplode}};
  \code{\link{xFromChars}}, \code{\link{xFromChars_}};
  \code{\link{xFromLines}}, \code{\link{xFromLines_}};
  \code{\link{xFromWords}}, \code{\link{xFromWords_}};
  \code{\link{xImplode}}, \code{\link{xImplode_}};
  \code{\link{xIsMatch}}; \code{\link{xRead}};
  \code{\link{xSliceString}}, \code{\link{xSliceString_}};
  \code{\link{xToChars}}; \code{\link{xToLines}};
  \code{\link{xToWords}}; \code{\link{xWrite}}
}

