% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xNotMember}
\alias{xNotMember}
\alias{xNotMember_}
\title{xNotMember}
\usage{
xNotMember(val, coll)

xNotMember_(val, ...)
}
\arguments{
\item{val}{an arbitrary value. The value to test for
   non-membership in a collection.}

\item{coll}{a collection. The collection to test elements from.}

\item{...}{see above.}
}
\value{
A boolean value.
}
\description{
Check if a collection doesn't contain a value.
}
\section{Type Signature}{

    any -> |any| -> &lt;logical>
}

\section{Corner Cases}{

   Various types of \code{Na} are not-distinguished between.
   Type conversion is not carried out.
}
\examples{

# 1. A toy example with the letters

xNotMember('a', LETTERS)

# False
}
\seealso{
Other set_functions: \code{\link{xDuplicatesOf}},
  \code{\link{xDuplicatesOf_}}; \code{\link{xInter}},
  \code{\link{xInter_}}; \code{\link{xIsMember}},
  \code{\link{xIsMember_}}; \code{\link{xIsSubset}},
  \code{\link{xIsSubset_}}; \code{\link{xNotSubset}},
  \code{\link{xNotSubset_}}; \code{\link{xPowerSetOf}},
  \code{\link{xPowerSetOf_}}; \code{\link{xProdSetOf}},
  \code{\link{xProdSetOf_}}; \code{\link{xReorder}},
  \code{\link{xReorder_}}; \code{\link{xUnionOf}},
  \code{\link{xUnionOf_}}; \code{\link{xUniqueOf}},
  \code{\link{xUniqueOf_}}
}

