% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xNotSubsetOf}
\alias{xNotSubsetOf}
\alias{xNotSubsetOf_}
\title{xNotSubsetOf}
\usage{
xNotSubsetOf(coll1, coll2)

xNotSubsetOf_(coll1, ...)
}
\arguments{
\item{coll1}{a collection. The set to test for non-membership in a superset.}

\item{coll2}{a collection. The superset to test.}

\item{...}{see above.}
}
\value{
A boolean value.
}
\description{
Test if a collection is not a subset of a second collection.
}
\section{Type Signature}{

    |any| -> |any| -> <logical>
}

\section{Corner Cases}{

   Returns logical(0) if \bold{coll1} or \bold{coll2} is length-zero.
}

\section{Rcpp}{


    This function is written using C++, and should be quite performant.
}
\examples{

NULL
}
\seealso{
Other set_functions: \code{\link{xDistinctOf}},
  \code{\link{xDistinctOf_}}; \code{\link{xDuplicatesOf}},
  \code{\link{xDuplicatesOf_}}; \code{\link{xIntersect}},
  \code{\link{xIntersect_}}; \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsIn}},
  \code{\link{xIsIn_}}; \code{\link{xIsSubsetOf}},
  \code{\link{xIsSubsetOf_}}; \code{\link{xNotEmpty}},
  \code{\link{xNotEmpty_}}; \code{\link{xNotIn}},
  \code{\link{xNotIn_}}; \code{\link{xPowerSetOf}},
  \code{\link{xPowerSetOf_}}; \code{\link{xProdSetOf}},
  \code{\link{xProdSetOf_}}; \code{\link{xUnite}},
  \code{\link{xUnite_}}
}

