\name{xNotTrue}
\alias{xNotTrue}
\alias{xNotTrue...}
\title{xNotTrue}
\usage{
xNotTrue(coll)

xNotTrue...(...)
}
\arguments{
  \item{coll}{a collection.}

  \item{...}{see above.}
}
\value{
A vector of boolean values.
}
\description{
Is an element of a collection not true?
}
\section{Corner Cases}{
  Returns logical(0) if \bold{coll} is length-zero.
}

