% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xOneOf}
\alias{xOneOf}
\alias{xOneOf...}
\title{xOneOf}
\usage{
xOneOf(coll)

xOneOf...(...)
}
\arguments{
\item{coll}{a collection. The collection to take a value from.}

\item{...}{see above.}
}
\value{
A value from \bold{coll}.
}
\description{
Select a random value from a collection.
}
\section{Corner Cases}{

     Returns the empty list if \bold{coll} is length-zero.
}
\examples{

NULL
}
\seealso{
Other reshaping_functions: \code{\link{xChop}},
  \code{\link{xChop...}}; \code{\link{xChunk}},
  \code{\link{xChunk...}}; \code{\link{xFlatten}},
  \code{\link{xFlatten...}}; \code{\link{xGroupBy}},
  \code{\link{xGroupBy...}}; \code{\link{xJoin}},
  \code{\link{xJoin...}}; \code{\link{xRecycle}},
  \code{\link{xRecycle...}}; \code{\link{xRepeat}},
  \code{\link{xRepeat...}}; \code{\link{xReverse}},
  \code{\link{xReverse...}}; \code{\link{xShuffle}},
  \code{\link{xShuffle...}}; \code{\link{xSortBy}},
  \code{\link{xSortBy...}}; \code{\link{xSplitAt}},
  \code{\link{xSplitAt...}}; \code{\link{xSplitWith}},
  \code{\link{xSplitWith...}}; \code{\link{xTabulate}},
  \code{\link{xTabulate...}}; \code{\link{xUnzipKeys}},
  \code{\link{xUnzipKeys...}}; \code{\link{xZipKeys}},
  \code{\link{xZipKeys...}}; \code{\link{xZip}},
  \code{\link{xZip...}}
}

