\name{xOrder}
\alias{xOrder}
\alias{xOrder...}
\title{xOrder}
\usage{
xOrder(nums)

xOrder...(...)
}
\arguments{
  \item{nums}{a vector of numbers. The numbers to rank in
  order of size.}

  \item{...}{see above.}
}
\value{
A vector of numbers.
}
\description{
Return a permutation of indices giving the order of an
integer vector.
}
\section{Corner Cases}{
  If nums is empty then \bold{integer(0)} is returned.
}
\examples{

NULL
}

