% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xParamsOf}
\alias{xParamsOf}
\title{xParamsOf}
\usage{
xParamsOf(fn)
}
\arguments{
\item{fn}{an arbitrary function or primitive function. The function to
  find the parametre names of.}
}
\value{
A character vector.
}
\description{
Get the parametre names of a function or primitive function.
}
\section{Type Signature}{

    function -> &lt;character>
}

\section{Corner Cases}{

	 If \bold{fn} is a primitive function a heuristic is used to obtain
	 its parametre names, which may not work for all functions. If a function
  has no parametres the empty character vector is returned.
}

\section{Stability - \bold{Uncertain}}{


This feature may be removed or altered significantly in future versions.
If you find this feature useful, please leave an issue at https://github.com/rgrannell1/kea/issues.
}
\examples{

# 1. Get the parametre names of a function.

xParamsOf(xMap)

# c('fn', 'coll')
}
\seealso{
Other parametre_functions: \code{\link{xArityOf}};
  \code{\link{xFix}}, \code{\link{xFix_}};
  \code{\link{xFormalsOf}}; \code{\link{xIsVariadic}};
  \code{\link{xSpread}}; \code{\link{xUnspread}}
}

