% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xPowerSetOf}
\alias{xPowerSetOf}
\alias{xPowerSetOf_}
\title{xPowerSetOf}
\usage{
xPowerSetOf(coll)

xPowerSetOf_(...)
}
\arguments{
\item{coll}{a collection. The collection to return the subsets of.}

\item{...}{see above.}
}
\value{
A list of lists.
}
\description{
Enumerate every way of subsetting a collection.
}
\details{
\bold{xPowerSetOf} generates the set of all subsets of a collection.
    This set has length \bold{2^length(coll)}, so inputs longer than
    twenty elements will take a very long time to compute.
}
\section{Type Signature}{

    |any| -> [[any]]
}

\section{Corner Cases}{

}
\examples{

# 1. generate all integer partitions of a number with distinct parts
#    The partition of a number n is a way or
#    writing n as the sum of positive integers.
#    In this case we are looking for partitions with distinct parts;
#    no number can be repeated.

# The partitions of 5 with distinct parts are
#
# 5
# 4 + 1
# 3 + 2

# These partitions are a subset of the product set
# of 1...5.

x_(1:5) $
xPowerSetOf() $
xSelect(nums := {
	(unlist \%then\% sum)(nums) == 5
})

# list(
#     list(2, 3),
#     list(1, 4),
#     list(5)
# )
}
\seealso{
Other combinatoric_functions: \code{\link{xChoose}},
  \code{\link{xChoose_}}; \code{\link{xCycle}},
  \code{\link{xCycle_}}; \code{\link{xProdSetOf}},
  \code{\link{xProdSetOf_}}

Other set_functions: \code{\link{xDuplicatesOf}},
  \code{\link{xDuplicatesOf_}}; \code{\link{xInterOf}},
  \code{\link{xInterOf_}}; \code{\link{xIsMember}},
  \code{\link{xIsMember_}}; \code{\link{xIsSubset}},
  \code{\link{xIsSubset_}}; \code{\link{xNotMember}},
  \code{\link{xNotMember_}}; \code{\link{xNotSubset}},
  \code{\link{xNotSubset_}}; \code{\link{xProdSetOf}},
  \code{\link{xProdSetOf_}}; \code{\link{xUnionOf}},
  \code{\link{xUnionOf_}}; \code{\link{xUniqueOf}},
  \code{\link{xUniqueOf_}}
}

