\name{xPredOf}
\alias{xPredOf}
\alias{xPredOf...}
\title{xPredOf}
\usage{
xPredOf(nums)

xPredOf...(...)
}
\arguments{
  \item{nums}{a vector of non-complex numbers.}

  \item{...}{see above.}
}
\value{
A vector of non-complex numbers.
}
\description{
Increment a vector of numbers.
}
\section{Corner Cases}{
  Returns the unit of the vector if \code{nums} is
  length-zero.
}
\examples{

NULL

}

