% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xPrepend}
\alias{xPrepend}
\alias{xPrepend_}
\title{xPrepend}
\usage{
xPrepend(val, coll)

xPrepend_(val, ...)
}
\arguments{
\item{val}{an arbitrary value. The value to prepend.}

\item{coll}{a collection. The collection to add the element to.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Add a value to the front of a collection.
}
\section{Type Signature}{

   any -> |any| -> [any]
}

\section{Corner Cases}{

   Returns a list containing \bold{val} when \bold{coll}
   is length-zero.
}
\examples{


# 1. generate the fibonacct numbers under 100.
#
# prepend to make accessing the previous two numbers easy.

fibs <-
	x__(1, 1) $
	xIterate(nums := {

		if (xFirstOf(nums) > 100) {
			Return(xSelect(x. < 100, nums))
		} else {
			xPrepend(xFirstOf(nums) + xSecondOf(nums), nums)
		}

	}) $
	x_Reverse()

# list(1, 1, 2, 3, 5, 8, 13, 21, 34, 55, 89)
}
\seealso{
Other reshaping_functions: \code{\link{xAppend}},
  \code{\link{xAppend_}}; \code{\link{xChop}},
  \code{\link{xChop_}}; \code{\link{xChunk}},
  \code{\link{xChunk_}}; \code{\link{xFlatten}},
  \code{\link{xFlatten_}}; \code{\link{xGroupBy}},
  \code{\link{xGroupBy_}}; \code{\link{xJoin}},
  \code{\link{xJoin_}}; \code{\link{xOneOf}},
  \code{\link{xOneOf_}}; \code{\link{xRepeat}},
  \code{\link{xRepeat_}}; \code{\link{xReverse}},
  \code{\link{xReverse_}}; \code{\link{xShuffle}},
  \code{\link{xShuffle_}}; \code{\link{xSortBy}},
  \code{\link{xSortBy_}}; \code{\link{xTabulate}},
  \code{\link{xTabulate_}}; \code{\link{xUnzipIndices}},
  \code{\link{xUnzipIndices_}}; \code{\link{xUnzipKeys}},
  \code{\link{xUnzipKeys_}}; \code{\link{xZipKeys}},
  \code{\link{xZipKeys_}}; \code{\link{xZip}},
  \code{\link{xZip_}}
}

