% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xProdBy}
\alias{xProdBy}
\alias{xProdBy_}
\title{xProdBy}
\usage{
xProdBy(fn, coll)

xProdBy_(fn, ...)
}
\arguments{
\item{fn}{a unary function. The function to measure the size of
    an element.}

\item{coll}{a non-empty collection. The collection to find the
   product of.}

\item{...}{see above.}
}
\value{
A double.
}
\description{
Get the product of a collection according to a measure function.
}
\section{Type Signature}{

    (any -> &lt;number>) -> |any| -> &lt;number>
}

\section{Corner Cases}{

   If an empty collection is given, numeric(0) is returned.
}
\examples{

# 1. compounding the numbers themselves.

xSumBy(xIdentity, 1:10)

# 3628800
}

