% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xRead}
\alias{xRead}
\title{xRead}
\usage{
xRead(str)
}
\arguments{
\item{str}{a length-one character vector. The path to
   import files from.}
}
\value{
A length-one character vector.
}
\description{
Read the contents of a file as a single string.
}
\section{Type Signature}{

    |character| -> &lt;character>
}

\section{Corner Cases}{

   xRead ignores missing terminal newlines. If \bold{str} is
   is length-zero the empty character-vector is returned.
}
\examples{

# 1. Read the first line of kiwi's description.

x_(system.file(package = 'kiwi', 'DESCRIPTION')) $
xRead() $ xToLines() $ x_Take(1)

# list("Package: kiwi")
}
\seealso{
Other inpure_functions: \code{\link{xDelay}};
  \code{\link{xDo}}, \code{\link{xDo_}};
  \code{\link{xExecute}}; \code{\link{xStopwatch}};
  \code{\link{xVersion}}; \code{\link{xWrite}}

Other text_processing_functions: \code{\link{xExplode}};
  \code{\link{xFromChars}}, \code{\link{xFromChars_}};
  \code{\link{xFromLines}}, \code{\link{xFromLines_}};
  \code{\link{xFromWords}}, \code{\link{xFromWords_}};
  \code{\link{xImplode}}, \code{\link{xImplode_}};
  \code{\link{xIsMatch}}; \code{\link{xNotMatch}};
  \code{\link{xSliceString}}, \code{\link{xSliceString_}};
  \code{\link{xToChars}}; \code{\link{xToLines}};
  \code{\link{xToWords}}; \code{\link{xWrite}}
}

