% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xReadChars}
\alias{xReadChars}
\title{xReadChars}
\usage{
xReadChars(str)
}
\arguments{
\item{str}{a length-one character vector. The path to
   import files from.}
}
\value{
A character vector, with one or more elements.
}
\description{
Import each line in a file to an element in a character vector.
}
\section{Type Signature}{

    |character| -> &lt;character>
}

\section{Corner Cases}{

   xReadChars ignores missing terminal newlines.
}
\examples{

# 1. Read the first 100 letters of arrow's description.

x_(system.file(package = 'arrow', 'DESCRIPTION')) $
xReadChars() $
x_Take(100)

# list("P", "a", "c", "k", "a", "g", "e", ":", " ", "a")
}
\seealso{
Other inpure_functions: \code{\link{xDelay}};
  \code{\link{xDo}}, \code{\link{xDo_}};
  \code{\link{xExecute}}; \code{\link{xReadLines}};
  \code{\link{xReadWords}}; \code{\link{xStopwatch}};
  \code{\link{xVersion}}; \code{\link{xWriteChars}};
  \code{\link{xWriteLines}}; \code{\link{xWriteWords}}

Other text_processing_functions: \code{\link{xExplode}};
  \code{\link{xFromChars}}, \code{\link{xFromChars_}};
  \code{\link{xFromLines}}, \code{\link{xFromLines_}};
  \code{\link{xFromWords}}, \code{\link{xFromWords_}};
  \code{\link{xImplode}}, \code{\link{xImplode_}};
  \code{\link{xIsMatch}}; \code{\link{xNotMatch}};
  \code{\link{xReadLines}}; \code{\link{xReadWords}};
  \code{\link{xSliceString}}, \code{\link{xSliceString_}};
  \code{\link{xToChars}}; \code{\link{xToLines}};
  \code{\link{xToWords}}; \code{\link{xWriteChars}};
  \code{\link{xWriteLines}}; \code{\link{xWriteWords}}
}

