% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xRejectEmpty}
\alias{xPack}
\alias{xPack...}
\alias{xRejectEmpty}
\alias{xRejectEmpty...}
\title{xRejectEmpty}
\usage{
xRejectEmpty(coll)

xRejectEmpty...(...)

xPack(coll)

xPack...(...)
}
\arguments{
\item{coll}{a collection. The collection to remove
   length-zero values from.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Remove the empty values from a collection.
}
\examples{

NULL

}
\seealso{
Other filtering_functions: \code{\link{xPartition}},
  \code{\link{xPartition...}}; \code{\link{xRejectNan}};
  \code{\link{xRejectNa}}, \code{\link{xRejectNa...}};
  \code{\link{xRejectNull}}, \code{\link{xRejectNull...}};
  \code{\link{xReject}}, \code{\link{xReject...}};
  \code{\link{xSelect}}, \code{\link{xSelect...}}
}

