\name{xRejectNa}
\alias{xRejectNa}
\alias{xRejectNa...}
\title{xRejectNa}
\usage{
xRejectNa(coll)

xRejectNa...(...)
}
\arguments{
  \item{coll}{a collection. The collection to remove na
  values from.}

  \item{...}{see above.}
}
\value{
A list.
}
\description{
Remove the Na values from a collection.
}
\examples{

NULL

}
\seealso{
Other filtering_functions: \code{\link{xPack}},
\code{\link{xPack...}}, \code{\link{xRejectEmpty}},
\code{\link{xRejectEmpty...}}; \code{\link{xPartition}};
\code{\link{xRejectNan}}; \code{\link{xRejectNull}},
\code{\link{xRejectNull...}}; \code{\link{xReject}},
\code{\link{xReject...}}; \code{\link{xSelect}},
\code{\link{xSelect...}}
}

