% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xRejectNan}
\alias{xRejectNan}
\alias{xRejectNan_}
\title{xRejectNan}
\usage{
xRejectNan(coll)

xRejectNan_(...)
}
\arguments{
\item{coll}{a collection. The collection to remove
   NaN values from.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Remove the NaN values from a collection.
}
\examples{

NULL

}
\seealso{
Other filtering_functions: \code{\link{xPack}},
  \code{\link{xPack_}}, \code{\link{xRejectEmpty}},
  \code{\link{xRejectEmpty_}}; \code{\link{xPartition}},
  \code{\link{xPartition_}}; \code{\link{xRejectNa}},
  \code{\link{xRejectNa_}}; \code{\link{xRejectNull}},
  \code{\link{xRejectNull_}}; \code{\link{xReject}},
  \code{\link{xReject_}}; \code{\link{xSelect}},
  \code{\link{xSelect_}}
}

