\name{xRejectNull}
\alias{xRejectNull}
\alias{xRejectNull...}
\title{xRejectNull}
\usage{
xRejectNull(coll)

xRejectNull...(...)
}
\arguments{
  \item{coll}{a collection. The collection to remove null
  values from.}

  \item{...}{see above.}
}
\value{
A list.
}
\description{
Remove the null values from a collection.
}
\examples{

NULL

}
\seealso{
Other filtering_functions: \code{\link{xPack}},
\code{\link{xPack...}}, \code{\link{xRejectEmpty}},
\code{\link{xRejectEmpty...}}; \code{\link{xPartition}};
\code{\link{xRejectNan}}; \code{\link{xRejectNa}},
\code{\link{xRejectNa...}}; \code{\link{xReject}},
\code{\link{xReject...}}; \code{\link{xSelect}},
\code{\link{xSelect...}}
}

