% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xRelate}
\alias{xRelate}
\alias{xRelate_}
\title{xRelate}
\usage{
xRelate(fn, coll)

xRelate_(fn, ...)
}
\arguments{
\item{fn}{a unary function. The function to apply to each element of the collection.}

\item{coll}{a collection. The collection to be mapped over.}

\item{...}{see above.}
}
\value{
A list of two-element lists.
}
\description{
Map a function over each element of a collection, keeping both the element and the result.
}
\section{Type Signature}{

    (any -> any) -> |any| -> [[any, any]]
}

\section{Corner Cases}{

   Returns the empty list if \bold{coll} is length-zero.
}
\examples{

NULL
}
\seealso{
Other function_application_functions: \code{\link{xApply}},
  \code{\link{xApply_}}; \code{\link{xDeepMap}},
  \code{\link{xDeepMap_}}; \code{\link{xFlatMap}},
  \code{\link{xFlatMap_}}; \code{\link{xMap}},
  \code{\link{xMap_}}; \code{\link{xSpread}};
  \code{\link{xThread}}, \code{\link{xThread_}};
  \code{\link{xUnspread}}
}

