\name{xRemoveEmpty}
\alias{xRemoveEmpty}
\alias{xRemoveEmpty...}
\title{xRemoveEmpty}
\usage{
xRemoveEmpty(coll)

xRemoveEmpty...(...)
}
\arguments{
  \item{coll}{a collection.}

  \item{...}{see above.}
}
\value{
A list.
}
\description{
Remove the null values from a collection.
}
\section{Variadic Functions

\link{ help_variadic }}{

}

