\name{xRemoveNan}
\alias{xRemoveNan}
\title{xRemoveNan}
\usage{
xRemoveNan(coll)
}
\arguments{
  \item{coll}{a collection. The collection to remove NaN
  values from.}

  \item{...}{see above.}
}
\value{
A list.
}
\description{
Remove the NaN values from a collection.
}
\examples{

NULL

}
\seealso{
Other filtering_functions: \code{\link{xPartition}};
\code{\link{xReject}}, \code{\link{xReject...}};
\code{\link{xRemoveEmpty}}, \code{\link{xRemoveEmpty...}};
\code{\link{xRemoveNa}}; \code{\link{xRemoveNull}},
\code{\link{xRemoveNull...}}; \code{\link{xSelect}},
\code{\link{xSelect...}}
}

