\name{xRemoveNull}
\alias{xRemoveNull}
\alias{xRemoveNull...}
\title{xRemoveNull}
\usage{
xRemoveNull(coll)

xRemoveNull...(...)
}
\arguments{
  \item{coll}{a collection.}

  \item{...}{see above.}
}
\value{
A list.
}
\description{
Remove the null values from a collection.
}
\section{Variadic Functions

\link{ help_variadic }}{

}

