% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xReorder}
\alias{xReorder}
\alias{xReorder_}
\title{xReorder}
\usage{
xReorder(nums, colls)

xReorder_(nums, ...)
}
\arguments{
\item{nums}{a collection of whole numbers. The indices
   by which to permute the input collections.}

\item{colls}{several collections of equal length to \bold{nums}. The
   collections to permute.}

\item{...}{see above.}
}
\value{
A list of collections.
}
\description{
Permute a collection of collections using indices.
}
\details{
\bold{xReorder} allows several collections to be rearranged
   in the same way simultaneously.

   \code{colls <- list(ids = list(15, 1, 12), ages = list(17, 29, 24), weights = list(80, 82, 76))}

   \code{xReorder(xOrderOf( xFirstOf(colls) ), colls)}
}
\section{Type Signature}{

    <number> -> ||any|| -> ||any||
}

\section{Corner Cases}{

   returns the empty list is \bold{nums} is length-zero.
}
\examples{

NULL

}
\seealso{
Other combinatoric_functions: \code{\link{xChoose}},
  \code{\link{xChoose_}}; \code{\link{xCycle}},
  \code{\link{xCycle_}}; \code{\link{xPowerSetOf}},
  \code{\link{xPowerSetOf_}}; \code{\link{xProdSetOf}},
  \code{\link{xProdSetOf_}}

Other set_functions: \code{\link{xDuplicatesOf}},
  \code{\link{xDuplicatesOf_}}; \code{\link{xInter}},
  \code{\link{xInter_}}; \code{\link{xIsMember}},
  \code{\link{xIsMember_}}; \code{\link{xIsSubset}},
  \code{\link{xIsSubset_}}; \code{\link{xNotMember}},
  \code{\link{xNotMember_}}; \code{\link{xNotSubset}},
  \code{\link{xNotSubset_}}; \code{\link{xPowerSetOf}},
  \code{\link{xPowerSetOf_}}; \code{\link{xProdSetOf}},
  \code{\link{xProdSetOf_}}; \code{\link{xUnionOf}},
  \code{\link{xUnionOf_}}; \code{\link{xUniqueOf}},
  \code{\link{xUniqueOf_}}
}

