\name{xRowUnit}
\alias{xRowUnit}
\title{$ xRowUnit}
\usage{
x_(  ) $ xRowUnit()
}
\value{
An empty matrix or data.frame.
}
\description{
Return the concatenative unit for a matrix or data.frame
row.
}
\seealso{
Other exclusively_methods: \code{\link{xByColnames}};
\code{\link{xByCols}}; \code{\link{xByLevels}},
\code{\link{xByLevels}}; \code{\link{xByRows}};
\code{\link{xColUnit}}; \code{\link{xElemsByCols}};
\code{\link{xElemsByRows}}; \code{\link{xExecute}};
\code{\link{xTap}}; \code{\link{xTranspose}}
}

