% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xSecondAs}
\alias{xSecondAs}
\alias{xSecondAs_}
\title{xSecondAs}
\usage{
xSecondAs(val, coll)

xSecondAs_(val, ...)
}
\arguments{
\item{val}{an arbitrary value. The value to use.}

\item{coll}{a collection. The collection to set the
   second element of.}

\item{...}{see above.}
}
\value{
A collection.
}
\description{
Set the second value in a collection.
}
\section{Type Signature}{

    any -> |any| -> [any]
}

\section{Corner Cases}{

    Throws an error for collections shorter than two
    elements.
}
\examples{

NULL
}

