% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xSpread}
\alias{xSpread}
\title{xSpread}
\usage{
xSpread(fn)
}
\arguments{
\item{fn}{a unary function. The function to convert
   to a variadic function.}
}
\value{
A variadic function.
}
\description{
Return a variadic function that passes its arguments
as a list to its underlying function.
}
\details{
\bold{xSpread} takes a unary function and
   returns a function with ellipsis parametres.
   The returned function passes its ellipsis arguments
   as a single list to the underlying function.
}
\section{Type Signature}{

   ([any] -> any) -> (...any -> any)
}

\section{Corner Cases}{

}
\examples{

# 1.
# Convert a function that takes a single value as an argument to
# a function that takes many arguments.
#

summate <- xSpread(xFold("+", 0) )

summate(1, 2, 3, 4, 5, 6)

# 21.
#

}
\seealso{
Other function_application_functions: \code{\link{xApply}},
  \code{\link{xApply_}}; \code{\link{xDeepMap}},
  \code{\link{xDeepMap_}}; \code{\link{xFlatMap}},
  \code{\link{xFlatMap_}}; \code{\link{xMap}},
  \code{\link{xMap_}}; \code{\link{xRelate}},
  \code{\link{xRelate_}}; \code{\link{xThread}},
  \code{\link{xThread_}}; \code{\link{xUnspread}}

Other function_modifying_functions: \code{\link{\%and\%}},
  \code{\link{\%or\%}}, \code{\link{xLift}},
  \code{\link{xLift_}}; \code{\link{\%then\%}},
  \code{\link{xCompose}}, \code{\link{xCompose_}};
  \code{\link{xCapture}}, \code{\link{xK}};
  \code{\link{xFix}}, \code{\link{xFix_}};
  \code{\link{xJuxtapose}}, \code{\link{xJuxtapose_}};
  \code{\link{xNegate}}; \code{\link{xThread}},
  \code{\link{xThread_}}; \code{\link{xUnspread}}

Other parametre_functions: \code{\link{xArityOf}};
  \code{\link{xFix}}, \code{\link{xFix_}};
  \code{\link{xFormalsOf}}; \code{\link{xParamsOf}};
  \code{\link{xUnspread}}
}

