\name{xStopwatch}
\alias{xStopwatch}
\title{xStopwatch}
\usage{
xStopwatch(num)
}
\arguments{
  \item{num}{a positive number. The number of seconds the
  function should return true for.}
}
\value{
A variadic function that ignores its arguments.
}
\description{
Create a function that returns true for a preset time after
creation.
}
\examples{

NULL

}
\seealso{
Other inpure_functions: \code{\link{xDelay}};
\code{\link{xDo}}, \code{\link{xDo...}};
\code{\link{xExecute}}; \code{\link{xVersion}}

Other time_functions: \code{\link{xDelay}};
\code{\link{xLimit}}
}

