\name{xSucc}
\alias{xSucc}
\title{xSucc}
\usage{
  xSucc(nums)
}
\arguments{
  \item{nums}{a collection of numbers.}
}
\value{
  a vector of numbers.
}
\description{
  Decrement a collection of numbers.
}
\section{Corner Cases}{
  If \code{nums} is empty then the unit of that vector is
  returned.
}

\section{Glossary}{
  function: a closure or primitive function.  A symbol or
  string refering to such a function may also be used.

  predicate: A function that returns a logical value
  (unary, unless stated otherwise).  A symbol or string
  refering to such a function may also be used.

  collection: A list, pairlist or vector of any type.

  arity: The number of arguments a function can accept.
  Variadic functions have an arity of +Infinity,

  number: a double, integer or complex number, unless
  stated otherwise.
}
\examples{
inst/examples/blank.R
}

