% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xSumBy}
\alias{xSumBy}
\alias{xSumBy_}
\title{xSumBy}
\usage{
xSumBy(fn, coll)

xSumBy_(fn, ...)
}
\arguments{
\item{fn}{a unary function. The function to measure the size of
    an element.}

\item{coll}{a collection. The collection to find the sum of.}

\item{...}{see above.}
}
\value{
A double.
}
\description{
Get the sum of a collection according to a measure function.
}
\section{Type Signature}{

    (any -> <number>) -> |any| -> <number>
}

\section{Corner Cases}{

   If an empty collection is given, numeric(0) is returned.
}
\examples{

# 1. summing the numbers themselves.

xSumBy(xIdentity, 1:10)

# 55

# 2. get the total characters in a character vector

xSumBy_(nchar, 'this', 'is', 'a', 'sentence.')
16
}
\seealso{
Other math_functions: \code{\link{xLenOf}},
  \code{\link{xLenOf_}}; \code{\link{xMaxBy}},
  \code{\link{xMaxBy_}}; \code{\link{xMeanBy}},
  \code{\link{xMeanBy_}}; \code{\link{xMinBy}},
  \code{\link{xMinBy_}}; \code{\link{xOrderOf}},
  \code{\link{xOrderOf_}}; \code{\link{xPoll}},
  \code{\link{xPoll_}}; \code{\link{xProdBy}},
  \code{\link{xProdBy_}}; \code{\link{xRank}},
  \code{\link{xRank_}}
}

