% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xTabulate}
\alias{xTabulate}
\alias{xTabulate...}
\title{xTabulate}
\usage{
xTabulate(coll)

xTabulate...(...)
}
\arguments{
\item{coll}{a collection. The values to find the frequency of.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Tabulate a collection into pairs of value:frequency lists.
}
\details{
\bold{xTabulate} is superficially similar to the base
    function \bold{table}: given a collection that may or
    may not contain duplicates, it calculates the frequencies
    of each unique element.

    \code{xTabulate(c('y', 'n', 'y', 'y', 'n'))}

    \code{list(list("y", 3), list("n", 2))}

    The result of the tabulation is unsorted for efficiency;
    if sorting is required \bold{xSortBy} can be used.
}
\section{Corner Cases}{

   If \bold{coll} is length-zero the empty list is returned.
}
\examples{

# a subset of the 1999 impeachment vote

raw_clinton_vote <- "
R	Guilty		Guilty
D	NotGuilty	NotGuilty
R	Guilty		Guilty
R	Guilty		Guilty
D	NotGuilty	NotGuilty
D	NotGuilty	NotGuilty
R	Guilty		Guilty
D	NotGuilty	NotGuilty
D	NotGuilty	NotGuilty
R	Guilty		Guilty
D	NotGuilty	NotGuilty
D	NotGuilty	NotGuilty
R	Guilty		Guilty
D	NotGuilty	NotGuilty
R	Guilty		Guilty
R	Guilty		Guilty
D	NotGuilty	NotGuilty
R	Guilty		Guilty
R	NotGuilty	NotGuilty
D	NotGuilty	NotGuilty
R	Guilty		Guilty
R	NotGuilty	NotGuilty
D	NotGuilty	NotGuilty
R	Guilty		Guilty
R	Guilty		Guilty
D	NotGuilty	NotGuilty"

clinton_vote <-
	x_(raw_clinton_vote) $ xToLines() $ xMap(xToWords) $
	xZip() $ xAddKeys(c('party', 'perjury', 'obstruction'))

# 1.
# Tabulate the overall results for the perjury vote

clinton_vote $ xAtKey('perjury') $ x_Tabulate()

# list( list("Guilty", 12), list("NotGuilty", 14) )

# 2. # Tabulate the overall results for the obstruction vote

clinton_vote $ xAtKey('obstruction') $ x_Tabulate()

# list( list("Guilty", 12), list("NotGuilty", 14) )
}
\seealso{
Other reshaping_functions: \code{\link{xChop}},
  \code{\link{xChop...}}; \code{\link{xChunk}},
  \code{\link{xChunk...}}; \code{\link{xFlatten}},
  \code{\link{xFlatten...}}; \code{\link{xGroupBy}},
  \code{\link{xGroupBy...}}; \code{\link{xJoin}},
  \code{\link{xJoin...}}; \code{\link{xOneOf}},
  \code{\link{xOneOf...}}; \code{\link{xRecycle}},
  \code{\link{xRecycle...}}; \code{\link{xRepeat}},
  \code{\link{xRepeat...}}; \code{\link{xReverse}},
  \code{\link{xReverse...}}; \code{\link{xShuffle}},
  \code{\link{xShuffle...}}; \code{\link{xSortBy}},
  \code{\link{xSortBy...}}; \code{\link{xSplitAt}},
  \code{\link{xSplitAt...}}; \code{\link{xSplitWith}},
  \code{\link{xSplitWith...}}; \code{\link{xUnzipKeys}},
  \code{\link{xUnzipKeys...}}; \code{\link{xZipKeys}},
  \code{\link{xZipKeys...}}; \code{\link{xZip}},
  \code{\link{xZip...}}
}

