% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xTake}
\alias{xTake}
\alias{xTake...}
\title{xTake}
\usage{
xTake(num, coll)

xTake...(num, ...)
}
\arguments{
\item{num}{a nonnegative whole number. The number of elements to
     return from the input collection.}

\item{coll}{a collection. The collection to subset.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Take several elements from the head of a collection.
}
\section{Corner Cases}{

   If \bold{coll} is empty the empty list is returned.
}
\examples{

NULL

}
\seealso{
Other selection_functions: \code{\link{xAtCol}},
  \code{\link{xAtCol...}}; \code{\link{xAt}},
  \code{\link{xAt...}}; \code{\link{xDropWhile}},
  \code{\link{xDropWhile...}}; \code{\link{xDrop}},
  \code{\link{xDrop...}}; \code{\link{xFirstOf}},
  \code{\link{xFirstOf...}}; \code{\link{xFourthOf}},
  \code{\link{xFourthOf...}}; \code{\link{xInitOf}},
  \code{\link{xInitOf...}}; \code{\link{xLastOf}},
  \code{\link{xLastOf...}}; \code{\link{xPluck}},
  \code{\link{xPluck...}}; \code{\link{xRestOf}},
  \code{\link{xRestOf...}}; \code{\link{xSecondOf}},
  \code{\link{xSecondOf...}}; \code{\link{xSlice}},
  \code{\link{xSlice...}}; \code{\link{xTakeWhile}},
  \code{\link{xTakeWhile...}}; \code{\link{xThirdOf}},
  \code{\link{xThirdOf...}}
}

