% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xThirdAs}
\alias{xThirdAs}
\alias{xThirdAs_}
\title{xThirdAs}
\usage{
xThirdAs(val, coll)

xThirdAs_(val, ...)
}
\arguments{
\item{val}{an arbitrary value. The value to set.}

\item{coll}{a collection. The collection to set the
   third element of.}

\item{...}{see above.}
}
\value{
A collection.
}
\description{
Set the third value in a collection.
}
\section{Type Signature}{

    any -> |any| -> [any]
}

\section{Corner Cases}{

    Throws an error for collections shorter than three
    elements.
}
\examples{

NULL
}

