\name{xTranspose}
\alias{xTranspose}
\title{$ xTranspose}
\usage{
x_(  ) $ xTranspose()
}
\value{
an arrow object containing a matrix.
}
\description{
Transpose a matrix.
}
\seealso{
Other exclusively_methods: \code{\link{xByColnames}};
\code{\link{xByCols}}; \code{\link{xByLevels}},
\code{\link{xByLevels}}; \code{\link{xByRows}};
\code{\link{xColUnit}}; \code{\link{xElemsByCols}};
\code{\link{xElemsByRows}}; \code{\link{xExecute}};
\code{\link{xRowUnit}}; \code{\link{xTap}}
}

