\name{xTruth}
\alias{xTruth}
\title{xTruth}
\usage{
xTruth(...)
}
\arguments{
  \item{...}{arguments to be dropped.}
}
\value{
true
}
\description{
A function that always returns true.
}
\examples{

# 1. Using truth as an identity
#    with xSelect.

xSelect(xTruth, 1:10)
}
\seealso{
Other logical_constant_functions: \code{\link{xFalsity}};
\code{\link{xIrrelevance}}
}

