\name{xUnchars}
\alias{xUnchars}
\title{Collapase a character vector with empty strs as delimiters.}
\usage{
  xUnchars(...)
}
\arguments{
  \item{...}{several character vectors.}
}
\value{
  a length-one character vector.
}
\description{
  Collapase a character vector with empty strs as
  delimiters.
}

