% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xUnionOf}
\alias{xUnionOf}
\alias{xUnionOf...}
\title{xUnionOf
``
Get the set union of several collections.}
\usage{
xUnionOf(colls)

xUnionOf...(...)
}
\arguments{
\item{colls}{a collection of collections. The collections
   to take the union of.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
xUnionOf
``
Get the set union of several collections.
}
\details{
\code{xUnionOf} returns a list of the unique
   elements in the combination of each collection
   in \code{colls}.
}
\section{Corner Cases}{

   Returns the empty list if \code{colls} is length-zero.
}
\examples{

NULL

}
\seealso{
Other set_functions: \code{\link{xDuplicatesOf}},
  \code{\link{xDuplicatesOf...}}; \code{\link{xInter}},
  \code{\link{xInter...}}; \code{\link{xIsMember}},
  \code{\link{xIsMember...}}; \code{\link{xIsSubset}},
  \code{\link{xIsSubset...}}; \code{\link{xNotMember}},
  \code{\link{xNotMember...}}; \code{\link{xReorder}},
  \code{\link{xReorder...}}; \code{\link{xPowerSetOf}},
  \code{\link{xPowerSetOf...}}; \code{\link{xProdSetOf}},
  \code{\link{xProdSetOf...}}; \code{\link{xUniqueOf}},
  \code{\link{xUniqueOf...}}
}

