\name{xUnit}
\alias{xUnit}
\title{xUnit}
\usage{
xUnit(coll)
}
\arguments{
  \item{coll}{an list, pairlist, or vector of any length.}
}
\value{
Returns null if \bold{coll} is a pairlist, a typed vector
of length zero if \bold{coll} is a vector, and the empty
list if \bold{coll} is a list.
}
\description{
Return the empty version of a collection.
}

