% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xUnite}
\alias{xUnite}
\alias{xUnite_}
\title{xUnite
``
Get the set union of two collections.}
\usage{
xUnite(coll1, coll2)

xUnite_(coll1, ...)
}
\arguments{
\item{coll1}{a collection. The first collection to get elements from.}

\item{coll2}{a collection. The second collection to get elements from.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
xUnite
``
Get the set union of two collections.
}
\details{
\code{xUnite} returns a list of the unique
   elements in both \bold{coll1} and \bold{coll2}.
}
\section{Type Signature}{

    ||any|| -> [any]
}

\section{Corner Cases}{

   Returns the empty list if \code{coll1} or \bold{coll2} is length-zero.
}
\examples{

NULL
}
\seealso{
Other set_functions: \code{\link{xDistinctOf}},
  \code{\link{xDistinctOf_}}; \code{\link{xDuplicatesOf}},
  \code{\link{xDuplicatesOf_}}; \code{\link{xIntersect}},
  \code{\link{xIntersect_}}; \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsIn}},
  \code{\link{xIsIn_}}; \code{\link{xIsSubsetOf}},
  \code{\link{xIsSubsetOf_}}; \code{\link{xNotEmpty}},
  \code{\link{xNotEmpty_}}; \code{\link{xNotIn}},
  \code{\link{xNotIn_}}; \code{\link{xNotSubsetOf}},
  \code{\link{xNotSubsetOf_}}; \code{\link{xPowerSetOf}},
  \code{\link{xPowerSetOf_}}; \code{\link{xProdSetOf}},
  \code{\link{xProdSetOf_}}
}

