\name{xUnlines}
\alias{xUnlines}
\title{xUnlines}
\usage{
xUnlines(strs)
}
\arguments{
  \item{strs}{a character vector.}
}
\value{
a length-one character vector.
}
\description{
Collapase a character vector with newlines as delimiters.
}
\seealso{
Other character_vector_functions: \code{\link{xChars}};
\code{\link{xCollapse}}; \code{\link{xLines}};
\code{\link{xSplitString}}; \code{\link{xSubString}};
\code{\link{xUnchars}}; \code{\link{xUnwords}};
\code{\link{xWords}}
}

