\name{xUnwords}
\alias{xUnwords}
\title{xUnwords}
\usage{
xUnwords(strs)
}
\arguments{
  \item{strs}{an arbitrary number of character vectors.}
}
\value{
a length-one character vector.
}
\description{
Collapase a character vector with spaces as delimiters.
}
\seealso{
Other character_vector_functions: \code{\link{xChars}};
\code{\link{xCollapse}}; \code{\link{xLines}};
\code{\link{xSplitString}}; \code{\link{xSubString}};
\code{\link{xUnchars}}; \code{\link{xUnlines}};
\code{\link{xWords}}
}

