% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xUnzipIndices}
\alias{xUnzipIndices}
\alias{xUnzipIndices_}
\title{xUnzipIndices}
\usage{
xUnzipIndices(coll)

xUnzipIndices_(...)
}
\arguments{
\item{coll}{a collection. The collection to split into index, element pairs.}

\item{...}{see above.}
}
\value{
A named list.
}
\description{
Split a named collection into a list of name: value list pairs.
}
\section{Type Signature}{

    |any| -> [[ <number>, ...any ]]
}

\section{Corner Cases}{

     Returns \code{list()} if \code{coll} is length-zero.
}
\examples{

NULL
}
\seealso{
Other key_functions: \code{\link{xAddKeys}},
  \code{\link{xAddKeys_}}; \code{\link{xPluck}},
  \code{\link{xPluck_}}; \code{\link{xUnzipKeys}},
  \code{\link{xUnzipKeys_}}; \code{\link{xZipKeys}},
  \code{\link{xZipKeys_}}

Other reshaping_functions: \code{\link{xChop}},
  \code{\link{xChop_}}; \code{\link{xChunk}},
  \code{\link{xChunk_}}; \code{\link{xFlatten}},
  \code{\link{xFlatten_}}; \code{\link{xGroupBy}},
  \code{\link{xGroupBy_}}; \code{\link{xJoin}},
  \code{\link{xJoin_}}; \code{\link{xOneOf}},
  \code{\link{xOneOf_}}; \code{\link{xRecycle}},
  \code{\link{xRecycle_}}; \code{\link{xRepeat}},
  \code{\link{xRepeat_}}; \code{\link{xReverse}},
  \code{\link{xReverse_}}; \code{\link{xShuffle}},
  \code{\link{xShuffle_}}; \code{\link{xSortBy}},
  \code{\link{xSortBy_}}; \code{\link{xSplitAt}},
  \code{\link{xSplitAt_}}; \code{\link{xSplitWith}},
  \code{\link{xSplitWith_}}; \code{\link{xTabulate}},
  \code{\link{xTabulate_}}; \code{\link{xUnzipKeys}},
  \code{\link{xUnzipKeys_}}; \code{\link{xZipKeys}},
  \code{\link{xZipKeys_}}; \code{\link{xZip}},
  \code{\link{xZip_}}
}

