% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xUnzipKeys}
\alias{xUnzipKeys}
\alias{xUnzipKeys...}
\title{xUnzipKeys}
\usage{
xUnzipKeys(coll)

xUnzipKeys...(...)
}
\arguments{
\item{coll}{a named collection. The collection to split into name, value pairs.}

\item{...}{see above.}
}
\value{
A named list.
}
\description{
Split a named collection into a list of name: value list pairs.
}
\details{
\bold{xUnzipKeys} is the inverse function to \bold{xZipKeys} - it
    takes a named collection and converts it into a list of
    key, value list pairs. This is primarily intended for allowing functions that
    cannot read collection names - like map and select - to be able to operate
    on both a value and its key simultaneouslty.
}
\section{Corner Cases}{

     Returns \code{list()} if \code{coll} is length-zero.
}
\examples{

NULL

}
\seealso{
Other key_functions: \code{\link{xAddKeys}};
  \code{\link{xPluck}}, \code{\link{xPluck...}};
  \code{\link{xZipKeys}}, \code{\link{xZipKeys...}}

Other reshaping_functions: \code{\link{xChop}},
  \code{\link{xChop...}}; \code{\link{xChunk}},
  \code{\link{xChunk...}}; \code{\link{xFlatten}},
  \code{\link{xFlatten...}}; \code{\link{xGroupBy}},
  \code{\link{xGroupBy...}}; \code{\link{xJoin}},
  \code{\link{xJoin...}}; \code{\link{xOneOf}},
  \code{\link{xOneOf...}}; \code{\link{xRecycle}},
  \code{\link{xRecycle...}}; \code{\link{xRepeat}},
  \code{\link{xRepeat...}}; \code{\link{xReverse}},
  \code{\link{xReverse...}}; \code{\link{xShuffle}},
  \code{\link{xShuffle...}}; \code{\link{xSortBy}},
  \code{\link{xSortBy...}}; \code{\link{xSplitAt}},
  \code{\link{xSplitAt...}}; \code{\link{xSplitWith}},
  \code{\link{xSplitWith...}}; \code{\link{xTabulate}},
  \code{\link{xTabulate...}}; \code{\link{xZipKeys}},
  \code{\link{xZipKeys...}}; \code{\link{xZip}},
  \code{\link{xZip...}}
}

