% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xUnzipNames}
\alias{xUnzipNames}
\alias{xUnzipNames...}
\title{xUnzipNames}
\usage{
xUnzipNames(coll)

xUnzipNames...(...)
}
\arguments{
\item{coll}{a named collection. The collection to split into name, value pairs.}

\item{...}{see above.}
}
\value{
A named list.
}
\description{
Split a named collection into a list of name: value list pairs.
}
\details{
\bold{xUnzipNames} is the inverse function to \bold{xZipNames} - it
    takes a named collection and converts it into a list of
    key, value list pairs. This is primarily intended for allowing functions that
    cannot read collection names - like map and select - to be able to operate
    on both a value and its key simultaneouslty.
}
\section{Corner Cases}{

     Returns \code{list()} if \code{coll} is length-zero.
}
\examples{

NULL

}
\seealso{
Other key_functions: \code{\link{xAsNamed}};
  \code{\link{xPluck}}, \code{\link{xPluck...}};
  \code{\link{xZipNames}}, \code{\link{xZipNames...}}

Other reshaping_functions: \code{\link{xChop}},
  \code{\link{xChop...}}; \code{\link{xChunk}},
  \code{\link{xChunk...}}; \code{\link{xFlatten}},
  \code{\link{xFlatten...}}; \code{\link{xJoin}},
  \code{\link{xJoin...}}; \code{\link{xOneOf}},
  \code{\link{xOneOf...}}; \code{\link{xRecycle}},
  \code{\link{xRecycle...}}; \code{\link{xRepeat}},
  \code{\link{xRepeat...}}; \code{\link{xReverse}},
  \code{\link{xReverse...}}; \code{\link{xShuffle}},
  \code{\link{xShuffle...}}; \code{\link{xSortBy}},
  \code{\link{xSortBy...}}; \code{\link{xSplitAt}},
  \code{\link{xSplitAt...}}; \code{\link{xSplitBy}},
  \code{\link{xSplitBy...}}; \code{\link{xTabulate}},
  \code{\link{xTabulate...}}; \code{\link{xZipNames}},
  \code{\link{xZipNames...}}; \code{\link{xZip}},
  \code{\link{xZip...}}
}

