% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xValuesOf}
\alias{xValuesOf}
\alias{xValuesOf_}
\title{xValuesOf}
\usage{
xValuesOf(coll)

xValuesOf_(...)
}
\arguments{
\item{coll}{a collection. The collection to remove names from.}

\item{...}{see above.}
}
\value{
a list.
}
\description{
Remove the names from a collection.
}
\section{Type Signature}{

    |any| -> [any]
}

\section{Corner Cases}{

    Returns length-zero when \bold{coll} is length-zero.
}
\examples{

# 1. A simple example.

xValuesOf(list('fruit' = 'bannana', 'colour' = 'yellow'))

# list("bannana", "yellow")
}
\seealso{
Other key_functions: \code{\link{xAddKeys}},
  \code{\link{xAddKeys_}}; \code{\link{xIndicesOf}},
  \code{\link{xIndicesOf_}}; \code{\link{xIndicesTo}};
  \code{\link{xKeysOf}}, \code{\link{xKeysOf_}};
  \code{\link{xUnzipIndices}},
  \code{\link{xUnzipIndices_}}; \code{\link{xUnzipKeys}},
  \code{\link{xUnzipKeys_}}; \code{\link{xZipKeys}},
  \code{\link{xZipKeys_}}
}

