\name{xVersion}
\alias{xVersion}
\title{xVersion}
\usage{
xVersion(...)
}
\arguments{
  \item{...}{arguments to be dropped.}
}
\value{
A three-number vector, containing the major release number,
minor release number and patch number.
}
\description{
Get the current version number of Arrow.
}
\examples{

# 1. xVersion
#    the most important (and only) example

xVersion( )
}
\seealso{
Other inpure_functions: \code{\link{xDelay}};
\code{\link{xDo}}, \code{\link{xDo...}};
\code{\link{xExecute}}; \code{\link{xStopwatch}}
}

