\name{xWords}
\alias{xWords}
\title{xWords}
\usage{
xWords(str)
}
\arguments{
  \item{str}{a length-one character vector.}
}
\value{
a character vector, with one or more elements.
}
\description{
Split a str at whitespace.
}
\seealso{
Other character_functions: \code{\link{xChars}};
\code{\link{xExplode}}; \code{\link{xImplode}};
\code{\link{xLines}}; \code{\link{xSubString}};
\code{\link{xUnchars}}; \code{\link{xUnlines}};
\code{\link{xUnwords}}
}

