% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xWrite.R
\name{xWrite}
\alias{xWrite}
\title{xWrite}
\usage{
xWrite(str1, str2)
}
\arguments{
\item{str1}{a length-one character vector. The path to
   export to.}

\item{str2}{a length-one character vector. The text to write.}
}
\value{
Null, as this is a purely side-effectual function.
}
\description{
Write a string to a file.
}
\section{Type Signature}{

    |character| -> { }
}

\section{Corner Cases}{

   xWrite does not write to a file if \bold{str1} is empty.
}
\examples{

NULL
}
\seealso{
Other inpure_functions: \code{\link{xDo}},
  \code{\link{xDo_}}; \code{\link{xExecute}};
  \code{\link{xOneOf}}, \code{\link{xOneOf_}};
  \code{\link{xRead}}; \code{\link{xShuffle}},
  \code{\link{xShuffle_}}; \code{\link{xStopwatch}};
  \code{\link{xVersion}}

Other text_processing_functions: \code{\link{xAmend}};
  \code{\link{xCarve}}; \code{\link{xCharsOf}};
  \code{\link{xExplode}}; \code{\link{xFromChars}},
  \code{\link{xFromChars_}}; \code{\link{xFromLines}},
  \code{\link{xFromLines_}}; \code{\link{xFromWords}},
  \code{\link{xFromWords_}}; \code{\link{xImplode}},
  \code{\link{xImplode_}}; \code{\link{xIsMatch}};
  \code{\link{xNotMatch}}; \code{\link{xRead}};
  \code{\link{xToChars}}; \code{\link{xToLines}};
  \code{\link{xToWords}}
}

