\name{xWriteChars}
\alias{xWriteChars}
\alias{xWriteWords}
\title{xWriteChars}
\usage{
xWriteChars(str, strs)

xWriteWords(str, strs)
}
\arguments{
  \item{str}{a length-one character vector. The path to
  import files from.}

  \item{strs}{a character vector. Lines to write to a
  file.}

  \item{str}{a length-one character vector. The path to
  import files from.}

  \item{strs}{a character vector. Lines to write to a
  file.}
}
\value{
null -

A character vector, with one or more elements.
}
\description{
Export a collection of strings as a single line in a file.

Export a collection of strings as a single line in a file.
}
\examples{

NULL

NULL
}
\seealso{
Other character_functions: \code{\link{xExplode}},
\code{\link{xExplode...}}; \code{\link{xFromChars}},
\code{\link{xFromChars...}}; \code{\link{xFromLines}},
\code{\link{xFromLines...}}; \code{\link{xFromWords}},
\code{\link{xFromWords...}}; \code{\link{xImplode}},
\code{\link{xImplode...}}; \code{\link{xReadChars}},
\code{\link{xReadLines}}; \code{\link{xReadWords}};
\code{\link{xSubstring}}, \code{\link{xSubstring...}};
\code{\link{xToChars}}; \code{\link{xToLines}};
\code{\link{xToWords}}; \code{\link{xWriteLines}}

Other character_functions: \code{\link{xExplode}},
\code{\link{xExplode...}}; \code{\link{xFromChars}},
\code{\link{xFromChars...}}; \code{\link{xFromLines}},
\code{\link{xFromLines...}}; \code{\link{xFromWords}},
\code{\link{xFromWords...}}; \code{\link{xImplode}},
\code{\link{xImplode...}}; \code{\link{xReadChars}},
\code{\link{xReadLines}}; \code{\link{xReadWords}};
\code{\link{xSubstring}}, \code{\link{xSubstring...}};
\code{\link{xToChars}}; \code{\link{xToLines}};
\code{\link{xToWords}}; \code{\link{xWriteLines}}
}

