% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{Fix}
\alias{Fix}
\title{An efficient implementation of xFix, lacking safe-guards and some features.
This is used internally when a function is called with only some of its arguments.}
\usage{
Fix(FN, SYM1, SYM2, SYM3, PRE1, PRE2, PRE3, PRE, FINAL)
}
\arguments{
\item{FN}{a symbol, used as a reference to the function to fix.}

\item{SYM1}{a symbol, the name of the first parametre.}

\item{SYM2}{a symbol, the name of the first parametre. Optional.}

\item{SYM3}{a symbol, the name of the first parametre. Optional.}

\item{PRE1}{an expression. Input checks that are run exclusively
on the first argument. Optional.}

\item{PRE2}{an expression. Input checks that are run exclusively
on the second argument. Optional.}

\item{PRE3}{an expression. Input checks that are run exclusively
on the third argument. Optional.}

\item{PRE}{an expression. Input checks that need more than one
argument at once (checking if two lengths are equal for example). Optional.}

\item{FINAL}{an expression. An arbitrary code block to run before running
the function body.}
}
\description{
An efficient implementation of xFix, lacking safe-guards and some features.
This is used internally when a function is called with only some of its arguments.
}

