% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{Na}
\alias{False}
\alias{Na}
\alias{Null}
\alias{True}
\title{Special Values}
\format{\preformatted{ logi NA
}}
\usage{
Na

Null

True

False
}
\description{
These are aliases to values already in R.
}
\details{
The value Nothing is rarely seen in R; the only place
    it is typically used is for constructing parametre lists for functions without default arguments.
}
\keyword{datasets}

