\docType{data}
\name{Na}
\alias{False}
\alias{Na}
\alias{Null}
\alias{True}
\title{Special Values}
\format{\preformatted{ logi NA
}}
\usage{
Na

Null

True

False
}
\description{
Arrow includes Pascal-case locked-versions of R's uppercase
special values.
}
\details{
These values are pascal-case varients of na, null, true and
false. These values cannot be overwritten, like the
uppercase varients but unlike the variables \code{T} and
\code{F}.
}
\keyword{datasets}

