% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{x.}
\alias{!=.xlambda_builder}
\alias{$.xlambda_builder}
\alias{&&.xlambda_builder}
\alias{&.xlambda_builder}
\alias{*.xlambda_builder}
\alias{+.xlambda_builder}
\alias{-.xlambda_builder}
\alias{/.xlambda_builder}
\alias{:.xlambda_builder}
\alias{<.xlambda_builder}
\alias{<=.xlambda_builder}
\alias{==.xlambda_builder}
\alias{>.xlambda_builder}
\alias{>=.xlambda_builder}
\alias{@.xlambda_builder}
\alias{\%/\%.xlambda_builder}
\alias{\%\%.xlambda_builder}
\alias{^.xlambda_builder}
\alias{x.}
\alias{|.xlambda_builder}
\alias{||.xlambda_builder}
\alias{~.xlambda_builder}
\title{x.}
\format{\preformatted{function (...)  
 - attr(*, "class")= chr "xlambda_builder"
}}
\usage{
x.

\\method{\\\%\\\%}{xlambda_builder}(dropped, val2)

\\method{\\\%/\\\%}{xlambda_builder}(dropped, val2)

\\method{$}{xlambda_builder}(dropped, val2)

"@.xlambda_builder"(dropped, val2)

\\method{^}{xlambda_builder}(dropped, val2)

":.xlambda_builder"(dropped, val2)

\\method{*}{xlambda_builder}(dropped, val2)

\\method{/}{xlambda_builder}(dropped, val2)

\\method{-}{xlambda_builder}(dropped, val2)

\\method{+}{xlambda_builder}(dropped, val2)

\\method{>}{xlambda_builder}(dropped, val2)

\\method{>=}{xlambda_builder}(dropped, val2)

\\method{<}{xlambda_builder}(dropped, val2)

\\method{<=}{xlambda_builder}(dropped, val2)

\\method{==}{xlambda_builder}(dropped, val2)

\\method{!=}{xlambda_builder}(dropped, val2)

\\method{&}{xlambda_builder}(dropped, val2)

"&&.xlambda_builder"(dropped, val2)

\\method{|}{xlambda_builder}(dropped, val2)

"||.xlambda_builder"(dropped, val2)

"~.xlambda_builder"(dropped, val2)
}
\arguments{
\item{dropped}{the value x.. This isn't used internally.}

\item{val2}{the right hand side of an infix function call.}
}
\value{
an arbitrary value.
}
\description{
Partially apply an infix function.
}
\details{
One of the most common uses of anonymous functions is
    to partially apply a binary function.
}
\examples{

# 1.

NULL
}
\keyword{datasets}

