\name{x_}
\alias{x_}
\title{x_}
\usage{
  x_(val)
}
\arguments{
  \item{val}{a function, collection, or arbitrary value.}

  \item{val}{any arbitrary value.}
}
\value{
  an arrow object.
}
\description{
  Generate a chainable arrow object
}
\section{Corner Cases}{
  asd.
}

\section{Glossary}{
  function: a closure or primitive function.  A symbol or
  string refering to such a function may also be used.

  predicate: A function that returns a logical value
  (unary, unless stated otherwise).  A symbol or string
  refering to such a function may also be used.

  collection: A list, pairlist or vector of any type.

  arity: The number of arguments a function can accept.
  Variadic functions have an arity of +Infinity,

  number: a double, integer or complex number, unless
  stated otherwise.
}
\examples{
inst/examples/blank.R
}

