% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xAllOf}
\alias{xAllOf}
\alias{xAllOf_}
\title{xAllOf}
\usage{
xAllOf(pred, coll)

xAllOf_(pred, ...)
}
\arguments{
\item{pred}{a predicate. The function used to test each element of
   the input collection.}

\item{coll}{a collection. The collection to test each element of.}

\item{...}{see above.}
}
\value{
A boolean value.
}
\description{
Is a predicate true for all elements of a collection?
}
\section{Type Signature}{

    (any -> logical) -> |any| -> &lt;boolean>
}

\section{Corner Cases}{

   If coll is length zero then logical(0) is returned. Na values
   are treated as False.
}
\examples{

# 1. Test if all the elements of a vector are true.

xAllOf(xI, c(True, False, True))
xAllOf_(xI, True, False, True)

# False
}
\seealso{
Other quantifier_functions: \code{\link{xAnyOf}},
  \code{\link{xAnyOf_}}; \code{\link{xNoneOf}},
  \code{\link{xNoneOf_}}; \code{\link{xPoll}},
  \code{\link{xPoll_}}
}

