% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xAny}
\alias{xAny}
\alias{xAny...}
\title{xAny}
\usage{
xAny(pred, coll)

xAny...(pred, ...)
}
\arguments{
\item{pred}{a predicate. The function used to test each element of
   the input collection.}

\item{coll}{a collection. The collection to test each element of.}

\item{...}{see above.}
}
\value{
A boolean value.
}
\description{
Is a predicate true for any element of a collection?
}
\section{Type Signature}{

    (any -> logical) -> |any| -> <boolean>
}

\section{Corner Cases}{

   If coll is length zero then logical(0) is returned.
}
\examples{

# 1. Using any to test if there are any true values in a vector.

xAny(xI, c(True, True, False, True))

# True

# 2. Does a function name with more that 25 characters exist
#    in the base R libraries?

xAny(name := nchar(name) > 25, ls('package:base'))

# several matches exist, including the terse 'suppressPackageStartupMessages'
}
\seealso{
Other quantifier_functions: \code{\link{xAll}},
  \code{\link{xAll...}}; \code{\link{xExists}},
  \code{\link{xExists...}}; \code{\link{xForall}},
  \code{\link{xForall...}}; \code{\link{xPoll}},
  \code{\link{xPoll...}}
}

