\name{xAsCharacter}
\alias{xAsCharacter}
\alias{xAsCharacter...}
\title{xAsCharacter}
\usage{
xAsCharacter(strs)

xAsCharacter...(...)
}
\arguments{
  \item{strs}{a collection of strings.}

  \item{...}{see above.}
}
\value{
A character vector.
}
\description{
Convert a collection to a double vector.
}
\section{Variadic Functions

\link{ help_variadic }}{

}

