\name{xAsCharacter}
\alias{xAsCharacter}
\alias{xAsCharacter...}
\title{xAsCharacter}
\usage{
xAsCharacter(strs)

xAsCharacter...(...)
}
\arguments{
  \item{strs}{a collection of strings. A list, pairlist or
  vector of length-one character vectors to convert to a
  character vector.}

  \item{...}{see above.}
}
\value{
A character vector.
}
\description{
Convert a collection to a double vector.
}
\details{
\code{xAsCharacter} converts a list, pairlist or vector of
length-one strings to a character vector. It does not
attempt to convert non-character collections to character
vectors - it simply converts character lists to character
typed vectors.
}
\examples{

NULL

}

