% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xAsCharacter}
\alias{xAsCharacter}
\alias{xAsCharacter_}
\title{xAsCharacter}
\usage{
xAsCharacter(strs)

xAsCharacter_(...)
}
\arguments{
\item{strs}{a collection of strings. A list, pairlist or vector
   of length-one character vectors to convert to a character vector.}

\item{...}{see above.}
}
\value{
A character vector.
}
\description{
Convert a collection of characters to a character vector.
}
\details{
\bold{xAsCharacter} converts a list, pairlist or vector of
   length-one strings to a character vector. It does not attempt
   to convert non-character collections to character vectors - it
   simply converts character lists to character typed vectors.
}
\section{Type Signature}{

    |character| -> &lt;character>
}

\section{Corner Cases}{

    xAsCharacter will throw an error if its input is
    not a collection of characters. All length-zero collections
    can be converted to character.
}
\examples{

# 1. Convert a list of character values to a character vector.

xAsCharacter(list('this', 'is', 'much', 'words'))

# c('this', 'is', 'much', 'words')
}
\seealso{
Other container_conversion_functions: \code{\link{xAsComplex}},
  \code{\link{xAsComplex_}}; \code{\link{xAsDouble}},
  \code{\link{xAsDouble_}}; \code{\link{xAsInteger}},
  \code{\link{xAsInteger_}}; \code{\link{xAsLogical}},
  \code{\link{xAsLogical_}}; \code{\link{xAsRaw}},
  \code{\link{xAsRaw_}}
}

