\name{xAsComplex}
\alias{xAsComplex}
\alias{xAsComplex...}
\title{xAsComplex}
\usage{
xAsComplex(ims)

xAsComplex...(...)
}
\arguments{
  \item{ims}{a collection of complex values. A list,
  pairlist or vector of length-one complex vectors to
  convert to a complex vector.}

  \item{...}{see above.}
}
\value{
An integer vector.
}
\description{
Convert a collection to a complex vector.
}
\details{
\code{xAsComplex} converts a list, pairlist or vector of
length-one complex values to a complex vector.
}
\section{Corner Cases}{
  Corner
}

