% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xAsComplex}
\alias{xAsComplex}
\alias{xAsComplex_}
\title{xAsComplex}
\usage{
xAsComplex(ims)

xAsComplex_(...)
}
\arguments{
\item{ims}{a collection of complex values. A list, pairlist or vector
   of length-one complex vectors to convert to a complex vector.}

\item{...}{see above.}
}
\value{
An integer vector.
}
\description{
Convert a collection to a complex vector.
}
\details{
\code{xAsComplex} converts a list, pairlist or vector of
   length-one complex values to a complex vector.
}
\section{Type Signature}{

    |complex| -> <complex>
}

\section{Corner Cases}{

   Corner
}
\examples{

NULL

}
\seealso{
Other container_conversion_functions: \code{\link{xAsCharacter}},
  \code{\link{xAsCharacter_}}; \code{\link{xAsDouble}},
  \code{\link{xAsDouble_}}; \code{\link{xAsInteger}},
  \code{\link{xAsInteger_}}; \code{\link{xAsLogical}},
  \code{\link{xAsLogical_}}; \code{\link{xAsRaw}},
  \code{\link{xAsRaw_}}
}

