\name{xAsComplex}
\alias{xAsComplex}
\alias{xAsComplex...}
\title{xAsComplex}
\usage{
xAsComplex(comps)

xAsComplex...(...)
}
\arguments{
  \item{comps}{a collection of complex values.}

  \item{...}{see above.}
}
\value{
A complex vector.
}
\description{
Convert a collection to a complex vector.
}
\details{
\code{xAsComplex} converts a list, pairlist or vector of
length-one complex numbers to a complex vector. It does not
attempt to convert non-complex collections to complex
vectors.
}
\section{Variadic Functions

\link{ help_variadic }}{

}

