% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xAsComplex.R
\name{xAsComplex}
\alias{xAsComplex}
\alias{xAsComplex_}
\title{xAsComplex}
\usage{
xAsComplex(ims)

xAsComplex_(...)
}
\arguments{
\item{ims}{a collection of complex values. A list, pairlist or vector
   of length-one complex vectors to convert to a complex vector.}

\item{...}{see above.}
}
\value{
An complex vector.
}
\description{
Convert a collection of complex values to a complex vector.
}
\details{
\bold{xAsComplex} converts a list, pairlist or vector of
   length-one complex values to a complex vector.
}
\section{Type Signature}{

    |complex| -> <complex>
}

\section{Corner Cases}{

    xAsComplex will throw an error if its input is
    not a collection of complex numbers. All length-zero collections
    can be converted to complex.
}

\section{Stability - \bold{Uncertain}}{


This feature may be removed or altered significantly in future versions.
If you find this feature useful, please leave an issue at https://github.com/rgrannell1/kea/issues.
}
\examples{

# 1. Convert a list of complex values to a complex vector.

xAsComplex(list(1 + 1i, 2 - 3i))
xAsComplex_(1 + 1i, 2 - 3i)

# c(1 + 1i, 2 - 3i)
}
\seealso{
Other container_conversion_functions: \code{\link{xAsCharacter}},
  \code{\link{xAsCharacter_}}; \code{\link{xAsDouble}},
  \code{\link{xAsDouble_}}; \code{\link{xAsInteger}},
  \code{\link{xAsInteger_}}; \code{\link{xAsList}},
  \code{\link{xAsList_}}; \code{\link{xAsLogical}},
  \code{\link{xAsLogical_}}; \code{\link{xAsRaw}},
  \code{\link{xAsRaw_}}

Other unboxing_functions: \code{\link{xAsCharacter}},
  \code{\link{xAsCharacter_}}; \code{\link{xAsDouble}},
  \code{\link{xAsDouble_}}; \code{\link{xAsInteger}},
  \code{\link{xAsInteger_}}; \code{\link{xAsLogical}},
  \code{\link{xAsLogical_}}; \code{\link{xAsRaw}},
  \code{\link{xAsRaw_}}
}

