\name{xAsComplex}
\alias{xAsComplex}
\alias{xAsComplex...}
\title{xAsComplex}
\usage{
xAsComplex(ims)

xAsComplex...(...)
}
\arguments{
  \item{ims}{a collection of complex values. A list,
  pairlist or vector of length-one complex vectors to
  convert to a complex vector.}

  \item{...}{see above.}
}
\value{
An integer vector.
}
\description{
Convert a collection to a complex vector.
}
\details{
\code{xAsComplex} converts a list, pairlist or vector of
length-one complex values to a complex vector.
}
\section{Corner Cases}{
  Corner
}
\examples{

NULL

}
\seealso{
Other container_conversion_functions:
\code{\link{xAsCharacter}}, \code{\link{xAsCharacter...}};
\code{\link{xAsDouble}}, \code{\link{xAsDouble...}};
\code{\link{xAsInteger}}, \code{\link{xAsInteger...}};
\code{\link{xAsLogical}}, \code{\link{xAsLogical...}};
\code{\link{xAsNamed}}; \code{\link{xAsRaw}},
\code{\link{xAsRaw...}}
}

