\name{xAsDouble}
\alias{xAsDouble}
\alias{xAsDouble...}
\title{xAsDouble}
\usage{
xAsDouble(nums)

xAsDouble...(...)
}
\arguments{
  \item{nums}{a collection of boolean values.}

  \item{...}{see above.}
}
\description{
Convert a collection to a double vector.
}
\section{Variadic Functions

\link{ help_variadic }}{

}

