% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xAsDouble}
\alias{xAsDouble}
\alias{xAsDouble_}
\title{xAsDouble}
\usage{
xAsDouble(nums)

xAsDouble_(...)
}
\arguments{
\item{nums}{a collection of double values. A list, pairlist or vector
   of length-one double vectors to convert to a double vector.}

\item{...}{see above.}
}
\value{
A double vector.
}
\description{
Convert a collection of doubles or integers to a double vector.
}
\details{
\bold{xAsDouble} converts a list, pairlist or vector of
   length-one double precision numbers to a double vector. It does not attempt
   to convert non-double collections to double vectors.
}
\section{Type Signature}{

    |number| -> &lt;numeric>
}

\section{Corner Cases}{

   Integer vectors are coerced to double vectors freely, since the
   difference is predominantly internal.
}

\section{Stability - \bold{Uncertain}}{


This feature may be removed or altered significantly in future versions.
If you find this feature useful, please leave an issue at https://github.com/rgrannell1/kea/issues.
}
\examples{

# 1. Convert a list of double values to a double vector.
#
# This also converts integers to doubles, as there is no
# information loss (all values can be coerced without failure).

xAsDouble(list(1, 2, 3, 4L, 5L))

# c(1, 2, 3, 4, 5)
}
\seealso{
Other container_conversion_functions: \code{\link{xAsCharacter}},
  \code{\link{xAsCharacter_}}; \code{\link{xAsComplex}},
  \code{\link{xAsComplex_}}; \code{\link{xAsInteger}},
  \code{\link{xAsInteger_}}; \code{\link{xAsLogical}},
  \code{\link{xAsLogical_}}; \code{\link{xAsRaw}},
  \code{\link{xAsRaw_}}
}

