% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xAsInteger}
\alias{xAsInteger}
\alias{xAsInteger_}
\title{xAsInteger}
\usage{
xAsInteger(ints)

xAsInteger_(...)
}
\arguments{
\item{ints}{a collection of integer values. A list, pairlist or vector
   of length-one integer vectors to convert to a integer vector.}

\item{...}{see above.}
}
\value{
An integer vector.
}
\description{
Convert a collection of integers to a integer vector.
}
\details{
\bold{xAsInteger} converts a list, pairlist or vector of
   length-one integers to a integer vector. It does not attempt
   to convert non-integer collections to integer vectors.
}
\section{Type Signature}{

    |integer| -> &lt;integer>
}

\section{Corner Cases}{

   Double vectors are not coerced to integer vectors:
   doubles can be decimal or infinite numbers, which
   would be lost upon coercion.
}
\examples{

# 1. Convert a list of integers to an integer vector.

xAsInteger(list(1L, 2L, 3L))

# c(1, 2, 3)
}
\seealso{
Other container_conversion_functions: \code{\link{xAsCharacter}},
  \code{\link{xAsCharacter_}}; \code{\link{xAsComplex}},
  \code{\link{xAsComplex_}}; \code{\link{xAsDouble}},
  \code{\link{xAsDouble_}}; \code{\link{xAsLogical}},
  \code{\link{xAsLogical_}}; \code{\link{xAsRaw}},
  \code{\link{xAsRaw_}}
}

