% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xAsInteger}
\alias{xAsInteger}
\alias{xAsInteger_}
\title{xAsInteger}
\usage{
xAsInteger(nums)

xAsInteger_(...)
}
\arguments{
\item{nums}{a collection of integer values. A list, pairlist or vector
   of length-one integer vectors to convert to a integer vector.}

\item{...}{see above.}
}
\value{
An integer vector.
}
\description{
Convert a collection to a integer vector.
}
\details{
\code{xAsInteger} converts a list, pairlist or vector of
   length-one integers to a integer vector. It does not attempt
   to convert non-integer collections to integer vectors.
}
\section{Type Signature}{

    |integer| -> <integer>
}

\section{Corner Cases}{

   Double vectors are not coerced to integer vectors:
   doubles can be decimal or infinite numbers, which
   would be lost upon coersion.
}
\examples{

NULL

}
\seealso{
Other container_conversion_functions: \code{\link{xAsCharacter}},
  \code{\link{xAsCharacter_}}; \code{\link{xAsComplex}},
  \code{\link{xAsComplex_}}; \code{\link{xAsDouble}},
  \code{\link{xAsDouble_}}; \code{\link{xAsLogical}},
  \code{\link{xAsLogical_}}; \code{\link{xAsRaw}},
  \code{\link{xAsRaw_}}
}

