% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xAsLogical}
\alias{xAsLogical}
\alias{xAsLogical_}
\title{xAsLogical}
\usage{
xAsLogical(bools)

xAsLogical_(...)
}
\arguments{
\item{bools}{a collection of logical values.}

\item{...}{see above.}
}
\value{
A logical vector.
}
\description{
Convert a collection to a logical vector.
}
\details{
\code{xAsLogical} converts a list, pairlist or vector of
   length-one logical values to a logical vector. It does not attempt
   to convert non-logical collections to logical vectors.
}
\examples{

NULL

}
\seealso{
Other container_conversion_functions: \code{\link{xAsCharacter}},
  \code{\link{xAsCharacter_}}; \code{\link{xAsComplex}},
  \code{\link{xAsComplex_}}; \code{\link{xAsDouble}},
  \code{\link{xAsDouble_}}; \code{\link{xAsInteger}},
  \code{\link{xAsInteger_}}; \code{\link{xAsRaw}},
  \code{\link{xAsRaw_}}
}

