% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xAsRaw}
\alias{xAsRaw}
\alias{xAsRaw_}
\title{xAsRaw}
\usage{
xAsRaw(raws)

xAsRaw_(...)
}
\arguments{
\item{raws}{a collection of raw values. A list, pairlist or vector
   of length-one raw vectors to convert to a raw vector.}

\item{...}{see above.}
}
\value{
A raw vector.
}
\description{
Convert a collection to a raw vector.
}
\details{
\code{xAsRaw} converts a list, pairlist or vector of
   length-one raw numbers to a raw vector. It does not attempt
   to convert non-raw collections to raw vectors.
}
\examples{

NULL

}
\seealso{
Other container_conversion_functions: \code{\link{xAsCharacter}},
  \code{\link{xAsCharacter_}}; \code{\link{xAsComplex}},
  \code{\link{xAsComplex_}}; \code{\link{xAsDouble}},
  \code{\link{xAsDouble_}}; \code{\link{xAsInteger}},
  \code{\link{xAsInteger_}}; \code{\link{xAsLogical}},
  \code{\link{xAsLogical_}}
}

