% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xAsRaw}
\alias{xAsRaw}
\alias{xAsRaw_}
\title{xAsRaw}
\usage{
xAsRaw(raws)

xAsRaw_(...)
}
\arguments{
\item{raws}{a collection of raw values. A list, pairlist or vector
   of length-one raw vectors to convert to a raw vector.}

\item{...}{see above.}
}
\value{
A raw vector.
}
\description{
Convert a collection of raw values to a raw vector.
}
\details{
\bold{xAsRaw} converts a list, pairlist or vector of
   length-one raw numbers to a raw vector. It does not attempt
   to convert non-raw collections to raw vectors.
}
\section{Type Signature}{

    |raw| -> &lt;raw>
}

\section{Corner Cases}{

}
\examples{

# 1. convert a raw list to a vector.

raws <- as.list(as.raw(1:10))

xAsRaw()
}
\seealso{
Other container_conversion_functions: \code{\link{xAsCharacter}},
  \code{\link{xAsCharacter_}}; \code{\link{xAsComplex}},
  \code{\link{xAsComplex_}}; \code{\link{xAsDouble}},
  \code{\link{xAsDouble_}}; \code{\link{xAsInteger}},
  \code{\link{xAsInteger_}}; \code{\link{xAsLogical}},
  \code{\link{xAsLogical_}}
}

