\name{xAsUnary}
\alias{xAsUnary}
\title{xAsUnary}
\usage{
  xAsUnary(fn)
}
\arguments{
  \item{fn}{an arbitrary function.}
}
\value{
  a unary function of x.
}
\description{
  Create a function that takes one argument, and applies
  each element of that argument as a parameter of its
  underlying function.
}
\seealso{
  Other higher_order_functions: \code{\link{xArity}},
  \code{\link{xChars}}, \code{\link{xCompose}},
  \code{\link{xConst}}, \code{\link{xFlip}},
  \code{\link{xIterate}}, \code{\link{xJuxtapose}},
  \code{\link{xLimit}}, \code{\link{xNot}},
  \code{\link{xParams}}, \code{\link{xUnfold}},
  \code{\link{xWait}}
}

