% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xAsVar}
\alias{xAsVar}
\title{xAsVar}
\usage{
xAsVar(sym)
}
\arguments{
\item{sym}{a symbol or string. The name of the variable
    to be converted to a variable.}
}
\value{
Null; used for side-effect.
}
\description{
Convert a constant value back into a normal R value.
}
\details{
\bold{xAsVar} takes a variable in the calling
    environment, and unlocks it, converting it back
    to a normal R variable.
}
\section{Corner Cases}{

    Throws an error if attempting to convert a
    variable that doesn't exist (in the parent frame).
    Non-locked variables are also allowed.
}

\section{Stability - \bold{Uncertain}}{


This feature may be removed or altered significantly in future versions.
If you find this feature useful, please leave an issue at https://github.com/rgrannell1/kea/issues.
}
\examples{

# 1. xAsVar can be used to unlock values back to
#    variables, since every function needs an inverse.

const <- list(
    letters =
        letters
)

xAsVal(const)
xAsVar(const)

const <- NULL
const

# NULL
}
\seealso{
Other immutable_value_functions: \code{\link{xAsVal}};
  \code{\link{xVal}}
}

