% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xAtCol}
\alias{xAtCol}
\alias{xAtCol...}
\title{xAtCol}
\usage{
xAtCol(num, colls)

xAtCol...(num, ...)
}
\arguments{
\item{num}{a whole number. The number of the column to select.}

\item{colls}{a collection of collections. The collection of
    rows to select a column from.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Select a column of values from a collection of rows.
}
\examples{

# 1. Selecting columns in data-frame like structures
#    Arrow eshews data frames, but works well with lists of rows.
#    All

df <- xList[ list(ith, letters[ith], LETTERS[ith]), ith <- 1:26 ]

# [
#   [1, a, A]
#   [2, b, A]
#      .
#      .
#   [26, z, Z]
# ]

x_(df) $ x_AtCol(2)

# list("a", "b", "c", "d", "e", "f", "g", "h", "i", "j", "k", "l",
#      "m", "n", "o", "p", "q", "r", "s", "t", "u", "v", "w", "x",
#      "y", "z")
}
\seealso{
Other selection_functions: \code{\link{xAtKey}},
  \code{\link{xAtKey...}}; \code{\link{xAt}},
  \code{\link{xAt...}}; \code{\link{xDropWhile}},
  \code{\link{xDropWhile...}}; \code{\link{xDrop}},
  \code{\link{xDrop...}}; \code{\link{xFirstOf}},
  \code{\link{xFirstOf...}}; \code{\link{xFourthOf}},
  \code{\link{xFourthOf...}}; \code{\link{xInitOf}},
  \code{\link{xInitOf...}}; \code{\link{xLastOf}},
  \code{\link{xLastOf...}}; \code{\link{xMaxBy}},
  \code{\link{xMaxBy...}}; \code{\link{xMinBy}},
  \code{\link{xMinBy...}}; \code{\link{xPluck}},
  \code{\link{xPluck...}}; \code{\link{xRestOf}},
  \code{\link{xRestOf...}}; \code{\link{xSecondOf}},
  \code{\link{xSecondOf...}}; \code{\link{xSlice}},
  \code{\link{xSlice...}}; \code{\link{xTakeWhile}},
  \code{\link{xTakeWhile...}}; \code{\link{xTake}},
  \code{\link{xTake...}}; \code{\link{xThirdOf}},
  \code{\link{xThirdOf...}}
}

