% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xByColkeys}
\alias{xByColkeys}
\title{xByColkeys}
\usage{
xByColkeys(tab)
}
\arguments{
\item{tab}{a matrix or data frame. The tabular structure to get the column names of.}
}
\value{
A kiwi containing a character vector.
}
\description{
Convert a matrix or data.frame to a collection of column names.
}
\details{
\bold{xByColkeys} returns the column names of a data frame or matrix.
}
\section{Corner Cases}{

    Returns character(0) when \bold{tab} is length-zero or has no keys.
}
\seealso{
Other methods: \code{\link{xByCols}};
  \code{\link{xByLevels}}; \code{\link{xByRowkeys}};
  \code{\link{xByRows}}; \code{\link{xByValues}};
  \code{\link{xElemsByCols}}; \code{\link{xElemsByRows}};
  \code{\link{xExecute}}; \code{\link{xTap}};
  \code{\link{x_}}, \code{\link{x__}}
}

