% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xByColkeys}
\alias{xByColkeys}
\title{$ xByColkeys}
\usage{
x_(  ) $ xByColkeys()
}
\value{
An arrow containing a character vector.
}
\description{
Convert a matrix or data.frame to a collection of column names.
}
\details{
\bold{xByColkeys} returns the column names of a data frame or matrix.
}
\seealso{
Other methods: \code{\link{xByCols}};
  \code{\link{xByLevels}}, \code{\link{xByLevels}};
  \code{\link{xByRows}}; \code{\link{xColUnit}};
  \code{\link{xElemsByCols}}; \code{\link{xElemsByRows}};
  \code{\link{xExecute}}; \code{\link{xRowUnit}};
  \code{\link{xTap}}; \code{\link{xTranspose}};
  \code{\link{x_}}
}

