\name{xByColnames}
\alias{xByColnames}
\title{$ xByColnames}
\usage{
x_(  ) $ xByColnames()
}
\value{
a list of strings.
}
\description{
Convert a matrix or data.frame to a collection of column
names.
}
\seealso{
Other exclusively_methods: \code{\link{xByCols}};
\code{\link{xByLevels}}; \code{\link{xByRows}};
\code{\link{xExecute}}
}

