\name{xByColnames}
\alias{xByColnames}
\title{$ xByColnames}
\usage{
x_(  ) $ xByColnames()
}
\value{
an arrow containing a list of strings.
}
\description{
Convert a matrix or data.frame to a collection of column
names.
}
\seealso{
Other exclusively_methods: \code{\link{xByCols}};
\code{\link{xByLevels}}, \code{\link{xByLevels}};
\code{\link{xByRows}}; \code{\link{xColUnit}};
\code{\link{xElemsByCols}}; \code{\link{xElemsByRows}};
\code{\link{xExecute}}; \code{\link{xRowUnit}};
\code{\link{xTap}}; \code{\link{xTranspose}}
}

