% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xByCols}
\alias{xByCols}
\title{$ xByCols}
\usage{
x_(  ) $ xByCols()
}
\arguments{
\item{colls}{a matrix or data frame. The object to convert to column-lists.}
}
\value{
An kiwi object containing a list of lists.
}
\description{
Convert a matrix or data.frame to a collection of columns.
}
\seealso{
Other methods: \code{\link{xByColkeys}};
  \code{\link{xByLevels}}, \code{\link{xByLevels}},
  \code{\link{xByValues}}; \code{\link{xByRowkeys}},
  \code{\link{xByRownames}}; \code{\link{xByRows}};
  \code{\link{xElemsByCols}}; \code{\link{xElemsByRows}};
  \code{\link{xExecute}}; \code{\link{xTap}};
  \code{\link{x_}}, \code{\link{x__}}
}

