% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xByCols---.R
\name{xByCols}
\alias{xByCols}
\title{xByCols}
\usage{
xByCols(tab)
}
\arguments{
\item{tab}{a matrix or data frame. The object to convert to column-lists.}
}
\value{
A kea object containing a list of lists.
}
\description{
Convert a matrix or data.frame to a collection of columns.
}
\section{Corner Cases}{

    Returns the empty list if \bold{tab} has no columns or have zero x zero dimensions.
}

\section{Stability - \bold{Uncertain}}{


This feature may be removed or altered significantly in future versions.
If you find this feature useful, please leave an issue at https://github.com/rgrannell1/kea/issues.
}
\seealso{
Other methods: \code{\link{xByLevels}};
  \code{\link{xByRows}}; \code{\link{xByValues}};
  \code{\link{xElemsByCols}}; \code{\link{xElemsByRows}};
  \code{\link{xExecute}}; \code{\link{x_}},
  \code{\link{x__}}
}

