% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xByCols}
\alias{xByCols}
\title{xByCols}
\usage{
xByCols(tab)
}
\arguments{
\item{tab}{a matrix or data frame. The object to convert to column-lists.}
}
\value{
A kiwi object containing a list of lists.
}
\description{
Convert a matrix or data.frame to a collection of columns.
}
\section{Corner Cases}{

    Returns the empty list if \bold{tab} has no columns or have zero x zero dimensions.
}
\seealso{
Other methods: \code{\link{xByColkeys}};
  \code{\link{xByLevels}}; \code{\link{xByRowkeys}};
  \code{\link{xByRows}}; \code{\link{xByValues}};
  \code{\link{xElemsByCols}}; \code{\link{xElemsByRows}};
  \code{\link{xExecute}}; \code{\link{xTap}};
  \code{\link{x_}}, \code{\link{x__}}
}

