% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xByCols}
\alias{xByCols}
\title{$ xByCols}
\usage{
x_(  ) $ xByCols()
}
\value{
An arrow object containing a list of lists.
}
\description{
Convert a matrix or data.frame to a collection of columns.
}
\seealso{
Other methods: \code{\link{xByColkeys}};
  \code{\link{xByLevels}}, \code{\link{xByLevels}};
  \code{\link{xByRows}}; \code{\link{xElemsByCols}};
  \code{\link{xElemsByRows}}; \code{\link{xExecute}};
  \code{\link{xTap}}; \code{\link{x_}}
}

