\name{xByLevels}
\alias{xByLevels}
\title{$ xByLevels}
\usage{
x_(  ) $ xByLevels()

x_(  ) $ xByLevels()
}
\value{
An arrow object containing a character vector.

An arrow object containing a list.
}
\description{
Convert a factor to a collection of levels.

Convert a factor to a collection of the factor's data.
}
\seealso{
Other exclusively_methods: \code{\link{xByColnames}};
\code{\link{xByCols}}; \code{\link{xByRows}};
\code{\link{xColUnit}}; \code{\link{xElemsByCols}};
\code{\link{xElemsByRows}}; \code{\link{xExecute}};
\code{\link{xRowUnit}}; \code{\link{xTap}};
\code{\link{xTranspose}}

Other exclusively_methods: \code{\link{xByColnames}};
\code{\link{xByCols}}; \code{\link{xByRows}};
\code{\link{xColUnit}}; \code{\link{xElemsByCols}};
\code{\link{xElemsByRows}}; \code{\link{xExecute}};
\code{\link{xRowUnit}}; \code{\link{xTap}};
\code{\link{xTranspose}}
}

