% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xByLevels}
\alias{xByLevels}
\alias{xByValues}
\title{xByLevels}
\usage{
x_(  ) $ xByLevels()

x_(  ) $ xByLevels()
}
\arguments{
\item{fact}{a factor. The factor to get the levels of.}

\item{fact}{a factor. The factor to get the underlying values from.}
}
\value{
A kiwi object containing a character vector.

A kiwi object containing a list.
}
\description{
Convert a factor to a collection of levels.

Convert a factor to a collection of the factor's data.
}
\seealso{
Other methods: \code{\link{xByColkeys}};
  \code{\link{xByCols}}; \code{\link{xByRowkeys}};
  \code{\link{xByRows}}; \code{\link{xElemsByCols}};
  \code{\link{xElemsByRows}}; \code{\link{xExecute}};
  \code{\link{xTap}}; \code{\link{x_}}, \code{\link{x__}}

Other methods: \code{\link{xByColkeys}};
  \code{\link{xByCols}}; \code{\link{xByRowkeys}};
  \code{\link{xByRows}}; \code{\link{xElemsByCols}};
  \code{\link{xElemsByRows}}; \code{\link{xExecute}};
  \code{\link{xTap}}; \code{\link{x_}}, \code{\link{x__}}
}

