\name{xByLevels}
\alias{xByLevels}
\title{$ xByLevels}
\usage{
x_(  ) $ xByLevels()
}
\value{
a character vector.
}
\description{
Convert a factor to a collection of levels.
}
\seealso{
Other exclusively_methods: \code{\link{xByColnames}};
\code{\link{xByCols}}; \code{\link{xByRows}};
\code{\link{xColUnit}}; \code{\link{xExecute}};
\code{\link{xRowUnit}}
}

