% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xByLevels}
\alias{xByLevels}
\alias{xByValues}
\title{$ xByLevels}
\usage{
x_(  ) $ xByLevels()

x_(  ) $ xByLevels()
}
\value{
An arrow object containing a character vector.

An arrow object containing a list.
}
\description{
Convert a collor to a collection of levels.

Convert a factor to a collection of the factor's data.
}
\seealso{
Other methods: \code{\link{xByColkeys}};
  \code{\link{xByCols}}; \code{\link{xByRowkeys}},
  \code{\link{xByRownames}}; \code{\link{xByRows}};
  \code{\link{xElemsByCols}}; \code{\link{xElemsByRows}};
  \code{\link{xExecute}}; \code{\link{xTap}};
  \code{\link{x_}}, \code{\link{x__}}

Other methods: \code{\link{xByColkeys}};
  \code{\link{xByCols}}; \code{\link{xByRowkeys}},
  \code{\link{xByRownames}}; \code{\link{xByRows}};
  \code{\link{xElemsByCols}}; \code{\link{xElemsByRows}};
  \code{\link{xExecute}}; \code{\link{xTap}};
  \code{\link{x_}}, \code{\link{x__}}
}

