% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xByLevels}
\alias{xByLevels}
\title{xByLevels}
\usage{
xByLevels(fact)
}
\arguments{
\item{fact}{a factor. The factor to get the levels of.}
}
\value{
A kea object containing a character vector.
}
\description{
Convert a factor to a collection of levels.
}
\section{Corner Cases}{

    Returns character(0) for empty factors.
}

\section{Stability 1 - \bold{Uncertain}}{


This feature may be removed or altered significantly in future versions.
If you find this feature useful, please leave an issue at https://github.com/rgrannell1/kea/issues.
}
\seealso{
Other methods: \code{\link{xByColkeys}};
  \code{\link{xByCols}}; \code{\link{xByRowkeys}};
  \code{\link{xByRows}}; \code{\link{xByValues}};
  \code{\link{xElemsByCols}}; \code{\link{xElemsByRows}};
  \code{\link{xExecute}}; \code{\link{xTap}};
  \code{\link{x_}}, \code{\link{x__}}
}

