% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xByRowkeys}
\alias{xByRowkeys}
\title{xByRowkeys}
\usage{
xByRowkeys(tab)
}
\arguments{
\item{tab}{a matrix or data frame. The tabular structure to get the row names of.}
}
\value{
A kea containing a character vector.
}
\description{
Convert a matrix or data.frame to a collection of column names.
}
\details{
\bold{xByRowkeys} returns the column names of a data frame or matrix.
}
\section{Corner Cases}{

    Returns character(0) when \bold{tab} is length-zero or has no keys.
}

\section{Stability 1 - \bold{Uncertain}}{


This feature may be removed or altered significantly in future versions.
If you find this feature useful, please leave an issue at https://github.com/rgrannell1/kea/issues.
}
\seealso{
Other methods: \code{\link{xByColkeys}};
  \code{\link{xByCols}}; \code{\link{xByLevels}};
  \code{\link{xByRows}}; \code{\link{xByValues}};
  \code{\link{xElemsByCols}}; \code{\link{xElemsByRows}};
  \code{\link{xExecute}}; \code{\link{xTap}};
  \code{\link{x_}}, \code{\link{x__}}
}

