\name{xByRows}
\alias{xByRows}
\title{$ xByRows}
\usage{
x_(  ) $ xByRows()
}
\value{
a list of lists.
}
\description{
Convert a matrix or data.frame to a collection of rows.
}
\seealso{
Other exclusively_methods: \code{\link{xByColnames}};
\code{\link{xByCols}}; \code{\link{xByLevels}};
\code{\link{xExecute}}
}

