% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xByRows---.R
\name{xByRows}
\alias{xByRows}
\title{xByRows}
\usage{
xByRows(tab)
}
\arguments{
\item{tab}{a matrix or data frame. The object to convert to column-lists.}
}
\value{
A kea object containing a list of lists.
}
\description{
Convert a matrix or data.frame to a collection of rows.
}
\details{
\bold{xByRows} is the most common way in which data-frames
    are reformatted for use by kea. Internally data frames are
    represented as lists of lists, and kea prefers this explicit representation.

    Column names are preserved by \bold{xByRows}.
}
\section{Corner Cases}{

    Returns the empty list if \bold{tab} has no rows or have zero x zero dimensions.
}

\section{Stability - \bold{Uncertain}}{


This feature may be removed or altered significantly in future versions.
If you find this feature useful, please leave an issue at https://github.com/rgrannell1/kea/issues.
}
\seealso{
Other methods: \code{\link{xByCols}};
  \code{\link{xByLevels}}; \code{\link{xByValues}};
  \code{\link{xElemsByCols}}; \code{\link{xElemsByRows}};
  \code{\link{xExecute}}; \code{\link{x_}},
  \code{\link{x__}}
}

