% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xByRows}
\alias{xByRows}
\title{$ xByRows}
\usage{
x_(  ) $ xByRows()
}
\value{
An arrow object containing a list of lists.
}
\description{
Convert a matrix or data.frame to a collection of rows.
}
\seealso{
Other methods: \code{\link{xByColnames}};
  \code{\link{xByCols}}; \code{\link{xByLevels}},
  \code{\link{xByLevels}}; \code{\link{xColUnit}};
  \code{\link{xElemsByCols}}; \code{\link{xElemsByRows}};
  \code{\link{xExecute}}; \code{\link{xRowUnit}};
  \code{\link{xTap}}; \code{\link{xTranspose}};
  \code{\link{x_}}
}

