% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xByRows}
\alias{xByRows}
\title{xByRows}
\usage{
xByRows(tab)
}
\arguments{
\item{tab}{a matrix or data frame. The object to convert to column-lists.}
}
\value{
A kea object containing a list of lists.
}
\description{
Convert a matrix or data.frame to a collection of rows.
}
\details{
\bold{xByRows} is the most common way in which data-frames
    are reformatted for use by kea. Internally data frames are
    represented as lists of lists, and kea prefers this explicit representation.

    Column names are preserved by \bold{xByRows}.
}
\section{Corner Cases}{

    Returns the empty list if \bold{tab} has no rows or have zero x zero dimensions.
}
\seealso{
Other methods: \code{\link{xByColkeys}};
  \code{\link{xByCols}}; \code{\link{xByLevels}};
  \code{\link{xByRowkeys}}; \code{\link{xByValues}};
  \code{\link{xElemsByCols}}; \code{\link{xElemsByRows}};
  \code{\link{xExecute}}; \code{\link{xTap}};
  \code{\link{x_}}, \code{\link{x__}}
}

