% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xByRows}
\alias{xByRows}
\title{$ xByRows}
\usage{
x_(  ) $ xByRows()
}
\value{
An arrow object containing a list of lists.
}
\description{
Convert a matrix or data.frame to a collection of rows.
}
\details{
\bold{xByRows} is the most common way in which data-frames
    are reformatted for use by arrow. Internally dataframes are
    represented as lists of lists, and arrow prefers this explicit representation.

    Column names are preserved by \bold{xByRows}.
}
\seealso{
Other methods: \code{\link{xByColkeys}};
  \code{\link{xByCols}}; \code{\link{xByLevels}},
  \code{\link{xByLevels}}, \code{\link{xByValues}};
  \code{\link{xByRowkeys}}, \code{\link{xByRownames}};
  \code{\link{xElemsByCols}}; \code{\link{xElemsByRows}};
  \code{\link{xExecute}}; \code{\link{xTap}};
  \code{\link{x_}}, \code{\link{x__}}
}

