% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xByValues---.R
\name{xByValues}
\alias{xByValues}
\title{xByValues}
\usage{
xByValues(fact)
}
\arguments{
\item{fact}{a factor. The factor to get the underlying values from.}
}
\value{
A kea object containing a list.
}
\description{
Convert a factor to a collection of the factor's data.
}
\section{Corner Cases}{

}

\section{Stability - \bold{Uncertain}}{


This feature may be removed or altered significantly in future versions.
If you find this feature useful, please leave an issue at https://github.com/rgrannell1/kea/issues.
}
\seealso{
Other methods: \code{\link{xByCols}};
  \code{\link{xByLevels}}; \code{\link{xByRows}};
  \code{\link{xElemsByCols}}; \code{\link{xElemsByRows}};
  \code{\link{xExecute}}; \code{\link{x_}},
  \code{\link{x__}}
}

