\name{xChars}
\alias{xChars}
\title{xChars}
\usage{
xChars(str)
}
\arguments{
  \item{str}{a length-one character vector.}
}
\value{
a character vector of length \code{nchar(str)}, with each
element being a single character.
}
\description{
Split a single string into a vector of characters.
}
\seealso{
Other character_functions: \code{\link{xExplode}},
\code{\link{xExplode...}}; \code{\link{xImplode}},
\code{\link{xImplode...}}; \code{\link{xLines}};
\code{\link{xSubstring}}, \code{\link{xSubstring...}};
\code{\link{xUnchars}}, \code{\link{xUnchars...}};
\code{\link{xUnlines}}, \code{\link{xUnlines...}};
\code{\link{xUnwords}}, \code{\link{xUnwords...}};
\code{\link{xWords}}
}

