\name{xChars}
\alias{xChars}
\title{xChars}
\usage{
xChars(str)
}
\arguments{
  \item{str}{a length-one character vector.}
}
\value{
a character vector of length \code{nchar(str)}, with each
element being a single character.
}
\description{
Split a single string into a vector of characters.
}
\seealso{
Other character_vector_functions: \code{\link{xCollapse}};
\code{\link{xLines}}; \code{\link{xSplitString}};
\code{\link{xSubString}}; \code{\link{xUnchars}};
\code{\link{xUnlines}}; \code{\link{xUnwords}};
\code{\link{xWords}}
}

