\name{xChars}
\alias{xChars}
\title{xChars}
\usage{
  xChars(str)
}
\arguments{
  \item{str}{a length-one character vector.}
}
\value{
  a character vector of length \code{nchar(str)}.
}
\description{
  Split a single string into a vector of characters.
}
\seealso{
  Other higher_order_functions: \code{\link{xArity}},
  \code{\link{xAsUnary}}, \code{\link{xCompose}},
  \code{\link{xConst}}, \code{\link{xFlip}},
  \code{\link{xIterate}}, \code{\link{xJuxtapose}},
  \code{\link{xLimit}}, \code{\link{xNot}},
  \code{\link{xParams}}, \code{\link{xUnfold}},
  \code{\link{xWait}}
}

