% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xChop}
\alias{xChop}
\alias{xChop...}
\title{xChop}
\usage{
xChop(num, coll)

xChop...(num, ...)
}
\arguments{
\item{num}{a nonnegative whole number. The desired number
   of output collections to generate.}

\item{coll}{a collection. The collection to chop up.}

\item{...}{see above.}
}
\value{
A list of lists.
}
\description{
Divide a collection into a fixed number of segments.
}
\details{
\bold{xChop} is used to divide a collection into several
segments. This differs from \bold{xChunk} in that \bold{xChop}
creates a fixed number of segments or arbitrary size, while
\bold{xChunk} creates segments of fixed size, but not a fixed
number of such segments.

\code{xChop(2, 1:5)}

\code{list(list(1, 2, 3), list(4, 5))}

\bold{xChop} can be used for dividing a task for parallel execution.
For example, a very large collection could be chopped into
four subcollections by using \bold{xChop(4, coll)}, before applying a
function to each sublist using \bold{mclapply( )}.
}
\section{Corner Cases}{

   The final sublist returned by \bold{xChop} may have less elements
   than the other sublists, depending on whether or not the length of \bold{coll}
   is evenly divisible by \bold{num}. If \bold{coll} is length-zero,
   the empty list is returned.
}
\examples{

NULL
}
\seealso{
Other reshaping_functions: \code{\link{xChunk}},
  \code{\link{xChunk...}}; \code{\link{xFlatten}},
  \code{\link{xFlatten...}}; \code{\link{xGroupBy}},
  \code{\link{xGroupBy...}}; \code{\link{xJoin}},
  \code{\link{xJoin...}}; \code{\link{xOneOf}},
  \code{\link{xOneOf...}}; \code{\link{xRecycle}},
  \code{\link{xRecycle...}}; \code{\link{xRepeat}},
  \code{\link{xRepeat...}}; \code{\link{xReverse}},
  \code{\link{xReverse...}}; \code{\link{xShuffle}},
  \code{\link{xShuffle...}}; \code{\link{xSortBy}},
  \code{\link{xSortBy...}}; \code{\link{xSplitAt}},
  \code{\link{xSplitAt...}}; \code{\link{xSplitWith}},
  \code{\link{xSplitWith...}}; \code{\link{xTabulate}},
  \code{\link{xTabulate...}}; \code{\link{xUnzipNames}},
  \code{\link{xUnzipNames...}}; \code{\link{xZipNames}},
  \code{\link{xZipNames...}}; \code{\link{xZip}},
  \code{\link{xZip...}}
}

