% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xChop}
\alias{xChop}
\alias{xChop_}
\title{xChop}
\usage{
xChop(num, coll)

xChop_(num, ...)
}
\arguments{
\item{num}{a nonnegative whole number. The desired number
   of output collections to generate.}

\item{coll}{a collection. The collection to chop up.}

\item{...}{see above.}
}
\value{
A list of lists.
}
\description{
Divide a collection into a fixed number of segments.
}
\details{
\bold{xChop} is used to divide a collection into several
    segments. This differs from \bold{xChunk} in that \bold{xChop}
    creates a fixed number of segments or arbitrary size, while
    \bold{xChunk} creates segments of fixed size, but not a fixed
    number of such segments.

    \code{xChop(2, 1:5)}

    \code{list(list(1, 2, 3), list(4, 5))}

    \bold{xChop} can be used for dividing a task for parallel execution.
    For example, a very large collection could be chopped into
    four sub-collections by using \bold{xChop(4, coll)}, before applying a
    function to each sublist using \bold{mclapply( )}.
}
\section{Type Signature}{

    |numeric| -> |any| -> [[any]]
}

\section{Corner Cases}{

   The final sublist returned by \bold{xChop} may have less elements
   than the other sublists, depending on whether or not the length of \bold{coll}
   is evenly divisible by \bold{num}. If \bold{coll} or \bold{num} is length-zero,
   the empty list is returned.
}
\examples{

NULL
}
\seealso{
Other reshaping_functions: \code{\link{xAppend}},
  \code{\link{xAppend_}}; \code{\link{xChunk}},
  \code{\link{xChunk_}}; \code{\link{xFlatten}},
  \code{\link{xFlatten_}}; \code{\link{xGather}},
  \code{\link{xGather_}}; \code{\link{xGroupBy}},
  \code{\link{xGroupBy_}}; \code{\link{xJoin}},
  \code{\link{xJoin_}}; \code{\link{xOneOf}},
  \code{\link{xOneOf_}}; \code{\link{xPrepend}},
  \code{\link{xPrepend_}}; \code{\link{xRecycle}},
  \code{\link{xRecycle_}}; \code{\link{xRepeat}},
  \code{\link{xRepeat_}}; \code{\link{xReverse}},
  \code{\link{xReverse_}}; \code{\link{xShuffle}},
  \code{\link{xShuffle_}}; \code{\link{xSortBy}},
  \code{\link{xSortBy_}}; \code{\link{xTabulate}},
  \code{\link{xTabulate_}}; \code{\link{xUnzipIndices}},
  \code{\link{xUnzipIndices_}}; \code{\link{xUnzipKeys}},
  \code{\link{xUnzipKeys_}}; \code{\link{xZipKeys}},
  \code{\link{xZipKeys_}}; \code{\link{xZip}},
  \code{\link{xZip_}}
}

