% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xChunk}
\alias{xChunk}
\alias{xChunk_}
\title{xChunk}
\usage{
xChunk(num, coll)

xChunk_(num, ...)
}
\arguments{
\item{num}{a nonnegative whole number. The desired
   length of each group of elements.}

\item{coll}{a collection. The collections to divide into groups.}

\item{...}{see above.}
}
\value{
A list of lists.
}
\description{
Divide a collection into segments of fixed length.
}
\section{Corner Cases}{

   The final list in the return value will have less than \bold{num}
   elements if \code{length(coll)} is not evenly divisible by \bold{num}.
   if \bold{coll} is length-zero, the empty list is returned.
}
\examples{

NULL

}
\seealso{
Other reshaping_functions: \code{\link{xChop}},
  \code{\link{xChop_}}; \code{\link{xFlatten}},
  \code{\link{xFlatten_}}; \code{\link{xGroupBy}},
  \code{\link{xGroupBy_}}; \code{\link{xJoin}},
  \code{\link{xJoin_}}; \code{\link{xOneOf}},
  \code{\link{xOneOf_}}; \code{\link{xRecycle}},
  \code{\link{xRecycle_}}; \code{\link{xRepeat}},
  \code{\link{xRepeat_}}; \code{\link{xReverse}},
  \code{\link{xReverse_}}; \code{\link{xShuffle}},
  \code{\link{xShuffle_}}; \code{\link{xSortBy}},
  \code{\link{xSortBy_}}; \code{\link{xSplitAt}},
  \code{\link{xSplitAt_}}; \code{\link{xSplitWith}},
  \code{\link{xSplitWith_}}; \code{\link{xTabulate}},
  \code{\link{xTabulate_}}; \code{\link{xUnzipKeys}},
  \code{\link{xUnzipKeys_}}; \code{\link{xZipKeys}},
  \code{\link{xZipKeys_}}; \code{\link{xZip}},
  \code{\link{xZip_}}
}

