% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xChunk}
\alias{xChunk}
\alias{xChunk_}
\title{xChunk}
\usage{
xChunk(num, coll)

xChunk_(num, ...)
}
\arguments{
\item{num}{a nonnegative whole number. The desired
   length of each group of elements.}

\item{coll}{a collection. The collections to divide into groups.}

\item{...}{see above.}
}
\value{
A list of lists.
}
\description{
Divide a collection into segments of fixed length.
}
\section{Type Signature}{

    |numeric| -> |any| -> [[any]]
}

\section{Corner Cases}{

   The final list in the return value will have less than \bold{num}
   elements if \code{length(coll)} is not evenly divisible by \bold{num}.
   if \bold{coll} or \bold{num} is length-zero, the empty list is returned.
}

\section{Rcpp}{


    This function is written using C++, and should be quite performant.
}
\examples{

# 1. Chunk a flat name,value .csv into pairs of names, values

csv <- "Monaco,48.9,Japan,44.6,Italy,44.3,Germany,43.7,Jersey,43.4,Hong Kong,42.8"

age_by_country <-
	x_(csv) $ xExplode(',') $
	xChunk(2) $ xMap( xUnspread((key : value) := {
		list(key, as.numeric(value))
	}) ) $
	xZipKeys()

# list(
#     Monaco = "48.9", Japan = "44.6", Italy = "44.3",
#     Germany = "43.7", Jersey = "43.4", `Hong Kong` = "42.8")

# 2. Select every third element in collection, starting from the
#    second value. (2nd, 5th, ...)

x_(letters) $ xChunk(3) $ xSelect(row := {
	# -- because the last might have less than three elements.
	xLenOf(row) == 3
}) $
xMap(xAt(2))

# list("b", "e", "h", "k", "n", "q", "t", "w")
}
\seealso{
Other reshaping_functions: \code{\link{xAppend}},
  \code{\link{xAppend_}}; \code{\link{xFlatten}},
  \code{\link{xFlatten_}}; \code{\link{xGather}},
  \code{\link{xGather_}}; \code{\link{xGroupBy}},
  \code{\link{xGroupBy_}}; \code{\link{xJoin}},
  \code{\link{xJoin_}}; \code{\link{xOneOf}},
  \code{\link{xOneOf_}}; \code{\link{xPrepend}},
  \code{\link{xPrepend_}}; \code{\link{xRecycle}},
  \code{\link{xRecycle_}}; \code{\link{xRepeat}},
  \code{\link{xRepeat_}}; \code{\link{xReverse}},
  \code{\link{xReverse_}}; \code{\link{xShuffle}},
  \code{\link{xShuffle_}}; \code{\link{xSortBy}},
  \code{\link{xSortBy_}}; \code{\link{xTabulate}},
  \code{\link{xTabulate_}}; \code{\link{xUnzipIndices}},
  \code{\link{xUnzipIndices_}}; \code{\link{xUnzipKeys}},
  \code{\link{xUnzipKeys_}}; \code{\link{xZipKeys}},
  \code{\link{xZipKeys_}}; \code{\link{xZip}},
  \code{\link{xZip_}}
}

