\name{xCollapse}
\alias{xCollapse}
\title{xCollapse}
\usage{
  xCollapse(str, ...)
}
\arguments{
  \item{str}{a string to use as a delimiter.}

  \item{...}{a number of character vectors.}
}
\value{
  a length-one character vector.
}
\description{
  Concatenate a character vector into a string with a
  delimiter.
}
\section{Glossary}{
  \bold{Function}: a closure or primitive function.  A
  symbol or string refering to such a function may also be
  used.

  \bold{Predicate}: A function that returns a logical value
  (unary, unless stated otherwise).  A symbol or string
  refering to such a function may also be used.

  \bold{Collection}: A list, pairlist or a typed vector.

  \bold{Arity}: The number of arguments a function can
  accept.  Unary (1), Binary (2), Trinary (3), Quaternary
  (4), ...., Variadic (Inf).

  \bold{Number}: a double, integer or complex number,
  unless stated otherwise.

  \bold{Higher Order Function}: a function that takes a
  function as one of its arguments, and/or returns a
  function.
}
\examples{
inst/examples/blank.R
}

