\name{xCollapse}
\alias{xCollapse}
\title{xCollapse}
\usage{
  xCollapse(str, strs)
}
\arguments{
  \item{str}{a string to use as a delimiter.}

  \item{strs}{a number of character vectors.}
}
\value{
  a length-one character vector.
}
\description{
  Concatenate a character vector into a string with a
  delimiter.
}

