\name{xColUnit}
\alias{xColUnit}
\title{$ xColUnit}
\usage{
x_(  ) $ xColUnit()
}
\value{
an empty matrix or data.frame.
}
\description{
Return the concatenative unit for a matrix or data.frame
column.
}
\seealso{
Other exclusively_methods: \code{\link{xByColnames}};
\code{\link{xByCols}}; \code{\link{xByLevels}};
\code{\link{xByRows}}; \code{\link{xExecute}};
\code{\link{xRowUnit}}
}

