\name{xCompose}
\alias{xCompose}
\alias{xCompose...}
\title{xCompose}
\usage{
xCompose(fns)

xCompose...(...)
}
\arguments{
  \item{fns}{a collection of functions.}

  \item{...}{see above.}
}
\value{
A function with the same parametres as \code{fn2}.
}
\description{
Compose two functions.
}
\section{Variadic Functions

\link{ help_variadic }}{

}
\seealso{
Other function_combining_functions:
\code{\link{xJuxtapose}}, \code{\link{xJuxtapose...}};
\code{\link{xLift}}, \code{\link{xLift...}}

Other function_modifying_functions:
\code{\link{xAsClosure}}; \code{\link{xAsFunction}},
\code{\link{xAsFunction...}}; \code{\link{xAsUnary}};
\code{\link{xAsVariadic}}; \code{\link{xConst}},
\code{\link{xK}}; \code{\link{xFlip}};
\code{\link{xJuxtapose}}, \code{\link{xJuxtapose...}};
\code{\link{xLift}}, \code{\link{xLift...}};
\code{\link{xNot}}; \code{\link{xPartial}},
\code{\link{xPartial...}}; \code{\link{xDelay}};
\code{\link{xThread}}, \code{\link{xThread...}};
\code{\link{xVectorise}}, \code{\link{xVectorize}}
}

