% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xDeepMap}
\alias{xDeepMap}
\alias{xDeepMap...}
\title{xDeepMap}
\usage{
xDeepMap(fn, coll)

xDeepMap...(fn, ...)
}
\arguments{
\item{fn}{a unary function. A function to recursively apply
   into a collection.}

\item{coll}{a collection. The collection to be mapped into.}

\item{...}{see above.}
}
\value{
A list or pairlist.
}
\description{
Recursively map a function into a nested collection,
preserving its structure.
}
\examples{

NULL

}
\seealso{
Other function_application_functions: \code{\link{xApply}},
  \code{\link{xApply...}}; \code{\link{xAsUnary}};
  \code{\link{xAsVariadic}}; \code{\link{xFlatMap}},
  \code{\link{xFlatMap...}}; \code{\link{xMapIndexed}},
  \code{\link{xMapIndexed...}}; \code{\link{xMapply}},
  \code{\link{xMapply...}}; \code{\link{xMap}},
  \code{\link{xMap...}}; \code{\link{xThread}},
  \code{\link{xThread...}}; \code{\link{xVectorise}},
  \code{\link{xVectorize}}
}

