% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xDrop}
\alias{xDrop}
\alias{xDrop_}
\title{xDrop}
\usage{
xDrop(num, coll)

xDrop_(num, ...)
}
\arguments{
\item{num}{a nonnegative whole number. The number of elements to
     return from the input collection.}

\item{coll}{a collection. The collection to subset.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Take several elements from the head of a collection.
}
\section{Corner Cases}{

   If \bold{coll} is empty the empty list is returned.
}
\examples{

NULL
}
\seealso{
Other selection_functions: \code{\link{xAtCol}},
  \code{\link{xAtCol_}}; \code{\link{xAtKey}},
  \code{\link{xAtKey_}}; \code{\link{xAt}},
  \code{\link{xAt_}}; \code{\link{xDropWhile}},
  \code{\link{xDropWhile_}}; \code{\link{xFirstOf}},
  \code{\link{xFirstOf_}}; \code{\link{xFourthOf}},
  \code{\link{xFourthOf_}}; \code{\link{xInitOf}},
  \code{\link{xInitOf_}}; \code{\link{xLastOf}},
  \code{\link{xLastOf_}}; \code{\link{xMaxBy}},
  \code{\link{xMaxBy_}}; \code{\link{xMinBy}},
  \code{\link{xMinBy_}}; \code{\link{xPluck}},
  \code{\link{xPluck_}}; \code{\link{xRestOf}},
  \code{\link{xRestOf_}}; \code{\link{xSecondOf}},
  \code{\link{xSecondOf_}}; \code{\link{xSlice}},
  \code{\link{xSlice_}}; \code{\link{xTakeWhile}},
  \code{\link{xTakeWhile_}}; \code{\link{xTake}},
  \code{\link{xTake_}}; \code{\link{xThirdOf}},
  \code{\link{xThirdOf_}}
}

