% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xDropWhile}
\alias{xDropWhile}
\alias{xDropWhile...}
\title{xDropWhile}
\usage{
xDropWhile(pred, coll)

xDropWhile...(pred, ...)
}
\arguments{
\item{pred}{a predicate. The functionto test each element of
   the collection with.}

\item{coll}{a collection. The collection to drop elements from.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Take every element in a collection from the first time a predicate
is false or na until the end of the collection.
}
\section{Corner Cases}{

	  Returns the emty list if \bold{coll} is length-zero.
}
\examples{

NULL

}
\seealso{
Other selection_functions: \code{\link{xAtCol}},
  \code{\link{xAtCol...}}; \code{\link{xAt}},
  \code{\link{xAt...}}; \code{\link{xDrop}},
  \code{\link{xDrop...}}; \code{\link{xFirstOf}},
  \code{\link{xFirstOf...}}; \code{\link{xFourthOf}},
  \code{\link{xFourthOf...}}; \code{\link{xInitOf}},
  \code{\link{xInitOf...}}; \code{\link{xLastOf}},
  \code{\link{xLastOf...}}; \code{\link{xMaxBy}},
  \code{\link{xMaxBy...}}; \code{\link{xMinBy}},
  \code{\link{xMinBy...}}; \code{\link{xPluck}},
  \code{\link{xPluck...}}; \code{\link{xRestOf}},
  \code{\link{xRestOf...}}; \code{\link{xSecondOf}},
  \code{\link{xSecondOf...}}; \code{\link{xSlice}},
  \code{\link{xSlice...}}; \code{\link{xTakeWhile}},
  \code{\link{xTakeWhile...}}; \code{\link{xTake}},
  \code{\link{xTake...}}; \code{\link{xThirdOf}},
  \code{\link{xThirdOf...}}
}

