% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xDropWhile}
\alias{xDropWhile}
\alias{xDropWhile_}
\title{xDropWhile}
\usage{
xDropWhile(pred, coll)

xDropWhile_(pred, ...)
}
\arguments{
\item{pred}{a predicate. The function to test each element of
   the collection with.}

\item{coll}{a collection. The collection to drop elements from.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Take every element in a collection from the first time a predicate
is false or na until the end of the collection.
}
\section{Type Signature}{

    (any -> &lt;logical>) -> |any| -> [any]
}

\section{Corner Cases}{

	  Returns the empty list if \bold{coll} is length-zero.
}
\examples{

NULL
}
\seealso{
Other selection_functions: \code{\link{xAtKey}},
  \code{\link{xAtKey_}}; \code{\link{xAt}},
  \code{\link{xAt_}}; \code{\link{xDrop}},
  \code{\link{xDrop_}}; \code{\link{xFirstOf}},
  \code{\link{xFirstOf_}}; \code{\link{xFourthOf}},
  \code{\link{xFourthOf_}}; \code{\link{xInitOf}},
  \code{\link{xInitOf_}}; \code{\link{xLastOf}},
  \code{\link{xLastOf_}}; \code{\link{xMaxBy}},
  \code{\link{xMaxBy_}}; \code{\link{xMinBy}},
  \code{\link{xMinBy_}}; \code{\link{xRestOf}},
  \code{\link{xRestOf_}}; \code{\link{xSecondOf}},
  \code{\link{xSecondOf_}}; \code{\link{xSlice}},
  \code{\link{xSlice_}}; \code{\link{xTakeWhile}},
  \code{\link{xTakeWhile_}}; \code{\link{xTake}},
  \code{\link{xTake_}}; \code{\link{xThirdOf}},
  \code{\link{xThirdOf_}}
}

