% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xDuplicatesOf}
\alias{xDuplicatesOf}
\alias{xDuplicatesOf...}
\title{xDuplicatesOf}
\usage{
xDuplicatesOf(coll)

xDuplicatesOf...(...)
}
\arguments{
\item{coll}{a collection. The collection to return
   duplicated elements from.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Return the duplicated elements in a collection.
}
\section{Corner Cases}{

   Returns the empty list if \code{colls} is length-zero.
}
\examples{

NULL

}
\seealso{
Other set_functions: \code{\link{xInter}},
  \code{\link{xInter...}}; \code{\link{xIsMember}},
  \code{\link{xIsMember...}}; \code{\link{xIsSubset}},
  \code{\link{xIsSubset...}}; \code{\link{xNotMember}},
  \code{\link{xNotMember...}}; \code{\link{xPowerSetOf}},
  \code{\link{xPowerSetOf...}}; \code{\link{xProdSetOf}},
  \code{\link{xProdSetOf...}}; \code{\link{xReorder}},
  \code{\link{xReorder...}}; \code{\link{xUnionOf}},
  \code{\link{xUnionOf...}}; \code{\link{xUniqueOf}},
  \code{\link{xUniqueOf...}}
}

