% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xDuplicatesOf}
\alias{xDuplicatesOf}
\alias{xDuplicatesOf_}
\title{xDuplicatesOf}
\usage{
xDuplicatesOf(coll)

xDuplicatesOf_(...)
}
\arguments{
\item{coll}{a collection. The collection to return
   duplicated elements from.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Return the duplicated elements in a collection.
}
\section{Type Signature}{

    |any| -> [any]
}

\section{Corner Cases}{

   Returns the empty list if \bold{colls} is length-zero.
}
\examples{

NULL

}
\seealso{
Other set_functions: \code{\link{xInterOf}},
  \code{\link{xInterOf_}}; \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsIn}},
  \code{\link{xIsIn_}}; \code{\link{xIsSubsetOf}},
  \code{\link{xIsSubsetOf_}}; \code{\link{xNotEmpty}},
  \code{\link{xNotEmpty_}}; \code{\link{xNotIn}},
  \code{\link{xNotIn_}}; \code{\link{xNotSubsetOf}},
  \code{\link{xNotSubsetOf_}}; \code{\link{xPowerSetOf}},
  \code{\link{xPowerSetOf_}}; \code{\link{xProdSetOf}},
  \code{\link{xProdSetOf_}}; \code{\link{xUnionOf}},
  \code{\link{xUnionOf_}}; \code{\link{xUniqueOf}},
  \code{\link{xUniqueOf_}}
}

