\name{xElemIsFalse}
\alias{xElemIsFalse}
\alias{xElemIsFalse...}
\title{xElemIsFalse}
\usage{
xElemIsFalse(coll)

xElemIsFalse...(...)
}
\arguments{
  \item{coll}{a collection. The collection to test each
  element of for being false.}

  \item{...}{see above.}
}
\value{
A vector of boolean values.
}
\description{
Is an element of a collection false?
}
\examples{

NULL

}

