\name{xElemIsNan}
\alias{xElemIsNan}
\alias{xElemIsNan...}
\title{xElemIsNan}
\usage{
xElemIsNan(coll)

xElemIsNan...(...)
}
\arguments{
  \item{coll}{a collection}

  \item{...}{see above.}
}
\value{
A vector of boolean values.
}
\description{
Is an element in a collection NaN?
}
\section{Corner Cases}{
  Returns True if coll is length-zero.
}
\examples{

NULL

}

