\name{xElemIsTrue}
\alias{xElemIsTrue}
\alias{xElemIsTrue...}
\title{xElemIsTrue}
\usage{
xElemIsTrue(coll)

xElemIsTrue...(...)
}
\arguments{
  \item{coll}{a collection. The collection to test each
  element of for being true.}

  \item{...}{see above.}
}
\value{
A vector of boolean values.
}
\description{
Is an element of a collection true?
}
\examples{

NULL

}
\seealso{
Other value_testing_functions: \code{\link{xElemIsFalse}},
\code{\link{xElemIsFalse...}}; \code{\link{xElemIsNan}},
\code{\link{xElemIsNan...}}; \code{\link{xElemIsNa}},
\code{\link{xElemIsNa...}}; \code{\link{xElemIsNull}},
\code{\link{xElemIsNull...}}; \code{\link{xElemNotFalse}},
\code{\link{xElemNotFalse...}}; \code{\link{xElemNotNan}},
\code{\link{xElemNotNan...}}; \code{\link{xElemNotNa}},
\code{\link{xElemNotNa...}}; \code{\link{xElemNotNull}},
\code{\link{xElemNotNull...}}; \code{\link{xElemNotTrue}},
\code{\link{xElemNotTrue...}}; \code{\link{xIsFalse}};
\code{\link{xIsNan}}; \code{\link{xIsNa}};
\code{\link{xIsNull}}; \code{\link{xIsTrue}};
\code{\link{xNotFalse}}; \code{\link{xNotNan}};
\code{\link{xNotNa}}; \code{\link{xNotNull}};
\code{\link{xNotTrue}}
}

