% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xElemNot}
\alias{xElemNot}
\alias{xElemNot_}
\title{xElemNot}
\usage{
xElemNot(val, coll)

xElemNot_(val, ...)
}
\arguments{
\item{val}{an arbitrary value. The value to use for testing.}

\item{coll}{a collection. The collection to test each element of.}

\item{...}{see above.}
}
\description{
Test if the elements of a collection are equal to a value.
}
\section{Type Signature}{

    |any| -> &lt;logical>
}

\section{Corner Cases}{

   If \bold{coll} is length-zero \bold{logical(0)} is
   returned. If NA is used as \bold{val} then all forms of
   NA are treated identically.
}
\examples{

# 1. Test which elements in a collection are NaN.

not_nan_element <- x_(c(1, 2, 3, 4, NaN, 5, 6, NaN)) $ x_ElemNot(NaN)

# c(TRUE, TRUE, TRUE, TRUE, FALSE, TRUE, TRUE, FALSE)
}
\seealso{
Other value_testing_functions: \code{\link{xElemIs}},
  \code{\link{xElemIs_}}; \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsFalse}};
  \code{\link{xIsNan}}; \code{\link{xIsNa}};
  \code{\link{xIsNull}}; \code{\link{xIsTrue}};
  \code{\link{xIs}}; \code{\link{xNotEmpty}},
  \code{\link{xNotEmpty_}}; \code{\link{xNotFalse}};
  \code{\link{xNotNan}}; \code{\link{xNotNa}};
  \code{\link{xNotNull}}; \code{\link{xNotTrue}};
  \code{\link{xNot}}
}

