% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xElemNotNa}
\alias{xElemNotNa}
\alias{xElemNotNa...}
\title{xElemNotNa}
\usage{
xElemNotNa(coll)

xElemNotNa...(...)
}
\arguments{
\item{coll}{a collection. The collection to test each element
   of for being non-na.}

\item{...}{see above.}
}
\value{
A vector of boolean values.
}
\description{
Test every element in a collection for being \code{Na} of any type.
}
\section{Corner Cases}{

   Returns logical(0) if \bold{coll} is length-zero.
}
\examples{

NULL

}
\seealso{
Other value_testing_functions: \code{\link{xElemIsFalse}},
  \code{\link{xElemIsFalse...}}; \code{\link{xElemIsNan}},
  \code{\link{xElemIsNan...}}; \code{\link{xElemIsNa}},
  \code{\link{xElemIsNa...}}; \code{\link{xElemIsNull}},
  \code{\link{xElemIsNull...}}; \code{\link{xElemIsTrue}},
  \code{\link{xElemIsTrue...}};
  \code{\link{xElemNotFalse}},
  \code{\link{xElemNotFalse...}};
  \code{\link{xElemNotNan}}, \code{\link{xElemNotNan...}};
  \code{\link{xElemNotNull}},
  \code{\link{xElemNotNull...}};
  \code{\link{xElemNotTrue}},
  \code{\link{xElemNotTrue...}}; \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty...}}; \code{\link{xIsFalse}};
  \code{\link{xIsNan}}; \code{\link{xIsNa}};
  \code{\link{xIsNull}}; \code{\link{xIsTrue}};
  \code{\link{xNotEmpty}}, \code{\link{xNotEmpty...}};
  \code{\link{xNotFalse}}; \code{\link{xNotNan}};
  \code{\link{xNotNa}}; \code{\link{xNotNull}};
  \code{\link{xNotTrue}}
}

