\name{xElemNotNan}
\alias{xElemNotNan}
\alias{xElemNotNan...}
\title{xElemNotNan}
\usage{
xElemNotNan(coll)

xElemNotNan...(...)
}
\arguments{
  \item{coll}{a collection}

  \item{...}{see above.}
}
\value{
A vector of boolean values.
}
\description{
Is an element of a collection not NaN?
}
\section{Corner Cases}{
  Returns the empty list if \code{coll is length-zero}.
}
\examples{

NULL

}

