\name{xElemNotTrue}
\alias{xElemNotTrue}
\alias{xElemNotTrue...}
\title{xElemNotTrue}
\usage{
xElemNotTrue(coll)

xElemNotTrue...(...)
}
\arguments{
  \item{coll}{a collection.}

  \item{...}{see above.}
}
\value{
A vector of boolean values.
}
\description{
Is an element of a collection not true?
}
\section{Corner Cases}{
  Returns logical(0) if \bold{coll} is length-zero.
}
\examples{

NULL

}

