% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xElemNotTrue}
\alias{xElemNotTrue}
\alias{xElemNotTrue_}
\title{xElemNotTrue}
\usage{
xElemNotTrue(coll)

xElemNotTrue_(...)
}
\arguments{
\item{coll}{a collection. The collection to test each element
   of for being non-true.}

\item{...}{see above.}
}
\value{
A vector of boolean values.
}
\description{
Is an element of a collection not true?
}
\section{Type Signature}{

    |any| -> &lt;logical>
}

\section{Corner Cases}{

   Returns logical(0) if \bold{coll} is length-zero.
}
\examples{

NULL

}
\seealso{
Other value_testing_functions: \code{\link{xElemIsFalse}},
  \code{\link{xElemIsFalse_}}; \code{\link{xElemIsNan}},
  \code{\link{xElemIsNan_}}; \code{\link{xElemIsNa}},
  \code{\link{xElemIsNa_}}; \code{\link{xElemIsNull}},
  \code{\link{xElemIsNull_}}; \code{\link{xElemIsTrue}},
  \code{\link{xElemIsTrue_}}; \code{\link{xElemNotFalse}},
  \code{\link{xElemNotFalse_}}; \code{\link{xElemNotNan}},
  \code{\link{xElemNotNan_}}; \code{\link{xElemNotNa}},
  \code{\link{xElemNotNa_}}; \code{\link{xElemNotNull}},
  \code{\link{xElemNotNull_}}; \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsFalse}};
  \code{\link{xIsNan}}; \code{\link{xIsNa}};
  \code{\link{xIsNot}}; \code{\link{xIsNull}};
  \code{\link{xIsTrue}}; \code{\link{xIs}};
  \code{\link{xNotEmpty}}, \code{\link{xNotEmpty_}};
  \code{\link{xNotFalse}}; \code{\link{xNotNan}};
  \code{\link{xNotNa}}; \code{\link{xNotNull}};
  \code{\link{xNotTrue}}
}

