% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xElemsByCols}
\alias{xElemsByCols}
\title{xElemsByCols}
\usage{
xElemsByCols(tab)
}
\arguments{
\item{tab}{a matrix. The matrix to enumerate by elements.}
}
\value{
A kea object containing a list.
}
\description{
Enumerate the elements of a matrix down columns.
}
\section{Corner Cases}{

     Returns the empty list if either dimension of \bold{tab} is empty.
}
\seealso{
Other methods: \code{\link{xByColkeys}};
  \code{\link{xByCols}}; \code{\link{xByLevels}};
  \code{\link{xByRowkeys}}; \code{\link{xByRows}};
  \code{\link{xByValues}}; \code{\link{xElemsByRows}};
  \code{\link{xExecute}}; \code{\link{xTap}};
  \code{\link{x_}}, \code{\link{x__}}
}

