% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xElemsByCols}
\alias{xElemsByCols}
\title{$ xElemsByCols}
\usage{
x_(  ) $ xElemsByCols()
}
\value{
An arrow object containing a list.
}
\description{
Enumerate the elements of a matrix down columns.
}
\seealso{
Other methods: \code{\link{xByColkeys}};
  \code{\link{xByCols}}; \code{\link{xByLevels}},
  \code{\link{xByLevels}}; \code{\link{xByRows}};
  \code{\link{xElemsByRows}}; \code{\link{xExecute}};
  \code{\link{xTap}}; \code{\link{x_}}
}

