% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xElemsByRows}
\alias{xElemsByRows}
\title{xElemsByRows}
\usage{
xElemsByRows(tab)
}
\arguments{
\item{tab}{a matrix. The matrix to enumerate by elements.}
}
\value{
A kiwi object containing a list.
}
\description{
Enumerate the elements of a matrix across rows.
}
\section{Corner Cases}{

     Returns the empty list if either dimension of \bold{tab} is empty.
}
\seealso{
Other methods: \code{\link{xByColkeys}};
  \code{\link{xByCols}}; \code{\link{xByLevels}};
  \code{\link{xByRowkeys}}; \code{\link{xByRows}};
  \code{\link{xByValues}}; \code{\link{xElemsByCols}};
  \code{\link{xExecute}}; \code{\link{xTap}};
  \code{\link{x_}}, \code{\link{x__}}
}

