% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xElemsByRows}
\alias{xElemsByRows}
\title{$ xElemsByRows}
\usage{
x_(  ) $ xElemsByRows()
}
\value{
An arrow object containing a list.
}
\description{
Enumerate the elements of a matrix across rows.
}
\seealso{
Other methods: \code{\link{xByColkeys}};
  \code{\link{xByCols}}; \code{\link{xByLevels}},
  \code{\link{xByLevels}}, \code{\link{xByValues}};
  \code{\link{xByRowkeys}}, \code{\link{xByRownames}};
  \code{\link{xByRows}}; \code{\link{xElemsByCols}};
  \code{\link{xExecute}}; \code{\link{xTap}};
  \code{\link{x_}}, \code{\link{x__}}
}

