% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xExecute}
\alias{xExecute}
\title{xExecute}
\usage{
xExecute(fn, val)
}
\arguments{
\item{fn}{a unary function. The function to apply to the data in the kea object.}

\item{val}{an arbitrary value. The contents of the kea object.}
}
\value{
The return value of the previous method.
}
\description{
Execute a (possibly side-effectual) function before
continuing with the previous method result.
}
\details{
\bold{xExecute} is mainly intended to be used as a method; it allows
    a side-effectful function to be run between method calls, keeping the
    previous result.

    \code{x_(1:10) $ xExecute(function () print('doing something')) $ x_SumBy(xI)}

    The above call wraps 1...10 in a kea object, prints 'doing something', and proceeds
    to sum the numbers 1...10.
}
\section{Type Signature}{

    (any -> any) -> any -> any
}

\section{Corner Cases}{

    None.
}
\seealso{
Other inpure_functions: \code{\link{xDelay}};
  \code{\link{xDo}}, \code{\link{xDo_}};
  \code{\link{xOneOf}}, \code{\link{xOneOf_}};
  \code{\link{xRead}}; \code{\link{xShuffle}},
  \code{\link{xShuffle_}}; \code{\link{xStopwatch}};
  \code{\link{xVersion}}; \code{\link{xWrite}}

Other methods: \code{\link{xByColkeys}};
  \code{\link{xByCols}}; \code{\link{xByLevels}};
  \code{\link{xByRowkeys}}; \code{\link{xByRows}};
  \code{\link{xByValues}}; \code{\link{xElemsByCols}};
  \code{\link{xElemsByRows}}; \code{\link{xTap}};
  \code{\link{x_}}, \code{\link{x__}}
}

