% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xExecute}
\alias{xExecute}
\title{$ xExecute}
\usage{
x_(  ) $ xExecute(fn)
}
\arguments{
\item{fn}{a unary function. The function to apply to the data in the arrow object.}

\item{val}{an arbitrary value. The contents of the arrow object.}
}
\value{
The return value of the previous method.
}
\description{
Execute a (possibly side-effectful) function before
continuing with the previous method result.
}
\section{Type Signature}{

    (any -> any) -> any -> any
}
\seealso{
Other inpure_functions: \code{\link{xDelay}};
  \code{\link{xDo}}, \code{\link{xDo_}};
  \code{\link{xReadChars}}; \code{\link{xReadLines}};
  \code{\link{xReadWords}}; \code{\link{xStopwatch}};
  \code{\link{xVersion}}; \code{\link{xWriteChars}};
  \code{\link{xWriteLines}}; \code{\link{xWriteWords}}

Other methods: \code{\link{xByColkeys}};
  \code{\link{xByCols}}; \code{\link{xByLevels}},
  \code{\link{xByLevels}}, \code{\link{xByValues}};
  \code{\link{xByRowkeys}}, \code{\link{xByRownames}};
  \code{\link{xByRows}}; \code{\link{xElemsByCols}};
  \code{\link{xElemsByRows}}; \code{\link{xTap}};
  \code{\link{x_}}, \code{\link{x__}}
}

