\name{xExecute}
\alias{xExecute}
\title{$ xExecute}
\usage{
x_(  ) $ xExecute(fn)
}
\arguments{
  \item{fn}{a nullary function}
}
\value{
the return value of the previous method.
}
\description{
Execute a (possibly side-effectful) function before
continuing with the previous method result.
}
\seealso{
Other exclusively_methods: \code{\link{xByColnames}};
\code{\link{xByCols}}; \code{\link{xByLevels}};
\code{\link{xByRows}}
}

