% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xExecute}
\alias{xExecute}
\title{xExecute}
\usage{
x_(  ) $ xExecute(fn)
}
\arguments{
\item{fn}{a unary function. The function to apply to the data in the kiwi object.}

\item{val}{an arbitrary value. The contents of the kiwi object.}
}
\value{
The return value of the previous method.
}
\description{
Execute a (possibly side-effectful) function before
continuing with the previous method result.
}
\section{Type Signature}{

    (any -> any) -> any -> any
}

\section{Corner Cases}{

}
\seealso{
Other inpure_functions: \code{\link{xDelay}};
  \code{\link{xDo}}, \code{\link{xDo_}};
  \code{\link{xRead}}; \code{\link{xStopwatch}};
  \code{\link{xVersion}}; \code{\link{xWrite}}

Other methods: \code{\link{xByColkeys}};
  \code{\link{xByCols}}; \code{\link{xByLevels}},
  \code{\link{xByLevels}}, \code{\link{xByValues}};
  \code{\link{xByRowkeys}}; \code{\link{xByRows}};
  \code{\link{xElemsByCols}}; \code{\link{xElemsByRows}};
  \code{\link{xTap}}; \code{\link{x_}}, \code{\link{x__}}
}

