% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xFirstOf}
\alias{xFirstOf}
\alias{xFirstOf_}
\title{xFirstOf}
\usage{
xFirstOf(coll)

xFirstOf_(...)
}
\arguments{
\item{coll}{a collection. The collection to get the
   first element of.}

\item{...}{see above.}
}
\value{
The first element in \bold{coll}.
}
\description{
Return the first element of a collection.
}
\section{Type Signature}{

    |any| -> any
}

\section{Corner Cases}{

   Throws an error if \bold{coll} has less than one element; this is
   because any other corner case would violate the functions type-signature.
}
\examples{

xFirstOf_('a', 'b', 'c')

# a

x_(letters) $ xFirstOf()

# a

xFirstOf(letters)

# a, yet again
}
\seealso{
Other selection_functions: \code{\link{xAtCol}},
  \code{\link{xAtCol_}}; \code{\link{xAtKey}},
  \code{\link{xAtKey_}}; \code{\link{xAt}},
  \code{\link{xAt_}}; \code{\link{xDropWhile}},
  \code{\link{xDropWhile_}}; \code{\link{xDrop}},
  \code{\link{xDrop_}}; \code{\link{xFourthOf}},
  \code{\link{xFourthOf_}}; \code{\link{xInitOf}},
  \code{\link{xInitOf_}}; \code{\link{xLastOf}},
  \code{\link{xLastOf_}}; \code{\link{xMaxBy}},
  \code{\link{xMaxBy_}}; \code{\link{xMinBy}},
  \code{\link{xMinBy_}}; \code{\link{xRestOf}},
  \code{\link{xRestOf_}}; \code{\link{xSecondOf}},
  \code{\link{xSecondOf_}}; \code{\link{xSlice}},
  \code{\link{xSlice_}}; \code{\link{xTakeWhile}},
  \code{\link{xTakeWhile_}}; \code{\link{xTake}},
  \code{\link{xTake_}}; \code{\link{xThirdOf}},
  \code{\link{xThirdOf_}}
}

