% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xFormalsOf}
\alias{xFormalsOf}
\title{xFormalsOf}
\usage{
xFormalsOf(fn)
}
\arguments{
\item{fn}{an arbitrary function. The function to get the parametres of.}
}
\value{
A named list, where each element's name is a parametre in \bold{fn} and each value
	  is the default value of that parametre.
}
\description{
xFormalsOf
}
\section{Type Signature}{

    function -> |any|

Get the parametres and defaults of a function or primitive function.
}

\section{Corner Cases}{

   If \bold{fn} is a primitive function a heuristic is used to obtain
   its formals. If a parametre has no default, the value of the
   corresponding element in the retun value will be the empty symbol,
   which is identical to \code{quote(expr=)}.
}
\examples{

NULL

}
\seealso{
Other parametre_functions: \code{\link{xArityOf}};
  \code{\link{xFix}}, \code{\link{xFix_}};
  \code{\link{xIsVariadic}}; \code{\link{xParamsOf}};
  \code{\link{xSpread}}; \code{\link{xUnspread}}
}

