% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xFourthOf}
\alias{xFourthOf}
\alias{xFourthOf_}
\title{xFourthOf}
\usage{
xFourthOf(coll)

xFourthOf_(...)
}
\arguments{
\item{coll}{a collection. The collection to get the
   fourth element of.}

\item{...}{see above.}
}
\value{
The fourth value in \bold{coll}.
}
\description{
xFourthOf
}
\section{Type Signature}{

    |any| -> any

Return the fourth value in a collection.
}

\section{Corner Cases}{

   Throws an error if \bold{coll} has less than four element; this is
   because any other corner case would violate the functions type-signature.
}
\seealso{
Other selection_functions: \code{\link{xAtKey}},
  \code{\link{xAtKey_}}; \code{\link{xAt}},
  \code{\link{xAt_}}; \code{\link{xDropWhile}},
  \code{\link{xDropWhile_}}; \code{\link{xDrop}},
  \code{\link{xDrop_}}; \code{\link{xFirstOf}},
  \code{\link{xFirstOf_}}; \code{\link{xInitOf}},
  \code{\link{xInitOf_}}; \code{\link{xLastOf}},
  \code{\link{xLastOf_}}; \code{\link{xMaxBy}},
  \code{\link{xMaxBy_}}; \code{\link{xMinBy}},
  \code{\link{xMinBy_}}; \code{\link{xRestOf}},
  \code{\link{xRestOf_}}; \code{\link{xSecondOf}},
  \code{\link{xSecondOf_}}; \code{\link{xSlice}},
  \code{\link{xSlice_}}; \code{\link{xTakeWhile}},
  \code{\link{xTakeWhile_}}; \code{\link{xTake}},
  \code{\link{xTake_}}; \code{\link{xThirdOf}},
  \code{\link{xThirdOf_}}
}

