% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xFourthOf}
\alias{xFourthOf}
\alias{xFourthOf...}
\title{xFourthOf}
\usage{
xFourthOf(coll)

xFourthOf...(...)
}
\arguments{
\item{coll}{a collection. The collection to get the
   fourth element of.}

\item{...}{see above.}
}
\value{
The fourth value in \bold{coll}.
}
\description{
Return the fourth value in a collection.
}
\section{Corner Cases}{

   Throws an error if \bold{coll} has less than four element; this is
   because any other corner case would violate the functions type-signature.
}
\seealso{
Other selection_functions: \code{\link{xAtCol}},
  \code{\link{xAtCol...}}; \code{\link{xAtKey}},
  \code{\link{xAtKey...}}; \code{\link{xAt}},
  \code{\link{xAt...}}; \code{\link{xDropWhile}},
  \code{\link{xDropWhile...}}; \code{\link{xDrop}},
  \code{\link{xDrop...}}; \code{\link{xFirstOf}},
  \code{\link{xFirstOf...}}; \code{\link{xInitOf}},
  \code{\link{xInitOf...}}; \code{\link{xLastOf}},
  \code{\link{xLastOf...}}; \code{\link{xMaxBy}},
  \code{\link{xMaxBy...}}; \code{\link{xMinBy}},
  \code{\link{xMinBy...}}; \code{\link{xPluck}},
  \code{\link{xPluck...}}; \code{\link{xRestOf}},
  \code{\link{xRestOf...}}; \code{\link{xSecondOf}},
  \code{\link{xSecondOf...}}; \code{\link{xSlice}},
  \code{\link{xSlice...}}; \code{\link{xTakeWhile}},
  \code{\link{xTakeWhile...}}; \code{\link{xTake}},
  \code{\link{xTake...}}; \code{\link{xThirdOf}},
  \code{\link{xThirdOf...}}
}

