% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xFromChars}
\alias{xFromChars}
\alias{xFromChars_}
\title{xFromChars}
\usage{
xFromChars(strs)

xFromChars_(...)
}
\arguments{
\item{strs}{a collection of strings. The
   strings to collapse into one string.}

\item{...}{see above.}
}
\value{
A length-one character vector.
}
\description{
xFromChars
}
\section{Type Signature}{

    |character| -> &lt;character>

Concatentate a character vector with no delimiter.
}

\section{Corner Cases}{

   Returns the empty character-vector when \bold{strs}
   is length-zero.
}
\examples{

# 1. Remove Specific Characters from a string
#    I would just use the base function gsub() here,
#    but xFromChars and xReject work too.

x_('the letter e is overrated') $ xToChars() $ xReject(x. == 'e') $ x_FromChars()
}
\seealso{
Other text_processing_functions: \code{\link{xAmend}};
  \code{\link{xCarve}}; \code{\link{xCharsOf}};
  \code{\link{xExplode}}; \code{\link{xFromLines}},
  \code{\link{xFromLines_}}; \code{\link{xFromWords}},
  \code{\link{xFromWords_}}; \code{\link{xImplode}},
  \code{\link{xImplode_}}; \code{\link{xIsMatch}};
  \code{\link{xNotMatch}}; \code{\link{xRead}};
  \code{\link{xToChars}}; \code{\link{xToLines}};
  \code{\link{xToWords}}; \code{\link{xWrite}}
}

