% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xFromChars}
\alias{xFromChars}
\alias{xFromChars...}
\title{xFromChars}
\usage{
xFromChars(strs)

xFromChars...(...)
}
\arguments{
\item{strs}{a collection of strings. The
   strings to collapse into one string.}

\item{...}{see above.}
}
\value{
A length-one character vector.
}
\description{
Concatentate a character vector with empty strs as delimiters.
}
\examples{

# 1. Remove Specific Characters from a string
#    in all likelyhood I'd use the base function gsub() here,
#    but xFromChars and xReject work too.

x_('the letter e is overrated') $
xToChars() $
xReject(
	char := {
		char == 'e'
}) $
x_FromChars()
}
\seealso{
Other text_processing_functions: \code{\link{xExplode}},
  \code{\link{xExplode...}}; \code{\link{xFromLines}},
  \code{\link{xFromLines...}}; \code{\link{xFromWords}},
  \code{\link{xFromWords...}}; \code{\link{xImplode}},
  \code{\link{xImplode...}}; \code{\link{xReadChars}};
  \code{\link{xReadLines}}; \code{\link{xReadWords}};
  \code{\link{xSliceString}}, \code{\link{xSliceString...}};
  \code{\link{xToChars}}; \code{\link{xToLines}};
  \code{\link{xToWords}}; \code{\link{xWriteChars}};
  \code{\link{xWriteLines}}; \code{\link{xWriteWords}}
}

