\name{xFromLines}
\alias{xFromLines}
\alias{xFromLines...}
\title{xFromLines}
\usage{
xFromLines(strs)

xFromLines...(...)
}
\arguments{
  \item{strs}{a collection of length-one character vectors.
  The strings to collapse into one string.}

  \item{...}{see above.}
}
\value{
A length-one character vector.
}
\description{
Concatentate a character vector with newlines as
delimiters.
}
\examples{

NULL

}
\seealso{
Other character_functions: \code{\link{xExplode}},
\code{\link{xExplode...}}; \code{\link{xFromChars}},
\code{\link{xFromChars...}}; \code{\link{xFromWords}},
\code{\link{xFromWords...}}; \code{\link{xImplode}},
\code{\link{xImplode...}}; \code{\link{xSubstring}},
\code{\link{xSubstring...}}; \code{\link{xToChars}};
\code{\link{xToLines}}; \code{\link{xToWords}}
}

