% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xFromLines}
\alias{xFromLines}
\alias{xFromLines_}
\title{xFromLines}
\usage{
xFromLines(strs)

xFromLines_(...)
}
\arguments{
\item{strs}{a collection of strings. The
   strings to collapse into one string.}

\item{...}{see above.}
}
\value{
A length-one character vector.
}
\description{
Concatentate a character vector with newlines as delimiters.
}
\section{Type Signature}{

    |character| -> &lt;character>
}

\section{Corner Cases}{

   Returns the empty character-vector when \bold{strs}
   is length-zero.
}
\examples{

NULL
}
\seealso{
Other text_processing_functions: \code{\link{xExplode}};
  \code{\link{xFromChars}}, \code{\link{xFromChars_}};
  \code{\link{xFromWords}}, \code{\link{xFromWords_}};
  \code{\link{xImplode}}, \code{\link{xImplode_}};
  \code{\link{xIsMatch}}; \code{\link{xNotMatch}};
  \code{\link{xRead}}; \code{\link{xSliceString}},
  \code{\link{xSliceString_}}; \code{\link{xToChars}};
  \code{\link{xToLines}}; \code{\link{xToWords}};
  \code{\link{xWrite}}
}

