% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xFromLines}
\alias{xFromLines}
\alias{xFromLines_}
\title{xFromLines}
\usage{
xFromLines(strs)

xFromLines_(...)
}
\arguments{
\item{strs}{a collection of strings. The
   strings to collapse into one string.}

\item{...}{see above.}
}
\value{
A length-one character vector.
}
\description{
Concatentate a character vector with newlines as delimiters.
}
\section{Type Signature}{

    |character| -> <character>
}
\examples{

NULL

}
\seealso{
Other text_processing_functions: \code{\link{xExplode}};
  \code{\link{xFromChars}}, \code{\link{xFromChars_}};
  \code{\link{xFromWords}}, \code{\link{xFromWords_}};
  \code{\link{xImplode}}, \code{\link{xImplode_}};
  \code{\link{xReadChars}}; \code{\link{xReadLines}};
  \code{\link{xReadWords}}; \code{\link{xSliceString}},
  \code{\link{xSliceString_}}; \code{\link{xToChars}};
  \code{\link{xToLines}}; \code{\link{xToWords}};
  \code{\link{xWriteChars}}; \code{\link{xWriteLines}};
  \code{\link{xWriteWords}}
}

