% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xGroupBy}
\alias{xGroupBy}
\alias{xGroupBy...}
\title{xGroupBy}
\usage{
xGroupBy(fn, coll)

xGroupBy...(fn, ...)
}
\arguments{
\item{fn}{a unary function. The function returning the value to group its input by.}

\item{coll}{a collection. The elements to group by a property.}

\item{...}{see above.}
}
\value{
A list
}
\description{
Group elements of a collection by the output of a function applied to each element.
}
\details{
xGroupBy allows the grouping of a heterogenous dataset into groups of elements
    similar in some way.
}
\section{Corner Cases}{

   If \bold{coll} is a empty collection the empty list is returned.
}
\examples{

NULL
}
\seealso{
Other reshaping_functions: \code{\link{xChop}},
  \code{\link{xChop...}}; \code{\link{xChunk}},
  \code{\link{xChunk...}}; \code{\link{xFlatten}},
  \code{\link{xFlatten...}}; \code{\link{xJoin}},
  \code{\link{xJoin...}}; \code{\link{xOneOf}},
  \code{\link{xOneOf...}}; \code{\link{xRecycle}},
  \code{\link{xRecycle...}}; \code{\link{xRepeat}},
  \code{\link{xRepeat...}}; \code{\link{xReverse}},
  \code{\link{xReverse...}}; \code{\link{xShuffle}},
  \code{\link{xShuffle...}}; \code{\link{xSortBy}},
  \code{\link{xSortBy...}}; \code{\link{xSplitAt}},
  \code{\link{xSplitAt...}}; \code{\link{xSplitWith}},
  \code{\link{xSplitWith...}}; \code{\link{xTabulate}},
  \code{\link{xTabulate...}}; \code{\link{xUnzipKeys}},
  \code{\link{xUnzipKeys...}}; \code{\link{xZipKeys}},
  \code{\link{xZipKeys...}}; \code{\link{xZip}},
  \code{\link{xZip...}}
}

