% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xIdentity}
\alias{xI}
\alias{xIdentity}
\title{xIdentity}
\usage{
xIdentity(val)

xI(val)
}
\arguments{
\item{val}{an arbitrary value.}
}
\value{
Returns \bold{val}.
}
\description{
Return an argument without modification.
}
\section{Type Signature}{

    any -> any
}

\section{Corner Cases}{

   None.
}
\examples{

# 1. Identity can be useful with higher-order functions.

xAllOf(xI, c(True, True, False))
xAllOf_(xI, True, True, False)

# False

xMaxBy_(xI, 1, 2, 3, 4, 5, 6)

# 6

xSumBy_(xIdentity, 1, 2, 3)

# 6
}
\seealso{
Other basic_functions: \code{\link{xCapture}},
  \code{\link{xK}}; \code{\link{xFalsity}};
  \code{\link{xIrrelevance}}; \code{\link{xTruth}};
  \code{\link{xClear}}, \code{\link{xClear_}}
}

