% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xIdentity}
\alias{xI}
\alias{xIdentity}
\title{xIdentity}
\usage{
xIdentity(val)

xI(val)
}
\arguments{
\item{val}{an arbitrary value.}
}
\value{
Returns \code{val}.
}
\description{
Return an argument without modification.
}
\section{Type Signature}{

    any -> any
}
\examples{

# 1. Identity can be useful with higher-order functions.

xAllOf(xI, c(True, True, False))

# False

xMaxBy(xI, c(1, 2, 3, 4, 5, 6))

# 6
}
\seealso{
Other basic_functions: \code{\link{xCapture}},
  \code{\link{xK}}; \code{\link{xFalsity}};
  \code{\link{xIrrelevance}}; \code{\link{xTruth}}
}

