\name{xIdentity}
\alias{xI}
\alias{xIdentity}
\title{xIdentity}
\usage{
xIdentity(val)

xI(val)
}
\arguments{
  \item{val}{an arbitrary value.}
}
\value{
returns \code{val}.
}
\description{
Return an argument without modification.
}
\section{Uses}{
  The identity function is useful for testing identities
  (for example, that mapping identity over a collection
  should return the original collection).
}

