\name{xImplode}
\alias{xImplode}
\alias{xImplode...}
\title{xImplode}
\usage{
xImplode(str, strs)

xImplode...(str, ...)
}
\arguments{
  \item{str}{a string to use as a delimiter.}

  \item{strs}{several character vectors.}

  \item{...}{see above.}
}
\value{
A length-one character vector.
}
\description{
Concatenate a character vector into a single string using a
delimiter.
}
\seealso{
Other character_functions: \code{\link{xExplode}},
\code{\link{xExplode...}}; \code{\link{xFromChars}},
\code{\link{xFromChars...}}; \code{\link{xFromLines}},
\code{\link{xFromLines...}}; \code{\link{xFromWords}},
\code{\link{xFromWords...}}; \code{\link{xSubstring}},
\code{\link{xSubstring...}}; \code{\link{xToChars}};
\code{\link{xToLines}}; \code{\link{xToWords}}
}

