\name{xImplode}
\alias{xImplode}
\alias{xImplode...}
\title{xImplode}
\usage{
xImplode(str, strs)

xImplode...(str, ...)
}
\arguments{
  \item{str}{a string to use as a delimiter.}

  \item{strs}{several character vectors.}

  \item{...}{see above.}
}
\value{
a length-one character vector.
}
\description{
Concatenate a character vector into a single string using a
delimiter.
}
\section{Variadic Functions

\link{ help_variadic }}{

}
\seealso{
Other character_functions: \code{\link{xChars}};
\code{\link{xExplode}}, \code{\link{xExplode...}};
\code{\link{xLines}}; \code{\link{xSubstring}},
\code{\link{xSubstring...}}; \code{\link{xUnchars}},
\code{\link{xUnchars...}}; \code{\link{xUnlines}},
\code{\link{xUnlines...}}; \code{\link{xUnwords}},
\code{\link{xUnwords...}}; \code{\link{xWords}}
}

