% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xImplode}
\alias{xImplode}
\alias{xImplode_}
\title{xImplode}
\usage{
xImplode(str, strs)

xImplode_(str, ...)
}
\arguments{
\item{str}{a length one character vector. The
   string to use as a delimiter.}

\item{strs}{a collection of length one character vectors. The
   strings to concatenate.}

\item{...}{see above.}
}
\value{
A length-one character vector.
}
\description{
Concatenate a character vector into a single string using a delimiter.
}
\section{Type Signature}{

    |character| -> |character| -> &lt;character>
}

\section{Corner Cases}{

   Returns the empty character vector when \bold{str} or
  \bold{strs} is length-zero. \bold{xImplode} does not coerce
   its arguments to character.
}

\section{Rcpp}{


    This function is written using C++, and should be quite performant.
}
\examples{

# 1. create a simple csv writer.

to_csv <- xImplode(', ')
to_csv(xTake(10, letters))

x_(letters) $ xImplode(',') $ x_Take(10)

# 'a,b,c,d,e,f,g,h,i,j'
}
\seealso{
Other text_processing_functions: \code{\link{xAmend}};
  \code{\link{xCharsOf}}; \code{\link{xExplode}};
  \code{\link{xFromChars}}, \code{\link{xFromChars_}};
  \code{\link{xFromLines}}, \code{\link{xFromLines_}};
  \code{\link{xFromWords}}, \code{\link{xFromWords_}};
  \code{\link{xIsMatch}}; \code{\link{xNotMatch}};
  \code{\link{xRead}}; \code{\link{xCarve}},
  \code{\link{xCarve_}}; \code{\link{xToChars}};
  \code{\link{xToLines}}; \code{\link{xToWords}};
  \code{\link{xWrite}}
}

