% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xImplode}
\alias{xImplode}
\alias{xImplode...}
\title{xImplode}
\usage{
xImplode(str, strs)

xImplode...(str, ...)
}
\arguments{
\item{str}{a length one character vector. The
   string to use as a delimiter.}

\item{strs}{a collection of length one character vectors. The
   strings to concatenate.}

\item{...}{see above.}
}
\value{
A length-one character vector.
}
\description{
Concatenate a character vector into a single string using a delimiter.
}
\examples{

# 1. create a simple csv writer.

to_csv <- xPartial(xImplode, ', ')
to_csv(xTake(10, letters))

# c('a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j')
}
\seealso{
Other text_processing_functions: \code{\link{xExplode}},
  \code{\link{xExplode...}}; \code{\link{xFromChars}},
  \code{\link{xFromChars...}}; \code{\link{xFromLines}},
  \code{\link{xFromLines...}}; \code{\link{xFromWords}},
  \code{\link{xFromWords...}}; \code{\link{xReadChars}};
  \code{\link{xReadLines}}; \code{\link{xReadWords}};
  \code{\link{xSliceString}},
  \code{\link{xSliceString...}}; \code{\link{xToChars}};
  \code{\link{xToLines}}; \code{\link{xToWords}};
  \code{\link{xWriteChars}}; \code{\link{xWriteLines}};
  \code{\link{xWriteWords}}
}

