% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xInter}
\alias{xInter}
\alias{xInter...}
\title{xInter}
\usage{
xInter(colls)

xInter...(...)
}
\arguments{
\item{colls}{a collection of collections. The collections to
   take the intersection of.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Get the set intersection of several collections.
}
\section{Corner Cases}{

   Returns the empty list if \bold{coll} is length-zero.
}
\examples{

NULL

}
\seealso{
Other set_functions: \code{\link{xDuplicatesOf}},
  \code{\link{xDuplicatesOf...}}; \code{\link{xIsMember}},
  \code{\link{xIsMember...}}; \code{\link{xIsSubset}},
  \code{\link{xIsSubset...}}; \code{\link{xNotMember}},
  \code{\link{xNotMember...}}; \code{\link{xPowerSetOf}},
  \code{\link{xPowerSetOf...}}; \code{\link{xProdSetOf}},
  \code{\link{xProdSetOf...}}; \code{\link{xReorder}},
  \code{\link{xReorder...}}; \code{\link{xUnionOf}},
  \code{\link{xUnionOf...}}; \code{\link{xUniqueOf}},
  \code{\link{xUniqueOf...}}
}

