% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xInter}
\alias{xInter}
\alias{xInter_}
\title{xInter}
\usage{
xInter(colls)

xInter_(...)
}
\arguments{
\item{colls}{a collection of collections. The collections to
   take the intersection of.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Get the set intersection of several collections.
}
\section{Corner Cases}{

   Returns the empty list if \bold{coll} is length-zero.
}
\examples{

NULL

}
\seealso{
Other set_functions: \code{\link{xDuplicatesOf}},
  \code{\link{xDuplicatesOf_}}; \code{\link{xIsMember}},
  \code{\link{xIsMember_}}; \code{\link{xIsSubset}},
  \code{\link{xIsSubset_}}; \code{\link{xNotMember}},
  \code{\link{xNotMember_}}; \code{\link{xPowerSetOf}},
  \code{\link{xPowerSetOf_}}; \code{\link{xProdSetOf}},
  \code{\link{xProdSetOf_}}; \code{\link{xReorder}},
  \code{\link{xReorder_}}; \code{\link{xUnionOf}},
  \code{\link{xUnionOf_}}; \code{\link{xUniqueOf}},
  \code{\link{xUniqueOf_}}
}

