% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xIntersect.R
\name{xIntersect}
\alias{xIntersect}
\alias{xIntersect_}
\title{xIntersect}
\usage{
xIntersect(coll1, coll2)

xIntersect_(coll1, ...)
}
\arguments{
\item{coll1}{a collection. The collection to test for membership in.}

\item{coll2}{a collection. The collection to return elements from.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Get the elements of a collection also in another.
}
\section{Type Signature}{

    |any| -> |any| -> [any]
}

\section{Corner Cases}{

   Returns the empty list if \bold{coll1} or \bold{coll2} is length-zero. \bold{xIntersect}
   preserves the names of \bold{coll2} where possible.
}
\examples{

NULL

}
\seealso{
Other set_functions: \code{\link{xDistinctOf}},
  \code{\link{xDistinctOf_}}; \code{\link{xDuplicatesOf}},
  \code{\link{xDuplicatesOf_}}; \code{\link{xExclude}},
  \code{\link{xExclude_}}; \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsIn}},
  \code{\link{xIsIn_}}; \code{\link{xIsSubsetOf}},
  \code{\link{xIsSubsetOf_}}; \code{\link{xNotEmpty}},
  \code{\link{xNotEmpty_}}; \code{\link{xNotIn}},
  \code{\link{xNotIn_}}; \code{\link{xNotSubsetOf}},
  \code{\link{xNotSubsetOf_}}; \code{\link{xPowerSetOf}},
  \code{\link{xPowerSetOf_}}; \code{\link{xCross}},
  \code{\link{xCross_}}; \code{\link{xUnite}},
  \code{\link{xUnite_}}
}

