% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xIntersect}
\alias{xIntersect}
\alias{xIntersect_}
\title{xIntersect}
\usage{
xIntersect(coll1, coll2)

xIntersect_(coll1, ...)
}
\arguments{
\item{coll1}{a collection. The first collection to get the intersection of.}

\item{...}{see above.}

\item{coll1}{a collection. The second collection to get the intersection of.}
}
\value{
A list.
}
\description{
Get the set intersection of two collections.
}
\details{
The intersection of two sets is the collection of elements shared by both.
}
\section{Type Signature}{

    |any| -> [any]
}

\section{Corner Cases}{

   Returns the empty list if \bold{coll} is length-zero.
}
\examples{

NULL

}
\seealso{
Other set_functions: \code{\link{xDistinctOf}},
  \code{\link{xDistinctOf_}}; \code{\link{xDuplicatesOf}},
  \code{\link{xDuplicatesOf_}}; \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsIn}},
  \code{\link{xIsIn_}}; \code{\link{xIsSubsetOf}},
  \code{\link{xIsSubsetOf_}}; \code{\link{xNotEmpty}},
  \code{\link{xNotEmpty_}}; \code{\link{xNotIn}},
  \code{\link{xNotIn_}}; \code{\link{xNotSubsetOf}},
  \code{\link{xNotSubsetOf_}}; \code{\link{xPowerSetOf}},
  \code{\link{xPowerSetOf_}}; \code{\link{xProdSetOf}},
  \code{\link{xProdSetOf_}}; \code{\link{xUnite}},
  \code{\link{xUnite_}}
}

