% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xIrrelevance}
\alias{xIrrelevance}
\title{xIrrelevance}
\usage{
xIrrelevance(...)
}
\arguments{
\item{...}{arguments to be dropped.}
}
\value{
na
}
\description{
xIrrelevance
}
\section{Type Signature}{

    ..any -> <logical>

A function that always returns na.
}

\section{Corner Cases}{

   None.
}
\examples{

# 1. Using irrelevance as an identity
#    with xReject.

xReject(xIrrelevance, 1:10)
}
\seealso{
Other basic_functions: \code{\link{xCapture}},
  \code{\link{xK}}; \code{\link{xFalsity}};
  \code{\link{xI}}, \code{\link{xIdentity}};
  \code{\link{xTruth}}; \code{\link{xClear}},
  \code{\link{xClear_}}
}

