% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xIs}
\alias{xIs}
\title{xIs}
\usage{
xIs(val1, val2)
}
\arguments{
\item{val1}{an arbitrary value. The first value to test.}

\item{val2}{an arbitrary value. The second value to test.}
}
\value{
A true or false value.
}
\description{
Are two values equal?
}
\section{Type Signature}{

    any -> any -> &lt;logical>
}

\section{Corner Cases}{

   \bold{xIs} does not allow for minor numeric inequality (0.999999 != 1),
   and treats -0. and 0. as identical.
}
\examples{

# 1. check if a value is NA

x_(Na) $ xIs(Na)
xIs(Na, Na)

# true
}
\seealso{
Other value_testing_functions: \code{\link{xElemIs}},
  \code{\link{xElemIs_}}; \code{\link{xElemNot}},
  \code{\link{xElemNot_}}; \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsFalse}};
  \code{\link{xIsNan}}; \code{\link{xIsNa}};
  \code{\link{xIsNull}}; \code{\link{xIsTrue}};
  \code{\link{xNotEmpty}}, \code{\link{xNotEmpty_}};
  \code{\link{xNotFalse}}; \code{\link{xNotNan}};
  \code{\link{xNotNa}}; \code{\link{xNotNull}};
  \code{\link{xNotTrue}}; \code{\link{xNot}}
}

