\name{xIsEmpty}
\alias{xIsEmpty}
\alias{xIsEmpty...}
\title{xIsEmpty}
\usage{
xIsEmpty(val)

xIsEmpty...(...)
}
\arguments{
  \item{val}{an arbitrary value. A value to check for being
  length zero.}

  \item{...}{see above.}
}
\value{
A boolean value.
}
\description{
Is a collection length-zero?
}
\examples{

NULL

}

