\name{xIsEmpty}
\alias{xIsEmpty}
\alias{xIsEmpty...}
\title{xIsEmpty}
\usage{
xIsEmpty(coll)

xIsEmpty...(...)
}
\arguments{
  \item{coll}{a collection.}

  \item{...}{see above.}
}
\value{
a boolean value.
}
\description{
Is a collection length-zero?
}
\section{Variadic Functions

\link{ help_variadic }}{

}

