% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xIsEmpty}
\alias{xIsEmpty}
\alias{xIsEmpty_}
\title{xIsEmpty}
\usage{
xIsEmpty(coll)

xIsEmpty_(...)
}
\arguments{
\item{coll}{a collection. The collection to test for being length zero.}

\item{...}{see above.}
}
\value{
A boolean value.
}
\description{
Is a collection length-zero?
}
\section{Type Signature}{

    |any| -> &lt;logical>
}

\section{Corner Cases}{

}
\examples{

# 1.
# xIsEmpty is used to check for a recursive function's base case below.

map <- function (fn, coll) {

	if (xIsEmpty(coll)) {
		list()
	} else {
		xJoin_(
			fn(xFirstOf(coll)),
			map(fn, xRestOf(coll)) )
	}
}

map(sqrt, 1:4)

# list(1, 1.4142135623731, 1.73205080756888, 2)

# 2.
# xIsEmpty can be used as a predicate for reject.

xReject_(xIsEmpty, 'a', 'b', integer(0), 'c')

# list('a', 'b', 'c')
}
\seealso{
Other value_testing_functions: \code{\link{xIsFalse}};
  \code{\link{xIsNan}}; \code{\link{xIsNa}};
  \code{\link{xIsNull}}; \code{\link{xIsTrue}};
  \code{\link{xIs}}; \code{\link{xNotEmpty}},
  \code{\link{xNotEmpty_}}; \code{\link{xNotFalse}};
  \code{\link{xNotNan}}; \code{\link{xNotNa}};
  \code{\link{xNotNull}}; \code{\link{xNotTrue}};
  \code{\link{xNot}}
}

