% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xIsEmpty}
\alias{xIsEmpty}
\alias{xIsEmpty_}
\title{xIsEmpty}
\usage{
xIsEmpty(coll)

xIsEmpty_(...)
}
\arguments{
\item{coll}{a collection. The collection to test for being length zero.}

\item{...}{see above.}
}
\value{
A boolean value.
}
\description{
Is a collection length-zero?
}
\section{Type Signature}{

    |any| -> <logical>
}

\section{Corner Cases}{

   Returns false for non-empty collections.
}
\examples{

# 1.
# xIsEmpty is used to check for a recursive function's base case below.

map <- function (fn, coll) {

	if (xIsEmpty(coll)) {
		list()
	} else {
		xJoin_(
			fn(xFirstOf(coll)),
			map(fn, xRestOf(coll)) )
	}
}

map(sqrt, 1:4)

# list(1, 1.4142135623731, 1.73205080756888, 2)

# 2.
# xIsEmpty can be used as a predicate for reject.

xReject_(xIsEmpty, 'a', 'b', integer(0), 'c')

# list('a', 'b', 'c')
}
\seealso{
Other set_functions: \code{\link{xDistinctOf}},
  \code{\link{xDistinctOf_}}; \code{\link{xDuplicatesOf}},
  \code{\link{xDuplicatesOf_}}; \code{\link{xExclude}},
  \code{\link{xExclude_}}; \code{\link{xIntersect}},
  \code{\link{xIntersect_}}; \code{\link{xIsIn}},
  \code{\link{xIsIn_}}; \code{\link{xIsSubsetOf}},
  \code{\link{xIsSubsetOf_}}; \code{\link{xNotEmpty}},
  \code{\link{xNotEmpty_}}; \code{\link{xNotIn}},
  \code{\link{xNotIn_}}; \code{\link{xNotSubsetOf}},
  \code{\link{xNotSubsetOf_}}; \code{\link{xPowerSetOf}},
  \code{\link{xPowerSetOf_}}; \code{\link{xProdSetOf}},
  \code{\link{xProdSetOf_}}; \code{\link{xUnite}},
  \code{\link{xUnite_}}

Other value_testing_functions: \code{\link{xIsFalse}};
  \code{\link{xIsInfixOf}}, \code{\link{xIsInfixOf_}};
  \code{\link{xIsNan}}; \code{\link{xIsNa}};
  \code{\link{xIsNull}}; \code{\link{xIsPrefixOf}},
  \code{\link{xIsPrefixOf_}}; \code{\link{xIsSuffixOf}},
  \code{\link{xIsSuffixOf_}}; \code{\link{xIsTrue}};
  \code{\link{xIs}}; \code{\link{xNotEmpty}},
  \code{\link{xNotEmpty_}}; \code{\link{xNotFalse}};
  \code{\link{xNotInfixOf}}, \code{\link{xNotInfixOf_}};
  \code{\link{xNotNan}}; \code{\link{xNotNa}};
  \code{\link{xNotNull}}; \code{\link{xNotPrefixOf}},
  \code{\link{xNotPrefixOf_}}; \code{\link{xNotSuffixOf}},
  \code{\link{xNotSuffixOf_}}; \code{\link{xNotTrue}};
  \code{\link{xNot}}
}

