\name{xIsEmpty}
\alias{xIsEmpty}
\alias{xIsEmpty...}
\title{xIsEmpty}
\usage{
xIsEmpty(coll)

xIsEmpty...(...)
}
\arguments{
  \item{coll}{a collection. The collection to check.}

  \item{...}{see above.}
}
\value{
A boolean value.
}
\description{
Is a collection length-zero?
}
\examples{

NULL

}

