\name{xIsFalse}
\alias{xIsFalse}
\alias{xIsFalse...}
\title{xIsFalse}
\usage{
xIsFalse(coll)

xIsFalse...(...)
}
\arguments{
  \item{coll}{a collection. The collection to test each
  element of.}

  \item{...}{see above.}
}
\value{
A vector of boolean values.
}
\description{
Is an element of a collection false?
}

