% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xIsIn}
\alias{xIsIn}
\alias{xIsIn_}
\title{xIsIn}
\usage{
xIsIn(val, coll)

xIsIn_(val, ...)
}
\arguments{
\item{val}{an arbitrary value. The value to test for membership
   in \bold{coll}.}

\item{coll}{a collection. The collection of elements to check against
   \bold{val}.}

\item{...}{see above.}
}
\value{
A boolean value.
}
\description{
Check if a collection contains a value.
}
\section{Type Signature}{

    any -> |any| -> <logical>
}

\section{Corner Cases}{

   The equality check used by \bold{xIsIn} has the same corner cases as \bold{xIs};
   equal integers and doubles are treated as different numbers, -0 is equal to +0.
   Returns \bold{logical(0)} when \bold{coll} is empty.
}

\section{Rcpp}{


    This function is written using C++, and should be quite performant.
}
\examples{

# 1. A simple set membership example.

is_primary_colour <- is_primary_color <- val := {
	xIsIn_(val, 'red', 'green', 'blue')
}

is_primary_colour('red')

# True

is_primary_colour('magenta')

# False
}
\seealso{
Other set_functions: \code{\link{xDistinctOf}},
  \code{\link{xDistinctOf_}}; \code{\link{xDuplicatesOf}},
  \code{\link{xDuplicatesOf_}}; \code{\link{xExclude}},
  \code{\link{xExclude_}}; \code{\link{xIntersect}},
  \code{\link{xIntersect_}}; \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsSubsetOf}},
  \code{\link{xIsSubsetOf_}}; \code{\link{xNotEmpty}},
  \code{\link{xNotEmpty_}}; \code{\link{xNotIn}},
  \code{\link{xNotIn_}}; \code{\link{xNotSubsetOf}},
  \code{\link{xNotSubsetOf_}}; \code{\link{xPowerSetOf}},
  \code{\link{xPowerSetOf_}}; \code{\link{xProdSetOf}},
  \code{\link{xProdSetOf_}}; \code{\link{xUnite}},
  \code{\link{xUnite_}}
}

