% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xIsNa}
\alias{xIsNa}
\title{xIsNa}
\usage{
xIsNa(val)
}
\arguments{
\item{val}{an arbitrary value. The value to test for being na.}

\item{...}{see above.}
}
\value{
A true or false value.
}
\description{
Is an value na?
}
\section{Type Signature}{

    any -> &lt;logical>
}

\section{Corner Cases}{

    xIsNa returns either true or false, to make it
    safe for use with if statements. If val is length zero
    False is returned.
}
\examples{

# 1. Test if a single value is NA

xIsNa(Na)

# True; every other case will return false.
}
\seealso{
Other value_testing_functions: \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsFalse}};
  \code{\link{xIsNan}}; \code{\link{xIsNull}};
  \code{\link{xIsTrue}}; \code{\link{xIs}};
  \code{\link{xNotEmpty}}, \code{\link{xNotEmpty_}};
  \code{\link{xNotFalse}}; \code{\link{xNotNan}};
  \code{\link{xNotNa}}; \code{\link{xNotNull}};
  \code{\link{xNotTrue}}; \code{\link{xNot}}
}

