% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xIsNa}
\alias{xIsNa}
\title{xIsNa}
\usage{
xIsNa(val)
}
\arguments{
\item{val}{an arbitrary value. The value to test for being na.}

\item{...}{see above.}
}
\value{
A true or false value.
}
\description{
Is an value na?
}
\section{Corner Cases}{

    xIsNa returns either true or false, to make it
    safe for use with if statements. If val is length zero
    False is returned.
}
\examples{

NULL
}
\seealso{
Other value_testing_functions: \code{\link{xElemIsFalse}},
  \code{\link{xElemIsFalse...}}; \code{\link{xElemIsNan}},
  \code{\link{xElemIsNan...}}; \code{\link{xElemIsNa}},
  \code{\link{xElemIsNa...}}; \code{\link{xElemIsNull}},
  \code{\link{xElemIsNull...}}; \code{\link{xElemIsTrue}},
  \code{\link{xElemIsTrue...}};
  \code{\link{xElemNotFalse}},
  \code{\link{xElemNotFalse...}};
  \code{\link{xElemNotNan}}, \code{\link{xElemNotNan...}};
  \code{\link{xElemNotNa}}, \code{\link{xElemNotNa...}};
  \code{\link{xElemNotNull}},
  \code{\link{xElemNotNull...}};
  \code{\link{xElemNotTrue}},
  \code{\link{xElemNotTrue...}}; \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty...}}; \code{\link{xIsFalse}};
  \code{\link{xIsNan}}; \code{\link{xIsNull}};
  \code{\link{xIsTrue}}; \code{\link{xNotEmpty}},
  \code{\link{xNotEmpty...}}; \code{\link{xNotFalse}};
  \code{\link{xNotNan}}; \code{\link{xNotNa}};
  \code{\link{xNotNull}}; \code{\link{xNotTrue}}
}

