\name{xIsNan}
\alias{xIsNan}
\alias{xIsNan...}
\title{xIsNan}
\usage{
xIsNan(coll)

xIsNan...(...)
}
\arguments{
  \item{coll}{a collection}

  \item{...}{see above.}
}
\value{
A vector of boolean values.
}
\description{
Is an element in a collection NaN?
}
\section{Corner Cases}{
  Returns True if coll is length-zero.
}
\examples{

NULL

}

