\name{xIsNan}
\alias{xIsNan}
\alias{xIsNan...}
\title{xIsNan}
\usage{
xIsNan(coll)

xIsNan...(...)
}
\arguments{
  \item{coll}{a collection}

  \item{...}{see above.}
}
\value{
a list.
}
\description{
Is an element in a collection NaN?
}
\section{Corner Cases}{
  returns True if coll is length-zero.
}

\section{Variadic Functions

\link{ help_variadic }}{

}

