% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xIsNan}
\alias{xIsNan}
\title{xIsNan}
\usage{
xIsNan(val)
}
\arguments{
\item{val}{an arbitrary value. The value to test for being
   nan.}

\item{...}{see above.}
}
\value{
A true or false value.
}
\description{
Is an value nan?
}
\section{Type Signature}{

    any -> <logical>
}

\section{Corner Cases}{

    xIsNan returns either true or false, to make it
    safe for use with if statements. If val is length zero
    False is returned.
}
\examples{

# 1. Division by zero generates NaN,
# and the value may pop up (often erroneously) in other contexts.

val <- 0 / 0

xIsNan(val)

# True
}
\seealso{
Other value_testing_functions: \code{\link{xElemIsFalse}},
  \code{\link{xElemIsFalse_}}; \code{\link{xElemIsNan}},
  \code{\link{xElemIsNan_}}; \code{\link{xElemIsNa}},
  \code{\link{xElemIsNa_}}; \code{\link{xElemIsNull}},
  \code{\link{xElemIsNull_}}; \code{\link{xElemIsTrue}},
  \code{\link{xElemIsTrue_}}; \code{\link{xElemNotFalse}},
  \code{\link{xElemNotFalse_}}; \code{\link{xElemNotNan}},
  \code{\link{xElemNotNan_}}; \code{\link{xElemNotNa}},
  \code{\link{xElemNotNa_}}; \code{\link{xElemNotNull}},
  \code{\link{xElemNotNull_}}; \code{\link{xElemNotTrue}},
  \code{\link{xElemNotTrue_}}; \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsFalse}};
  \code{\link{xIsNa}}; \code{\link{xIsNull}};
  \code{\link{xIsTrue}}; \code{\link{xIs}};
  \code{\link{xNotEmpty}}, \code{\link{xNotEmpty_}};
  \code{\link{xNotFalse}}; \code{\link{xNotNan}};
  \code{\link{xNotNa}}; \code{\link{xNotNull}};
  \code{\link{xNotTrue}}
}

