% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xIsNull}
\alias{xIsNull}
\title{xIsNull}
\usage{
xIsNull(val)
}
\arguments{
\item{val}{an arbitrary value. The value to test for
   being null.}

\item{...}{see above.}
}
\value{
A true or false value.
}
\description{
Is an value null?
}
\section{Type Signature}{

    any -> &lt;logical>
}

\section{Corner Cases}{

    xIsNull returns either true or false, to make it
    safe for use with if statements.
}
\examples{

# 1. The obvious example

xIsNull(Null)

# True
}
\seealso{
Other value_testing_functions: \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsFalse}};
  \code{\link{xIsNan}}; \code{\link{xIsNa}};
  \code{\link{xIsTrue}}; \code{\link{xIs}};
  \code{\link{xNotEmpty}}, \code{\link{xNotEmpty_}};
  \code{\link{xNotFalse}}; \code{\link{xNotNan}};
  \code{\link{xNotNa}}; \code{\link{xNotNull}};
  \code{\link{xNotTrue}}; \code{\link{xNot}}
}

