\name{xIsNull}
\alias{xIsNull}
\alias{xIsNull...}
\title{xIsNull}
\usage{
xIsNull(coll)

xIsNull...(...)
}
\arguments{
  \item{coll}{a collection}

  \item{...}{see above.}
}
\value{
A vector of boolean values.
}
\description{
Is an element of a collection null?
}
\section{Corner Cases}{
  Returns True if coll is Null.
}

\section{Variadic Functions

\link{ help_variadic }}{

}

