% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xIsSubset}
\alias{xIsSubset}
\alias{xIsSubset...}
\title{xIsSubset}
\usage{
xIsSubset(coll1, coll2)

xIsSubset...(coll1, ...)
}
\arguments{
\item{coll1}{a collection. The set to test for membership in a superset.}

\item{coll1}{a collection. The superset to test.}

\item{...}{see above.}
}
\value{
A boolean value.
}
\description{
Test if a collection is a subset of a second collection.
}
\section{Corner Cases}{

   Returns logical(0) if \bold{coll1} or \bold{coll2} is length-zero.
}
\examples{

NULL
}
\seealso{
Other set_functions: \code{\link{xDuplicatesOf}},
  \code{\link{xDuplicatesOf...}}; \code{\link{xInter}},
  \code{\link{xInter...}}; \code{\link{xIsMember}},
  \code{\link{xIsMember...}}; \code{\link{xNotMember}},
  \code{\link{xNotMember...}}; \code{\link{xPermute}},
  \code{\link{xPermute...}}; \code{\link{xPowerSetOf}},
  \code{\link{xPowerSetOf...}}; \code{\link{xProdSetOf}},
  \code{\link{xProdSetOf...}}; \code{\link{xUnionOf}},
  \code{\link{xUnionOf...}}; \code{\link{xUniqueOf}},
  \code{\link{xUniqueOf...}}
}

