% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xIsSubset}
\alias{xIsSubset}
\alias{xIsSubset_}
\title{xIsSubset}
\usage{
xIsSubset(coll1, coll2)

xIsSubset_(coll1, ...)
}
\arguments{
\item{coll1}{a collection. The set to test for membership in a superset.}

\item{coll2}{a collection. The superset to test.}

\item{...}{see above.}
}
\value{
A boolean value.
}
\description{
Test if a collection is a subset of a second collection.
}
\section{Type Signature}{

    |any| -> |any| -> &lt;logical>
}

\section{Corner Cases}{

   Returns logical(0) if \bold{coll1} or \bold{coll2} is length-zero.
}
\examples{

NULL
}
\seealso{
Other set_functions: \code{\link{xDuplicatesOf}},
  \code{\link{xDuplicatesOf_}}; \code{\link{xInterOf}},
  \code{\link{xInterOf_}}; \code{\link{xIsMember}},
  \code{\link{xIsMember_}}; \code{\link{xNotMember}},
  \code{\link{xNotMember_}}; \code{\link{xNotSubset}},
  \code{\link{xNotSubset_}}; \code{\link{xPowerSetOf}},
  \code{\link{xPowerSetOf_}}; \code{\link{xProdSetOf}},
  \code{\link{xProdSetOf_}}; \code{\link{xUnionOf}},
  \code{\link{xUnionOf_}}; \code{\link{xUniqueOf}},
  \code{\link{xUniqueOf_}}
}

