% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xIsSuffixOf}
\alias{xIsSuffixOf}
\alias{xIsSuffixOf_}
\title{xIsSuffixOf}
\usage{
xIsSuffixOf(coll1, coll2)

xIsSuffixOf_(coll1, ...)
}
\arguments{
\item{coll1}{a collection. The suffix to test for.}

\item{coll2}{a collection. The collection to test for a suffix.}

\item{...}{see above.}
}
\value{
A boolean value.
}
\description{
Is a collection a suffix of another?
}
\section{Type Signature}{

    |any| -> |any| -> &lt;logical>
}

\section{Corner Cases}{

   If the first collection is longer than the second, False is returned.
   If either collection is length-zero, \bold{logical(0)} is returned.
}
\examples{

# 1. A toy example.

xIsSuffixOf(2:4, 1:4)

# true
}
\seealso{
Other value_testing_functions: \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsFalse}};
  \code{\link{xIsInfixOf}}, \code{\link{xIsInfixOf_}};
  \code{\link{xIsNan}}; \code{\link{xIsNa}};
  \code{\link{xIsNull}}; \code{\link{xIsPrefixOf}},
  \code{\link{xIsPrefixOf_}}; \code{\link{xIsTrue}};
  \code{\link{xIs}}; \code{\link{xNotEmpty}},
  \code{\link{xNotEmpty_}}; \code{\link{xNotFalse}};
  \code{\link{xNotInfixOf}}, \code{\link{xNotInfixOf_}};
  \code{\link{xNotNan}}; \code{\link{xNotNa}};
  \code{\link{xNotNull}}; \code{\link{xNotPrefixOf}},
  \code{\link{xNotPrefixOf_}}; \code{\link{xNotSuffixOf}},
  \code{\link{xNotSuffixOf_}}; \code{\link{xNotTrue}};
  \code{\link{xNot}}
}

