\name{xIsVal}
\alias{xIsVal}
\title{xIsVal}
\usage{
xIsVal(sym)
}
\arguments{
  \item{sym}{a symbol or string. The variable name to
  test.}
}
\value{
A boolean value.
}
\description{
Is a name binding locked?
}
\section{Corner Cases}{
  If a variable isn't assigned, then \code{false} is
  returned.
}
\examples{

NULL

}

