% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xIsVal}
\alias{xIsVal}
\title{xIsVal}
\usage{
xIsVal(sym)
}
\arguments{
\item{sym}{a symbol or string. The variable name to test.}
}
\value{
A boolean value.
}
\description{
Is a variable locked?
}
\section{Corner Cases}{

   If a variable isn't assigned, then \code{false} is returned.
}
\examples{

# 1. xIsVal has a very specific use case; testing if a value can
# be modified.

xVal(tau, 6.283185)

xIsVal(tau)

# True

}

