% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xIsVariadic}
\alias{xIsVariadic}
\title{xIsVariadic}
\usage{
xIsVariadic(fn)
}
\arguments{
\item{fn}{an function. The function to test for
   variadic parametres}
}
\value{
A boolean value.
}
\description{
Does a function have ellipsis arguments?
}
\section{Type Signature}{

    any -> &lt;logical>
}

\section{Corner Cases}{

   Returns true when a function has a mix of variadic
   and non-variadic parametres.
}

\section{Stability - \bold{Uncertain}}{


This feature may be removed or altered significantly in future versions.
If you find this feature useful, please leave an issue at https://github.com/rgrannell1/kea/issues.
}
\examples{

# 1. Test if a function is variadic

xIsVariadic(sum)

# true
}
\seealso{
Other parametre_functions: \code{\link{xArityOf}};
  \code{\link{xFix}}, \code{\link{xFix_}};
  \code{\link{xFormalsOf}}; \code{\link{xParamsOf}};
  \code{\link{xSpread}}; \code{\link{xUnspread}}
}

