\name{xIsVariadic}
\alias{xIsVariadic}
\title{xIsVariadic}
\usage{
xIsVariadic(fn)
}
\arguments{
  \item{fn}{an arbitrary function or primitive function.}
}
\value{
A boolean value.
}
\description{
Is a function variadic?
}
\seealso{
Other parametre_functions: \code{\link{xAsUnary}};
\code{\link{xAsVariadic}}; \code{\link{xFlip}};
\code{\link{xFormalsOf}}; \code{\link{xParamsOf}};
\code{\link{xPartial}}, \code{\link{xPartial...}}
}

