% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xIsVariadic}
\alias{xIsVariadic}
\title{xIsVariadic}
\usage{
xIsVariadic(fn)
}
\arguments{
\item{fn}{an function. The function to test for
   variadic parametres.}
}
\value{
A boolean value.
}
\description{
Does a function have ellipsis arguments?
}
\examples{

# 1. Test if a function is variadic

xIsVariadic(sum)

# true
}
\seealso{
Other parametre_functions: \code{\link{xArityOf}};
  \code{\link{xFormalsOf}}; \code{\link{xParamsOf}};
  \code{\link{xPartial}}, \code{\link{xPartial_}};
  \code{\link{xSpread}}; \code{\link{xUnspread}}
}

