\name{xIsVariadic}
\alias{xIsVariadic}
\title{xIsVariadic}
\usage{
xIsVariadic(fn)
}
\arguments{
  \item{fn}{an function. The function to test for variadic
  parametres.}
}
\value{
A boolean value.
}
\description{
Does a function have ellipsis arguments?
}
\examples{

NULL

}
\seealso{
Other parametre_functions: \code{\link{xArityOf}};
\code{\link{xAsUnary}}; \code{\link{xAsVariadic}};
\code{\link{xFormalsOf}}; \code{\link{xParamsOf}};
\code{\link{xPartial}}, \code{\link{xPartial...}}
}

