% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xJoin}
\alias{xJoin}
\alias{xJoin...}
\title{xJoin}
\usage{
xJoin(colls)

xJoin...(...)
}
\arguments{
\item{colls}{a collection of collections. The collections
   to concatenate to each other.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Concatenate several collections into one collection.
}
\section{Corner Cases}{

   Empty collections act as a unit for concatenation;
   concatenating the empty list to another list returns
   the second list, without modification.
}
\examples{

# 1.
# The variadic form of xJoin is very similar to c( )

xJoin...('a', 'b', 'c')
# list('a', 'b', 'c')
}
\seealso{
Other reshaping_functions: \code{\link{xChop}},
  \code{\link{xChop...}}; \code{\link{xChunk}},
  \code{\link{xChunk...}}; \code{\link{xFlatten}},
  \code{\link{xFlatten...}}; \code{\link{xGroupBy}},
  \code{\link{xGroupBy...}}; \code{\link{xOneOf}},
  \code{\link{xOneOf...}}; \code{\link{xRecycle}},
  \code{\link{xRecycle...}}; \code{\link{xRepeat}},
  \code{\link{xRepeat...}}; \code{\link{xReverse}},
  \code{\link{xReverse...}}; \code{\link{xShuffle}},
  \code{\link{xShuffle...}}; \code{\link{xSortBy}},
  \code{\link{xSortBy...}}; \code{\link{xSplitAt}},
  \code{\link{xSplitAt...}}; \code{\link{xSplitWith}},
  \code{\link{xSplitWith...}}; \code{\link{xTabulate}},
  \code{\link{xTabulate...}}; \code{\link{xUnzipKeys}},
  \code{\link{xUnzipKeys...}}; \code{\link{xZipKeys}},
  \code{\link{xZipKeys...}}; \code{\link{xZip}},
  \code{\link{xZip...}}
}

