% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xJoin.R
\name{xJoin}
\alias{xJoin}
\alias{xJoin_}
\title{xJoin}
\usage{
xJoin(colls)

xJoin_(...)
}
\arguments{
\item{colls}{a collection of collections. The collections
   to concatenate to each other.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Concatenate several collections into one collection.
}
\details{
xJoin is useful for flatting a collection by one
   level; it can take a list of lists, and return
   a list of each item joined end to end.
}
\section{Type Signature}{

   ||any|| -> [any]
}

\section{Corner Cases}{

   Empty collections act as a unit for concatenation;
   concatenating the empty list to another list returns
   the second, without modification.
}
\examples{

# 1.
# The variadic form of xJoin is very similar to c( ),
#

xJoin_('a', 'b', 'c')
# list('a', 'b', 'c')
}
\seealso{
Other reshaping_functions: \code{\link{xAppend}},
  \code{\link{xAppend_}}; \code{\link{xChunk}},
  \code{\link{xChunk_}}; \code{\link{xFlatten}},
  \code{\link{xFlatten_}}; \code{\link{xGather}},
  \code{\link{xGather_}}; \code{\link{xGroupBy}},
  \code{\link{xGroupBy_}}; \code{\link{xOneOf}},
  \code{\link{xOneOf_}}; \code{\link{xPrepend}},
  \code{\link{xPrepend_}}; \code{\link{xRecycle}},
  \code{\link{xRecycle_}}; \code{\link{xRepeat}},
  \code{\link{xRepeat_}}; \code{\link{xReverse}},
  \code{\link{xReverse_}}; \code{\link{xShuffle}},
  \code{\link{xShuffle_}}; \code{\link{xSortBy}},
  \code{\link{xSortBy_}}; \code{\link{xTabulate}},
  \code{\link{xTabulate_}}; \code{\link{xUnzipIndices}},
  \code{\link{xUnzipIndices_}}; \code{\link{xUnzipKeys}},
  \code{\link{xUnzipKeys_}}; \code{\link{xZipKeys}},
  \code{\link{xZipKeys_}}; \code{\link{xZip}},
  \code{\link{xZip_}}
}

