% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xJuxtapose}
\alias{xJuxtapose}
\alias{xJuxtapose...}
\title{xJuxtapose}
\usage{
xJuxtapose(fns)

xJuxtapose...(...)
}
\arguments{
\item{fns}{a collections of functions. The functions
   to call with a set of arguments.}

\item{...}{see above.}
}
\value{
A variadic function that returns a list.
}
\description{
Create a function that applies an argument to a list of underlying functions.
}
\details{
\bold{xJuxtapose} is a method of applying one value to several functions
   simultaneously. The function returned by \bold{xJuxtapose} returns
   a list - on element for each input function to \bold{xJuxtapose}.
   Each element of this list contains the result of calling an element of
   \bold{fns} with the arguments supplied to \bold{xJuxtapose}'s return function.

   A function to summarise a data set can be implemented with \bold{xJuxtapose}:

   \code{summarise <- xJuxtapose...(min, median, max)}

   \code{summarise(c(3, 1, 2, 4))}

   \code{list(1, 2.5, 4)}

   Each function passed to \bold{xJuxtapose} is invoked with the data set,
   and the results are collected in a list.
}
\section{Corner Cases}{

   If no functions are provided the empty list is returned.
}
\examples{

# 1. Collect several summary statistics of a dataset
#    simultaneously.

nums <- xList[c(a, a^2, a^3), a <- 1:10]

x_(nums) $ x_Map(xJuxtapose...(max, mean, min))
}
\seealso{
Other function_combining_functions: \code{\link{\%and\%}},
  \code{\link{\%or\%}}, \code{\link{xLift}},
  \code{\link{xLift...}}; \code{\link{\%of\%}},
  \code{\link{\%then\%}}, \code{\link{xCompose}},
  \code{\link{xCompose...}}

Other function_modifying_functions: \code{\link{\%and\%}},
  \code{\link{\%or\%}}, \code{\link{xLift}},
  \code{\link{xLift...}}; \code{\link{\%of\%}},
  \code{\link{\%then\%}}, \code{\link{xCompose}},
  \code{\link{xCompose...}}; \code{\link{xAsClosure}};
  \code{\link{xCapture}}, \code{\link{xK}};
  \code{\link{xDelay}}; \code{\link{xNot}};
  \code{\link{xPartial}}, \code{\link{xPartial...}};
  \code{\link{xSpread}}; \code{\link{xThread}},
  \code{\link{xThread...}}; \code{\link{xUnspread}};
  \code{\link{xVectorise}}, \code{\link{xVectorize}}
}

