% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xKeysOf}
\alias{xKeysOf}
\alias{xKeysOf_}
\title{xKeysOf}
\usage{
xKeysOf(coll)

xKeysOf_(...)
}
\arguments{
\item{coll}{a collection. The collection to get the names of.}

\item{...}{see above.}
}
\value{
a character vector.
}
\description{
Get the names of a collection.
}
\section{Type Signature}{

    |any| -> <character>
}

\section{Corner Cases}{

    Returns character(0) when \bold{coll} is length-zero.
}
\examples{

NULL
}
\seealso{
Other key_functions: \code{\link{xAddKeys}},
  \code{\link{xAddKeys_}}; \code{\link{xIndicesOf}},
  \code{\link{xIndicesOf_}}; \code{\link{xUnzipIndices}},
  \code{\link{xUnzipIndices_}}; \code{\link{xUnzipKeys}},
  \code{\link{xUnzipKeys_}}; \code{\link{xZipKeys}},
  \code{\link{xZipKeys_}}
}

