% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xLastOf}
\alias{xLastOf}
\alias{xLastOf...}
\title{xLastOf}
\usage{
xLastOf(coll)

xLastOf...(...)
}
\arguments{
\item{coll}{a collection. The collection to return
   the last element of.}

\item{...}{see above.}
}
\value{
The value of the last element in \bold{coll}.
}
\description{
Return the last element in a collection.
}
\section{Corner Cases}{

   Throws an error if \bold{coll} has less than one element; this is
   because any other corner case would violate the function's type-signature.
}
\examples{

NULL

}
\seealso{
Other selection_functions: \code{\link{xAtCol}},
  \code{\link{xAtCol...}}; \code{\link{xAt}},
  \code{\link{xAt...}}; \code{\link{xDropWhile}},
  \code{\link{xDropWhile...}}; \code{\link{xDrop}},
  \code{\link{xDrop...}}; \code{\link{xFirstOf}},
  \code{\link{xFirstOf...}}; \code{\link{xFourthOf}},
  \code{\link{xFourthOf...}}; \code{\link{xInitOf}},
  \code{\link{xInitOf...}}; \code{\link{xMaxBy}},
  \code{\link{xMaxBy...}}; \code{\link{xMinBy}},
  \code{\link{xMinBy...}}; \code{\link{xPluck}},
  \code{\link{xPluck...}}; \code{\link{xRestOf}},
  \code{\link{xRestOf...}}; \code{\link{xSecondOf}},
  \code{\link{xSecondOf...}}; \code{\link{xSlice}},
  \code{\link{xSlice...}}; \code{\link{xTakeWhile}},
  \code{\link{xTakeWhile...}}; \code{\link{xTake}},
  \code{\link{xTake...}}; \code{\link{xThirdOf}},
  \code{\link{xThirdOf...}}
}

