\name{xLength}
\alias{xLength}
\alias{xLength...}
\title{xLength}
\usage{
xLength(coll)

xLength...(...)
}
\arguments{
  \item{coll}{a collection}

  \item{...}{see above.}
}
\value{
a nonnegative integer.
}
\description{
Get the length of a collection
}
\section{Corner Cases}{
  returns zero if \code{coll} is empty.
}

\section{Variadic Functions

\link{ help_variadic }}{

}

