% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xLenOf}
\alias{xLenOf}
\alias{xLenOf_}
\title{xLenOf}
\usage{
xLenOf(coll)

xLenOf_(...)
}
\arguments{
\item{coll}{a collection. The collection to test the
   length of.}

\item{...}{see above.}
}
\value{
A nonnegative integer.
}
\description{
Get the length of a collection.
}
\section{Type Signature}{

    |any| -> &lt;number>
}

\section{Corner Cases}{

     Returns zero if \bold{coll} is empty.
}
\examples{

# 1. find the length of a list

xLenOf(letters)

# 26

x_(letters) $ x_LenOf()

# 26
}
\seealso{
Other math_functions: \code{\link{xMeanOf}},
  \code{\link{xMeanOf_}}; \code{\link{xPoll}},
  \code{\link{xPoll_}}; \code{\link{xProdBy}},
  \code{\link{xProdBy_}}; \code{\link{xSumBy}},
  \code{\link{xSumBy_}}
}

