\name{xLenOf}
\alias{xLenOf}
\alias{xLenOf...}
\title{xLenOf}
\usage{
xLenOf(coll)

xLenOf...(...)
}
\arguments{
  \item{coll}{a collection}

  \item{...}{see above.}
}
\value{
A nonnegative integer.
}
\description{
Get the length of a collection
}
\section{Corner Cases}{
  Returns zero if \bold{coll} is empty.
}

