\name{xLift}
\alias{xLift}
\title{xLift}
\usage{
xLift(fn, fns)
}
\arguments{
  \item{fn}{a binary function.}

  \item{fns}{a list or pairlist of binary functions.}
}
\value{
returns a unary function of x.
}
\description{
Compose a binary function with two other functions.
}
\seealso{
Other higher_order_functions: \code{\link{xApply}};
\code{\link{xArity}}; \code{\link{xAsUnary}};
\code{\link{xCompose}}; \code{\link{xConst}};
\code{\link{xFlip}}; \code{\link{xFormals}};
\code{\link{xIterate}}; \code{\link{xJuxtapose}};
\code{\link{xLimit}}; \code{\link{xNot}};
\code{\link{xParams}}; \code{\link{xStopWatch}};
\code{\link{xUnfold}}; \code{\link{xWait}}
}

