% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xLimit}
\alias{xLimit}
\title{xLimit}
\usage{
xLimit(fn, num)
}
\arguments{
\item{fn}{an arbitrary function. The function to
   limit the number of times it can be called.}

\item{num}{a nonnegative whole number. The number
   of times to allow the function to be called.}
}
\value{
A function with the same parametres as \bold{fn}.
}
\description{
Create a function that can call its underlying
function a limited number of times.
}
\details{
\bold{xLimit} limits the number of times a function
   can be called. It is primarily used for functions that
   makes request to online services with a rate limit.
   \bold{xLimit} makes it easy to respect a website's
   rate limits without adding a lot of boilerplate to your
   code.

   \code{require(RCurl)}

   \code{url <- "http://randomword.setgetgo.com/get.php"}

   \code{random_word <- xLimit(function () paste(httpGET(url)), 3)}

   \code{xMap(function (ith) random_word(), 1:3)}

   \code{list("mashy\r\n", "iridectomy\r\n", "cystopyelitis\r\n")}
}
\section{Type Signature}{

    function -> &lt;numeric> -> function
}

\section{Corner Cases}{

   xLimit throws an error if \bold{num} is not length-one.
}
\examples{

# 1. Get random words from an API.

# require(RCurl)

# url <- "http://randomword.setgetgo.com/get.php"

# random_word <-
#	x_( function (...) paste(httpGET(url)) ) $
#	x_Limit(3)


# xMap(random_word, 1:4)

# list("regulative", "pulverizator", "unsigmatic", NULL)

# the final request isn't executed, since the limit
# of three calls was exceeded.
}
\seealso{
Other time_functions: \code{\link{xDelay}};
  \code{\link{xStopwatch}}
}

