\name{xLines}
\alias{xLines}
\title{xLines}
\usage{
xLines(str)
}
\arguments{
  \item{str}{a string.}
}
\value{
a character vector.
}
\description{
Split a string at every newline character.
}
\seealso{
Other character_functions: \code{\link{xChars}};
\code{\link{xExplode}}; \code{\link{xImplode}},
\code{\link{xImplode...}}; \code{\link{xSubString}},
\code{\link{xSubString...}}; \code{\link{xUnchars}},
\code{\link{xUnchars...}}; \code{\link{xUnlines}},
\code{\link{xUnlines...}}; \code{\link{xUnwords}},
\code{\link{xUnwords...}}; \code{\link{xWords}}
}

