% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xLocate.R
\name{xLocate}
\alias{xLocate}
\alias{xLocate_}
\title{xLocate}
\usage{
xLocate(pred, coll)

xLocate_(pred, ...)
}
\arguments{
\item{pred}{a predicate function. The function to test each element
   of a collection with.}

\item{coll}{a collection. The collection with elements to test.}

\item{...}{see above.}
}
\value{
An length-one or length-zero whole number.
}
\description{
Get the indices of the elements that return true for a predicate.
}
\section{Type Signature}{

    (any -> <logical>) -> |any| -> <integer>
}

\section{Corner Cases}{

     returns integer(0) if no match is found, or \bold{coll} is length-zero.
}
\examples{

# 1. Use xLocate like xWhich

xLocate(xI, c(True, False, True, Na))

# c(1, 3)

# 2. Locate the solutions to a polynomial

solves_equation <- x := {
	x^2 - 6*x == 0
}

x_(-10:10) $ xLocate(solves_equation)

# the 11th and 17th elements of -10, -9, ..., 10 solve the equation. (x = 0, x = 6)
}
\seealso{
Other key_functions: \code{\link{xAddKeys}},
  \code{\link{xAddKeys_}}; \code{\link{xIndicesOf}},
  \code{\link{xIndicesOf_}}; \code{\link{xIndicesTo}};
  \code{\link{xKeysOf}}, \code{\link{xKeysOf_}};
  \code{\link{xUnzipIndices}},
  \code{\link{xUnzipIndices_}}; \code{\link{xUnzipKeys}},
  \code{\link{xUnzipKeys_}}; \code{\link{xValuesOf}},
  \code{\link{xValuesOf_}}; \code{\link{xZipKeys}},
  \code{\link{xZipKeys_}}
}

