% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xLocate}
\alias{xLocate}
\alias{xLocate...}
\title{xLocate}
\usage{
xLocate(pred, coll)

xLocate...(pred, ...)
}
\arguments{
\item{pred}{a predicate function. The function to test each element
   of a collection with.}

\item{coll}{a collection. The collection with elements to test.}

\item{...}{see above.}
}
\value{
An length-one or length-zero whole number.
}
\description{
Get the indices of the elements that return true for a predicate.
}
\section{Corner Cases}{

     returns integer(0) if no match is found.
}
\examples{

NULL
}

