% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xMapply}
\alias{xMapply}
\alias{xMapply...}
\title{xMapply}
\usage{
xMapply(fn, colls)

xMapply...(fn, ...)
}
\arguments{
\item{fn}{a function. The function to apply to each tuple
   of elements in \bold{colls}.}

\item{colls}{a collection of collections. The collection
   to have a function applied to each inner tuple.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Apply a function to each element of a collection.
}
\section{Corner Cases}{

   Returns the empty list is \bold{colls} is length-zero.
}
\examples{

# 1. CE Find the minimum distance
#    between any two points in a set of points.
#    xMapply could be swapped for xMap(xUnspread(fn) )
#    below, as was done with the call to reject, but
#    map is important enough to have a few variants for
#    convenience.

points <- list(
	c(12, 2),
	c(32, -4),
	c(5, -2),
	c(6, 7),
	c(9, 31)
)

# get the set product to generate all pairs of points.
x_( xProdSetOf...(points, points) ) $
# remove the pairs wich contain the same point twice.
xReject(xUnspread(identical)) $
xMapply(
	(xs : ys) := {
		# the distance function.

		sqrt( (xs[1] - xs[2])^2 + (ys[1] - ys[2])^2 )

	}
) $
x_Reduce(min)

# 7.071068
}
\seealso{
Other function_application_functions: \code{\link{xApply}},
  \code{\link{xApply...}}; \code{\link{xUnspread}};
  \code{\link{xSpread}}; \code{\link{xDeepMap}},
  \code{\link{xDeepMap...}}; \code{\link{xFlatMap}},
  \code{\link{xFlatMap...}}; \code{\link{xMapIndexed}},
  \code{\link{xMapIndexed...}}; \code{\link{xMap}},
  \code{\link{xMap...}}; \code{\link{xThread}},
  \code{\link{xThread...}}; \code{\link{xVectorise}},
  \code{\link{xVectorize}}
}

