% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xMeanOf}
\alias{xMeanOf}
\alias{xMeanOf_}
\title{xMeanOf}
\usage{
xMeanOf(nums)

xMeanOf_(...)
}
\arguments{
\item{nums}{a vector of numbers. The numbers to compute the mean of.}

\item{...}{see above.}
}
\value{
A single double vector.
}
\description{
Return the mean of a collection of numbers.
}
\section{Type Signature}{

    |numeric| -> &lt;double>
}

\section{Corner Cases}{

   If \bold{nums} is empty then \bold{double(0)} is returned. The mean
   of vectors containing infinite values is infinite. If \bold{nums}
   cannot contain NA or NaN values, as these values do not have a defined
   magnitude.
}
\examples{

# 1. Get the mean of random uniform numbers between 0...1

x_(xRepeat(1, 100)) $ xFlatMap(runif) $ x_MeanOf()

# 0.5496172
}

