% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xMinBy}
\alias{xMinBy}
\alias{xMinBy...}
\title{xMinBy}
\usage{
xMinBy(fn, coll)

xMinBy...(fn, ...)
}
\arguments{
\item{fn}{a unary function. The function to measure the size of
    an element.}

\item{coll}{a non-empty collection. The collection to find the
   smallest value of.}

\item{...}{see above.}
}
\value{
An element from \bold{coll}.
}
\description{
Get the smallest value in a collection according to a measure function.
}
\section{Corner Cases}{

   If an empty collection is given, an error is thrown. This is because
   is impossible to return an element from a list that has none.
}
\examples{

NULL
}
\seealso{
Other selection_functions: \code{\link{xAtCol}},
  \code{\link{xAtCol...}}; \code{\link{xAtKey}},
  \code{\link{xAtKey...}}; \code{\link{xAt}},
  \code{\link{xAt...}}; \code{\link{xDropWhile}},
  \code{\link{xDropWhile...}}; \code{\link{xDrop}},
  \code{\link{xDrop...}}; \code{\link{xFirstOf}},
  \code{\link{xFirstOf...}}; \code{\link{xFourthOf}},
  \code{\link{xFourthOf...}}; \code{\link{xInitOf}},
  \code{\link{xInitOf...}}; \code{\link{xLastOf}},
  \code{\link{xLastOf...}}; \code{\link{xMaxBy}},
  \code{\link{xMaxBy...}}; \code{\link{xPluck}},
  \code{\link{xPluck...}}; \code{\link{xRestOf}},
  \code{\link{xRestOf...}}; \code{\link{xSecondOf}},
  \code{\link{xSecondOf...}}; \code{\link{xSlice}},
  \code{\link{xSlice...}}; \code{\link{xTakeWhile}},
  \code{\link{xTakeWhile...}}; \code{\link{xTake}},
  \code{\link{xTake...}}; \code{\link{xThirdOf}},
  \code{\link{xThirdOf...}}
}

