\name{xNegate}
\alias{xNegate}
\title{Flip the sign of each number in a vector.}
\usage{
  xNegate(nums)
}
\arguments{
  \item{nums}{a vector of numbers.}
}
\value{
  a vector of number.
}
\description{
  xNegate
}
\section{Corner Cases}{
  If nums is empty then the unit of the collection is
  returned.
}

