\name{xNegate}
\alias{xNegate}
\alias{xNegate...}
\title{xNegate}
\usage{
xNegate(nums)

xNegate...(...)
}
\arguments{
  \item{nums}{a vector of numbers.}

  \item{...}{see above.}
}
\value{
A vector of number.
}
\description{
Flip the sign of each number in a vector.
}
\section{Corner Cases}{
  If nums is empty then the unit of the collection is
  returned.
}
\examples{

NULL

}

