% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xNegate}
\alias{xNegate}
\title{xNegate}
\usage{
xNegate(pred)
}
\arguments{
\item{pred}{a predicate. The predicate function to return a negation of.}
}
\value{
A predicate function with the same parametres as \code{pred}.
}
\description{
Return the logical negation of a function.
}
\details{
\bold{xNegate} takes a predicate function - such as \bold{is.integer( )} -
   and returns the negation of that function - \bold{not.integer( )}.

   \code{not.integer <- xNegate(is.integer)}

   \code{is.integer(1L)}

   \code{True}

   \code{not.integer(1L)}

   \code{False}

   In general, if a function \bold{fn} returns true \bold{xNegate(fn)} returns false
   for the same value. Similarily, if \bold{fn} returns false then \bold{xNegate(fn)}
   returns true for that value. Finally (R uses three-value logic) if \bold{fn} returns
   na for a value so does its negation.
}
\section{Type Signature}{

    (any -> &lt;logical>) -> (any -> &lt;logical>)
}
\examples{

# 1. Reverse a type checking predicate

xSelect_(xNegate(is.character), 'a', 'b', 3, 'd', 5)

# list(3, 5)
}
\seealso{
Other function_modifying_functions: \code{\link{\%and\%}},
  \code{\link{\%or\%}}, \code{\link{xLift}},
  \code{\link{xLift_}}; \code{\link{\%of\%}},
  \code{\link{\%then\%}}, \code{\link{xCompose}},
  \code{\link{xCompose_}}; \code{\link{xCapture}},
  \code{\link{xK}}; \code{\link{xDelay}};
  \code{\link{xFix}}, \code{\link{xFix_}};
  \code{\link{xJuxtapose}}, \code{\link{xJuxtapose_}};
  \code{\link{xSpread}}; \code{\link{xThread}},
  \code{\link{xThread_}}; \code{\link{xUnspread}};
  \code{\link{xVectorise}}, \code{\link{xVectorize}}
}

