% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xNot}
\alias{xNot}
\title{xNot}
\usage{
xNot(val1, val2)
}
\arguments{
\item{val1}{an arbitrary value. The first value to test.}

\item{val2}{an arbitrary value. The second value to test.}
}
\value{
A true or false value.
}
\description{
Are two values not equal?
}
\section{Type Signature}{

    any -> any -> <logical>
}

\section{Corner Cases}{

   Shares its corner-cases with the base function \link{identical}.
}

\section{Corner Cases}{

   \bold{xNot} does not allow for minor numeric inequality (0.999999 != 1),
   and treats -0. and 0. as identical.
}
\examples{

# 1. check if a value isn't NaN

x_(NaN) $ xNot(NaN)
xNot(NaN, NaN)

# False
}
\seealso{
Other value_testing_functions: \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsFalse}};
  \code{\link{xIsInfixOf}}, \code{\link{xIsInfixOf_}};
  \code{\link{xIsNan}}; \code{\link{xIsNa}};
  \code{\link{xIsNull}}; \code{\link{xIsPrefixOf}},
  \code{\link{xIsPrefixOf_}}; \code{\link{xIsSuffixOf}},
  \code{\link{xIsSuffixOf_}}; \code{\link{xIsTrue}};
  \code{\link{xIs}}; \code{\link{xNotEmpty}},
  \code{\link{xNotEmpty_}}; \code{\link{xNotFalse}};
  \code{\link{xNotInfixOf}}, \code{\link{xNotInfixOf_}};
  \code{\link{xNotNan}}; \code{\link{xNotNa}};
  \code{\link{xNotNull}}; \code{\link{xNotPrefixOf}},
  \code{\link{xNotPrefixOf_}}; \code{\link{xNotSuffixOf}},
  \code{\link{xNotSuffixOf_}}; \code{\link{xNotTrue}}
}

