% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xNotEmpty}
\alias{xNotEmpty}
\alias{xNotEmpty_}
\title{xNotEmpty}
\usage{
xNotEmpty(coll)

xNotEmpty_(...)
}
\arguments{
\item{coll}{a collection. The value to test for
   having non-zero length.}

\item{...}{see above.}
}
\value{
A boolean value.
}
\description{
Is a collection not length-zero?
}
\section{Type Signature}{

    any -> &lt;logical>
}

\section{Corner Cases}{

   Returns true for non-empty collections.
}
\examples{

# 1. Conditionally execute code on non-empty collections.
#
# Often code on lists will have two branches;
# the case for the empty list and the case for the
# non-empty list.

safeFirstOf <- function (coll, default = NA) {
	if (xNotEmpty(coll)) {
		xFirstOf(coll)
	} else {
		default
	}
}

safeFirstOf(letters)

# 'a'

safeFirstOf(character(0))

# Na - xFirstOf usually fails for empty collections.
}
\seealso{
Other value_testing_functions: \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsFalse}};
  \code{\link{xIsNan}}; \code{\link{xIsNa}};
  \code{\link{xIsNull}}; \code{\link{xIsTrue}};
  \code{\link{xIs}}; \code{\link{xNotFalse}};
  \code{\link{xNotNan}}; \code{\link{xNotNa}};
  \code{\link{xNotNull}}; \code{\link{xNotTrue}};
  \code{\link{xNot}}
}

