% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xNotEmpty}
\alias{xNotEmpty}
\alias{xNotEmpty_}
\title{xNotEmpty}
\usage{
xNotEmpty(coll)

xNotEmpty_(...)
}
\arguments{
\item{coll}{a collection. The value to test for
   having non-zero length.}

\item{...}{see above.}
}
\value{
A boolean value.
}
\description{
Is a collection not length-zero?
}
\section{Type Signature}{

    any -> &lt;logical>
}
\examples{

NULL

}
\seealso{
Other value_testing_functions: \code{\link{xElemIs}},
  \code{\link{xElemIs_}}; \code{\link{xElemNot}},
  \code{\link{xElemNot_}}; \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsFalse}};
  \code{\link{xIsNan}}; \code{\link{xIsNa}};
  \code{\link{xIsNull}}; \code{\link{xIsTrue}};
  \code{\link{xIs}}; \code{\link{xNotFalse}};
  \code{\link{xNotNan}}; \code{\link{xNotNa}};
  \code{\link{xNotNull}}; \code{\link{xNotTrue}};
  \code{\link{xNot}}
}

