% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xNotEmpty}
\alias{xNotEmpty}
\alias{xNotEmpty_}
\title{xNotEmpty}
\usage{
xNotEmpty(coll)

xNotEmpty_(...)
}
\arguments{
\item{coll}{a collection. The value to test for
   having non-zero length.}

\item{...}{see above.}
}
\value{
A boolean value.
}
\description{
Is a collection not length-zero?
}
\section{Type Signature}{

    any -> &lt;logical>
}

\section{Corner Cases}{

   Returns true for non-empty collections.
}
\examples{

# 1. Conditionally execute code on non-empty collections.
#
# Often code on lists will have two branches;
# the case for the empty list and the case for the
# non-empty list.

safeFirstOf <- function (coll, default = NA) {
	if (xNotEmpty(coll)) {
		xFirstOf(coll)
	} else {
		default
	}
}

safeFirstOf(letters)

# 'a'

safeFirstOf(character(0))

# Na - xFirstOf usually fails for empty collections.
}
\seealso{
Other set_functions: \code{\link{xDuplicatesOf}},
  \code{\link{xDuplicatesOf_}}; \code{\link{xIntersect}},
  \code{\link{xIntersect_}}; \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsIn}},
  \code{\link{xIsIn_}}; \code{\link{xIsSubsetOf}},
  \code{\link{xIsSubsetOf_}}; \code{\link{xNotIn}},
  \code{\link{xNotIn_}}; \code{\link{xNotSubsetOf}},
  \code{\link{xNotSubsetOf_}}; \code{\link{xPowerSetOf}},
  \code{\link{xPowerSetOf_}}; \code{\link{xProdSetOf}},
  \code{\link{xProdSetOf_}}; \code{\link{xUnionOf}},
  \code{\link{xUnionOf_}}; \code{\link{xUniqueOf}},
  \code{\link{xUniqueOf_}}

Other value_testing_functions: \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsFalse}};
  \code{\link{xIsInfixOf}}, \code{\link{xIsInfixOf_}};
  \code{\link{xIsNan}}; \code{\link{xIsNa}};
  \code{\link{xIsNull}}; \code{\link{xIsPrefixOf}},
  \code{\link{xIsPrefixOf_}}; \code{\link{xIsSuffixOf}},
  \code{\link{xIsSuffixOf_}}; \code{\link{xIsTrue}};
  \code{\link{xIs}}; \code{\link{xNotFalse}};
  \code{\link{xNotInfixOf}}, \code{\link{xNotInfixOf_}};
  \code{\link{xNotNan}}; \code{\link{xNotNa}};
  \code{\link{xNotNull}}; \code{\link{xNotPrefixOf}},
  \code{\link{xNotPrefixOf_}}; \code{\link{xNotSuffixOf}},
  \code{\link{xNotSuffixOf_}}; \code{\link{xNotTrue}};
  \code{\link{xNot}}
}

