% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xNotEmpty}
\alias{xNotEmpty}
\alias{xNotEmpty...}
\title{xNotEmpty}
\usage{
xNotEmpty(coll)

xNotEmpty...(...)
}
\arguments{
\item{coll}{a collection. The value to test for
   having non-zero length.}

\item{...}{see above.}
}
\value{
A boolean value.
}
\description{
Is a collection not length-zero?
}
\examples{

NULL

}
\seealso{
Other value_testing_functions: \code{\link{xElemIsFalse}},
  \code{\link{xElemIsFalse...}}; \code{\link{xElemIsNan}},
  \code{\link{xElemIsNan...}}; \code{\link{xElemIsNa}},
  \code{\link{xElemIsNa...}}; \code{\link{xElemIsNull}},
  \code{\link{xElemIsNull...}}; \code{\link{xElemIsTrue}},
  \code{\link{xElemIsTrue...}};
  \code{\link{xElemNotFalse}},
  \code{\link{xElemNotFalse...}};
  \code{\link{xElemNotNan}}, \code{\link{xElemNotNan...}};
  \code{\link{xElemNotNa}}, \code{\link{xElemNotNa...}};
  \code{\link{xElemNotNull}},
  \code{\link{xElemNotNull...}};
  \code{\link{xElemNotTrue}},
  \code{\link{xElemNotTrue...}}; \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty...}}; \code{\link{xIsFalse}};
  \code{\link{xIsNan}}; \code{\link{xIsNa}};
  \code{\link{xIsNull}}; \code{\link{xIsSubset}},
  \code{\link{xIsSubset...}}; \code{\link{xIsTrue}};
  \code{\link{xNotFalse}}; \code{\link{xNotNan}};
  \code{\link{xNotNa}}; \code{\link{xNotNull}};
  \code{\link{xNotTrue}}
}

