% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xNotFalse.R
\name{xNotFalse}
\alias{xNotFalse}
\title{xNotFalse}
\usage{
xNotFalse(val)
}
\arguments{
\item{val}{an arbitrary value.The value to test for
   not-falsity.}

\item{...}{see above.}
}
\value{
A true or false value.
}
\description{
Is an value not false?
}
\section{Type Signature}{

    any -> <logical>
}

\section{Corner Cases}{

    xNotFalse returns either true or false, to make it
    safe for use with if statements. If val is length zero
    True is returned.
}
\examples{

# 1. An obvious example.

xNotFalse(True)

# True
}
\seealso{
Other value_testing_functions: \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsFalse}};
  \code{\link{xIsInfixOf}}, \code{\link{xIsInfixOf_}};
  \code{\link{xIsNan}}; \code{\link{xIsNa}};
  \code{\link{xIsNull}}; \code{\link{xIsPrefixOf}},
  \code{\link{xIsPrefixOf_}}; \code{\link{xIsSuffixOf}},
  \code{\link{xIsSuffixOf_}}; \code{\link{xIsTrue}};
  \code{\link{xIs}}; \code{\link{xNotEmpty}},
  \code{\link{xNotEmpty_}}; \code{\link{xNotInfixOf}},
  \code{\link{xNotInfixOf_}}; \code{\link{xNotNan}};
  \code{\link{xNotNa}}; \code{\link{xNotNull}};
  \code{\link{xNotPrefixOf}}, \code{\link{xNotPrefixOf_}};
  \code{\link{xNotSuffixOf}}, \code{\link{xNotSuffixOf_}};
  \code{\link{xNotTrue}}; \code{\link{xNot}}
}

