% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xNotNa}
\alias{xNotNa}
\title{xNotNa}
\usage{
xNotNa(val)
}
\arguments{
\item{val}{an arbitrary value.The value to test for
   not-na.}

\item{...}{see above.}
}
\value{
A true or false value.
}
\description{
Is an value not na?
}
\section{Type Signature}{

    any -> &lt;logical>
}

\section{Corner Cases}{

    xNotNa returns either true or false, to make it
    safe for use with if statements. If val is length zero
    True is returned.
}
\examples{

NULL
}
\seealso{
Other value_testing_functions: \code{\link{xElemIs}},
  \code{\link{xElemIs_}}; \code{\link{xElemNot}},
  \code{\link{xElemNot_}}; \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsFalse}};
  \code{\link{xIsNan}}; \code{\link{xIsNa}};
  \code{\link{xIsNull}}; \code{\link{xIsTrue}};
  \code{\link{xIs}}; \code{\link{xNotEmpty}},
  \code{\link{xNotEmpty_}}; \code{\link{xNotFalse}};
  \code{\link{xNotNan}}; \code{\link{xNotNull}};
  \code{\link{xNotTrue}}; \code{\link{xNot}}
}

