\name{xNotNa}
\alias{xNotNa}
\alias{xNotNa...}
\title{xNotNa}
\usage{
xNotNa(coll)

xNotNa...(...)
}
\arguments{
  \item{coll}{a collection.}

  \item{...}{see above.}
}
\value{
A vector of boolean values.
}
\description{
Test every element in a collection for being \code{Na} of
any type.
}
\section{Corner Cases}{
  Returns logical(0) if \bold{coll} is length-zero.
}
\examples{

NULL

}

