% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xNotNan}
\alias{xNotNan}
\title{xNotNan}
\usage{
xNotNan(val)
}
\arguments{
\item{val}{an arbitrary value.The value to test for
   not-nan.}

\item{...}{see above.}
}
\value{
A true or false value.
}
\description{
Is an value not nan?
}
\section{Corner Cases}{

    xNotNan returns either true or false, to make it
    safe for use with if statements. If val is length zero
    True is returned.
}
\examples{

NULL
}
\seealso{
Other value_testing_functions: \code{\link{xElemIsFalse}},
  \code{\link{xElemIsFalse_}}; \code{\link{xElemIsNan}},
  \code{\link{xElemIsNan_}}; \code{\link{xElemIsNa}},
  \code{\link{xElemIsNa_}}; \code{\link{xElemIsNull}},
  \code{\link{xElemIsNull_}}; \code{\link{xElemIsTrue}},
  \code{\link{xElemIsTrue_}}; \code{\link{xElemNotFalse}},
  \code{\link{xElemNotFalse_}}; \code{\link{xElemNotNan}},
  \code{\link{xElemNotNan_}}; \code{\link{xElemNotNa}},
  \code{\link{xElemNotNa_}}; \code{\link{xElemNotNull}},
  \code{\link{xElemNotNull_}}; \code{\link{xElemNotTrue}},
  \code{\link{xElemNotTrue_}}; \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsFalse}};
  \code{\link{xIsNan}}; \code{\link{xIsNa}};
  \code{\link{xIsNull}}; \code{\link{xIsTrue}};
  \code{\link{xNotEmpty}}, \code{\link{xNotEmpty_}};
  \code{\link{xNotFalse}}; \code{\link{xNotNa}};
  \code{\link{xNotNull}}; \code{\link{xNotTrue}}
}

