\name{xNotNan}
\alias{xNotNan}
\alias{xNotNan...}
\title{xNotNan}
\usage{
xNotNan(coll)

xNotNan...(...)
}
\arguments{
  \item{coll}{a collection}

  \item{...}{see above.}
}
\value{
A list.
}
\description{
Is an element of a collection not NaN?
}
\section{Corner Cases}{
  Returns the empty list if \code{coll is length-zero}.
}

\section{Variadic Functions

\link{ help_variadic }}{

}

