% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xNotNan}
\alias{xNotNan}
\title{xNotNan}
\usage{
xNotNan(val)
}
\arguments{
\item{val}{an arbitrary value.The value to test for
   not-nan.}

\item{...}{see above.}
}
\value{
A true or false value.
}
\description{
Is an value not nan?
}
\section{Type Signature}{

    any -> <logical>
}

\section{Corner Cases}{

    xNotNan returns either true or false, to make it
    safe for use with if statements. If val is length zero
    True is returned.
}
\examples{

# 1. The most obvious example.

xNotNan(NaN)

# False
}
\seealso{
Other value_testing_functions: \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsFalse}};
  \code{\link{xIsInfixOf}}, \code{\link{xIsInfixOf_}};
  \code{\link{xIsNan}}; \code{\link{xIsNa}};
  \code{\link{xIsNull}}; \code{\link{xIsPrefixOf}},
  \code{\link{xIsPrefixOf_}}; \code{\link{xIsSuffixOf}},
  \code{\link{xIsSuffixOf_}}; \code{\link{xIsTrue}};
  \code{\link{xIs}}; \code{\link{xNotEmpty}},
  \code{\link{xNotEmpty_}}; \code{\link{xNotFalse}};
  \code{\link{xNotInfixOf}}, \code{\link{xNotInfixOf_}};
  \code{\link{xNotNa}}; \code{\link{xNotNull}};
  \code{\link{xNotPrefixOf}}, \code{\link{xNotPrefixOf_}};
  \code{\link{xNotSuffixOf}}, \code{\link{xNotSuffixOf_}};
  \code{\link{xNotTrue}}; \code{\link{xNot}}
}

