\name{xNotNull}
\alias{xNotNull}
\alias{xNotNull...}
\title{xNotNull}
\usage{
xNotNull(coll)

xNotNull...(...)
}
\arguments{
  \item{coll}{a collection}

  \item{...}{see above.}
}
\value{
A vector of boolean values.
}
\description{
Is an element of a collection null?
}
\section{Corner Cases}{
  Returns True if coll is Null.
}

\section{Variadic Functions

\link{ help_variadic }}{

}

