% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xNotNull}
\alias{xNotNull}
\title{xNotNull}
\usage{
xNotNull(val)
}
\arguments{
\item{val}{an arbitrary value. The value to test for
   not being null.}

\item{...}{see above.}
}
\value{
A true or false value.
}
\description{
Is an value not null?
}
\section{Type Signature}{

    any -> &lt;logical>
}

\section{Corner Cases}{

    xNotNull returns either true or false, to make it
    safe for use with if statements.
}
\examples{

NULL
}
\seealso{
Other value_testing_functions: \code{\link{xElemIsFalse}},
  \code{\link{xElemIsFalse_}}; \code{\link{xElemIsNan}},
  \code{\link{xElemIsNan_}}; \code{\link{xElemIsNa}},
  \code{\link{xElemIsNa_}}; \code{\link{xElemIsNull}},
  \code{\link{xElemIsNull_}}; \code{\link{xElemIsTrue}},
  \code{\link{xElemIsTrue_}}; \code{\link{xElemNotFalse}},
  \code{\link{xElemNotFalse_}}; \code{\link{xElemNotNan}},
  \code{\link{xElemNotNan_}}; \code{\link{xElemNotNa}},
  \code{\link{xElemNotNa_}}; \code{\link{xElemNotNull}},
  \code{\link{xElemNotNull_}}; \code{\link{xElemNotTrue}},
  \code{\link{xElemNotTrue_}}; \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsFalse}};
  \code{\link{xIsNan}}; \code{\link{xIsNa}};
  \code{\link{xIsNot}}; \code{\link{xIsNull}};
  \code{\link{xIsTrue}}; \code{\link{xIs}};
  \code{\link{xNotEmpty}}, \code{\link{xNotEmpty_}};
  \code{\link{xNotFalse}}; \code{\link{xNotNan}};
  \code{\link{xNotNa}}; \code{\link{xNotTrue}}
}

