% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xNotTrue}
\alias{xNotTrue}
\title{xNotTrue}
\usage{
xNotTrue(val)
}
\arguments{
\item{val}{an arbitrary value. The value to test for
   non-truth.}

\item{...}{see above.}
}
\value{
A true or false value.
}
\description{
Is an value not true?
}
\section{Type Signature}{

    any -> &lt;logical>
}

\section{Corner Cases}{

    xNotTrue returns either true or false, to make it
    safe for use with if statements. If val is length zero
    True is returned.
}
\examples{

# 1. An obvious example

xNotTrue(True)

# False
}
\seealso{
Other value_testing_functions: \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsFalse}};
  \code{\link{xIsNan}}; \code{\link{xIsNa}};
  \code{\link{xIsNull}}; \code{\link{xIsTrue}};
  \code{\link{xIs}}; \code{\link{xNotEmpty}},
  \code{\link{xNotEmpty_}}; \code{\link{xNotFalse}};
  \code{\link{xNotNan}}; \code{\link{xNotNa}};
  \code{\link{xNotNull}}; \code{\link{xNot}}
}

