% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xOrderOf.R
\name{xOrderOf}
\alias{xOrderOf}
\alias{xOrderOf_}
\title{xOrderOf}
\usage{
xOrderOf(nums)

xOrderOf_(...)
}
\arguments{
\item{nums}{a vector of numbers. The numbers to rank
   in order of size.}

\item{...}{see above.}
}
\value{
A vector of numbers.
}
\description{
Return a permutation of indices that reorders a collection of numbers.
}
\details{
\bold{xOrderOf} returns the indices for a collection that are required to
   re-order it. For example,

   \code{c(3, 1, 2)[ xOrderOf(c(3,1,2)) ]}

   re-arranges the collection as

   \code{c(1, 2, 3)}

   This is a trivial use of \bold{xOrderOf}; a more common use is to
   re-order a collection of collections by the indices of one row or column.
}
\section{Type Signature}{

    |numeric| -> <integer>
}

\section{Corner Cases}{

   If nums is empty then \bold{integer(0)} is returned. Nums cannot contain
   na or nan values, as then are unorderable.
}
\examples{

NULL
}
\seealso{
Other math_functions: \code{\link{xLenOf}},
  \code{\link{xLenOf_}}; \code{\link{xMaxBy}},
  \code{\link{xMaxBy_}}; \code{\link{xMeanBy}},
  \code{\link{xMeanBy_}}; \code{\link{xMedianBy}},
  \code{\link{xMedianBy_}}; \code{\link{xMinBy}},
  \code{\link{xMinBy_}}; \code{\link{xPoll}},
  \code{\link{xPoll_}}; \code{\link{xProdBy}},
  \code{\link{xProdBy_}}; \code{\link{xRank}},
  \code{\link{xRank_}}; \code{\link{xSumBy}},
  \code{\link{xSumBy_}}
}

