% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xParamsOf}
\alias{xParamsOf}
\title{xParamsOf}
\usage{
xParamsOf(fn)
}
\arguments{
\item{fn}{an arbitrary function or primitive function. The function to
  find the parametre names of.}
}
\value{
A character vector.
}
\description{
Get the parametre names of a function or primitive function.
}
\section{Type Signature}{

    function -> &lt;character>
}

\section{Corner Cases}{

	 If \bold{fn} is a primitive function a heuristic is used to obtain
	 its parametre names, which may not work for all functions. If a function
  has no parametres the empty character vector is returned.
}
\examples{

NULL

}
\seealso{
Other parametre_functions: \code{\link{xArityOf}};
  \code{\link{xFix}}, \code{\link{xFix_}};
  \code{\link{xFormalsOf}}; \code{\link{xIsVariadic}};
  \code{\link{xSpread}}; \code{\link{xUnspread}}
}

