% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xPartial}
\alias{xPartial}
\alias{xPartial_}
\title{xPartial}
\usage{
xPartial(fn, coll)

xPartial_(fn, ...)
}
\arguments{
\item{fn}{an arbitrary function. The function to have some
   of its arguments fixed.}

\item{coll}{a collection. The arguments to fix.}

\item{...}{see above.}
}
\value{
A function of equal or lesser arity to \bold{fn}.
}
\description{
Fix several arguments of a function, producing a function of smaller arity.
}
\details{
Partial application is a mechanism for specialising a general function
    to a particular purpose. For example, the very general function \bold{xMap}
    can be easily specialised to particular tasks using \code{xPartial}:

    \code{lengths <- xPartial(xMap, list(fn = length))}

    \code{sizes <- xPartial(xMap, object.size))}

    In this case the use of xPartial can match the argument to specialise by
    name, or by position. The use of \bold{xPartial} over \bold{xPartial_} is
    required in the first case, to avoid a parametre name class between the argument of
    \bold{xMap} to fix (fn) and the function passed to \bold{xPartial} (fn).
}
\section{Corner Cases}{

   Partial application also works for ellipses (eg list(... = 1)).
}
\examples{

NULL

}
\seealso{
Other function_modifying_functions: \code{\link{\%and\%}},
  \code{\link{\%or\%}}, \code{\link{xLift}},
  \code{\link{xLift_}}; \code{\link{\%of\%}},
  \code{\link{\%then\%}}, \code{\link{xCompose}},
  \code{\link{xCompose_}}; \code{\link{xAsClosure}};
  \code{\link{xCapture}}, \code{\link{xK}};
  \code{\link{xDelay}}; \code{\link{xJuxtapose}},
  \code{\link{xJuxtapose_}}; \code{\link{xNot}};
  \code{\link{xSpread}}; \code{\link{xThread}},
  \code{\link{xThread_}}; \code{\link{xUnspread}};
  \code{\link{xVectorise}}, \code{\link{xVectorize}}

Other parametre_functions: \code{\link{xArityOf}};
  \code{\link{xFormalsOf}}; \code{\link{xIsVariadic}};
  \code{\link{xParamsOf}}; \code{\link{xSpread}};
  \code{\link{xUnspread}}
}

