% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xPartition}
\alias{xPartition}
\alias{xPartition_}
\title{xPartition}
\usage{
xPartition(pred, coll)

xPartition_(pred, ...)
}
\arguments{
\item{pred}{a predicate. The predicate function with which to partition
   the input collection.}

\item{coll}{a collection. The collection to partition into
   two collections.}

\item{...}{see above.}
}
\value{
Returns two lists; a list of elements in a collection for which a
   predicate returns true, and a list of elements in a collection for which
   a predicate returns false or na.
}
\description{
Divide elements in a collection into two collections
based on a predicate function.
}
\section{Type Signature}{

    (any -> &lt;logical>) -> |any| -> [[any]]
}

\section{Corner Cases}{

	  If \bold{coll} is empty the empty list is returned. If all the
	  elements return only true/only false, then one of two sublists will be the
	  empty list.
}
\examples{

# 1. Partitioning

mixed <- xShuffle(c(letters, LETTERS))

# -- is the letter a lower-case letter?
xPartition(
	xFix_(xIsMember, coll = letters),
	mixed)

# shuffling will vary the order.
#
# list(
#     list(
#         "n", "w", "a", "z", "i", "y", "b", "t", "c", "u", "j",
#         "o", "f", "d", "v", "k", "g", "p", "r", "q", "x", "m", "s",
#         "e", "l", "h"),
#     list(
#         "C", "D", "V", "N", "Q", "X", "S", "M",
#         "G", "H", "O", "L", "E", "F", "W", "R", "Y", "T", "I", "K",
#         "A", "B", "J", "P", "U", "Z"))
#
}
\seealso{
Other filtering_functions: \code{\link{xReject}},
  \code{\link{xReject_}}; \code{\link{xSelect}},
  \code{\link{xSelect_}}
}

