% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xPartition}
\alias{xPartition}
\alias{xPartition...}
\title{xPartition}
\usage{
xPartition(pred, coll)

xPartition...(pred, ...)
}
\arguments{
\item{pred}{a predicate. The predicate function with which to partition
   the input collection.}

\item{coll}{a collection. The collection to partition into
   two collections.}

\item{...}{see above.}
}
\value{
Returns two lists; a list of elements in a collection for which a
   predicate returns true, and a list of elements in a collection for which
   a predicate returns false or na.
}
\description{
Divide elements in a collection into two collections
based on a predicate function.
}
\section{Corner Cases}{

	  If \bold{coll} is empty the empty list is returned. If all the
	  elements return only true/only false, then one of two sublists will be the
	  empty list.
}
\examples{

NULL

}
\seealso{
Other filtering_functions: \code{\link{xPack}},
  \code{\link{xPack...}}, \code{\link{xRejectEmpty}},
  \code{\link{xRejectEmpty...}}; \code{\link{xRejectNan}},
  \code{\link{xRejectNan...}}; \code{\link{xRejectNa}},
  \code{\link{xRejectNa...}}; \code{\link{xRejectNull}},
  \code{\link{xRejectNull...}}; \code{\link{xReject}},
  \code{\link{xReject...}}; \code{\link{xSelect}},
  \code{\link{xSelect...}}
}

