% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xPermute}
\alias{xPermute}
\alias{xPermute...}
\title{xPermute}
\usage{
xPermute(nums, colls)

xPermute...(nums, ...)
}
\arguments{
\item{nums}{a collection of whole numbers. The indices
   by which to permute the input collections.}

\item{colls}{several collections of equal length to \bold{nums}. The
   collections to permute.}

\item{...}{see above.}
}
\value{
A list of collections.
}
\description{
Permute a collection of collections using indices.
}
\details{
\bold{xPermute} allows several collections to be rearranged
   in the same way simultaneously.

   \code{colls <- list(ids = list(15, 1, 12), ages = list(17, 29, 24), weights = list(80, 82, 76))}

   \code{xPermute(xOrderOf( xFirstOf(colls) ), colls)}
}
\section{Corner Cases}{

   returns the empty list is \bold{nums} is length-zero.
}
\examples{

NULL

}
\seealso{
Other combinatoric_functions: \code{\link{xChoose}},
  \code{\link{xChoose...}}; \code{\link{xCycle}},
  \code{\link{xCycle...}}; \code{\link{xPowerSetOf}},
  \code{\link{xPowerSetOf...}}; \code{\link{xProdSetOf}},
  \code{\link{xProdSetOf...}}

Other set_functions: \code{\link{xDuplicatesOf}},
  \code{\link{xDuplicatesOf...}}; \code{\link{xInter}},
  \code{\link{xInter...}}; \code{\link{xIsMember}},
  \code{\link{xIsMember...}}; \code{\link{xIsSubset}},
  \code{\link{xIsSubset...}}; \code{\link{xNotMember}},
  \code{\link{xNotMember...}}; \code{\link{xPowerSetOf}},
  \code{\link{xPowerSetOf...}}; \code{\link{xProdSetOf}},
  \code{\link{xProdSetOf...}}; \code{\link{xUnionOf}},
  \code{\link{xUnionOf...}}; \code{\link{xUniqueOf}},
  \code{\link{xUniqueOf...}}
}

