% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xPluck}
\alias{xPluck}
\alias{xPluck...}
\title{xPluck}
\usage{
xPluck(str, colls)

xPluck...(str, ...)
}
\arguments{
\item{str}{a string. The key to select from each collection.}

\item{colls}{a collections of collections. The collection
   of collections to select keys from.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Map over a collection of lists or pairlists,
selecting fields in each element by name.
}
\section{Corner Cases}{

   Returns the empty list if \code{colls} is length-zero.
   If str is length-zero then the empty string "" is
   used to match key-names.
}
\examples{

# 1. selecting a key from each row in a data frame.

median_lethal_dose <- list(
	'toxicant' =
		c('ricin', 'dioxin', 'sarin', 'vx', 'batrachotoxin', 'maitotoxin', 'polonium-210', 'botox'),
	'mg_per_kg' =
		c(22, 20, 17, 2, 2, 0.13, 0.01, 0.001)
)

x_(median_lethal_dose) $ xZip() $ x_Pluck('mg_per_kg')

# list(22, 20, 17, 2, 2, 0.13, 0.01, 0.001)
}
\seealso{
Other key_functions: \code{\link{xAddKeys}};
  \code{\link{xUnzipKeys}}, \code{\link{xUnzipKeys...}};
  \code{\link{xZipKeys}}, \code{\link{xZipKeys...}}

Other selection_functions: \code{\link{xAtCol}},
  \code{\link{xAtCol...}}; \code{\link{xAtKey}},
  \code{\link{xAtKey...}}; \code{\link{xAt}},
  \code{\link{xAt...}}; \code{\link{xDropWhile}},
  \code{\link{xDropWhile...}}; \code{\link{xDrop}},
  \code{\link{xDrop...}}; \code{\link{xFirstOf}},
  \code{\link{xFirstOf...}}; \code{\link{xFourthOf}},
  \code{\link{xFourthOf...}}; \code{\link{xInitOf}},
  \code{\link{xInitOf...}}; \code{\link{xLastOf}},
  \code{\link{xLastOf...}}; \code{\link{xMaxBy}},
  \code{\link{xMaxBy...}}; \code{\link{xMinBy}},
  \code{\link{xMinBy...}}; \code{\link{xRestOf}},
  \code{\link{xRestOf...}}; \code{\link{xSecondOf}},
  \code{\link{xSecondOf...}}; \code{\link{xSlice}},
  \code{\link{xSlice...}}; \code{\link{xTakeWhile}},
  \code{\link{xTakeWhile...}}; \code{\link{xTake}},
  \code{\link{xTake...}}; \code{\link{xThirdOf}},
  \code{\link{xThirdOf...}}
}

