% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xPluck}
\alias{xPluck}
\alias{xPluck_}
\title{xPluck}
\usage{
xPluck(str, colls)

xPluck_(str, ...)
}
\arguments{
\item{str}{a string. The key to select from each collection.}

\item{colls}{a collections of collections. The collection
   of collections to select keys from.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Map over a collection of lists or pairlists,
selecting fields in each element by name.
}
\section{Type Signature}{

    <character> -> ||any|| -> [any]
}

\section{Corner Cases}{

   Returns the empty list if \code{colls} is length-zero.
   If str is length-zero then the empty string "" is
   used to match key-names.
}
\examples{

# 1. selecting a key from each row in a data frame.

median_lethal_dose <- list(
	'toxicant' =
		c('ricin', 'dioxin', 'sarin', 'vx', 'batrachotoxin', 'maitotoxin', 'polonium-210', 'botox'),
	'mg_per_kg' =
		c(22, 20, 17, 2, 2, 0.13, 0.01, 0.001)
)

x_(median_lethal_dose) $ xZip() $ x_Pluck('mg_per_kg')

# list(22, 20, 17, 2, 2, 0.13, 0.01, 0.001)
}
\seealso{
Other key_functions: \code{\link{xAddKeys}},
  \code{\link{xAddKeys_}}; \code{\link{xUnzipIndices}},
  \code{\link{xUnzipIndices_}}; \code{\link{xUnzipKeys}},
  \code{\link{xUnzipKeys_}}; \code{\link{xZipKeys}},
  \code{\link{xZipKeys_}}

Other selection_functions: \code{\link{xAtCol}},
  \code{\link{xAtCol_}}; \code{\link{xAtKey}},
  \code{\link{xAtKey_}}; \code{\link{xAt}},
  \code{\link{xAt_}}; \code{\link{xDropWhile}},
  \code{\link{xDropWhile_}}; \code{\link{xDrop}},
  \code{\link{xDrop_}}; \code{\link{xFirstOf}},
  \code{\link{xFirstOf_}}; \code{\link{xFourthOf}},
  \code{\link{xFourthOf_}}; \code{\link{xInitOf}},
  \code{\link{xInitOf_}}; \code{\link{xLastOf}},
  \code{\link{xLastOf_}}; \code{\link{xMaxBy}},
  \code{\link{xMaxBy_}}; \code{\link{xMinBy}},
  \code{\link{xMinBy_}}; \code{\link{xRestOf}},
  \code{\link{xRestOf_}}; \code{\link{xSecondOf}},
  \code{\link{xSecondOf_}}; \code{\link{xSlice}},
  \code{\link{xSlice_}}; \code{\link{xTakeWhile}},
  \code{\link{xTakeWhile_}}; \code{\link{xTake}},
  \code{\link{xTake_}}; \code{\link{xThirdOf}},
  \code{\link{xThirdOf_}}
}

