\name{xPred}
\alias{xPred}
\title{xPred}
\usage{
  xPred(nums)
}
\arguments{
  \item{nums}{a vector of non-complex numbers.}
}
\value{
  a vector of non-complex numbers.
}
\description{
  Increment a vector of numbers.
}
\section{Corner Cases}{
  returns the unit of the vector if \code{nums} is
  length-zero}.
}

\section{Glossary}{
  function: a closure or primitive function.  A symbol or
  string refering to such a function may also be used.

  predicate: A function that returns a logical value
  (unary, unless stated otherwise).  A symbol or string
  refering to such a function may also be used.

  collection: A list, pairlist or vector of any type.

  arity: The number of arguments a function can accept.
  Variadic functions have an arity of +Infinity,

  number: a double, integer or complex number, unless
  stated otherwise.
}
\examples{
inst/examples/blank.R
}

