\name{xPred}
\alias{xPred}
\alias{xPred...}
\title{xPred}
\usage{
xPred(nums)

xPred...(...)
}
\arguments{
  \item{nums}{a vector of non-complex numbers.}

  \item{...}{see above.}
}
\value{
a vector of non-complex numbers.
}
\description{
Increment a vector of numbers.
}
\section{Corner Cases}{
  returns the unit of the vector if \code{nums} is
  length-zero.
}

\section{Variadic Functions

\link{ help_variadic }}{

}

