\name{xPred}
\alias{xPred}
\title{xPred}
\usage{
  xPred(nums)
}
\arguments{
  \item{nums}{a vector of non-complex numbers.}
}
\value{
  a vector of non-complex numbers.
}
\description{
  Increment a vector of numbers.
}
\section{Corner Cases}{
  returns the unit of the vector if \code{nums} is
  length-zero.
}

