% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xProdBy}
\alias{xProdBy}
\alias{xProdBy_}
\title{xProdBy}
\usage{
xProdBy(fn, coll)

xProdBy_(fn, ...)
}
\arguments{
\item{fn}{a unary function. The function to measure the size of
    an element.}

\item{coll}{a non-empty collection. The collection to find the
   product of.}

\item{...}{see above.}
}
\value{
A double.
}
\description{
Get the product of a collection according to a measure function.
}
\section{Type Signature}{

    (any -> &lt;number>) -> |any| -> &lt;number>
}

\section{Corner Cases}{

   If an empty collection is given, numeric(0) is returned.
}
\examples{

# 1. compounding the numbers themselves.

xSumBy(xIdentity, 1:10)

# 3628800
}
\seealso{
Other selection_functions: \code{\link{xAtKey}},
  \code{\link{xAtKey_}}; \code{\link{xAt}},
  \code{\link{xAt_}}; \code{\link{xDropWhile}},
  \code{\link{xDropWhile_}}; \code{\link{xDrop}},
  \code{\link{xDrop_}}; \code{\link{xFirstOf}},
  \code{\link{xFirstOf_}}; \code{\link{xFourthOf}},
  \code{\link{xFourthOf_}}; \code{\link{xInitOf}},
  \code{\link{xInitOf_}}; \code{\link{xLastOf}},
  \code{\link{xLastOf_}}; \code{\link{xMaxBy}},
  \code{\link{xMaxBy_}}; \code{\link{xMinBy}},
  \code{\link{xMinBy_}}; \code{\link{xRestOf}},
  \code{\link{xRestOf_}}; \code{\link{xSecondOf}},
  \code{\link{xSecondOf_}}; \code{\link{xSlice}},
  \code{\link{xSlice_}}; \code{\link{xSumBy}},
  \code{\link{xSumBy_}}; \code{\link{xTakeWhile}},
  \code{\link{xTakeWhile_}}; \code{\link{xTake}},
  \code{\link{xTake_}}; \code{\link{xThirdOf}},
  \code{\link{xThirdOf_}}
}

