% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xProdBy}
\alias{xProdBy}
\alias{xProdBy_}
\title{xProdBy}
\usage{
xProdBy(fn, coll)

xProdBy_(fn, ...)
}
\arguments{
\item{fn}{a unary function. The function to measure the size of
    an element.}

\item{coll}{a non-empty collection. The collection to find the
   product of.}

\item{...}{see above.}
}
\value{
A double.
}
\description{
Get the product of a collection according to a measure function.
}
\section{Type Signature}{

    (any -> &lt;number>) -> |any| -> &lt;number>
}

\section{Corner Cases}{

   If an empty collection is given, numeric(0) is returned.
}
\examples{

# 1. compounding the numbers themselves.

xProdBy(xIdentity, 1:10)

# 3628800
}
\seealso{
Other math_functions: \code{\link{xLenOf}},
  \code{\link{xLenOf_}}; \code{\link{xMaxBy}},
  \code{\link{xMaxBy_}}; \code{\link{xMeanOf}},
  \code{\link{xMeanOf_}}; \code{\link{xMinBy}},
  \code{\link{xMinBy_}}; \code{\link{xOrderOf}},
  \code{\link{xOrderOf_}}; \code{\link{xPoll}},
  \code{\link{xPoll_}}; \code{\link{xRank}},
  \code{\link{xRank_}}; \code{\link{xSumBy}},
  \code{\link{xSumBy_}}
}

