\name{xRank}
\alias{xRank}
\alias{xRank...}
\title{xRank}
\usage{
xRank(nums)

xRank...(...)
}
\arguments{
  \item{nums}{a vector of numbers. The numbers to rank in
  order of size.}

  \item{...}{see above.}
}
\value{
A vector of numbers.
}
\description{
Rank numbers in order of size.
}
\section{Corner Cases}{
  If nums is empty then \bold{integer(0)} is returned.
}
\examples{

NULL
}

