% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xRank}
\alias{xRank}
\alias{xRank_}
\title{xRank}
\usage{
xRank(nums)

xRank_(...)
}
\arguments{
\item{nums}{a vector of numbers. The numbers to rank
   in order of size.}

\item{...}{see above.}
}
\value{
A vector of numbers.
}
\description{
Rank a collection of numbers from largest to smallest.
}
\section{Corner Cases}{

   If nums is empty then \bold{integer(0)} is returned. Tied values are
   ranked in order of appearance; the first occurrence is ranked lower than
   the second occurrence. Na or NaN values cannot be included, as they are
   unorderable.
}
\examples{

NULL
}
\seealso{
Other math_functions: \code{\link{xLenOf}},
  \code{\link{xLenOf_}}; \code{\link{xMaxBy}},
  \code{\link{xMaxBy_}}; \code{\link{xMeanBy}},
  \code{\link{xMeanBy_}}; \code{\link{xMinBy}},
  \code{\link{xMinBy_}}; \code{\link{xOrderOf}},
  \code{\link{xOrderOf_}}; \code{\link{xPoll}},
  \code{\link{xPoll_}}; \code{\link{xProdBy}},
  \code{\link{xProdBy_}}; \code{\link{xSumBy}},
  \code{\link{xSumBy_}}
}

