% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xReadChars}
\alias{xReadChars}
\title{xReadChars}
\usage{
xReadChars(str)
}
\arguments{
\item{str}{a length-one character vector. The path to
   import files from.}
}
\value{
A character vector, with one or more elements.
}
\description{
Import each line in a file to an element in a character vector.
}
\section{Corner Cases}{

   xReadChars ignores missing terminal newlines.
}
\examples{

NULL
}
\seealso{
Other inpure_functions: \code{\link{xDelay}};
  \code{\link{xDo}}, \code{\link{xDo_}};
  \code{\link{xExecute}}; \code{\link{xReadLines}};
  \code{\link{xReadWords}}; \code{\link{xStopwatch}};
  \code{\link{xVersion}}; \code{\link{xWriteChars}};
  \code{\link{xWriteLines}}; \code{\link{xWriteWords}}

Other text_processing_functions: \code{\link{xExplode}},
  \code{\link{xExplode_}}; \code{\link{xFromChars}},
  \code{\link{xFromChars_}}; \code{\link{xFromLines}},
  \code{\link{xFromLines_}}; \code{\link{xFromWords}},
  \code{\link{xFromWords_}}; \code{\link{xImplode}},
  \code{\link{xImplode_}}; \code{\link{xReadLines}};
  \code{\link{xReadWords}}; \code{\link{xSliceString}},
  \code{\link{xSliceString_}}; \code{\link{xToChars}};
  \code{\link{xToLines}}; \code{\link{xToWords}};
  \code{\link{xWriteChars}}; \code{\link{xWriteLines}};
  \code{\link{xWriteWords}}
}

