\name{xReadWords}
\alias{xReadWords}
\title{xReadWords}
\usage{
xReadWords(str)
}
\arguments{
  \item{str}{a length-one character vector. The path to
  import files from.}
}
\value{
A character vector, with one or more elements.
}
\description{
Import each word in a file to a string.
}
\details{
xReadWords ignores missing terminal newlines.
}
\examples{

NULL
}
\seealso{
Other character_functions: \code{\link{xExplode}},
\code{\link{xExplode...}}; \code{\link{xFromChars}},
\code{\link{xFromChars...}}; \code{\link{xFromLines}},
\code{\link{xFromLines...}}; \code{\link{xFromWords}},
\code{\link{xFromWords...}}; \code{\link{xImplode}},
\code{\link{xImplode...}}; \code{\link{xReadChars}},
\code{\link{xReadLines}}; \code{\link{xSubstring}},
\code{\link{xSubstring...}}; \code{\link{xToChars}};
\code{\link{xToLines}}; \code{\link{xToWords}};
\code{\link{xWriteChars}}, \code{\link{xWriteWords}};
\code{\link{xWriteLines}}
}

