% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xReadWords}
\alias{xReadWords}
\title{xReadWords}
\usage{
xReadWords(str)
}
\arguments{
\item{str}{a length-one character vector. The path to
   import files from.}
}
\value{
A character vector, with one or more elements.
}
\description{
Import each word in a file to an element in a character vector.
}
\section{Corner Cases}{

   xReadWords ignores missing terminal newlines.
}
\examples{

NULL
}
\seealso{
Other inpure_functions: \code{\link{xDelay}};
  \code{\link{xDo}}, \code{\link{xDo...}};
  \code{\link{xExecute}}; \code{\link{xReadChars}};
  \code{\link{xReadLines}}; \code{\link{xStopwatch}};
  \code{\link{xVersion}}; \code{\link{xWriteChars}};
  \code{\link{xWriteLines}}; \code{\link{xWriteWords}}

Other text_processing_functions: \code{\link{xExplode}},
  \code{\link{xExplode...}}; \code{\link{xFromChars}},
  \code{\link{xFromChars...}}; \code{\link{xFromLines}},
  \code{\link{xFromLines...}}; \code{\link{xFromWords}},
  \code{\link{xFromWords...}}; \code{\link{xImplode}},
  \code{\link{xImplode...}}; \code{\link{xReadChars}};
  \code{\link{xReadLines}}; \code{\link{xSubstring}},
  \code{\link{xSubstring...}}; \code{\link{xToChars}};
  \code{\link{xToLines}}; \code{\link{xToWords}};
  \code{\link{xWriteChars}}; \code{\link{xWriteLines}};
  \code{\link{xWriteWords}}
}

