\name{xRecurMap}
\alias{xRecurMap}
\title{xRecurMap}
\usage{
xRecurMap(fn, coll)
}
\arguments{
  \item{fn}{a unary function.}

  \item{coll}{a list or pairlist.}
}
\value{
a list or pairlist.
}
\description{
Recursively map a function into a nested collection,
preserving its structure.
}
\seealso{
Other higher_order_functions map_like_functions
collection_functions: \code{\link{xFlatMap}};
\code{\link{xMapIndexed}}; \code{\link{xMapMany}};
\code{\link{xMapWhen}}; \code{\link{xMapply}};
\code{\link{xMap}}
}

