\name{xRecycle}
\alias{xRecycle}
\alias{xRecycle...}
\title{xRecycle}
\usage{
xRecycle(colls)

xRecycle...(...)
}
\arguments{
  \item{colls}{a collection of collections.}

  \item{...}{see above.}
}
\value{
A list.
}
\description{
xRecycle
}
\section{Corner Cases}{
  Returns the empty list if \bold{colls} is length-zero, or
  if a collection in \bold{colls} is length-zero.
}

