% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xRecycle.R
\name{xRecycle}
\alias{xRecycle}
\alias{xRecycle_}
\title{xRecycle}
\usage{
xRecycle(num, coll)

xRecycle_(num, ...)
}
\arguments{
\item{num}{a nonnegative whole number. The length of the output collection.}

\item{coll}{a collection. The collection to recycle elements of.}

\item{...}{see above.}
}
\value{
a list.
}
\description{
Reuse elements of a collection in order until the collection is a certain length.
}
\section{Corner Cases}{

   If \bold{coll} is empty the empty list is returned,
   or if \bold{num} is length-zero. If \bold{num} is shorter than the length of
   \bold{coll}, \bold{coll} is truncated and returned.
}

\section{Rcpp}{


    This function is written using C++, and should be quite performant.
}
\examples{

NULL
}
\seealso{
Other reshaping_functions: \code{\link{xAmassBy}},
  \code{\link{xAmassBy_}}; \code{\link{xAppend}},
  \code{\link{xAppend_}}; \code{\link{xChunk}},
  \code{\link{xChunk_}}; \code{\link{xFlatten}},
  \code{\link{xFlatten_}}; \code{\link{xGather}},
  \code{\link{xGather_}}; \code{\link{xGroupBy}},
  \code{\link{xGroupBy_}}; \code{\link{xJoin}},
  \code{\link{xJoin_}}; \code{\link{xOneOf}},
  \code{\link{xOneOf_}}; \code{\link{xPrepend}},
  \code{\link{xPrepend_}}; \code{\link{xRepeat}},
  \code{\link{xRepeat_}}; \code{\link{xReverse}},
  \code{\link{xReverse_}}; \code{\link{xShuffle}},
  \code{\link{xShuffle_}}; \code{\link{xSortBy}},
  \code{\link{xSortBy_}}; \code{\link{xTabulate}},
  \code{\link{xTabulate_}}; \code{\link{xUnzipIndices}},
  \code{\link{xUnzipIndices_}}; \code{\link{xUnzipKeys}},
  \code{\link{xUnzipKeys_}}; \code{\link{xZipKeys}},
  \code{\link{xZipKeys_}}; \code{\link{xZip}},
  \code{\link{xZip_}}
}

