% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xRecycle}
\alias{xRecycle}
\alias{xRecycle_}
\title{xRecycle}
\usage{
xRecycle(colls)

xRecycle_(...)
}
\arguments{
\item{colls}{a collection of collections. The collection
    to convert to a non-ragged form.}

\item{...}{see above.}
}
\value{
A list of lists.
}
\description{
Make every subcollection in a collection the same
length by cyclically reusing elements.
}
\details{
xRecycle takes a ragged collection - a collection of collections
    of varying length - and recycles elements to make the shorter
    collections the same length as the longest collection.

    \code{xRecycle_( list('a', 'b'), 1:3 )}

    \code{list(list("a", "b", "a"), list(1L, 2L, 3L))}
}
\section{Corner Cases}{

   Returns the empty list if \bold{colls} is length-zero, or if a
   collection in \bold{colls} is length-zero.
}
\seealso{
Other reshaping_functions: \code{\link{xChop}},
  \code{\link{xChop_}}; \code{\link{xChunk}},
  \code{\link{xChunk_}}; \code{\link{xFlatten}},
  \code{\link{xFlatten_}}; \code{\link{xGroupBy}},
  \code{\link{xGroupBy_}}; \code{\link{xJoin}},
  \code{\link{xJoin_}}; \code{\link{xOneOf}},
  \code{\link{xOneOf_}}; \code{\link{xRepeat}},
  \code{\link{xRepeat_}}; \code{\link{xReverse}},
  \code{\link{xReverse_}}; \code{\link{xShuffle}},
  \code{\link{xShuffle_}}; \code{\link{xSortBy}},
  \code{\link{xSortBy_}}; \code{\link{xSplitAt}},
  \code{\link{xSplitAt_}}; \code{\link{xSplitWith}},
  \code{\link{xSplitWith_}}; \code{\link{xTabulate}},
  \code{\link{xTabulate_}}; \code{\link{xUnzipIndices}},
  \code{\link{xUnzipIndices_}}; \code{\link{xUnzipKeys}},
  \code{\link{xUnzipKeys_}}; \code{\link{xZipKeys}},
  \code{\link{xZipKeys_}}; \code{\link{xZip}},
  \code{\link{xZip_}}
}

