% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xRejectEmpty}
\alias{xPack}
\alias{xPack_}
\alias{xRejectEmpty}
\alias{xRejectEmpty_}
\title{Remove the empty values from a collection.}
\usage{
xRejectEmpty(coll)

xRejectEmpty_(...)

xPack(coll)

xPack_(...)
}
\arguments{
\item{coll}{a collection. The collection to remove
   length-zero values from.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Remove the empty values from a collection.
}
\examples{

# 1.
}
\seealso{
Other filtering_functions: \code{\link{xPartition}},
  \code{\link{xPartition_}}; \code{\link{xRejectNan}},
  \code{\link{xRejectNan_}}; \code{\link{xRejectNa}},
  \code{\link{xRejectNa_}}; \code{\link{xRejectNull}},
  \code{\link{xRejectNull_}}; \code{\link{xReject}},
  \code{\link{xReject_}}; \code{\link{xSelect}},
  \code{\link{xSelect_}}
}

