% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xRejectNa}
\alias{xRejectNa}
\alias{xRejectNa_}
\title{xRejectNa}
\usage{
xRejectNa(coll)

xRejectNa_(...)
}
\arguments{
\item{coll}{a collection. The collection to remove
   na values from.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Remove the Na values from a collection.
}
\examples{

NULL

}
\seealso{
Other filtering_functions: \code{\link{xPack}},
  \code{\link{xPack_}}, \code{\link{xRejectEmpty}},
  \code{\link{xRejectEmpty_}}; \code{\link{xPartition}},
  \code{\link{xPartition_}}; \code{\link{xRejectNan}},
  \code{\link{xRejectNan_}}; \code{\link{xRejectNull}},
  \code{\link{xRejectNull_}}; \code{\link{xReject}},
  \code{\link{xReject_}}; \code{\link{xSelect}},
  \code{\link{xSelect_}}
}

