% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xRejectNull}
\alias{xRejectNull}
\alias{xRejectNull_}
\title{xRejectNull}
\usage{
xRejectNull(coll)

xRejectNull_(...)
}
\arguments{
\item{coll}{a collection. The collection to remove
   null values from.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Remove the null values from a collection.
}
\examples{

NULL

}
\seealso{
Other filtering_functions: \code{\link{xPack}},
  \code{\link{xPack_}}, \code{\link{xRejectEmpty}},
  \code{\link{xRejectEmpty_}}; \code{\link{xPartition}},
  \code{\link{xPartition_}}; \code{\link{xRejectNan}},
  \code{\link{xRejectNan_}}; \code{\link{xRejectNa}},
  \code{\link{xRejectNa_}}; \code{\link{xReject}},
  \code{\link{xReject_}}; \code{\link{xSelect}},
  \code{\link{xSelect_}}
}

