\name{xRemoveEmpty}
\alias{xRemoveEmpty}
\alias{xRemoveEmpty...}
\title{xRemoveEmpty}
\usage{
xRemoveEmpty(coll)

xRemoveEmpty...(...)
}
\arguments{
  \item{coll}{a collection.}

  \item{...}{see above.}
}
\value{
A list.
}
\description{
Remove the null values from a collection.
}
\section{Variadic Functions

\link{ help_variadic }}{

}
\seealso{
Other filtering_functions: \code{\link{xPack}},
\code{\link{xPack...}}; \code{\link{xPartition}};
\code{\link{xReject}}, \code{\link{xReject...}};
\code{\link{xRemoveNan}}; \code{\link{xRemoveNa}};
\code{\link{xRemoveNull}}, \code{\link{xRemoveNull...}};
\code{\link{xSelect}}, \code{\link{xSelect...}}
}

