\name{xRemoveEmpty}
\alias{xRemoveEmpty}
\alias{xRemoveEmpty...}
\title{xRemoveEmpty}
\usage{
xRemoveEmpty(coll)

xRemoveEmpty...(...)
}
\arguments{
  \item{coll}{a collection. The collection to remove
  length-zero values from.}

  \item{...}{see above.}
}
\value{
A list.
}
\description{
Remove the empty values from a collection.
}
\examples{

NULL

}
\seealso{
Other filtering_functions: \code{\link{xPartition}};
\code{\link{xReject}}, \code{\link{xReject...}};
\code{\link{xRemoveNan}}; \code{\link{xRemoveNa}};
\code{\link{xRemoveNull}}, \code{\link{xRemoveNull...}};
\code{\link{xSelect}}, \code{\link{xSelect...}}
}

