\name{xRemoveNa}
\alias{xRemoveNa}
\title{xRemoveNa}
\usage{
xRemoveNa(coll)
}
\arguments{
  \item{coll}{a collection.}

  \item{...}{see above.}
}
\value{
A list.
}
\description{
Remove the NA values from a collection.
}
\section{Variadic Functions

\link{ help_variadic }}{

}
\seealso{
Other filtering_functions: \code{\link{xPack}},
\code{\link{xPack...}}; \code{\link{xPartition}};
\code{\link{xReject}}, \code{\link{xReject...}};
\code{\link{xRemoveEmpty}}, \code{\link{xRemoveEmpty...}};
\code{\link{xRemoveNan}}; \code{\link{xRemoveNull}},
\code{\link{xRemoveNull...}}; \code{\link{xSelect}},
\code{\link{xSelect...}}
}

