\name{xRemoveNa}
\alias{xRemoveNa}
\title{xRemoveNa}
\usage{
xRemoveNa(coll)
}
\arguments{
  \item{coll}{a collection. The collection to remove na
  values from.}

  \item{...}{see above.}
}
\value{
A list.
}
\description{
Remove the Na values from a collection.
}
\examples{

NULL

}
\seealso{
Other filtering_functions: \code{\link{xPartition}};
\code{\link{xReject}}, \code{\link{xReject...}};
\code{\link{xRemoveEmpty}}, \code{\link{xRemoveEmpty...}};
\code{\link{xRemoveNan}}; \code{\link{xRemoveNull}},
\code{\link{xRemoveNull...}}; \code{\link{xSelect}},
\code{\link{xSelect...}}
}

