\name{xRemoveNull}
\alias{xRemoveNull}
\alias{xRemoveNull...}
\title{xRemoveNull}
\usage{
xRemoveNull(coll)

xRemoveNull...(...)
}
\arguments{
  \item{coll}{a collection. The collection to remove null
  values from.}

  \item{...}{see above.}
}
\value{
A list.
}
\description{
Remove the null values from a collection.
}
\examples{

NULL

}
\seealso{
Other filtering_functions: \code{\link{xPartition}};
\code{\link{xReject}}, \code{\link{xReject...}};
\code{\link{xRemoveEmpty}}, \code{\link{xRemoveEmpty...}};
\code{\link{xRemoveNan}}; \code{\link{xRemoveNa}};
\code{\link{xSelect}}, \code{\link{xSelect...}}
}

