% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xRepeat}
\alias{xRepeat}
\alias{xRepeat...}
\title{xRepeat}
\usage{
xRepeat(num, coll)

xRepeat...(num, ...)
}
\arguments{
\item{num}{a nonnegative positive number. The number of
   times to repeat the collection.}

\item{coll}{a collection. The collection to repeat
   end to end.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Create a collection by repeating another collection several times.
}
\section{Corner Cases}{

   Returns the empty list if \bold{coll} is length-zero or num is zero.
}
\examples{

# Example 1
#
# Use xRepeat to repeat a sine wave.

range <- seq(0, 2*pi, len = 600)
wave <- xRepeat(5, sin(range))

plot(unlist(wave), type = 'l')

# Example 2
#
# Use xRepeat to build up a nested list.

}
\seealso{
Other reshaping_functions: \code{\link{xChop}},
  \code{\link{xChop...}}; \code{\link{xChunk}},
  \code{\link{xChunk...}}; \code{\link{xFlatten}},
  \code{\link{xFlatten...}}; \code{\link{xGroupBy}},
  \code{\link{xGroupBy...}}; \code{\link{xJoin}},
  \code{\link{xJoin...}}; \code{\link{xOneOf}},
  \code{\link{xOneOf...}}; \code{\link{xRecycle}},
  \code{\link{xRecycle...}}; \code{\link{xReverse}},
  \code{\link{xReverse...}}; \code{\link{xShuffle}},
  \code{\link{xShuffle...}}; \code{\link{xSortBy}},
  \code{\link{xSortBy...}}; \code{\link{xSplitAt}},
  \code{\link{xSplitAt...}}; \code{\link{xSplitBy}},
  \code{\link{xSplitBy...}}; \code{\link{xTabulate}},
  \code{\link{xTabulate...}}; \code{\link{xUnzipNames}},
  \code{\link{xUnzipNames...}}; \code{\link{xZipNames}},
  \code{\link{xZipNames...}}; \code{\link{xZip}},
  \code{\link{xZip...}}
}

