% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xRepeat}
\alias{xRepeat}
\alias{xRepeat_}
\title{xRepeat}
\usage{
xRepeat(num, coll)

xRepeat_(num, ...)
}
\arguments{
\item{num}{a nonnegative positive number. The number of
   times to repeat the collection.}

\item{coll}{a collection. The collection to repeat
   end to end.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Create a collection by repeating another collection several times.
}
\section{Type Signature}{

    |number| -> |any| -> [any]
}

\section{Corner Cases}{

   Returns the empty list if \bold{coll} is length-zero or num is zero.
}
\examples{

# 1.
#
# Use xRepeat to repeat a sine wave.

range <- seq(0, 2*pi, len = 600)
wave <- xRepeat(5, sin(range))

plot(unlist(wave), type = 'l')
}
\seealso{
Other reshaping_functions: \code{\link{xAppend}},
  \code{\link{xAppend_}}; \code{\link{xChop}},
  \code{\link{xChop_}}; \code{\link{xChunk}},
  \code{\link{xChunk_}}; \code{\link{xFlatten}},
  \code{\link{xFlatten_}}; \code{\link{xGroupBy}},
  \code{\link{xGroupBy_}}; \code{\link{xJoin}},
  \code{\link{xJoin_}}; \code{\link{xOneOf}},
  \code{\link{xOneOf_}}; \code{\link{xPrepend}},
  \code{\link{xPrepend_}}; \code{\link{xReverse}},
  \code{\link{xReverse_}}; \code{\link{xShuffle}},
  \code{\link{xShuffle_}}; \code{\link{xSortBy}},
  \code{\link{xSortBy_}}; \code{\link{xTabulate}},
  \code{\link{xTabulate_}}; \code{\link{xUnzipIndices}},
  \code{\link{xUnzipIndices_}}; \code{\link{xUnzipKeys}},
  \code{\link{xUnzipKeys_}}; \code{\link{xZipKeys}},
  \code{\link{xZipKeys_}}; \code{\link{xZip}},
  \code{\link{xZip_}}
}

