% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xReverse}
\alias{xReverse}
\alias{xReverse_}
\title{xReverse}
\usage{
xReverse(coll)

xReverse_(...)
}
\arguments{
\item{coll}{a collection. The collection to reverse.}

\item{...}{see above.}
}
\value{
A list of the same length as \bold{coll}
}
\description{
Reverse the order of elements in a collection.
}
\section{Corner Cases}{

   Reversing the empty list yields the empty list.
}
\examples{

# 1.
# select the third last letter in the english alphabet.

x_(letters) $ xReverse() $ xThirdOf()

# the above is more semantic than
# x_(letters) $ xAt(length(letters) - 3)
}
\seealso{
Other reshaping_functions: \code{\link{xChop}},
  \code{\link{xChop_}}; \code{\link{xChunk}},
  \code{\link{xChunk_}}; \code{\link{xFlatten}},
  \code{\link{xFlatten_}}; \code{\link{xGroupBy}},
  \code{\link{xGroupBy_}}; \code{\link{xJoin}},
  \code{\link{xJoin_}}; \code{\link{xOneOf}},
  \code{\link{xOneOf_}}; \code{\link{xRecycle}},
  \code{\link{xRecycle_}}; \code{\link{xRepeat}},
  \code{\link{xRepeat_}}; \code{\link{xShuffle}},
  \code{\link{xShuffle_}}; \code{\link{xSortBy}},
  \code{\link{xSortBy_}}; \code{\link{xSplitAt}},
  \code{\link{xSplitAt_}}; \code{\link{xSplitWith}},
  \code{\link{xSplitWith_}}; \code{\link{xTabulate}},
  \code{\link{xTabulate_}}; \code{\link{xUnzipKeys}},
  \code{\link{xUnzipKeys_}}; \code{\link{xZipKeys}},
  \code{\link{xZipKeys_}}; \code{\link{xZip}},
  \code{\link{xZip_}}
}

