% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xReverse}
\alias{xReverse}
\alias{xReverse...}
\title{xReverse}
\usage{
xReverse(coll)

xReverse...(...)
}
\arguments{
\item{coll}{a collection. The collection to reverse.}

\item{...}{see above.}
}
\value{
A list of the same length as \bold{coll}
}
\description{
Reverse the order of elements in a collection.
}
\section{Corner Cases}{

   Reversing the empty list yields the empty list.
}
\examples{

NULL

}
\seealso{
Other reshaping_functions: \code{\link{xChop}},
  \code{\link{xChop...}}; \code{\link{xChunk}},
  \code{\link{xChunk...}}; \code{\link{xFlatten}},
  \code{\link{xFlatten...}}; \code{\link{xGroupBy}},
  \code{\link{xGroupBy...}}; \code{\link{xJoin}},
  \code{\link{xJoin...}}; \code{\link{xOneOf}},
  \code{\link{xOneOf...}}; \code{\link{xRecycle}},
  \code{\link{xRecycle...}}; \code{\link{xRepeat}},
  \code{\link{xRepeat...}}; \code{\link{xShuffle}},
  \code{\link{xShuffle...}}; \code{\link{xSortBy}},
  \code{\link{xSortBy...}}; \code{\link{xSplitAt}},
  \code{\link{xSplitAt...}}; \code{\link{xSplitWith}},
  \code{\link{xSplitWith...}}; \code{\link{xTabulate}},
  \code{\link{xTabulate...}}; \code{\link{xUnzipKeys}},
  \code{\link{xUnzipKeys...}}; \code{\link{xZipKeys}},
  \code{\link{xZipKeys...}}; \code{\link{xZip}},
  \code{\link{xZip...}}
}

