\name{xReverse}
\alias{xReverse}
\title{xReverse}
\usage{
xReverse(coll)
}
\arguments{
  \item{coll}{a collection.}

  \item{...}{see above.}
}
\value{
A list of the same length as \code{coll}
}
\description{
Reverse a collection.
}
\section{Corner Cases}{
  Reversing the empty list yields the empty list.
}

\section{Variadic Functions

\link{ help_variadic }}{

}
\seealso{
Other reshaping_functions: \code{\link{xAssoc}},
\code{\link{xAssoc...}}; \code{\link{xChop}},
\code{\link{xChop...}}; \code{\link{xDissoc}},
\code{\link{xDissoc...}}; \code{\link{xFlatten}},
\code{\link{xFlatten...}}; \code{\link{xGroup}},
\code{\link{xGroup...}}; \code{\link{xJoin}},
\code{\link{xJoin...}}; \code{\link{xPermute}},
\code{\link{xPermute...}}; \code{\link{xRepeat}},
\code{\link{xRepeat...}}; \code{\link{xShuffle}},
\code{\link{xShuffle...}}; \code{\link{xSplitAt}},
\code{\link{xSplitAt...}}; \code{\link{xSplitWith}},
\code{\link{xSplitWith...}}; \code{\link{xZipWith}},
\code{\link{xZipWith...}}; \code{\link{xZip}},
\code{\link{xZip...}}
}

