% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xRiffle}
\alias{xRiffle}
\alias{xRiffle_}
\title{xRiffle}
\usage{
xRiffle(val, coll)

xRiffle_(val, ...)
}
\arguments{
\item{val}{an arbitrary value. The value to intersperse
    between elements in the list.}

\item{coll}{a collection. The collection to insert elements throughout.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Insert an element in between each element of a list.
}
\section{Type Signatures}{

    any -> |any| -> [any]
}

\section{Corner Cases}{

   Returns the empty list if \bold{coll} is empty.
}

\section{Corner Cases}{

   Returns the empty list if \bold{coll} is length-zero.
}

\section{Stability - \bold{Experimental}}{


This feature is newly introduced or experimental, may contain bugs, and may be removed
or altered in future versions if it is not sufficiently useful. Breaking changes
are likely in the near future.

If you find this feature useful, please leave an issue at https://github.com/rgrannell1/kea/issues.
}
\examples{

# 1. a toy use

xFromChars(xRiffle_(',', 'a', 'b', 'c'))

# "a,b,c"
#
# a more idiomatic way to achieve this would be with
#
# xImplode_(',', 'a', 'b', 'c')

# 2. inject elements into a list.
#
# by combining xRiffle with xJoin you can inject lists between
# each element of a collection. The following example isn't very
# practicaly, but demonstrates the complex lists you can build with
# xRiffle well.

x_(paste(1:5)) $ xRiffle(list('+', '10', '*')) $ xJoin() $ x_FromWords()

# "1 + 10 * 2 + 10 * 3 + 10 * 4 + 10 * 5"
}

