\name{xRowUnit}
\alias{xRowUnit}
\title{$ xRowUnit}
\usage{
x_(  ) $ xRowUnit()
}
\value{
an empty matrix or data.frame.
}
\description{
Return the concatenative unit for a matrix or data.frame
row.
}
\seealso{
Other exclusively_methods: \code{\link{xByColnames}};
\code{\link{xByCols}}; \code{\link{xByLevels}};
\code{\link{xByRows}}; \code{\link{xColUnit}};
\code{\link{xExecute}}
}

