% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xRowUnit}
\alias{xRowUnit}
\title{$ xRowUnit}
\usage{
x_(  ) $ xRowUnit()
}
\value{
An empty matrix or data.frame.
}
\description{
Return the concatenative unit for a matrix or data.frame row.
}
\seealso{
Other methods: \code{\link{xByColkeys}};
  \code{\link{xByCols}}; \code{\link{xByLevels}},
  \code{\link{xByLevels}}; \code{\link{xByRows}};
  \code{\link{xColUnit}}; \code{\link{xElemsByCols}};
  \code{\link{xElemsByRows}}; \code{\link{xExecute}};
  \code{\link{xTap}}; \code{\link{xTranspose}};
  \code{\link{x_}}
}

