% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xSecondOf}
\alias{xSecondOf}
\alias{xSecondOf...}
\title{xSecondOf}
\usage{
xSecondOf(coll)

xSecondOf...(...)
}
\arguments{
\item{coll}{a collection. The collection to get the
   second element of.}

\item{...}{see above.}
}
\value{
The second element in \bold{coll}.
}
\description{
Return the second element in a collection.
}
\section{Corner Cases}{

   Throws an error if \bold{coll} has less than two
   elements; this is because any other corner case
   would violate the functions type-signature.
}
\examples{

# 1. CE Return the penultimate string in a sentence.
#    Not the most efficient solution, but it's usually
#    sufficient. As always, using a right fold would be more
#    efficient but less expressive. Every problem can be solved
#    with a fold.

sentence <- "He just kept talking in one long incredibly unbroken sentence moving from
topic to topic so that no-one had a chance to
interrupt; it was really quite hypnotic"

x_(sentence) $
xToWords() $
xReverse() $
x_SecondOf()

# 'quite'
}
\seealso{
Other selection_functions: \code{\link{xAtCol}},
  \code{\link{xAtCol...}}; \code{\link{xAt}},
  \code{\link{xAt...}}; \code{\link{xDropWhile}},
  \code{\link{xDropWhile...}}; \code{\link{xDrop}},
  \code{\link{xDrop...}}; \code{\link{xFirstOf}},
  \code{\link{xFirstOf...}}; \code{\link{xFourthOf}},
  \code{\link{xFourthOf...}}; \code{\link{xInitOf}},
  \code{\link{xInitOf...}}; \code{\link{xLastOf}},
  \code{\link{xLastOf...}}; \code{\link{xMaxBy}},
  \code{\link{xMaxBy...}}; \code{\link{xMinBy}},
  \code{\link{xMinBy...}}; \code{\link{xPluck}},
  \code{\link{xPluck...}}; \code{\link{xRestOf}},
  \code{\link{xRestOf...}}; \code{\link{xSlice}},
  \code{\link{xSlice...}}; \code{\link{xTakeWhile}},
  \code{\link{xTakeWhile...}}; \code{\link{xTake}},
  \code{\link{xTake...}}; \code{\link{xThirdOf}},
  \code{\link{xThirdOf...}}
}

