% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xSecondOf}
\alias{xSecondOf}
\alias{xSecondOf_}
\title{xSecondOf}
\usage{
xSecondOf(coll)

xSecondOf_(...)
}
\arguments{
\item{coll}{a collection. The collection to get the
   second element of.}

\item{...}{see above.}
}
\value{
The second element in \bold{coll}.
}
\description{
Return the second element in a collection.
}
\section{Type Signature}{

    |any| -> any
}

\section{Corner Cases}{

   Throws an error if \bold{coll} has less than two
   elements; this is because any other corner case
   would violate the functions type-signature.
}
\examples{

# 1. CodeEvalReturn the penultimate string in a sentence.
#    Not the most efficient solution, but it's usually
#    sufficient. As always, using a right fold would be more
#    efficient but less expressive. Every problem can be solved
#    with a fold.

sentence <- "He just kept talking in one long incredibly unbroken sentence moving from
topic to topic so that no-one had a chance to
interrupt; it was really quite hypnotic"

x_(sentence) $ xToWords() $ xReverse() $ x_SecondOf()

# 'quite'
}
\seealso{
Other selection_functions: \code{\link{xAtKey}},
  \code{\link{xAtKey_}}; \code{\link{xAt}},
  \code{\link{xAt_}}; \code{\link{xDropWhile}},
  \code{\link{xDropWhile_}}; \code{\link{xDrop}},
  \code{\link{xDrop_}}; \code{\link{xFirstOf}},
  \code{\link{xFirstOf_}}; \code{\link{xFourthOf}},
  \code{\link{xFourthOf_}}; \code{\link{xInitOf}},
  \code{\link{xInitOf_}}; \code{\link{xLastOf}},
  \code{\link{xLastOf_}}; \code{\link{xMaxBy}},
  \code{\link{xMaxBy_}}; \code{\link{xMinBy}},
  \code{\link{xMinBy_}}; \code{\link{xProdBy}},
  \code{\link{xProdBy_}}; \code{\link{xRestOf}},
  \code{\link{xRestOf_}}; \code{\link{xSlice}},
  \code{\link{xSlice_}}; \code{\link{xSumBy}},
  \code{\link{xSumBy_}}; \code{\link{xTakeWhile}},
  \code{\link{xTakeWhile_}}; \code{\link{xTake}},
  \code{\link{xTake_}}; \code{\link{xThirdOf}},
  \code{\link{xThirdOf_}}
}

