\name{xSetProd}
\alias{xSetProd}
\title{xSetProd}
\usage{
  xSetProd(colls)
}
\arguments{
  \item{colls}{n collections.}
}
\value{
  a list of n-element lists.
}
\description{
  Get the cartesian product of several collections.
}
\section{Corner Cases}{
  returns the empty list if \code{coll} is length-zero.
}
\seealso{
  Other combinatoric_functions collection_functions:
  \code{\link{xCombos}}
}

