% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xShuffle}
\alias{xShuffle}
\alias{xShuffle...}
\title{xShuffle}
\usage{
xShuffle(coll)

xShuffle...(...)
}
\arguments{
\item{coll}{a collection. The collection to rearrange.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Randomly rearrange a collection.
}
\section{Corner Cases}{

     Returns the empty list if \bold{coll} is length-zero.
}
\examples{

NULL

}
\seealso{
Other reshaping_functions: \code{\link{xChop}},
  \code{\link{xChop...}}; \code{\link{xChunk}},
  \code{\link{xChunk...}}; \code{\link{xFlatten}},
  \code{\link{xFlatten...}}; \code{\link{xJoin}},
  \code{\link{xJoin...}}; \code{\link{xOneOf}},
  \code{\link{xOneOf...}}; \code{\link{xRecycle}},
  \code{\link{xRecycle...}}; \code{\link{xRepeat}},
  \code{\link{xRepeat...}}; \code{\link{xReverse}},
  \code{\link{xReverse...}}; \code{\link{xSortBy}},
  \code{\link{xSortBy...}}; \code{\link{xSplitAt}},
  \code{\link{xSplitAt...}}; \code{\link{xSplitBy}},
  \code{\link{xSplitBy...}}; \code{\link{xTabulate}},
  \code{\link{xTabulate...}}; \code{\link{xUnzipNames}},
  \code{\link{xUnzipNames...}}; \code{\link{xZipNames}},
  \code{\link{xZipNames...}}; \code{\link{xZip}},
  \code{\link{xZip...}}
}

