% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xShuffle}
\alias{xShuffle}
\alias{xShuffle_}
\title{xShuffle}
\usage{
xShuffle(coll)

xShuffle_(...)
}
\arguments{
\item{coll}{a collection. The collection to rearrange.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Randomly rearrange a collection.
}
\section{Corner Cases}{

     Returns the empty list if \bold{coll} is length-zero.
}
\examples{

# 1.

NULL
}
\seealso{
Other reshaping_functions: \code{\link{xChop}},
  \code{\link{xChop_}}; \code{\link{xChunk}},
  \code{\link{xChunk_}}; \code{\link{xFlatten}},
  \code{\link{xFlatten_}}; \code{\link{xGroupBy}},
  \code{\link{xGroupBy_}}; \code{\link{xJoin}},
  \code{\link{xJoin_}}; \code{\link{xOneOf}},
  \code{\link{xOneOf_}}; \code{\link{xRecycle}},
  \code{\link{xRecycle_}}; \code{\link{xRepeat}},
  \code{\link{xRepeat_}}; \code{\link{xReverse}},
  \code{\link{xReverse_}}; \code{\link{xSortBy}},
  \code{\link{xSortBy_}}; \code{\link{xSplitAt}},
  \code{\link{xSplitAt_}}; \code{\link{xSplitWith}},
  \code{\link{xSplitWith_}}; \code{\link{xTabulate}},
  \code{\link{xTabulate_}}; \code{\link{xUnzipIndices}},
  \code{\link{xUnzipIndices_}}; \code{\link{xUnzipKeys}},
  \code{\link{xUnzipKeys_}}; \code{\link{xZipKeys}},
  \code{\link{xZipKeys_}}; \code{\link{xZip}},
  \code{\link{xZip_}}
}

