% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xSlice}
\alias{xSlice}
\alias{xSlice...}
\title{xSlice}
\usage{
xSlice(nums, coll)

xSlice...(nums, ...)
}
\arguments{
\item{nums}{a vector of whole numbers. Indices to select values
   in \bold{coll}.}

\item{coll}{a collection. The collection to subset.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Select a collection at certain indices.
}
\details{
\bold{xSlice} is similar to base R's subsetting operator '[',
    except that it performs more validation on the input indices,
    and it acts as a normal function.
}
\section{Corner Cases}{

   \bold{xSlice} does not allow subscripting values out of bounds, unlike
   base R. If an index larger than the maximum value in \bold{coll} is given
   an error is thrown.
}
\examples{

NULL
}
\seealso{
Other selection_functions: \code{\link{xAtCol}},
  \code{\link{xAtCol...}}; \code{\link{xAt}},
  \code{\link{xAt...}}; \code{\link{xDropWhile}},
  \code{\link{xDropWhile...}}; \code{\link{xDrop}},
  \code{\link{xDrop...}}; \code{\link{xFirstOf}},
  \code{\link{xFirstOf...}}; \code{\link{xFourthOf}},
  \code{\link{xFourthOf...}}; \code{\link{xInitOf}},
  \code{\link{xInitOf...}}; \code{\link{xLastOf}},
  \code{\link{xLastOf...}}; \code{\link{xMaxBy}},
  \code{\link{xMaxBy...}}; \code{\link{xMinBy}},
  \code{\link{xMinBy...}}; \code{\link{xPluck}},
  \code{\link{xPluck...}}; \code{\link{xRestOf}},
  \code{\link{xRestOf...}}; \code{\link{xSecondOf}},
  \code{\link{xSecondOf...}}; \code{\link{xTakeWhile}},
  \code{\link{xTakeWhile...}}; \code{\link{xTake}},
  \code{\link{xTake...}}; \code{\link{xThirdOf}},
  \code{\link{xThirdOf...}}
}

