% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xSlice}
\alias{xSlice}
\alias{xSlice_}
\title{xSlice}
\usage{
xSlice(nums, coll)

xSlice_(nums, ...)
}
\arguments{
\item{nums}{a vector of whole numbers. Indices to select values
   in \bold{coll}.}

\item{coll}{a collection. The collection to subset.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Select a collection at certain indices.
}
\details{
\bold{xSlice} is similar to base R's subsetting operator '[',
    except that it performs more validation on the input indices,
    and it acts as a normal function.
}
\section{Type Signature}{

    |numeric| -> |any| -> [any]
}

\section{Corner Cases}{

   \bold{xSlice} does not allow subscripting values out of bounds, unlike
   base R. If an index larger than the maximum value in \bold{coll} is given
   an error is thrown. The empty list is returned when \bold{nums} is length zero.
}
\examples{

# 1. Subsetting a collection
#
# Slice is the replacement for '[[' in Kiwi. It behaves similarly.

nums <- 1:10

x_(nums) $ x_Slice(xWhere(nums \%\% 3 == 0))

# list(3, 6, 9)
#
# which is similar to

# nums[w nums \%\% 3 == 0 ]

# using standard subsetting.
}
\seealso{
Other selection_functions: \code{\link{xAtKey}},
  \code{\link{xAtKey_}}; \code{\link{xAt}},
  \code{\link{xAt_}}; \code{\link{xDropWhile}},
  \code{\link{xDropWhile_}}; \code{\link{xDrop}},
  \code{\link{xDrop_}}; \code{\link{xFirstOf}},
  \code{\link{xFirstOf_}}; \code{\link{xFourthOf}},
  \code{\link{xFourthOf_}}; \code{\link{xInitOf}},
  \code{\link{xInitOf_}}; \code{\link{xLastOf}},
  \code{\link{xLastOf_}}; \code{\link{xMaxBy}},
  \code{\link{xMaxBy_}}; \code{\link{xMinBy}},
  \code{\link{xMinBy_}}; \code{\link{xRestOf}},
  \code{\link{xRestOf_}}; \code{\link{xSecondOf}},
  \code{\link{xSecondOf_}}; \code{\link{xTakeWhile}},
  \code{\link{xTakeWhile_}}; \code{\link{xTake}},
  \code{\link{xTake_}}; \code{\link{xThirdOf}},
  \code{\link{xThirdOf_}}
}

