% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xSplitAt}
\alias{xSplitAt}
\alias{xSplitAt...}
\title{xSplitAt}
\usage{
xSplitAt(nums, coll)

xSplitAt...(nums, ...)
}
\arguments{
\item{nums}{a vector of nonnegative whole numbers. The indices
   to split a collection \bold{after}.}

\item{coll}{a collection. The collection to split.}

\item{...}{see above.}
}
\value{
A list of lists.
}
\description{
Split a collection into at several indices.
}
\section{Corner Cases}{

   If \code{nums} is zero then the first list in the
   returned value is empty.
   Likewise, if \code{nums} is equal or larger than the
   length of \bold{coll} then
   the second return list is empty. If \bold{coll} is
   length zero both lists are empty.
}
\examples{

NULL

}
\seealso{
Other reshaping_functions: \code{\link{xChop}},
  \code{\link{xChop...}}; \code{\link{xChunk}},
  \code{\link{xChunk...}}; \code{\link{xFlatten}},
  \code{\link{xFlatten...}}; \code{\link{xGroupBy}},
  \code{\link{xGroupBy...}}; \code{\link{xJoin}},
  \code{\link{xJoin...}}; \code{\link{xOneOf}},
  \code{\link{xOneOf...}}; \code{\link{xRecycle}},
  \code{\link{xRecycle...}}; \code{\link{xRepeat}},
  \code{\link{xRepeat...}}; \code{\link{xReverse}},
  \code{\link{xReverse...}}; \code{\link{xShuffle}},
  \code{\link{xShuffle...}}; \code{\link{xSortBy}},
  \code{\link{xSortBy...}}; \code{\link{xSplitBy}},
  \code{\link{xSplitBy...}}; \code{\link{xTabulate}},
  \code{\link{xTabulate...}}; \code{\link{xUnzipNames}},
  \code{\link{xUnzipNames...}}; \code{\link{xZipNames}},
  \code{\link{xZipNames...}}; \code{\link{xZip}},
  \code{\link{xZip...}}
}

