% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xSplitWith}
\alias{xSplitWith}
\alias{xSplitWith_}
\title{xSplitWith}
\usage{
xSplitWith(pred, coll)

xSplitWith_(pred, ...)
}
\arguments{
\item{pred}{a binary predicate. The predicate to test whether
     a split should be introduced between the two elements.}

\item{coll}{a collection. The collection to split.}

\item{...}{see above.}
}
\value{
A list of lists.
}
\description{
Divide a collection into groups determined by a predicate.
}
\section{Corner Cases}{

     Returns the empty list if \bold{coll} is length-zero.
}
\examples{

NULL
}
\seealso{
Other reshaping_functions: \code{\link{xChop}},
  \code{\link{xChop_}}; \code{\link{xChunk}},
  \code{\link{xChunk_}}; \code{\link{xFlatten}},
  \code{\link{xFlatten_}}; \code{\link{xGroupBy}},
  \code{\link{xGroupBy_}}; \code{\link{xJoin}},
  \code{\link{xJoin_}}; \code{\link{xOneOf}},
  \code{\link{xOneOf_}}; \code{\link{xRecycle}},
  \code{\link{xRecycle_}}; \code{\link{xRepeat}},
  \code{\link{xRepeat_}}; \code{\link{xReverse}},
  \code{\link{xReverse_}}; \code{\link{xShuffle}},
  \code{\link{xShuffle_}}; \code{\link{xSortBy}},
  \code{\link{xSortBy_}}; \code{\link{xSplitAt}},
  \code{\link{xSplitAt_}}; \code{\link{xTabulate}},
  \code{\link{xTabulate_}}; \code{\link{xUnzipIndices}},
  \code{\link{xUnzipIndices_}}; \code{\link{xUnzipKeys}},
  \code{\link{xUnzipKeys_}}; \code{\link{xZipKeys}},
  \code{\link{xZipKeys_}}; \code{\link{xZip}},
  \code{\link{xZip_}}
}

