% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xSplitWith}
\alias{xSplitWith}
\alias{xSplitWith...}
\title{xSplitWith}
\usage{
xSplitWith(pred, coll)

xSplitWith...(pred, ...)
}
\arguments{
\item{pred}{a binary predicate. The predicate to test whether
     a split should be introduced between the two elements.}

\item{coll}{a collection. The collection to split.}

\item{...}{see above.}
}
\value{
A list of lists.
}
\description{
Divide a collection into groups determined by a predicate.
}
\section{Corner Cases}{

     Returns the empty list if \bold{coll} is length-zero.
}
\examples{

NULL
}
\seealso{
Other reshaping_functions: \code{\link{xChop}},
  \code{\link{xChop...}}; \code{\link{xChunk}},
  \code{\link{xChunk...}}; \code{\link{xFlatten}},
  \code{\link{xFlatten...}}; \code{\link{xGroupBy}},
  \code{\link{xGroupBy...}}; \code{\link{xJoin}},
  \code{\link{xJoin...}}; \code{\link{xOneOf}},
  \code{\link{xOneOf...}}; \code{\link{xRecycle}},
  \code{\link{xRecycle...}}; \code{\link{xRepeat}},
  \code{\link{xRepeat...}}; \code{\link{xReverse}},
  \code{\link{xReverse...}}; \code{\link{xShuffle}},
  \code{\link{xShuffle...}}; \code{\link{xSortBy}},
  \code{\link{xSortBy...}}; \code{\link{xSplitAt}},
  \code{\link{xSplitAt...}}; \code{\link{xTabulate}},
  \code{\link{xTabulate...}}; \code{\link{xUnzipNames}},
  \code{\link{xUnzipNames...}}; \code{\link{xZipNames}},
  \code{\link{xZipNames...}}; \code{\link{xZip}},
  \code{\link{xZip...}}
}

