% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xSpread}
\alias{xSpread}
\title{xSpread}
\usage{
xSpread(fn)
}
\arguments{
\item{fn}{a unary function. The function to convert
   to a variadic function.}
}
\value{
A variadic function.
}
\description{
Return a variadic function that passes its arguments
as a list to its underyling function.
}
\details{
\code{xSpread} takes a unary function and
   returns a function with ellipsis parametres.
   The returned function passes its ellipsis arguments
   as a single list to the underlying function.
}
\examples{

# 1.
# Convert a function that takes a single value as an argument to
# a function that takes many arguments.
#

summate <- xSpread(nums := xFold("+", 0, nums) )

summate(1, 2, 3, 4, 5, 6)

# 2.
#

}
\seealso{
Other function_application_functions: \code{\link{xApply}},
  \code{\link{xApply_}}; \code{\link{xDeepMap}},
  \code{\link{xDeepMap_}}; \code{\link{xFlatMap}},
  \code{\link{xFlatMap_}}; \code{\link{xMapIndexed}},
  \code{\link{xMapIndexed_}}; \code{\link{xMapply}},
  \code{\link{xMapply_}}; \code{\link{xMap}},
  \code{\link{xMap_}}; \code{\link{xThread}},
  \code{\link{xThread_}}; \code{\link{xUnspread}};
  \code{\link{xVectorise}}, \code{\link{xVectorize}}

Other function_modifying_functions: \code{\link{\%and\%}},
  \code{\link{\%or\%}}, \code{\link{xLift}},
  \code{\link{xLift_}}; \code{\link{\%of\%}},
  \code{\link{\%then\%}}, \code{\link{xCompose}},
  \code{\link{xCompose_}}; \code{\link{xAsClosure}};
  \code{\link{xCapture}}, \code{\link{xK}};
  \code{\link{xDelay}}; \code{\link{xJuxtapose}},
  \code{\link{xJuxtapose_}}; \code{\link{xNot}};
  \code{\link{xFix}}, \code{\link{xFix_}};
  \code{\link{xThread}}, \code{\link{xThread_}};
  \code{\link{xUnspread}}; \code{\link{xVectorise}},
  \code{\link{xVectorize}}

Other parametre_functions: \code{\link{xArityOf}};
  \code{\link{xFormalsOf}}; \code{\link{xIsVariadic}};
  \code{\link{xParamsOf}}; \code{\link{xFix}},
  \code{\link{xFix_}}; \code{\link{xUnspread}}
}

