\name{xStopwatch}
\alias{xStopwatch}
\title{xStopwatch}
\usage{
xStopwatch(num)
}
\arguments{
  \item{num}{a positive number.}
}
\value{
A variadic function that ignores its arguments.
}
\description{
Create a function that returns true for a preset time after
creation.
}
\examples{

NULL

}
\seealso{
Other time_functions: \code{\link{xDelay}};
\code{\link{xLimit}}
}

