\name{xStopWatch}
\alias{xStopWatch}
\title{xStopWatch}
\usage{
  xStopWatch(num)
}
\arguments{
  \item{num}{a positive number.}
}
\value{
  a variadic function that ignores its arguments.
}
\description{
  Create a function that returns true for a preset time
  after creation.
}
\seealso{
  Other higher_order_functions: \code{\link{xArity}},
  \code{\link{xAsUnary}}, \code{\link{xCompose}},
  \code{\link{xConst}}, \code{\link{xFlip}},
  \code{\link{xFormals}}, \code{\link{xIterate}},
  \code{\link{xJuxtapose}}, \code{\link{xLift}},
  \code{\link{xLimit}}, \code{\link{xNot}},
  \code{\link{xParams}}, \code{\link{xUnfold}},
  \code{\link{xWait}}
}

