% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xSubstring}
\alias{xSubstring}
\alias{xSubstring...}
\title{xSubstring}
\usage{
xSubstring(str, nums)

xSubstring...(str, ...)
}
\arguments{
\item{str}{a string. The string to subset.}

\item{nums}{a collection of whole numbers. The indices to select
   characters with. Negative, positive and zero indices are allowed,
   but normal R index semantics dictate how they can intermix.}

\item{...}{see above.}
}
\value{
A character vector.
}
\description{
Subset a string using normal R vector indexing.
}
\section{Corner Cases}{

     Returns the empty list if \bold{coll} is length-zero. If any number in
     nums is larger than the length of \bold{coll} an error is thrown.
}
\examples{

NULL

}
\seealso{
Other text_processing_functions: \code{\link{xExplode}},
  \code{\link{xExplode...}}; \code{\link{xFromChars}},
  \code{\link{xFromChars...}}; \code{\link{xFromLines}},
  \code{\link{xFromLines...}}; \code{\link{xFromWords}},
  \code{\link{xFromWords...}}; \code{\link{xImplode}},
  \code{\link{xImplode...}}; \code{\link{xReadChars}};
  \code{\link{xReadLines}}; \code{\link{xReadWords}};
  \code{\link{xToChars}}; \code{\link{xToLines}};
  \code{\link{xToWords}}; \code{\link{xWriteChars}};
  \code{\link{xWriteLines}}; \code{\link{xWriteWords}}
}

