\name{xSubString}
\alias{xSubString}
\title{xSubString}
\usage{
xSubString(str, nums)
}
\arguments{
  \item{str}{a string.}

  \item{nums}{indices of \code{str}.}
}
\value{
a character vector.
}
\description{
Subset a string using normal R vector indexing.
}
\section{Corner Cases}{
  returns the empty list if \code{coll} is length-zero.
}
\seealso{
Other character_vector_functions: \code{\link{xChars}};
\code{\link{xExplode}}; \code{\link{xImplode}};
\code{\link{xLines}}; \code{\link{xUnchars}};
\code{\link{xUnlines}}; \code{\link{xUnwords}};
\code{\link{xWords}}
}

