\name{xSucc}
\alias{xSucc}
\title{xSucc}
\usage{
  xSucc(nums)
}
\arguments{
  \item{nums}{a collection of numbers.}
}
\value{
  a vector of numbers.
}
\description{
  Decrement a collection of numbers.
}
\section{Corner Cases}{
  If \code{nums} is empty then the unit of that vector is
  returned.
}

