\name{xSuccOf}
\alias{xSuccOf}
\alias{xSuccOf...}
\title{xSuccOf}
\usage{
xSuccOf(nums)

xSuccOf...(...)
}
\arguments{
  \item{nums}{a collection of numbers. The numbers to
  increment.}

  \item{...}{see above.}
}
\value{
A vector of numbers.
}
\description{
Increment a collection of numbers.
}
\section{Corner Cases}{
  If \code{nums} is empty then the unit of that vector is
  returned.
}
\examples{

NULL

}

