% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xSwap}
\alias{xSwap}
\alias{xSwap_}
\title{xSwap}
\usage{
xSwap(val1, val2, coll)

xSwap_(val1, val2, ...)
}
\arguments{
\item{val1}{an arbitrary value. The value to replace.}

\item{val2}{an arbitrary value. The replacement value.}

\item{coll}{a collection. The collection to replace elements in.}

\item{...}{see above.}
}
\description{
Replace each occurence of a certain value in a collection with another value.
}
\section{Type Signature}{

    any -> any -> [any] -> [any]
}

\section{Rcpp}{


    This function is written using C++, and should be quite performant.
}

\section{Corner Cases}{

   The equality test that \bold{xSwap} distinguishes between numbers with very small
   differences (0.999999 != 1) and treats -0 as equal to 0.
}
\examples{

# 1. Replace NaN and NA values with 0.

xSwap( NaN, 0, xSwap(Na, 0, c(NaN, 1, 2, 3, 4, Na, -10)) )

# or the much more readable.
#
x__(NaN, 1, 2, 3, 4, Na, -10) $ xSwap(NaN, 0) $ x_Swap(Na, 0)

# list(0, 1, 2, 3, 4, 0, -10)
}

