% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xTabulate}
\alias{xTabulate}
\alias{xTabulate...}
\title{xTabulate}
\usage{
xTabulate(coll)

xTabulate...(...)
}
\arguments{
\item{coll}{a collection. The values to find the frequency of.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Tabulate a collection into pairs of value:frequency lists.
}
\details{
\bold{xTabulate} is superficially similar to the base
    function \bold{table}: given a collection that may or
    may not contain duplicates, it calculates the frequencies
    of each unique element.

    \code{xTabulate(c('y', 'n', 'y', 'y', 'n'))}

    \code{list(list("y", 3), list("n", 2))}

    The result of the tabulation is unsorted for efficiency;
    if sorting is required \bold{xSortBy} can be used.
}
\section{Corner Cases}{

   If \bold{coll} is length-zero the empty list is returned.
}
\examples{

NULL
}
\seealso{
Other reshaping_functions: \code{\link{xChop}},
  \code{\link{xChop...}}; \code{\link{xChunk}},
  \code{\link{xChunk...}}; \code{\link{xFlatten}},
  \code{\link{xFlatten...}}; \code{\link{xJoin}},
  \code{\link{xJoin...}}; \code{\link{xOneOf}},
  \code{\link{xOneOf...}}; \code{\link{xRecycle}},
  \code{\link{xRecycle...}}; \code{\link{xRepeat}},
  \code{\link{xRepeat...}}; \code{\link{xReverse}},
  \code{\link{xReverse...}}; \code{\link{xShuffle}},
  \code{\link{xShuffle...}}; \code{\link{xSortBy}},
  \code{\link{xSortBy...}}; \code{\link{xSplitAt}},
  \code{\link{xSplitAt...}}; \code{\link{xSplitBy}},
  \code{\link{xSplitBy...}}; \code{\link{xUnzipNames}},
  \code{\link{xUnzipNames...}}; \code{\link{xZipNames}},
  \code{\link{xZipNames...}}; \code{\link{xZip}},
  \code{\link{xZip...}}
}

