% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xTake}
\alias{xTake}
\alias{xTake_}
\title{xTake}
\usage{
xTake(num, coll)

xTake_(num, ...)
}
\arguments{
\item{num}{a nonnegative whole number. The number of elements to
     return from the input collection.}

\item{coll}{a collection. The collection to subset.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Take several elements from the head of a collection.
}
\section{Type Signature}{

   <number> -> |any| -> |any|
}

\section{Corner Cases}{

   If \bold{coll} is empty the empty list is returned,
   or if \bold{num} is length-zero.
}

\section{Rcpp}{


    This function is written using C++, and should be quite performant.
}
\examples{

# 1. Create a selector to take a fixed number of values.

nine <- 1:9
sevenOf <- xTake(7)

sevenOf(nine)

# list(1, 2, 3, 4, 5, 6, 7, 8, 9)
}
\seealso{
Other selection_functions: \code{\link{xAtKey}},
  \code{\link{xAtKey_}}; \code{\link{xAt}},
  \code{\link{xAt_}}; \code{\link{xDropWhile}},
  \code{\link{xDropWhile_}}; \code{\link{xDrop}},
  \code{\link{xDrop_}}; \code{\link{xFirstOf}},
  \code{\link{xFirstOf_}}; \code{\link{xFourthOf}},
  \code{\link{xFourthOf_}}; \code{\link{xInitOf}},
  \code{\link{xInitOf_}}; \code{\link{xLastOf}},
  \code{\link{xLastOf_}}; \code{\link{xMaxBy}},
  \code{\link{xMaxBy_}}; \code{\link{xMinBy}},
  \code{\link{xMinBy_}}; \code{\link{xRestOf}},
  \code{\link{xRestOf_}}; \code{\link{xSecondOf}},
  \code{\link{xSecondOf_}}; \code{\link{xSlice}},
  \code{\link{xSlice_}}; \code{\link{xTakeWhile}},
  \code{\link{xTakeWhile_}}; \code{\link{xThirdOf}},
  \code{\link{xThirdOf_}}
}

