\name{xTake}
\alias{xTake}
\title{xTake}
\usage{
  xTake(num, coll)
}
\arguments{
  \item{num}{a nonnegative whole number.}

  \item{coll}{a collection}
}
\value{
  a list.
}
\description{
  Take several elements from the head of a collection.
}
\section{Corner Cases}{
  If \code{coll} is empty the empty list is returned.
}
\seealso{
  Other collection_functions: \code{\link{xAsFunction}},
  \code{\link{xAssoc}}, \code{\link{xContains}},
  \code{\link{xDiffer}}, \code{\link{xDissoc}},
  \code{\link{xDrop}}, \code{\link{xFirst}},
  \code{\link{xFlatten}}, \code{\link{xFourth}},
  \code{\link{xInit}}, \code{\link{xInter}},
  \code{\link{xIsEmpty}}, \code{\link{xIsFalse}},
  \code{\link{xIsNa}}, \code{\link{xIsNan}},
  \code{\link{xIsNull}}, \code{\link{xIsTrue}},
  \code{\link{xLast}}, \code{\link{xLength}},
  \code{\link{xName}}, \code{\link{xNotFalse}},
  \code{\link{xNotNa}}, \code{\link{xNotNan}},
  \code{\link{xNotTrue}}, \code{\link{xPack}},
  \code{\link{xRepeat}}, \code{\link{xRest}},
  \code{\link{xSecond}}, \code{\link{xSegment}},
  \code{\link{xShuffle}}, \code{\link{xSplit}},
  \code{\link{xThird}}, \code{\link{xUnion}},
  \code{\link{xUnit}}
}

