% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xTakeWhile.R
\name{xTakeWhile}
\alias{xTakeWhile}
\alias{xTakeWhile_}
\title{xTakeWhile}
\usage{
xTakeWhile(pred, coll)

xTakeWhile_(pred, ...)
}
\arguments{
\item{pred}{a predicate. The function to test each element of
   the collection with.}

\item{coll}{a collection. The collection to drop elements from.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Take every element in a collection from the start
until a predicate returns false.
}
\section{Type Signature}{

   (any -> <logical>) -> |any| -> |any|
}

\section{Corner Cases}{

   Returns the empty list if \bold{coll} is
   length-zero or the first element of
   \bold{coll} returns false for the predicate.
   Na values are considered false.
}
\examples{

NULL

}
\seealso{
Other selection_functions: \code{\link{xAtKey}},
  \code{\link{xAtKey_}}; \code{\link{xAt}},
  \code{\link{xAt_}}; \code{\link{xDropWhile}},
  \code{\link{xDropWhile_}}; \code{\link{xDrop}},
  \code{\link{xDrop_}}; \code{\link{xFirstOf}},
  \code{\link{xFirstOf_}}; \code{\link{xFourthOf}},
  \code{\link{xFourthOf_}}; \code{\link{xInitOf}},
  \code{\link{xInitOf_}}; \code{\link{xLastOf}},
  \code{\link{xLastOf_}}; \code{\link{xMaxBy}},
  \code{\link{xMaxBy_}}; \code{\link{xMinBy}},
  \code{\link{xMinBy_}}; \code{\link{xRestOf}},
  \code{\link{xRestOf_}}; \code{\link{xSecondOf}},
  \code{\link{xSecondOf_}}; \code{\link{xSlice}},
  \code{\link{xSlice_}}; \code{\link{xTake}},
  \code{\link{xTake_}}; \code{\link{xThirdOf}},
  \code{\link{xThirdOf_}}
}

