\name{xTap}
\alias{xTap}
\title{$ xTap}
\usage{
x_(  ) $ xTap(fn)
}
\value{
An arrow object.
}
\description{
Apply an anonymous function to the contents of an arrow
object.
}
\seealso{
Other exclusively_methods: \code{\link{xByColnames}};
\code{\link{xByCols}}; \code{\link{xByLevels}},
\code{\link{xByLevels}}; \code{\link{xByRows}};
\code{\link{xColUnit}}; \code{\link{xElemsByCols}};
\code{\link{xElemsByRows}}; \code{\link{xExecute}};
\code{\link{xRowUnit}}; \code{\link{xTranspose}}
}

