% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xTap}
\alias{xTap}
\title{$ xTap}
\usage{
x_(  ) $ xTap(fn)
}
\value{
An arrow object.
}
\description{
Apply an anonymous function to the contents of an arrow object.
}
\details{
\bold{xTap} applies non-arrow functions to be applied to arrow objects.
   These functions can be named or unnamed. This allows base functions
   or external libraries to interoperate with arrow's chaining methods.

   \code{x_(letters) $ xShuffle() $ x_Tap(sort)}

   \bold{xTap} also allows the use of anonymous methods.

   \code{x_(1:10) $ x_Tap(nums := all(nums > 0))}
}
\seealso{
Other methods: \code{\link{xByColnames}};
  \code{\link{xByCols}}; \code{\link{xByLevels}},
  \code{\link{xByLevels}}; \code{\link{xByRows}};
  \code{\link{xColUnit}}; \code{\link{xElemsByCols}};
  \code{\link{xElemsByRows}}; \code{\link{xExecute}};
  \code{\link{xRowUnit}}; \code{\link{xTranspose}};
  \code{\link{x_}}
}

