% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xTap}
\alias{xTap}
\title{xTap}
\usage{
x_(  ) $ xTap(fn)
}
\arguments{
\item{fn}{a unary function. The function to apply to the data in the kiwi object.}

\item{val}{an arbitrary value. The contents of the kiwi object.}
}
\value{
A kiwi object.
}
\description{
Apply an anonymous function to the contents of an kiwi object.
}
\details{
\bold{xTap} applies non-kiwi functions to be applied to kiwi objects.
   These functions can be named or unnamed. This allows base functions
   or external libraries to interoperate with kiwi's chaining methods.

   \code{x_(letters) $ xShuffle() $ x_Tap(sort)}

   \bold{xTap} also allows the use of anonymous methods.

   \code{x_(1:10) $ x_Tap(nums := all(nums > 0))}
}
\section{Type Signature}{

    (any -> any) -> any -> any
}

\section{Corner Cases}{

}
\seealso{
Other methods: \code{\link{xByColkeys}};
  \code{\link{xByCols}}; \code{\link{xByLevels}},
  \code{\link{xByLevels}}, \code{\link{xByValues}};
  \code{\link{xByRowkeys}}; \code{\link{xByRows}};
  \code{\link{xElemsByCols}}; \code{\link{xElemsByRows}};
  \code{\link{xExecute}}; \code{\link{x_}},
  \code{\link{x__}}
}

