% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xThirdOf}
\alias{xThirdOf}
\alias{xThirdOf_}
\title{xThirdOf}
\usage{
xThirdOf(coll)

xThirdOf_(...)
}
\arguments{
\item{coll}{a collection. The collection to get the
   third element of.}

\item{...}{see above.}
}
\value{
The third element in \bold{coll}.
}
\description{
Return the third value in a collection.
}
\section{Type Signature}{

    |any| -> any
}

\section{Corner Cases}{

   Throws an error if \bold{coll} has less than
   three elements; this is because any other corner
   case would violate the function's type-signature.
}
\examples{

# 1. A simple example.

xThirdOf(letters)

# c

x_(letters) $ x_ThirdOf()

# c
}
\seealso{
Other selection_functions: \code{\link{xAtCol}},
  \code{\link{xAtCol_}}; \code{\link{xAtKey}},
  \code{\link{xAtKey_}}; \code{\link{xAt}},
  \code{\link{xAt_}}; \code{\link{xDropWhile}},
  \code{\link{xDropWhile_}}; \code{\link{xDrop}},
  \code{\link{xDrop_}}; \code{\link{xFirstOf}},
  \code{\link{xFirstOf_}}; \code{\link{xFourthOf}},
  \code{\link{xFourthOf_}}; \code{\link{xInitOf}},
  \code{\link{xInitOf_}}; \code{\link{xLastOf}},
  \code{\link{xLastOf_}}; \code{\link{xMaxBy}},
  \code{\link{xMaxBy_}}; \code{\link{xMinBy}},
  \code{\link{xMinBy_}}; \code{\link{xPluck}},
  \code{\link{xPluck_}}; \code{\link{xRestOf}},
  \code{\link{xRestOf_}}; \code{\link{xSecondOf}},
  \code{\link{xSecondOf_}}; \code{\link{xSlice}},
  \code{\link{xSlice_}}; \code{\link{xTakeWhile}},
  \code{\link{xTakeWhile_}}; \code{\link{xTake}},
  \code{\link{xTake_}}
}

