% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xThread}
\alias{xThread}
\alias{xThread...}
\title{xThread}
\usage{
xThread(val, fns)

xThread...(val, ...)
}
\arguments{
\item{val}{an arbitrary value. The value to feed to
   the input functions.}

\item{fns}{a collection of unary functions. The functions
  to successively pipe the value through.}

\item{...}{see above.}
}
\value{
An arbitrary value.
}
\description{
Iteratively apply a list of functions to a value.
}
\section{Corner Cases}{

   Returns the empty list if \bold{coll} is length-zero.
}
\examples{

NULL

}
\seealso{
Other function_application_functions: \code{\link{xApply}},
  \code{\link{xApply...}}; \code{\link{xUnspread}};
  \code{\link{xSpread}}; \code{\link{xDeepMap}},
  \code{\link{xDeepMap...}}; \code{\link{xFlatMap}},
  \code{\link{xFlatMap...}}; \code{\link{xMapIndexed}},
  \code{\link{xMapIndexed...}}; \code{\link{xMapply}},
  \code{\link{xMapply...}}; \code{\link{xMap}},
  \code{\link{xMap...}}; \code{\link{xVectorise}},
  \code{\link{xVectorize}}

Other function_modifying_functions: \code{\link{\%and\%}},
  \code{\link{\%or\%}}, \code{\link{xLift}},
  \code{\link{xLift...}}; \code{\link{\%of\%}},
  \code{\link{\%then\%}}, \code{\link{xCompose}},
  \code{\link{xCompose...}}; \code{\link{xAsClosure}};
  \code{\link{xUnspread}}; \code{\link{xSpread}};
  \code{\link{xCapture}}, \code{\link{xK}};
  \code{\link{xDelay}}; \code{\link{xJuxtapose}},
  \code{\link{xJuxtapose...}}; \code{\link{xNot}};
  \code{\link{xPartial}}, \code{\link{xPartial...}};
  \code{\link{xVectorise}}, \code{\link{xVectorize}}
}

