% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xThread}
\alias{xThread}
\alias{xThread_}
\title{xThread}
\usage{
xThread(val, fns)

xThread_(val, ...)
}
\arguments{
\item{val}{an arbitrary value. The value to feed to
   the input functions.}

\item{fns}{a collection of unary functions. The functions
  to successively pipe the value through.}

\item{...}{see above.}
}
\value{
An arbitrary value.
}
\description{
Iteratively apply a list of functions to a value.
}
\section{Type Signature}{

    any -> |(any -> any)| -> any
}

\section{Corner Cases}{

   Returns the empty list if \bold{coll} is length-zero.
}
\examples{

# 1. xThread can be used to emulate methods for the
#    base functions in R

# get the scalar square of a matrix, and sum the rows
xThread_(matrix(1:4, 2, 2),
	t,
	X := X^2,
	rowSums
)

# c(5, 25)
# is conceptually similar to

# x_(matrix(1:4, 2, 2)) $ xTranspose() $ xTap(X := X^2) $ x_Tap(rowSums)
#
# but without all the tapping.
}
\seealso{
Other function_application_functions: \code{\link{xApply}},
  \code{\link{xApply_}}; \code{\link{xDeepMap}},
  \code{\link{xDeepMap_}}; \code{\link{xFlatMap}},
  \code{\link{xFlatMap_}}; \code{\link{xMapIndexed}},
  \code{\link{xMapIndexed_}}; \code{\link{xMapply}},
  \code{\link{xMapply_}}; \code{\link{xMap}},
  \code{\link{xMap_}}; \code{\link{xSpread}};
  \code{\link{xUnspread}}; \code{\link{xVectorise}},
  \code{\link{xVectorize}}

Other function_modifying_functions: \code{\link{\%and\%}},
  \code{\link{\%or\%}}, \code{\link{xLift}},
  \code{\link{xLift_}}; \code{\link{\%of\%}},
  \code{\link{\%then\%}}, \code{\link{xCompose}},
  \code{\link{xCompose_}}; \code{\link{xAsClosure}};
  \code{\link{xCapture}}, \code{\link{xK}};
  \code{\link{xDelay}}; \code{\link{xFix}},
  \code{\link{xFix_}}; \code{\link{xJuxtapose}},
  \code{\link{xJuxtapose_}}; \code{\link{xNot}};
  \code{\link{xSpread}}; \code{\link{xUnspread}};
  \code{\link{xVectorise}}, \code{\link{xVectorize}}
}

