% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xToChars}
\alias{xToChars}
\title{xToChars}
\usage{
xToChars(str)
}
\arguments{
\item{str}{a length-one character vector. The string to split into
   characters.}
}
\value{
A character vector of length \code{nchar(str)}, with each element being a
   single character.
}
\description{
Split a single string into a vector of characters.
}
\section{Type Signature}{

    <character> -> &lt;character>
}
\examples{

# 1. CE: Sum of Digits
# sum the digits of a positive integer

x_(toString(12358134L)) $ xToChars() $ xMap(as.numeric) $ x_Reduce("+")

# 27

# 2. Read in protein sequence data, and then find the most frequent
# amino acid.

dgcr8 <- xFromLines_(
	"METDESPSPLPCGPAGEAVMESRARPFQALPREQSPPPPLQTSSGAEVMDVGSGGDGQSELPAEDPFNFYGASLLSKGSF",
	"SKGRLLIDPNCSGHSPRTARHAPAVRKFSPDLKLLKDVKISVSFTESCRSKDRKVLYTGAERDVRAECGLLLSPVSGDVHA",
	"CPFGGSVGDGVGIGGESADKKDEENELDQEKRVEYAVLDELEDFTDNLELDEEGAGGFTAKAIVQRDRVDEEALNFPYEDDF",
	"DNDVDALLEEGLCAPKKRRTEEKYGGDSDHPSDGETSVQPMMTKIKTVLKSRGRPPTEPLPDGWIMTFHNSGVPVYLHRESRVV",
	"TWSRPYFLGTGSIRKHDPPLSSIPCLHYKKMKDNEEREQSSDLTPSGDVSPVKPLSRSAELEFPLDEPDSMGADPGPPDEKDPLG",
	"AEAAPGALGQVKAKVEVCKDESVDLEEFRSYLEKRFDFEQVTVKKFRTWAERRQFNREMKRKQAESERPILPANQKLITLSVQDAP",
	"TKKEFVINPNGKSEVCILHEYMQRVLKVRPVYNFFECENPSEPFGASVTIDGVTYGSGTASSKKLAKNKAARATLEILIPDFVKQTSE",
	"EKPKDSEELEYFNHISIEDSRVYELTSKAGLLSPYQILHECLKRNHGMGDTSIKFEVVPGKNQKSEYVMACGKHTVRGWCKNKRVGKQLA",
	"SQKILQLLHPHVKNWGSLLRMYGRESSKMVKQETSDKSVIELQQYAKKNKPNLHILSKLQEEMKRLAEEREETRKKPKMSIVASAQPGGEPL",
	"CTVDV")

x_(dgcr8) $ xToChars() $ xTabulate() $ x_MaxBy(xSecondOf)

# list('E', 77)
# Glutamic Acid is the most common residue in dgcr8
}
\seealso{
Other text_processing_functions: \code{\link{xExplode}};
  \code{\link{xFromChars}}, \code{\link{xFromChars_}};
  \code{\link{xFromLines}}, \code{\link{xFromLines_}};
  \code{\link{xFromWords}}, \code{\link{xFromWords_}};
  \code{\link{xImplode}}, \code{\link{xImplode_}};
  \code{\link{xIsMatch}}; \code{\link{xNotMatch}};
  \code{\link{xReadChars}}; \code{\link{xReadLines}};
  \code{\link{xReadWords}}; \code{\link{xSliceString}},
  \code{\link{xSliceString_}}; \code{\link{xToLines}};
  \code{\link{xToWords}}; \code{\link{xWriteChars}};
  \code{\link{xWriteLines}}; \code{\link{xWriteWords}}
}

