\name{xToLines}
\alias{xToLines}
\title{xToLines}
\usage{
xToLines(str)
}
\arguments{
  \item{str}{a string.}
}
\value{
A character vector.
}
\description{
Split a string at every newline character.
}
\details{
The input string is split at newline characters.  Multiple
newlines are treated the same as one newline.
}
\seealso{
Other character_functions: \code{\link{xExplode}},
\code{\link{xExplode...}}; \code{\link{xFromChars}},
\code{\link{xFromChars...}}; \code{\link{xFromLines}},
\code{\link{xFromLines...}}; \code{\link{xFromWords}},
\code{\link{xFromWords...}}; \code{\link{xImplode}},
\code{\link{xImplode...}}; \code{\link{xSubstring}},
\code{\link{xSubstring...}}; \code{\link{xToChars}};
\code{\link{xToWords}}
}

