% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xToLines}
\alias{xToLines}
\title{xToLines}
\usage{
xToLines(str)
}
\arguments{
\item{str}{a length-one character vector. The string to split into
   lines.}
}
\value{
A character vector.
}
\description{
Split a string at every newline character.
}
\details{
The input string is split at newline characters.
   Multiple newlines are treated the same as one newline.
}
\section{Type Signature}{

    &lt;character> -> &lt;character>
}

\section{Corner Cases}{

   Returns the empty character-vector when \bold{str}
   is length-zero.
}
\examples{

# 1. Lexically sort a list.
#

info <-
"
a: a for aardvark.
z: z for zebra.
d: d for deer.
f: f for ferret.
g: g for green snake.
y: y for yapok.
"

x_(info) $ xToLines() $ x_Tap(sort)
}
\seealso{
Other text_processing_functions: \code{\link{xAmend}};
  \code{\link{xCarve}}; \code{\link{xCharsOf}};
  \code{\link{xExplode}}; \code{\link{xFromChars}},
  \code{\link{xFromChars_}}; \code{\link{xFromLines}},
  \code{\link{xFromLines_}}; \code{\link{xFromWords}},
  \code{\link{xFromWords_}}; \code{\link{xImplode}},
  \code{\link{xImplode_}}; \code{\link{xIsMatch}};
  \code{\link{xNotMatch}}; \code{\link{xRead}};
  \code{\link{xToChars}}; \code{\link{xToWords}};
  \code{\link{xWrite}}
}

