\name{xToWords}
\alias{xToWords}
\title{xToWords}
\usage{
xToWords(str)
}
\arguments{
  \item{str}{a length-one character vector.}
}
\value{
a character vector, with one or more elements.
}
\description{
Split a str at whitespace.
}
\seealso{
Other character_functions: \code{\link{xExplode}},
\code{\link{xExplode...}}; \code{\link{xFromChars}},
\code{\link{xFromChars...}}; \code{\link{xFromLines}},
\code{\link{xFromLines...}}; \code{\link{xFromWords}},
\code{\link{xFromWords...}}; \code{\link{xImplode}},
\code{\link{xImplode...}}; \code{\link{xSubstring}},
\code{\link{xSubstring...}}; \code{\link{xToChars}};
\code{\link{xToLines}}
}

