% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xToWords}
\alias{xToWords}
\title{xToWords}
\usage{
xToWords(str)
}
\arguments{
\item{str}{a length-one character vector. The string to split into
   whitespace-seperated words.}
}
\value{
A character vector, with one or more elements.
}
\description{
Split a string at whitespaces.
}
\details{
The input string is split at spaces, tabs, and newlines,
   or a mix of the above. Multiple whitespaces are treated
   the same as one whitespace.
}
\section{Type Signature}{

    &lt;character> -> &lt;character>
}

\section{Corner Cases}{

   Returns the empty character-vector when \bold{str}
   is length-zero.
}
\examples{

# 1. Find the longest word in a sentence.
# searching for a minimum is a classic use of
# reduce over fold.

lorem_ipsum <-
" Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nullam gravida dui a odio
 commodo, eu porta ante pulvinar. Nulla nec adipiscing nibh. Donec sagittis ligula in
 lacus placerat, eget hendrerit justo sagittis. In varius dui blandit dui hendrerit
 mattis. Vestibulum ante ipsum primis in faucibus orci luctus et ultrices posuere
 cubilia Curae; Mauris at luctus erat. Aliquam eget mi vehicula, tincidunt orci vitae,
 facilisis eros. Mauris quam est, iaculis gravida mi vitae, semper lobortis magna. Donec
 mattis, ipsum vel rhoncus pretium, tortor libero accumsan lectus, vitae convallis nulla
 mauris non purus. Morbi nisi risus, rutrum condimentum sem at, tristique dictum justo.
 Duis tincidunt dictum odio ac dignissim. Sed eget auctor odio."

x_(lorem_ipsum) $ xToWords() $
x_Reduce(
	(best : new) := {
		ifelse(nchar(best) < nchar(new), new, best)
	}
)
}
\seealso{
Other text_processing_functions: \code{\link{xExplode}};
  \code{\link{xFromChars}}, \code{\link{xFromChars_}};
  \code{\link{xFromLines}}, \code{\link{xFromLines_}};
  \code{\link{xFromWords}}, \code{\link{xFromWords_}};
  \code{\link{xImplode}}, \code{\link{xImplode_}};
  \code{\link{xIsMatch}}; \code{\link{xNotMatch}};
  \code{\link{xRead}}; \code{\link{xSliceString}},
  \code{\link{xSliceString_}}; \code{\link{xToChars}};
  \code{\link{xToLines}}; \code{\link{xWrite}}
}

