% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xTruth}
\alias{xTruth}
\title{xTruth}
\usage{
xTruth(...)
}
\arguments{
\item{...}{arguments to be dropped.}
}
\value{
true
}
\description{
A function that always returns true.
}
\section{Type Signature}{

    ...any -> <logical>
}
\examples{

# 1.
# xTruth is useful for functions that take predicates .

xSelect(xTruth, 1:3)

# c(1, 2, 3)
}
\seealso{
Other basic_functions: \code{\link{xCapture}},
  \code{\link{xK}}; \code{\link{xFalsity}};
  \code{\link{xIrrelevance}}; \code{\link{xI}},
  \code{\link{xIdentity}}
}

