\name{xTruth}
\alias{xTruth}
\title{xTruth}
\usage{
xTruth(...)
}
\arguments{
  \item{...}{arguments to be dropped.}
}
\value{
true
}
\description{
A function that always returns true.
}
\examples{

NULL

}
\seealso{
Other basic_functions: \code{\link{xCapture}},
\code{\link{xK}}; \code{\link{xFalsity}};
\code{\link{xIrrelevance}}; \code{\link{xI}},
\code{\link{xIdentity}}
}

