\name{xUnchars}
\alias{xUnchars}
\alias{xUnchars...}
\title{xUnchars}
\usage{
xUnchars(strs)

xUnchars...(...)
}
\arguments{
  \item{strs}{several character vectors.}

  \item{...}{see above.}
}
\value{
a length-one character vector.
}
\description{
Collapase a character vector with empty strs as delimiters.
}
\section{Variadic Functions

\link{ help_variadic }}{

}
\seealso{
Other character_functions: \code{\link{xChars}};
\code{\link{xExplode}}, \code{\link{xExplode...}};
\code{\link{xImplode}}, \code{\link{xImplode...}};
\code{\link{xLines}}; \code{\link{xSubstring}},
\code{\link{xSubstring...}}; \code{\link{xUnlines}},
\code{\link{xUnlines...}}; \code{\link{xUnwords}},
\code{\link{xUnwords...}}; \code{\link{xWords}}
}

