\name{xUnchars}
\alias{xUnchars}
\title{xUnchars}
\usage{
xUnchars(strs)
}
\arguments{
  \item{strs}{several character vectors.}
}
\value{
a length-one character vector.
}
\description{
Collapase a character vector with empty strs as delimiters.
}
\seealso{
Other character_vector_functions: \code{\link{xChars}};
\code{\link{xCollapse}}; \code{\link{xLines}};
\code{\link{xSplitString}}; \code{\link{xSubString}};
\code{\link{xUnlines}}; \code{\link{xUnwords}};
\code{\link{xWords}}
}

