\name{xUnion}
\alias{xUnion}
\title{xUnion}
\usage{
  xUnion(coll1, coll2)
}
\arguments{
  \item{coll1}{a collection}

  \item{coll2}{a collection}
}
\value{
  a list.
}
\description{
  Get the set union of two collections.
}
\section{Corner Cases}{
  returns the empty list if \code{coll} is length-zero.
}
\seealso{
  Other collection_functions: \code{\link{xAsFunction}},
  \code{\link{xAssoc}}, \code{\link{xContains}},
  \code{\link{xDiffer}}, \code{\link{xDissoc}},
  \code{\link{xDrop}}, \code{\link{xFirst}},
  \code{\link{xFlatten}}, \code{\link{xFourth}},
  \code{\link{xInit}}, \code{\link{xInter}},
  \code{\link{xIsEmpty}}, \code{\link{xIsFalse}},
  \code{\link{xIsNa}}, \code{\link{xIsNan}},
  \code{\link{xIsNull}}, \code{\link{xIsTrue}},
  \code{\link{xLast}}, \code{\link{xLength}},
  \code{\link{xName}}, \code{\link{xNotFalse}},
  \code{\link{xNotNa}}, \code{\link{xNotNan}},
  \code{\link{xNotTrue}}, \code{\link{xPack}},
  \code{\link{xRepeat}}, \code{\link{xRest}},
  \code{\link{xSecond}}, \code{\link{xSegment}},
  \code{\link{xShuffle}}, \code{\link{xSplit}},
  \code{\link{xTake}}, \code{\link{xThird}},
  \code{\link{xUnit}}
}

