\name{xUnion}
\alias{xUnion}
\title{xUnion}
\usage{
xUnion(colls)
}
\arguments{
  \item{colls}{a collection of collections.}
}
\value{
a list.
}
\description{
Get the set union of several collections.
}
\section{Corner Cases}{
  returns the empty list if \code{colls} is length-zero.
}
\seealso{
Other collection_functions: \code{\link{xAsFunction}};
\code{\link{xAssoc}}; \code{\link{xContains}};
\code{\link{xDiffer}}; \code{\link{xDissoc}};
\code{\link{xDrop}}; \code{\link{xFirst}};
\code{\link{xFlatten}}; \code{\link{xFourth}};
\code{\link{xInit}}; \code{\link{xInter}};
\code{\link{xIsEmpty}}; \code{\link{xIsFalse}};
\code{\link{xIsNan}}; \code{\link{xIsNa}};
\code{\link{xIsNull}}; \code{\link{xIsTrue}};
\code{\link{xLast}}; \code{\link{xLength}};
\code{\link{xName}}; \code{\link{xNotFalse}};
\code{\link{xNotNan}}; \code{\link{xNotNa}};
\code{\link{xNotTrue}}; \code{\link{xPack}};
\code{\link{xRepeat}}; \code{\link{xRest}};
\code{\link{xReverse}}; \code{\link{xSecond}};
\code{\link{xSegment}}; \code{\link{xShuffle}};
\code{\link{xSplit}}; \code{\link{xTake}};
\code{\link{xThird}}; \code{\link{xUnit}}
}

