% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xUnionOf}
\alias{xUnionOf}
\alias{xUnionOf_}
\title{xUnionOf
``
Get the set union of several collections.}
\usage{
xUnionOf(colls)

xUnionOf_(...)
}
\arguments{
\item{colls}{a collection of collections. The collections
   to take the union of.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
xUnionOf
``
Get the set union of several collections.
}
\details{
\code{xUnionOf} returns a list of the unique
   elements in the combination of each collection
   in \code{colls}.
}
\section{Type Signature}{

    ||any|| -> [any]
}

\section{Corner Cases}{

   Returns the empty list if \code{colls} is length-zero.
}
\examples{

NULL
}
\seealso{
Other set_functions: \code{\link{xDuplicatesOf}},
  \code{\link{xDuplicatesOf_}}; \code{\link{xInterOf}},
  \code{\link{xInterOf_}}; \code{\link{xIsEmpty}},
  \code{\link{xIsEmpty_}}; \code{\link{xIsIn}},
  \code{\link{xIsIn_}}; \code{\link{xIsSubset}},
  \code{\link{xIsSubset_}}; \code{\link{xNotEmpty}},
  \code{\link{xNotEmpty_}}; \code{\link{xNotIn}},
  \code{\link{xNotIn_}}; \code{\link{xNotSubset}},
  \code{\link{xNotSubset_}}; \code{\link{xPowerSetOf}},
  \code{\link{xPowerSetOf_}}; \code{\link{xProdSetOf}},
  \code{\link{xProdSetOf_}}; \code{\link{xUniqueOf}},
  \code{\link{xUniqueOf_}}
}

