% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xUniqueOf}
\alias{xUniqueOf}
\alias{xUniqueOf_}
\title{xUniqueOf}
\usage{
xUniqueOf(coll)

xUniqueOf_(...)
}
\arguments{
\item{coll}{a collection. The collection to return unique
   elements from.}

\item{...}{see above.}
}
\value{
A list.
}
\description{
Return the unique elements in a collection.
}
\section{Type Signature}{

    |any| -> [any]
}

\section{Corner Cases}{

   Returns the empty list if \code{colls} is length-zero.
}
\examples{

# 1. Find the smallest unique number in a list.

lowest <-
x_( list(1, 1, 3, 2, 2, 2, 5, 4, 9, 9) ) $
xUniqueOf() $
x_Tap(xAsDouble \%then\% min)

# 2. CE is a string a pangram?
#    does a string contain each letter.

is_pangram <- line := {

	x_(line) $
	xToChars() $ xMap(tolower) $ xUniqueOf() $
	xSelect(
		xFix_(xIsMember, coll = letters)
	) $
	x_LenOf() == 26
}

is_pangram('A quick brown fox jumps over the lazy dog')

# True

is_pangram('not pangramic thank you mam.')

# False
}
\seealso{
Other set_functions: \code{\link{xDuplicatesOf}},
  \code{\link{xDuplicatesOf_}}; \code{\link{xInter}},
  \code{\link{xInter_}}; \code{\link{xIsMember}},
  \code{\link{xIsMember_}}; \code{\link{xIsSubset}},
  \code{\link{xIsSubset_}}; \code{\link{xNotMember}},
  \code{\link{xNotMember_}}; \code{\link{xPowerSetOf}},
  \code{\link{xPowerSetOf_}}; \code{\link{xProdSetOf}},
  \code{\link{xProdSetOf_}}; \code{\link{xReorder}},
  \code{\link{xReorder_}}; \code{\link{xUnionOf}},
  \code{\link{xUnionOf_}}
}

