\name{xUnit}
\alias{xUnit}
\title{Return the neutral element of a collection.}
\usage{
  xUnit(coll)
}
\arguments{
  \item{coll}{an list, pairlist, or vector of any length.}
}
\value{
  Returns null if \code{coll} is a pairlist, a typed vector
  of length zero if \code{coll} is a vector, and the empty
  list if \code{coll} is a list.
}
\description{
  Return the neutral element of a collection.
}

