\name{xUnlines}
\alias{xUnlines}
\alias{xUnlines...}
\title{xUnlines}
\usage{
xUnlines(strs)

xUnlines...(...)
}
\arguments{
  \item{strs}{a character vector.}

  \item{...}{see above.}
}
\value{
a length-one character vector.
}
\description{
Collapase a character vector with newlines as delimiters.
}
\section{Variadic Functions

\link{ help_variadic }}{

}
\seealso{
Other character_functions: \code{\link{xChars}};
\code{\link{xExplode}}; \code{\link{xImplode}},
\code{\link{xImplode...}}; \code{\link{xLines}};
\code{\link{xSubstring}}, \code{\link{xSubstring...}};
\code{\link{xUnchars}}, \code{\link{xUnchars...}};
\code{\link{xUnwords}}, \code{\link{xUnwords...}};
\code{\link{xWords}}
}

