\name{xUnwords}
\alias{xUnwords}
\alias{xUnwords...}
\title{xUnwords}
\usage{
xUnwords(strs)

xUnwords...(...)
}
\arguments{
  \item{strs}{an arbitrary number of character vectors.}

  \item{...}{see above.}
}
\value{
a length-one character vector.
}
\description{
Collapase a character vector with spaces as delimiters.
}
\section{Variadic Functions

\link{ help_variadic }}{

}
\seealso{
Other character_functions: \code{\link{xChars}};
\code{\link{xExplode}}, \code{\link{xExplode...}};
\code{\link{xImplode}}, \code{\link{xImplode...}};
\code{\link{xLines}}; \code{\link{xSubstring}},
\code{\link{xSubstring...}}; \code{\link{xUnchars}},
\code{\link{xUnchars...}}; \code{\link{xUnlines}},
\code{\link{xUnlines...}}; \code{\link{xWords}}
}

