\name{xUnwords}
\alias{xUnwords}
\title{Collapase a character vector with spaces as delimiters.}
\usage{
  xUnwords(...)
}
\arguments{
  \item{...}{an arbitrary number of character vectors.}
}
\value{
  a length-one character vector.
}
\description{
  Collapase a character vector with spaces as delimiters.
}

