\name{xUnwords}
\alias{xUnwords}
\title{xUnwords}
\usage{
xUnwords(strs)
}
\arguments{
  \item{strs}{an arbitrary number of character vectors.}
}
\value{
a length-one character vector.
}
\description{
Collapase a character vector with spaces as delimiters.
}
\seealso{
Other character_functions: \code{\link{xChars}};
\code{\link{xExplode}}; \code{\link{xImplode}};
\code{\link{xLines}}; \code{\link{xSubString}};
\code{\link{xUnchars}}; \code{\link{xUnlines}};
\code{\link{xWords}}
}

