% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xUnzipIndices}
\alias{xUnzipIndices}
\alias{xUnzipIndices_}
\title{xUnzipIndices}
\usage{
xUnzipIndices(coll)

xUnzipIndices_(...)
}
\arguments{
\item{coll}{a collection. The collection to split into index, element pairs.}

\item{...}{see above.}
}
\value{
A named list.
}
\description{
Split a collection into a list of index: value pairs
}
\details{
\bold{xUnzipIndices} reshapes a collection to allow you to pass
    both the indices and elements to another higher order function.
    More generally, \bold{xUnzipIndices} allows you to selectively work on
    elements based on their position in a collection, which functions like
    \bold{xMap} and \bold{xSelect} otherwise do not support.
}
\section{Type Signature}{

    |any| -> [[ <number>, ...any ]]
}

\section{Corner Cases}{

     Returns \code{list()} if \code{coll} is length-zero.
}

\section{Rcpp}{


    This function is written using C++, and should be quite performant.
}
\examples{

# 1. enumerate the letters with their indices.

x_(letters) $ xUnzipIndices() $ xMap( xUnspread((ith : val) := {
	xFromWords_('the ', paste(ith), ' letter is ', val)
}) )
}
\seealso{
Other key_functions: \code{\link{xAddKeys}},
  \code{\link{xAddKeys_}}; \code{\link{xIndicesOf}},
  \code{\link{xIndicesOf_}}; \code{\link{xIndicesTo}};
  \code{\link{xKeysOf}}, \code{\link{xKeysOf_}};
  \code{\link{xLocate}}, \code{\link{xLocate_}};
  \code{\link{xUnzipKeys}}, \code{\link{xUnzipKeys_}};
  \code{\link{xValuesOf}}, \code{\link{xValuesOf_}};
  \code{\link{xZipKeys}}, \code{\link{xZipKeys_}}

Other reshaping_functions: \code{\link{xAppend}},
  \code{\link{xAppend_}}; \code{\link{xChop}},
  \code{\link{xChop_}}; \code{\link{xChunk}},
  \code{\link{xChunk_}}; \code{\link{xFlatten}},
  \code{\link{xFlatten_}}; \code{\link{xGather}},
  \code{\link{xGather_}}; \code{\link{xGroupBy}},
  \code{\link{xGroupBy_}}; \code{\link{xJoin}},
  \code{\link{xJoin_}}; \code{\link{xOneOf}},
  \code{\link{xOneOf_}}; \code{\link{xPrepend}},
  \code{\link{xPrepend_}}; \code{\link{xRepeat}},
  \code{\link{xRepeat_}}; \code{\link{xReverse}},
  \code{\link{xReverse_}}; \code{\link{xShuffle}},
  \code{\link{xShuffle_}}; \code{\link{xSortBy}},
  \code{\link{xSortBy_}}; \code{\link{xTabulate}},
  \code{\link{xTabulate_}}; \code{\link{xUnzipKeys}},
  \code{\link{xUnzipKeys_}}; \code{\link{xZipKeys}},
  \code{\link{xZipKeys_}}; \code{\link{xZip}},
  \code{\link{xZip_}}
}

