% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xVal}
\alias{xVal}
\title{xVal}
\usage{
xVal(sym, val)
}
\arguments{
\item{sym}{a symbol or string. The variable name
   to assign a value.}

\item{val}{an arbitrary value. The value to be assigned.}
}
\value{
Null; this function is used for side-effects.
}
\description{
Assign a constant value to the calling environment.
}
\details{
xVal works like normal R assignment, with the exception
   that if any attempt to update the assigned variable
   will result in an error. This is analogous to const in
   other languages.

   \code{the_letter_a <- "a"}
   \code{xVal(the_letter_a, "a")}

   As the above call shows \code{xVal} is a standard function,
   not an infix function like normal assignment.

   Kea's value functions are built on top of base R's \bold{lockBinding}
   mechanism, which is self-described as being 'experimental'.
}
\section{Corner Cases}{

   \bold{xVal} overwrites the value referenced by \code{sym} if the variable
   already exists in the parent frame.
}

\section{Stability 1 - \bold{Uncertain}}{


This feature may be removed or altered significantly in future versions.
If you find this feature useful, please leave an issue at https://github.com/rgrannell1/kea/issues.
}
\examples{

# 1. assign a constant value,
# that can't later be altered.

xVal(tau, 6.283185)
}
\seealso{
Other immutable_value_functions: \code{\link{xAsVal}};
  \code{\link{xAsVar}}
}

