% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xValuesOf}
\alias{xValuesOf}
\title{xValuesOf}
\usage{
xValuesOf(coll)
}
\arguments{
\item{coll}{a collection. The collection to remove names from.}

\item{...}{see above.}
}
\value{
a list.
}
\description{
Remove the names from a collection.
}
\section{Type Signature}{

    |any| -> [any]
}

\section{Corner Cases}{

    Returns length-zero when \bold{coll} is length-zero.
}
\examples{

NULL
}
\seealso{
Other key_functions: \code{\link{xAddKeys}},
  \code{\link{xAddKeys_}}; \code{\link{xIndicesOf}},
  \code{\link{xIndicesOf_}}; \code{\link{xKeysOf}},
  \code{\link{xKeysOf_}}; \code{\link{xUnzipIndices}},
  \code{\link{xUnzipIndices_}}; \code{\link{xUnzipKeys}},
  \code{\link{xUnzipKeys_}}; \code{\link{xZipKeys}},
  \code{\link{xZipKeys_}}
}

