% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xVectorise}
\alias{xVectorise}
\alias{xVectorize}
\title{xVectorise}
\usage{
xVectorise(fn)

xVectorize(fn)
}
\arguments{
\item{fn}{a unary function. The function to vectorise.}
}
\value{
A unary function of coll.
}
\description{
Convert a function on one value to a function on collections of values.
}
\details{
\bold{xVectorise} is a shorthand function for partially applying
   \bold{xMap} with a function, returning a vectorised function.
   This can be useful for taking a function that only works on single
   values - like \bold{is.null( )} - and extending them to work on
   collections of values.

   \code{is.nulls <- xVectorise(is.null)}

   \code{is.nulls(list(1, 2, Null, 4))}

   \code{list(False, False, True, False)}
}
\section{Type Signature}{

    (any -> any) -> (|any| -> [any])
}

\section{Corner Cases}{

   Returns the empty list if \bold{coll} is length-zero.
}
\examples{

# 1. HR Functional Programming
# Replaces a list of numbers with their absolute values

# This is fairly perfect use of xVectorise - a function which takes a function that
# works on one value and returns a function that works on a collection of values.
# xVectorise is really just a shorthand for partially applying xMap with a function.

absolutes <- xVectorise(abs)
absolutes(c(-1, 0, +1))

# c(1, 0, 1)
}
\seealso{
Other function_application_functions: \code{\link{xApply}},
  \code{\link{xApply_}}; \code{\link{xDeepMap}},
  \code{\link{xDeepMap_}}; \code{\link{xFlatMap}},
  \code{\link{xFlatMap_}}; \code{\link{xMapply}},
  \code{\link{xMapply_}}; \code{\link{xMap}},
  \code{\link{xMap_}}; \code{\link{xSpread}};
  \code{\link{xThread}}, \code{\link{xThread_}};
  \code{\link{xUnspread}}

Other function_modifying_functions: \code{\link{\%and\%}},
  \code{\link{\%or\%}}, \code{\link{xLift}},
  \code{\link{xLift_}}; \code{\link{\%of\%}},
  \code{\link{\%then\%}}, \code{\link{xCompose}},
  \code{\link{xCompose_}}; \code{\link{xAsClosure}};
  \code{\link{xCapture}}, \code{\link{xK}};
  \code{\link{xDelay}}; \code{\link{xFix}},
  \code{\link{xFix_}}; \code{\link{xJuxtapose}},
  \code{\link{xJuxtapose_}}; \code{\link{xNot}};
  \code{\link{xSpread}}; \code{\link{xThread}},
  \code{\link{xThread_}}; \code{\link{xUnspread}}
}

