\name{xVersion}
\alias{xVersion}
\title{xVersion}
\usage{
xVersion()
}
\value{
A three-number vector, containing the major release number,
minor release number and patch number.
}
\description{
Get the current version number of Arrow.
}
\details{
\code{xZipWith} takes multiple collections (columns), calls
the input function with each row, and returns a a list of
equal length to the input columns.
}

