% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xVersion}
\alias{xVersion}
\title{xVersion}
\usage{
xVersion(...)
}
\arguments{
\item{...}{arguments to be dropped.}
}
\value{
A three-number vector, containing the
   major release number, minor release number and patch number.
}
\description{
Get the current version number of Arrow.
}
\examples{

# 1. xVersion
#    the most important (and only) example

xVersion( )
}
\seealso{
Other inpure_functions: \code{\link{xDelay}};
  \code{\link{xDo}}, \code{\link{xDo...}};
  \code{\link{xExecute}}; \code{\link{xReadChars}};
  \code{\link{xReadLines}}; \code{\link{xReadWords}};
  \code{\link{xStopwatch}}; \code{\link{xWriteChars}};
  \code{\link{xWriteLines}}; \code{\link{xWriteWords}}
}

