% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xVersion}
\alias{xVersion}
\title{xVersion}
\usage{
xVersion(...)
}
\arguments{
\item{...}{arguments to be dropped.}
}
\value{
A three-number vector, containing the
   major release number, minor release number and patch number.
}
\description{
Get the current version number of Kiwi.
}
\details{
Kiwi versioning follows the semantic version
    standard (2.0.0), which is summarised by the author as

 "Given a version number MAJOR.MINOR.PATCH, increment the:

 MAJOR version when you make incompatible API changes,
 MINOR version when you add functionality in a backwards-compatible manner, and
 PATCH version when you make backwards-compatible bug fixes."

 For pre 1.0.0 versions Kiwi's minor version is bumped anytime a backwards-
 incompatable change is made, or new functionality is added.
}
\section{Type Signature}{

    ...any -> &lt;integer>
}

\section{Corner Cases}{

   Any arguments passed to \bold{xVersion} are dropped.
}
\examples{

# 1. xVersion
#    the most important (and only) example

xVersion()
}
\seealso{
Other inpure_functions: \code{\link{xDelay}};
  \code{\link{xDo}}, \code{\link{xDo_}};
  \code{\link{xExecute}}; \code{\link{xRead}};
  \code{\link{xStopwatch}}; \code{\link{xWrite}}
}

