% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xWriteChars}
\alias{xWriteChars}
\title{xWriteChars}
\usage{
xWriteChars(str, strs)
}
\arguments{
\item{str}{a length-one character vector. The path to
   export to.}

\item{strs}{a character vector. Lines to write to a file.}
}
\value{
null -
}
\description{
Export a collection of strings as a single line in a file.
}
\section{Type Signature}{

    |character| -> {}
}
\examples{

NULL
}
\seealso{
Other inpure_functions: \code{\link{xDelay}};
  \code{\link{xDo}}, \code{\link{xDo_}};
  \code{\link{xExecute}}; \code{\link{xReadChars}};
  \code{\link{xReadLines}}; \code{\link{xReadWords}};
  \code{\link{xStopwatch}}; \code{\link{xVersion}};
  \code{\link{xWriteLines}}; \code{\link{xWriteWords}}

Other text_processing_functions: \code{\link{xExplode}};
  \code{\link{xFromChars}}, \code{\link{xFromChars_}};
  \code{\link{xFromLines}}, \code{\link{xFromLines_}};
  \code{\link{xFromWords}}, \code{\link{xFromWords_}};
  \code{\link{xImplode}}, \code{\link{xImplode_}};
  \code{\link{xIsMatch}}; \code{\link{xNotMatch}};
  \code{\link{xReadChars}}; \code{\link{xReadLines}};
  \code{\link{xReadWords}}; \code{\link{xSliceString}},
  \code{\link{xSliceString_}}; \code{\link{xToChars}};
  \code{\link{xToLines}}; \code{\link{xToWords}};
  \code{\link{xWriteLines}}; \code{\link{xWriteWords}}
}

