% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xWriteWords}
\alias{xWriteWords}
\title{xWriteWords}
\usage{
xWriteWords(str, strs)
}
\arguments{
\item{str}{a length-one character vector. The path to
   export to.}

\item{strs}{a character vector. Lines to write to a file.}
}
\value{
A character vector, with one or more elements.
}
\description{
Export a collection of strings as a single line in a file.
}
\examples{

NULL
}
\seealso{
Other inpure_functions: \code{\link{xDelay}};
  \code{\link{xDo}}, \code{\link{xDo...}};
  \code{\link{xExecute}}; \code{\link{xReadChars}};
  \code{\link{xReadLines}}; \code{\link{xReadWords}};
  \code{\link{xStopwatch}}; \code{\link{xVersion}};
  \code{\link{xWriteChars}}; \code{\link{xWriteLines}}

Other text_processing_functions: \code{\link{xExplode}},
  \code{\link{xExplode...}}; \code{\link{xFromChars}},
  \code{\link{xFromChars...}}; \code{\link{xFromLines}},
  \code{\link{xFromLines...}}; \code{\link{xFromWords}},
  \code{\link{xFromWords...}}; \code{\link{xImplode}},
  \code{\link{xImplode...}}; \code{\link{xReadChars}};
  \code{\link{xReadLines}}; \code{\link{xReadWords}};
  \code{\link{xSliceString}}, \code{\link{xSliceString...}};
  \code{\link{xToChars}}; \code{\link{xToLines}};
  \code{\link{xToWords}}; \code{\link{xWriteChars}};
  \code{\link{xWriteLines}}
}

