% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xZipKeys}
\alias{xZipKeys}
\alias{xZipKeys_}
\title{xZipKeys}
\usage{
xZipKeys(colls)

xZipKeys_(...)
}
\arguments{
\item{colls}{a list or pairlist of two-element lists or pairlists,
   with the first element being a string and the second
   element being an arbitrary value. The string part of each
   sublist is used as the name in the output list, and the
   remaining part is used as the value associated with that key.}

\item{...}{see above.}
}
\value{
A named list.
}
\description{
Convert a collection of name, value pairs into a named collection.
}
\section{Corner Cases}{

   Returns the empty list if \bold{coll} is length-zero.
}
\examples{

# 1. turn a csv of country names and median ages into
#    a named list of countires, with values being the median age.

csv <- "Monaco,48.9,Japan,44.6,Italy,44.3,Germany,43.7,Jersey,43.4"

x_(csv) $ xExplode(',') $ xChunk(2) $ x_ZipKeys()

# list(
#    Monaco =
#        48.9,
#    Japan =
#        44.6,
#    Italy =
#        44.3,
#    Germany =
#        43.7,
#    Jersey =
#        43.4
# )

x_(csv) $ xExplode(',') $ xChunk(2) $ x_ZipKeys()
}
\seealso{
Other key_functions: \code{\link{xAddKeys}};
  \code{\link{xPluck}}, \code{\link{xPluck_}};
  \code{\link{xUnzipKeys}}, \code{\link{xUnzipKeys_}}

Other reshaping_functions: \code{\link{xChop}},
  \code{\link{xChop_}}; \code{\link{xChunk}},
  \code{\link{xChunk_}}; \code{\link{xFlatten}},
  \code{\link{xFlatten_}}; \code{\link{xGroupBy}},
  \code{\link{xGroupBy_}}; \code{\link{xJoin}},
  \code{\link{xJoin_}}; \code{\link{xOneOf}},
  \code{\link{xOneOf_}}; \code{\link{xRecycle}},
  \code{\link{xRecycle_}}; \code{\link{xRepeat}},
  \code{\link{xRepeat_}}; \code{\link{xReverse}},
  \code{\link{xReverse_}}; \code{\link{xShuffle}},
  \code{\link{xShuffle_}}; \code{\link{xSortBy}},
  \code{\link{xSortBy_}}; \code{\link{xSplitAt}},
  \code{\link{xSplitAt_}}; \code{\link{xSplitWith}},
  \code{\link{xSplitWith_}}; \code{\link{xTabulate}},
  \code{\link{xTabulate_}}; \code{\link{xUnzipKeys}},
  \code{\link{xUnzipKeys_}}; \code{\link{xZip}},
  \code{\link{xZip_}}
}

